Steady-State Mean Vector for the Linear Stochastic Differential Equation Model

The steady-state mean vector is given by $$ -\boldsymbol{\Phi}^{-1} \boldsymbol{\iota} $$.

Usage

LinSDEMean(phi, iota)

Arguments

phi

Numeric matrix. The drift matrix which represents the rate of change of the solution in the absence of any random fluctuations (\(\boldsymbol{\Phi}\)).

iota

Numeric vector. An unobserved term that is constant over time (\(\boldsymbol{\iota}\)).

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

iota <- c(0.317, 0.230)
phi <- matrix(
  data = c(
    -0.10,
    0.05,
    0.05,
    -0.10
  ),
  nrow = 2
)
LinSDEMean(phi = phi, iota = iota)
#> [1] 5.76 5.18