Steady-State Mean Vector for the Observed Variables in the Linear Stochastic Differential Equation Model

The steady-state mean vector for the observed variables in the linear stochastic differential equation model \(\mathrm{Mean} \left( \mathbf{y} \right)\) is given by $$ \mathrm{Mean} \left( \mathbf{y} \right) = \boldsymbol{\nu} + \boldsymbol{\Lambda} \mathrm{Mean} \left( \boldsymbol{\eta} \right) $$ where \(\boldsymbol{\nu}\) is the vector of intercept values for the measurement model, \(\boldsymbol{\Lambda}\) is the matrix of factor loadings, and \(\mathrm{Mean} \left( \boldsymbol{\eta} \right)\) is the steady-state mean vector for the latent variables.

Usage

LinSDEMeanY(nu, lambda, mean_eta)

Arguments

nu

Numeric vector. Vector of intercept values for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

mean_eta

Numeric vector. Steady-state mean vector of the latent variables \(\mathrm{Mean} \left( \boldsymbol{\eta} \right)\).

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0.0, 0.0, -0.693
  ),
  nrow = 3
)
iota <- rep(x = 1, times = 3)
lambda <- diag(3)
nu <- rep(x = 1, times = 3)
mean_eta <- LinSDEMeanEta(
  phi = phi,
  iota = iota
)
LinSDEMeanY(
  nu = nu,
  lambda = lambda,
  mean_eta = mean_eta
)
#> [1] 3.801120 7.183295 7.128597