Convert Parameters from the Ornstein–Uhlenbeck Model to the Vector Autoregressive Model

This function converts parameters from the Ornstein–Uhlenbeck model to the vector autoregressive model parameterization.

Usage

OU2VAR(mu, phi, sigma, delta_t)

Arguments

mu

Numeric vector. The long-term mean or equilibrium level (\(\boldsymbol{\mu}\)).

phi

Numeric matrix. The rate of mean reversion, determining how quickly the variable returns to its mean (\(\boldsymbol{\Phi}\)).

sigma

Numeric matrix. The matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

delta_t

Positive integer. Discrete time interval (\(\delta_t\)).

Value

Returns a list of vector autoregressive model parameters:

alpha: Numeric vector. Vector of intercepts for the dynamic model (\(\boldsymbol{\alpha}\)).

beta: Numeric matrix. Transition matrix relating the values of the latent variables at time t - delta_t to those at time t where delta_t is a positive integer. (\(\boldsymbol{\beta}\)).

psi: Numeric matrix. The process noise covariance matrix (\(\boldsymbol{\Psi}\)).

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

p <- k <- 2
mu <- c(5.76, 5.18)
phi <- matrix(
  data = c(0.10, -0.05, -0.05, 0.10),
  nrow = p
)
sigma <- matrix(
  data = c(2.79, 0.06, 0.06, 3.27),
  nrow = p
)
delta_t <- 1

OU2VAR(
  mu = mu,
  phi = phi,
  sigma = sigma,
  delta_t = delta_t
)
#> $alpha
#>           [,1]
#> [1,] 0.3071697
#> [2,] 0.2263804
#> 
#> $beta
#>            [,1]       [,2]
#> [1,] 0.90596870 0.04526072
#> [2,] 0.04526072 0.90596870
#> 
#> $psi
#>           [,1]      [,2]
#> [1,] 2.5356886 0.1873121
#> [2,] 0.1873121 2.9707348
#> 

p <- k <- 3
mu <- c(0, 0, 0)
phi <- matrix(
  data = c(-0.357, 0.771, -0.450, 0, -0.511, 0.729, 0, 0, -0.693),
  nrow = p
)
sigma <- diag(p)
delta_t <- 1

OU2VAR(
  mu = mu,
  phi = phi,
  sigma = sigma,
  delta_t = delta_t
)
#> $alpha
#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0
#> 
#> $beta
#>           [,1]      [,2]     [,3]
#> [1,]  1.429036  0.000000 0.000000
#> [2,] -1.191152  1.666957 0.000000
#> [3,]  1.238265 -1.332822 1.999706
#> 
#> $psi
#>            [,1]       [,2]       [,3]
#> [1,]  1.4595848 -0.6645845  0.5983138
#> [2,] -0.6645845  2.1260283 -1.1515623
#> [3,]  0.5983138 -1.1515623  2.9517140
#>