Steady-State Covariance Matrix for the Latent Variables in the State Space Model

The steady-state covariance matrix for the latent variables in the state space model \(\mathrm{Cov} \left( \boldsymbol{\eta} \right)\) is given by $$ \mathrm{vec} \left( \mathrm{Cov} \left( \boldsymbol{\eta} \right) \right) = \left( \mathbf{I} - \boldsymbol{\beta} \otimes \boldsymbol{\beta} \right)^{-1} \mathrm{vec} \left( \boldsymbol{\Psi} \right) $$ where \(\boldsymbol{\beta}\) is the transition matrix relating the values of the latent variables at the previous to the current time point and \(\boldsymbol{\Psi}\) is the covariance matrix of volatility or randomness in the process.

Usage

SSMCovEta(beta, psi)

Arguments

beta

Numeric matrix. Transition matrix relating the values of the latent variables at the previous to the current time point (\(\boldsymbol{\beta}\)).

psi

Numeric matrix. The covariance matrix of the process noise (\(\boldsymbol{\Psi}\)).

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

beta <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0.0, 0.0, 0.5
  ),
  nrow = 3
)
psi <- 0.1 * diag(3)
SSMCovEta(
  beta = beta,
  psi = psi
)
#>            [,1]      [,2]       [,3]
#> [1,] 0.19607843 0.1183232 0.02985385
#> [2,] 0.11832319 0.3437711 0.13818551
#> [3,] 0.02985385 0.1381855 0.26638284