Steady-State Mean Vector for the Latent Variables in the State Space Model

The steady-state mean vector for the latent variables in the state space model \(\mathrm{Mean} \left( \boldsymbol{\eta} \right)\) is given by $$ \mathrm{Mean} \left( \boldsymbol{\eta} \right) = \left( \mathbf{I} - \boldsymbol{\beta} \right)^{-1} \boldsymbol{\alpha} $$ where \(\boldsymbol{\beta}\) is the transition matrix relating the values of the latent variables at the previous to the current time point, \(\boldsymbol{\alpha}\) is a vector of constant values for the dynamic model, and \(\mathbf{I}\) is an identity matrix.

Usage

SSMMeanEta(beta, alpha)

Arguments

beta

Numeric matrix. Transition matrix relating the values of the latent variables at the previous to the current time point (\(\boldsymbol{\beta}\)).

alpha

Numeric vector. Vector of constant values for the dynamic model (\(\boldsymbol{\alpha}\)).

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

beta <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0.0, 0.0, 0.5
  ),
  nrow = 3
)
alpha <- rep(x = 1, times = 3)
SSMMeanEta(
  beta = beta,
  alpha = alpha
)
#> [1] 3.333333 6.666667 6.666667