Simulate Covariance Matrices from the Multivariate Normal Distribution

This function simulates random covariance matrices from the multivariate normal distribution. The function ensures that the generated covariance matrices are positive semi-definite.

Usage

SimCovN(n, sigma, vcov_sigma_vech_l)

Arguments

n

Positive integer. Number of replications.

sigma

Numeric matrix. The covariance matrix (\(\boldsymbol{\Sigma}\)).

vcov_sigma_vech_l

Numeric matrix. Cholesky factorization (t(chol(vcov_sigma_vech))) of the sampling variance-covariance matrix of \(\mathrm{vech} \left( \boldsymbol{\Sigma} \right)\).

Value

Returns a list of random covariance matrices.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

n <- 10
sigma <- matrix(
  data = c(
    1.0, 0.5, 0.3,
    0.5, 1.0, 0.4,
    0.3, 0.4, 1.0
  ),
  nrow = 3
)
vcov_sigma_vech_l <- t(
  chol(
    0.001 * diag(3 * (3 + 1) / 2)
  )
)
SimCovN(
  n = n,
  sigma = sigma,
  vcov_sigma_vech_l = vcov_sigma_vech_l
)
#> [[1]]
#>           [,1]      [,2]      [,3]
#> [1,] 1.0029123 0.5383928 0.2692902
#> [2,] 0.5383928 0.9895781 0.3453100
#> [3,] 0.2692902 0.3453100 0.9995039
#> 
#> [[2]]
#>           [,1]      [,2]      [,3]
#> [1,] 0.9295634 0.4846987 0.3474894
#> [2,] 0.4846987 0.9983945 0.4595277
#> [3,] 0.3474894 0.4595277 1.0006539
#> 
#> [[3]]
#>           [,1]      [,2]      [,3]
#> [1,] 1.0065987 0.4467486 0.3189609
#> [2,] 0.4467486 1.0348725 0.4249734
#> [3,] 0.3189609 0.4249734 1.0172577
#> 
#> [[4]]
#>           [,1]      [,2]      [,3]
#> [1,] 0.9286775 0.5125949 0.2977968
#> [2,] 0.5125949 0.9706371 0.4288240
#> [3,] 0.2977968 0.4288240 0.9096426
#> 
#> [[5]]
#>           [,1]      [,2]      [,3]
#> [1,] 1.0213520 0.5645402 0.2742215
#> [2,] 0.5645402 0.9844067 0.4289707
#> [3,] 0.2742215 0.4289707 1.0065213
#> 
#> [[6]]
#>           [,1]      [,2]      [,3]
#> [1,] 1.0237028 0.4572151 0.2597157
#> [2,] 0.4572151 0.9643785 0.3954098
#> [3,] 0.2597157 0.3954098 0.9975186
#> 
#> [[7]]
#>           [,1]      [,2]      [,3]
#> [1,] 1.0238913 0.4484577 0.2898989
#> [2,] 0.4484577 1.0065099 0.4439838
#> [3,] 0.2898989 0.4439838 0.9841791
#> 
#> [[8]]
#>           [,1]      [,2]      [,3]
#> [1,] 1.0070641 0.4440822 0.2471224
#> [2,] 0.4440822 1.0097071 0.3638273
#> [3,] 0.2471224 0.3638273 1.0174320
#> 
#> [[9]]
#>           [,1]      [,2]      [,3]
#> [1,] 1.0290599 0.4734041 0.2846425
#> [2,] 0.4734041 0.9959778 0.4286602
#> [3,] 0.2846425 0.4286602 0.9919400
#> 
#> [[10]]
#>           [,1]      [,2]      [,3]
#> [1,] 0.9803007 0.5068224 0.3009730
#> [2,] 0.5068224 1.0796971 0.3974445
#> [3,] 0.3009730 0.3974445 1.0318089
#>