Simulate Data using a State Space Model Parameterization for n > 1 Individuals (Varying Parameters)

This function simulates data using a state space model parameterization for n > 1 individuals. In this model, the parameters can vary across individuals.

Usage

SimSSM0Vary(
  n,
  mu0,
  sigma0_sqrt,
  alpha,
  beta,
  psi_sqrt,
  nu,
  lambda,
  theta_sqrt,
  time,
  burn_in
)

Arguments

n

Positive integer. Number of individuals.

mu0

List of numeric vectors. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_sqrt

List of numeric matrices. Cholesky decomposition of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

alpha

List of numeric vectors. Vector of intercepts for the dynamic model (\(\boldsymbol{\alpha}\)).

beta

List of numeric matrices. Transition matrix relating the values of the latent variables at time t - 1 to those at time t (\(\boldsymbol{\beta}\)).

psi_sqrt

List of numeric matrices. Cholesky decomposition of the process noise covariance matrix (\(\boldsymbol{\Psi}\)).

nu

List of numeric vectors. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

List of numeric matrices. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta_sqrt

List of numeric matrices. Cholesky decomposition of the measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

time: A vector of discrete time points from 1 to t.

id: A vector of ID numbers of length t.

n: Number of individuals.

Details

Parameters can vary across individuals by providing a list of parameter values. If the length of any of the parameters (mu0, sigma0_sqrt, alpha, beta, psi_sqrt, nu, lambda, and theta_sqrt) is less the n, the function will cycle through the available values.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Shumway, R. H., & Stoffer, D. S. (2017). Time series analysis and its applications: With R examples. Springer International Publishing. doi:10.1007/978-3-319-52452-8

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
# In this example, beta varies across individuals
set.seed(42)
k <- p <- 3
iden <- diag(k)
iden_sqrt <- chol(iden)
null_vec <- rep(x = 0, times = k)
n <- 5
mu0 <- list(null_vec)
sigma0_sqrt <- list(iden_sqrt)
alpha <- list(null_vec)
beta <- list(
  diag(x = 0.1, nrow = k),
  diag(x = 0.2, nrow = k),
  diag(x = 0.3, nrow = k),
  diag(x = 0.4, nrow = k),
  diag(x = 0.5, nrow = k)
)
psi_sqrt <- list(iden_sqrt)
nu <- list(null_vec)
lambda <- list(iden)
theta_sqrt <- list(chol(diag(x = 0.50, nrow = k)))
time <- 50
burn_in <- 0

ssm <- SimSSM0Vary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:3] -0.245 -0.436 -1.023 -2.457 -0.204 ...
#>   ..$ eta : num [1:50, 1:3] -0.532 -0.393 -0.886 -1.035 0.723 ...
#>   ..$ time: num [1:50, 1] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ id  : int [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:3] -1.959 0.696 0.72 0.731 -1.144 ...
#>   ..$ eta : num [1:50, 1:3] -1.846 0.796 0.233 0.456 -1.22 ...
#>   ..$ time: num [1:50, 1] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ id  : int [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:3] -1.0675 -0.2109 2.5609 -0.0386 -0.5492 ...
#>   ..$ eta : num [1:50, 1:3] -0.0319 0.2846 2.2576 0.7544 -0.6037 ...
#>   ..$ time: num [1:50, 1] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ id  : int [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:3] -0.93 -1.268 -0.816 -0.594 -0.355 ...
#>   ..$ eta : num [1:50, 1:3] -0.37 -1.067 0.331 -0.351 -0.501 ...
#>   ..$ time: num [1:50, 1] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ id  : int [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:3] 0.00235 -0.20966 0.33209 0.44063 1.29961 ...
#>   ..$ eta : num [1:50, 1:3] 0.337 0.787 -0.653 -1.09 0.958 ...
#>   ..$ time: num [1:50, 1] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ id  : int [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...