Simulate Data from a Linear Growth Curve Model (Individual-Varying Parameters)

This function simulates data from a linear growth curve model for n > 1 individuals. In this model, the parameters can vary across individuals.

Usage

SimSSMLinGrowthIVary(
  n,
  mu0,
  sigma0,
  theta,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  time
)

Arguments

n

Positive integer. Number of individuals.

mu0

A list of numeric vectors. Each element of the list is a vector of length two. The first element is the mean of the intercept, and the second element is the mean of the slope.

sigma0

A list of numeric matrices. Each element of the list is the covariance matrix of the intercept and the slope.

theta

A list numeric values. Each element of the list is the common measurement error variance.

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to y at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables (intercept and slope) at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

A list of length n of numeric matrices. Each element of the list is a matrix of observed covariates in type = 1 or type = 2. The number of rows in each matrix should be equal to time.

type

Integer. State space model type.

time

Positive integer. Number of time points to simulate.

Value

Returns a list with the following elements:

call: Function call.

args: Function arguments.

data: Generated data which is a list of length n.

fun: Function used.

Each element of data is a list with the following elements:

id: A vector of ID numbers of length t.

time: A vector of discrete time points from 0 to t - 1.

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

Details

Parameters can vary across individuals by providing a list of parameter values. If the length of any of the parameters (mu0, sigma0, mu, theta, gamma_y, or gamma_eta) is less the n, the function will cycle through the available values.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
# In this example, the mean vector of the intercept and slope vary.
# Specifically,
# there are two sets of values representing two latent classes.
set.seed(42)
n <- 10
mu0_1 <- c(0.615, 1.006) # lower starting point, higher growth
mu0_2 <- c(1.000, 0.500) # higher starting point, lower growth
mu0 <- list(mu0_1, mu0_2)
sigma0 <- list(
  matrix(
    data = c(
      1.932,
      0.618,
      0.618,
      0.587
    ),
    nrow = 2
  )
)
theta <- list(0.6)
time <- 10
gamma_y <- list(matrix(data = 0.10, nrow = 1, ncol = 2))
gamma_eta <- list(matrix(data = 0.10, nrow = 2, ncol = 2))
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = 2 * time),
        ncol = 2
      )
    )
  }
)

# Type 0
SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  theta = theta,
  type = 0,
  time = time
)
#>        id time           y1
#>   [1,]  1    0  1.446828794
#>   [2,]  1    1  0.265832953
#>   [3,]  1    2  0.885457154
#>   [4,]  1    3  1.759457886
#>   [5,]  1    4  3.208233036
#>   [6,]  1    5  3.569153966
#>   [7,]  1    6  4.867919022
#>   [8,]  1    7  4.987793082
#>   [9,]  1    8  6.659625543
#>  [10,]  1    9  7.086521651
#>  [11,]  2    0  4.462830841
#>  [12,]  2    1  5.104637158
#>  [13,]  2    2  6.301068415
#>  [14,]  2    3  7.382545421
#>  [15,]  2    4  9.746663664
#>  [16,]  2    5 10.743721926
#>  [17,]  2    6 11.044543992
#>  [18,]  2    7 13.675923157
#>  [19,]  2    8 15.562694155
#>  [20,]  2    9 15.867421211
#>  [21,]  3    0  1.780046577
#>  [22,]  3    1  3.714089652
#>  [23,]  3    2  6.700330453
#>  [24,]  3    3  8.417742240
#>  [25,]  3    4 10.845229310
#>  [26,]  3    5 13.281923787
#>  [27,]  3    6 15.132063291
#>  [28,]  3    7 18.111943245
#>  [29,]  3    8 19.745361706
#>  [30,]  3    9 21.265618922
#>  [31,]  4    0  2.648501520
#>  [32,]  4    1  3.467242739
#>  [33,]  4    2  4.879707699
#>  [34,]  4    3  4.809801196
#>  [35,]  4    4  6.504870523
#>  [36,]  4    5  6.341021883
#>  [37,]  4    6  7.446747315
#>  [38,]  4    7  8.408988440
#>  [39,]  4    8  9.383986198
#>  [40,]  4    9  9.478374505
#>  [41,]  5    0 -0.642785450
#>  [42,]  5    1 -1.290516844
#>  [43,]  5    2 -2.093990442
#>  [44,]  5    3  0.230494583
#>  [45,]  5    4 -0.628377360
#>  [46,]  5    5 -0.014801876
#>  [47,]  5    6  0.557883296
#>  [48,]  5    7 -0.390213583
#>  [49,]  5    8 -0.294400261
#>  [50,]  5    9 -0.008011172
#>  [51,]  6    0  1.389888864
#>  [52,]  6    1  1.353084557
#>  [53,]  6    2 -0.332644197
#>  [54,]  6    3  0.572360504
#>  [55,]  6    4  0.298331668
#>  [56,]  6    5 -0.009000885
#>  [57,]  6    6 -0.123862094
#>  [58,]  6    7 -1.890050128
#>  [59,]  6    8 -0.500751584
#>  [60,]  6    9 -1.354813882
#>  [61,]  7    0  0.441908472
#>  [62,]  7    1  3.391417549
#>  [63,]  7    2  2.277688923
#>  [64,]  7    3  3.952283473
#>  [65,]  7    4  2.270595768
#>  [66,]  7    5  3.596885693
#>  [67,]  7    6  4.027575579
#>  [68,]  7    7  6.008156350
#>  [69,]  7    8  5.819168973
#>  [70,]  7    9  5.436869417
#>  [71,]  8    0  1.766198340
#>  [72,]  8    1  1.857902893
#>  [73,]  8    2  0.060807984
#>  [74,]  8    3 -3.030858567
#>  [75,]  8    4 -3.344924617
#>  [76,]  8    5 -1.208030967
#>  [77,]  8    6 -3.065781384
#>  [78,]  8    7 -3.664380945
#>  [79,]  8    8 -4.823121598
#>  [80,]  8    9 -5.509158077
#>  [81,]  9    0 -1.760397325
#>  [82,]  9    1  1.795981320
#>  [83,]  9    2  0.768906306
#>  [84,]  9    3  1.683818627
#>  [85,]  9    4  3.612700310
#>  [86,]  9    5  3.374502357
#>  [87,]  9    6  5.647805552
#>  [88,]  9    7  5.729849482
#>  [89,]  9    8  7.141955214
#>  [90,]  9    9  8.017086409
#>  [91,] 10    0  1.558812578
#>  [92,] 10    1  1.604560756
#>  [93,] 10    2 -0.820251335
#>  [94,] 10    3 -1.879686743
#>  [95,] 10    4 -2.392558075
#>  [96,] 10    5 -3.827834508
#>  [97,] 10    6 -4.887399910
#>  [98,] 10    7 -3.972721823
#>  [99,] 10    8 -5.468339729
#> [100,] 10    9 -7.095880453

# Type 1
SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  theta = theta,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  time = time
)
#>        id time         y1           x1            x2
#>   [1,]  1    0 -0.8540579  1.370958447  1.304870e+00
#>   [2,]  1    1 -0.3039822 -0.564698171  2.286645e+00
#>   [3,]  1    2  1.7054477  0.363128411 -1.388861e+00
#>   [4,]  1    3  2.2505514  0.632862605 -2.787888e-01
#>   [5,]  1    4  4.9273997  0.404268323 -1.333213e-01
#>   [6,]  1    5  4.8214032 -0.106124516  6.359504e-01
#>   [7,]  1    6  5.6693302  1.511521997 -2.842529e-01
#>   [8,]  1    7  6.6343079 -0.094659038 -2.656455e+00
#>   [9,]  1    8  8.0734270  2.018423714 -2.440467e+00
#>  [10,]  1    9  9.5755720 -0.062714099  1.320113e+00
#>  [11,]  2    0  0.5405206 -0.306638594  4.554501e-01
#>  [12,]  2    1 -0.5249165 -1.781308434  7.048373e-01
#>  [13,]  2    2 -0.7195642 -0.171917356  1.035104e+00
#>  [14,]  2    3  0.7211198  1.214674699 -6.089264e-01
#>  [15,]  2    4  2.7343883  1.895193461  5.049551e-01
#>  [16,]  2    5  1.9803674 -0.430469132 -1.717009e+00
#>  [17,]  2    6  2.8112614 -0.257269383 -7.844590e-01
#>  [18,]  2    7  2.4257039 -1.763163085 -8.509076e-01
#>  [19,]  2    8  2.6336702  0.460097355 -2.414208e+00
#>  [20,]  2    9  2.9148115 -0.639994876  3.612261e-02
#>  [21,]  3    0 -3.0991132  0.205998600  3.219253e-01
#>  [22,]  3    1 -2.7605804 -0.361057299 -7.838389e-01
#>  [23,]  3    2 -3.6356736  0.758163236  1.575728e+00
#>  [24,]  3    3 -4.2005811 -0.726704827  6.428993e-01
#>  [25,]  3    4 -3.4439972 -1.368281044  8.976065e-02
#>  [26,]  3    5 -3.6193090  0.432818026  2.765507e-01
#>  [27,]  3    6 -4.4287284 -0.811393176  6.792888e-01
#>  [28,]  3    7 -3.9940270  1.444101262  8.983289e-02
#>  [29,]  3    8 -4.7490089 -0.431446203 -2.993090e+00
#>  [30,]  3    9 -5.8829719  0.655647883  2.848830e-01
#>  [31,]  4    0 -0.1749907 -0.367234643 -1.043119e+00
#>  [32,]  4    1 -0.1345276  0.185230565 -9.018639e-02
#>  [33,]  4    2  0.1902095  0.581823727  6.235182e-01
#>  [34,]  4    3 -0.5028283  1.399736827 -9.535234e-01
#>  [35,]  4    4 -3.4746297 -0.727292059 -5.428288e-01
#>  [36,]  4    5 -2.0989519  1.302542632  5.809965e-01
#>  [37,]  4    6 -2.5525575  0.335848120  7.681787e-01
#>  [38,]  4    7 -1.3200290  1.038506099  4.637676e-01
#>  [39,]  4    8 -2.4500255  0.920728568 -8.857763e-01
#>  [40,]  4    9 -2.1264322  0.720878163 -1.099781e+00
#>  [41,]  5    0 -1.1994105  1.512707010  1.392116e+00
#>  [42,]  5    1  0.9076637  0.257921438 -4.761739e-01
#>  [43,]  5    2  1.9321668  0.088440229  6.503486e-01
#>  [44,]  5    3  2.5881072 -0.120896538  1.391110e+00
#>  [45,]  5    4  4.1844658 -1.194328895 -1.110789e+00
#>  [46,]  5    5  4.0805317  0.611996898 -8.607926e-01
#>  [47,]  5    6  6.2259802 -0.217139846 -1.131739e+00
#>  [48,]  5    7  7.0992685 -0.182756706 -1.459214e+00
#>  [49,]  5    8  6.5579687  0.933346329  7.998255e-02
#>  [50,]  5    9  9.0514436  0.821773111  6.532043e-01
#>  [51,]  6    0  2.7242830  1.200965376 -2.509255e-02
#>  [52,]  6    1  3.8900325  1.044751087  1.080727e-01
#>  [53,]  6    2  6.4007688 -1.003208647 -4.854352e-01
#>  [54,]  6    3  7.1724581  1.848481902 -5.042171e-01
#>  [55,]  6    4 10.9138089 -0.666773409 -1.661099e+00
#>  [56,]  6    5 10.4113319  0.105513812 -3.823337e-01
#>  [57,]  6    6 13.1747940 -0.422255882 -5.126503e-01
#>  [58,]  6    7 13.8229201 -0.122350172  2.701891e+00
#>  [59,]  6    8 16.0571335  0.188193035 -1.362116e+00
#>  [60,]  6    9 16.5262651  0.119160958  1.372562e-01
#>  [61,]  7    0  1.2191182 -1.493625067  5.676206e-01
#>  [62,]  7    1  0.4043738 -1.470435741 -4.928774e-01
#>  [63,]  7    2  2.4488822  0.124702386  6.288407e-05
#>  [64,]  7    3  3.1168901 -0.996639135  1.122890e+00
#>  [65,]  7    4  3.7516937 -0.001822614  1.439856e+00
#>  [66,]  7    5  3.9594847 -0.428258881 -1.097114e+00
#>  [67,]  7    6  4.7086381 -0.613671606 -1.173196e-01
#>  [68,]  7    7  5.3694665 -2.024677845  1.201498e+00
#>  [69,]  7    8  6.0241612 -1.224747950 -4.697296e-01
#>  [70,]  7    9  5.4399818  0.179516441 -5.246948e-02
#>  [71,]  8    0  0.1320784 -0.086107298 -4.069848e-02
#>  [72,]  8    1 -0.4813830 -0.887679018 -1.551545e+00
#>  [73,]  8    2 -0.4760116 -0.444684005  1.167170e+00
#>  [74,]  8    3 -0.7373041 -0.029444879 -2.736457e-01
#>  [75,]  8    4 -2.7545988 -0.413868849 -4.678453e-01
#>  [76,]  8    5 -0.6347585  1.113386023 -1.238252e+00
#>  [77,]  8    6 -2.2548629 -0.480992842 -7.762034e-03
#>  [78,]  8    7 -2.7121886 -0.433169033 -8.002822e-01
#>  [79,]  8    8 -2.1961884  0.696862577 -5.334923e-01
#>  [80,]  8    9 -3.5167530 -1.056368413  1.287675e+00
#>  [81,]  9    0  0.1310433 -0.175525870 -2.297781e-01
#>  [82,]  9    1  2.1819043 -1.071782384  8.366191e-01
#>  [83,]  9    2  2.4831638  0.163206882 -1.745056e+00
#>  [84,]  9    3  5.2944402 -0.362738416  1.689459e+00
#>  [85,]  9    4  5.9324218  0.590013548  8.647780e-01
#>  [86,]  9    5  8.0799657  1.432421928 -1.507760e-01
#>  [87,]  9    6  8.4668620 -0.992692511 -1.449007e+00
#>  [88,]  9    7  9.7843341  0.454650298  6.430087e-01
#>  [89,]  9    8 12.2509604  0.084898059  4.831939e-01
#>  [90,]  9    9 14.0278010  0.895565582 -6.355626e-03
#>  [91,] 10    0 -0.6184515  0.151455893  2.049586e-01
#>  [92,] 10    1  0.2839775 -0.584108970 -3.450880e-01
#>  [93,] 10    2  2.4835580  0.368806733  2.526117e-01
#>  [94,] 10    3  2.0644507  0.294654340 -1.294002e+00
#>  [95,] 10    4  1.0645266 -0.279259373 -9.591704e-01
#>  [96,] 10    5  2.1894677 -1.336236655  1.085775e+00
#>  [97,] 10    6  2.2907872  0.700748818  4.037749e-01
#>  [98,] 10    7  4.4449343  0.554196622  5.864875e-01
#>  [99,] 10    8  5.6869904 -0.836306593  1.815228e+00
#> [100,] 10    9  4.6796070 -1.594588162  1.288214e-01

# Type 2
SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  theta = theta,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  time = time
)
#>        id time          y1           x1            x2
#>   [1,]  1    0  0.35316117  1.370958447  1.304870e+00
#>   [2,]  1    1  4.68225901 -0.564698171  2.286645e+00
#>   [3,]  1    2  3.01646966  0.363128411 -1.388861e+00
#>   [4,]  1    3  5.82674595  0.632862605 -2.787888e-01
#>   [5,]  1    4  6.73754184  0.404268323 -1.333213e-01
#>   [6,]  1    5  7.96328100 -0.106124516  6.359504e-01
#>   [7,]  1    6  9.51272991  1.511521997 -2.842529e-01
#>   [8,]  1    7 10.14641557 -0.094659038 -2.656455e+00
#>   [9,]  1    8 11.96206946  2.018423714 -2.440467e+00
#>  [10,]  1    9 13.41029222 -0.062714099  1.320113e+00
#>  [11,]  2    0  1.13266159 -0.306638594  4.554501e-01
#>  [12,]  2    1  2.04246840 -1.781308434  7.048373e-01
#>  [13,]  2    2  1.31456682 -0.171917356  1.035104e+00
#>  [14,]  2    3  1.77729307  1.214674699 -6.089264e-01
#>  [15,]  2    4  1.57639482  1.895193461  5.049551e-01
#>  [16,]  2    5  0.24876766 -0.430469132 -1.717009e+00
#>  [17,]  2    6  1.62234906 -0.257269383 -7.844590e-01
#>  [18,]  2    7  1.34640103 -1.763163085 -8.509076e-01
#>  [19,]  2    8  0.87567465  0.460097355 -2.414208e+00
#>  [20,]  2    9  1.72866030 -0.639994876  3.612261e-02
#>  [21,]  3    0 -1.49999724  0.205998600  3.219253e-01
#>  [22,]  3    1 -1.31014080 -0.361057299 -7.838389e-01
#>  [23,]  3    2 -2.66002351  0.758163236  1.575728e+00
#>  [24,]  3    3 -0.93344856 -0.726704827  6.428993e-01
#>  [25,]  3    4 -1.75090923 -1.368281044  8.976065e-02
#>  [26,]  3    5 -1.33028438  0.432818026  2.765507e-01
#>  [27,]  3    6 -3.08634202 -0.811393176  6.792888e-01
#>  [28,]  3    7 -1.31616098  1.444101262  8.983289e-02
#>  [29,]  3    8 -2.44255484 -0.431446203 -2.993090e+00
#>  [30,]  3    9 -2.08218376  0.655647883  2.848830e-01
#>  [31,]  4    0 -0.68846380 -0.367234643 -1.043119e+00
#>  [32,]  4    1  1.84704870  0.185230565 -9.018639e-02
#>  [33,]  4    2 -0.09627845  0.581823727  6.235182e-01
#>  [34,]  4    3  1.35712582  1.399736827 -9.535234e-01
#>  [35,]  4    4  0.06045555 -0.727292059 -5.428288e-01
#>  [36,]  4    5  2.03277856  1.302542632  5.809965e-01
#>  [37,]  4    6  2.47305500  0.335848120  7.681787e-01
#>  [38,]  4    7  2.43003303  1.038506099  4.637676e-01
#>  [39,]  4    8  3.85118204  0.920728568 -8.857763e-01
#>  [40,]  4    9  3.95071462  0.720878163 -1.099781e+00
#>  [41,]  5    0  2.18117426  1.512707010  1.392116e+00
#>  [42,]  5    1  4.59093076  0.257921438 -4.761739e-01
#>  [43,]  5    2  3.83278551  0.088440229  6.503486e-01
#>  [44,]  5    3  7.09128052 -0.120896538  1.391110e+00
#>  [45,]  5    4  8.89056991 -1.194328895 -1.110789e+00
#>  [46,]  5    5  9.66927523  0.611996898 -8.607926e-01
#>  [47,]  5    6 11.45905612 -0.217139846 -1.131739e+00
#>  [48,]  5    7 11.85204231 -0.182756706 -1.459214e+00
#>  [49,]  5    8 15.09182163  0.933346329  7.998255e-02
#>  [50,]  5    9 15.94112736  0.821773111  6.532043e-01
#>  [51,]  6    0  3.11666048  1.200965376 -2.509255e-02
#>  [52,]  6    1  3.78074013  1.044751087  1.080727e-01
#>  [53,]  6    2  3.49501935 -1.003208647 -4.854352e-01
#>  [54,]  6    3  4.27452714  1.848481902 -5.042171e-01
#>  [55,]  6    4  4.02924200 -0.666773409 -1.661099e+00
#>  [56,]  6    5  4.02435837  0.105513812 -3.823337e-01
#>  [57,]  6    6  3.26795803 -0.422255882 -5.126503e-01
#>  [58,]  6    7  3.81464739 -0.122350172  2.701891e+00
#>  [59,]  6    8  4.05406312  0.188193035 -1.362116e+00
#>  [60,]  6    9  2.64301258  0.119160958  1.372562e-01
#>  [61,]  7    0 -0.86325469 -1.493625067  5.676206e-01
#>  [62,]  7    1  0.98030544 -1.470435741 -4.928774e-01
#>  [63,]  7    2  2.94072441  0.124702386  6.288407e-05
#>  [64,]  7    3  2.14015740 -0.996639135  1.122890e+00
#>  [65,]  7    4  6.67010784 -0.001822614  1.439856e+00
#>  [66,]  7    5  8.38940606 -0.428258881 -1.097114e+00
#>  [67,]  7    6  7.81408188 -0.613671606 -1.173196e-01
#>  [68,]  7    7 10.14537694 -2.024677845  1.201498e+00
#>  [69,]  7    8 13.37702561 -1.224747950 -4.697296e-01
#>  [70,]  7    9 13.03690813  0.179516441 -5.246948e-02
#>  [71,]  8    0  0.88315466 -0.086107298 -4.069848e-02
#>  [72,]  8    1  1.10768138 -0.887679018 -1.551545e+00
#>  [73,]  8    2  2.25029856 -0.444684005  1.167170e+00
#>  [74,]  8    3  2.91980728 -0.029444879 -2.736457e-01
#>  [75,]  8    4  0.81380485 -0.413868849 -4.678453e-01
#>  [76,]  8    5  3.06679266  1.113386023 -1.238252e+00
#>  [77,]  8    6  4.01614759 -0.480992842 -7.762034e-03
#>  [78,]  8    7  3.49527397 -0.433169033 -8.002822e-01
#>  [79,]  8    8  4.45996253  0.696862577 -5.334923e-01
#>  [80,]  8    9  5.29153683 -1.056368413  1.287675e+00
#>  [81,]  9    0 -1.79308920 -0.175525870 -2.297781e-01
#>  [82,]  9    1 -1.85719059 -1.071782384  8.366191e-01
#>  [83,]  9    2 -4.97027196  0.163206882 -1.745056e+00
#>  [84,]  9    3 -4.98035536 -0.362738416  1.689459e+00
#>  [85,]  9    4 -4.15601440  0.590013548  8.647780e-01
#>  [86,]  9    5 -7.02756753  1.432421928 -1.507760e-01
#>  [87,]  9    6 -7.09248408 -0.992692511 -1.449007e+00
#>  [88,]  9    7 -6.57318773  0.454650298  6.430087e-01
#>  [89,]  9    8 -7.66529900  0.084898059  4.831939e-01
#>  [90,]  9    9 -7.59212655  0.895565582 -6.355626e-03
#>  [91,] 10    0  0.93983552  0.151455893  2.049586e-01
#>  [92,] 10    1  2.18186749 -0.584108970 -3.450880e-01
#>  [93,] 10    2  1.40557637  0.368806733  2.526117e-01
#>  [94,] 10    3  2.55139487  0.294654340 -1.294002e+00
#>  [95,] 10    4  1.69568249 -0.279259373 -9.591704e-01
#>  [96,] 10    5  3.08455735 -1.336236655  1.085775e+00
#>  [97,] 10    6  2.75661320  0.700748818  4.037749e-01
#>  [98,] 10    7  4.61298825  0.554196622  5.864875e-01
#>  [99,] 10    8  4.57340429 -0.836306593  1.815228e+00
#> [100,] 10    9  5.13846873 -1.594588162  1.288214e-01