Simulate Data from the Ornstein–Uhlenbeck Model using a State Space Model Parameterization (n = 1)

This function simulates data from the Ornstein–Uhlenbeck model using a state space model parameterization. See details for more information.

Usage

SimSSMOU(
  mu0,
  sigma0,
  mu,
  phi,
  sigma,
  nu,
  lambda,
  theta,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  delta_t,
  time,
  burn_in = 0
)

Arguments

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0

Numeric matrix. The covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

mu

Numeric vector. The long-term mean or equilibrium level (\(\boldsymbol{\mu}\)).

phi

Numeric matrix. The rate of mean reversion, determining how quickly the variable returns to its mean (\(\boldsymbol{\Phi}\)).

sigma

Numeric matrix. The matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

nu

Numeric vector. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta

Numeric matrix. The measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the observed variables at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

Numeric matrix. The matrix of observed covariates in type = 1 or type = 2. The number of rows should be equal to time + burn_in.

type

Integer. State space model type.

delta_t

Numeric. Time interval (\(\delta_t\)).

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

time: A vector of continuous time points of length t starting from 0 with delta_t increments.

id: A vector of ones.

Details

Type 0

The measurement model is given by $$ \mathbf{y}_{t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{t} + \boldsymbol{\varepsilon}_{t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\mathbf{y}_{t}\), \(\boldsymbol{\eta}_{t}\), and \(\boldsymbol{\varepsilon}_{t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{t}\) is a vector of observed random variables, \(\boldsymbol{\eta}_{t}\) is a vector of latent random variables, and \(\boldsymbol{\varepsilon}_{t}\) is a vector of random measurement errors, at time \(t\). \(\boldsymbol{\nu}\) is a vector of intercepts, \(\boldsymbol{\Lambda}\) is a matrix of factor loadings, and \(\boldsymbol{\Theta}\) is the covariance matrix of \(\boldsymbol{\varepsilon}\).

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{t} \right) \mathrm{d}t + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{t} $$ where \(\boldsymbol{\mu}\) is the long-term mean or equilibrium level, \(\boldsymbol{\Phi}\) is the rate of mean reversion, determining how quickly the variable returns to its mean, \(\boldsymbol{\Sigma}\) is the matrix of volatility or randomness in the process, and \(\mathrm{d}\boldsymbol{W}\) is a Wiener process or Brownian motion, which represents random fluctuations.

Type 1

The measurement model is given by $$ \mathbf{y}_{t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{t} + \boldsymbol{\varepsilon}_{t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) . $$

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{t} \right) \mathrm{d}t + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{t} + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{t} $$ where \(\mathbf{x}_{t}\) is a vector of covariates at time \(t\), and \(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\) is the coefficient matrix linking the covariates to the latent variables.

Type 2

The measurement model is given by $$ \mathbf{y}_{t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{t} + \boldsymbol{\Gamma}_{\mathbf{y}} \mathbf{x}_{t} + \boldsymbol{\varepsilon}_{t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\boldsymbol{\Gamma}_{\mathbf{y}}\) is the coefficient matrix linking the covariates to the observed variables.

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{t} \right) \mathrm{d}t + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{t} + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{t} . $$

References

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Uhlenbeck, G. E., & Ornstein, L. S. (1930). On the theory of the brownian motion. Physical Review, 36(5), 823–841. doi:10.1103/physrev.36.823

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
p <- k <- 2
iden <- diag(p)
mu0 <- c(-3.0, 1.5)
sigma0 <- iden
mu <- c(5.76, 5.18)
phi <- matrix(data = c(0.10, -0.05, -0.05, 0.10), nrow = p)
sigma <- matrix(
  data = c(2.79, 0.06, 0.06, 3.27),
  nrow = p
)
nu <- rep(x = 0, times = k)
lambda <- diag(k)
theta <- diag(x = 0.50, nrow = k)
delta_t <- 0.10
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- 0.10 * diag(k)
x <- matrix(
  data = rnorm(n = k * (time + burn_in)),
  ncol = k
)

# Type 0
SimSSMOU(
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  type = 0,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>       id time          y1          y2
#>  [1,]  1  0.0 -1.16935506   1.4133361
#>  [2,]  1  0.1 -2.18801660   1.4167640
#>  [3,]  1  0.2 -2.54664204   2.4131708
#>  [4,]  1  0.3 -3.11756527   1.5312020
#>  [5,]  1  0.4 -2.37086795   0.2814600
#>  [6,]  1  0.5 -2.40184853  -0.7324589
#>  [7,]  1  0.6 -3.07086035  -1.2247631
#>  [8,]  1  0.7 -1.81374803  -2.6054593
#>  [9,]  1  0.8 -1.18462177  -1.7567735
#> [10,]  1  0.9 -3.23475706  -0.6361786
#> [11,]  1  1.0 -4.70922766  -1.5499605
#> [12,]  1  1.1 -3.11669490  -0.2598166
#> [13,]  1  1.2 -3.78881316  -1.7532555
#> [14,]  1  1.3 -3.38477391  -2.8110487
#> [15,]  1  1.4 -3.91877659  -3.1533039
#> [16,]  1  1.5 -4.53675819  -3.1618123
#> [17,]  1  1.6 -3.56008572  -4.2435208
#> [18,]  1  1.7 -3.76382107  -4.5749548
#> [19,]  1  1.8 -3.28880315  -4.0534850
#> [20,]  1  1.9 -4.70868015  -5.2888362
#> [21,]  1  2.0 -4.29590129  -6.4539873
#> [22,]  1  2.1 -4.14848374  -5.2905552
#> [23,]  1  2.2 -4.44836877  -5.7686819
#> [24,]  1  2.3 -5.97675314  -5.8962052
#> [25,]  1  2.4 -5.31383552  -4.2170666
#> [26,]  1  2.5 -4.93922534  -3.4287727
#> [27,]  1  2.6 -4.62843921  -4.4121571
#> [28,]  1  2.7 -5.20317835  -5.7421100
#> [29,]  1  2.8 -5.04860079  -6.1357398
#> [30,]  1  2.9 -6.68834440  -4.3464122
#> [31,]  1  3.0 -4.53786046  -5.6132265
#> [32,]  1  3.1 -4.22291078  -6.2746491
#> [33,]  1  3.2 -3.62541910  -6.5082472
#> [34,]  1  3.3 -4.59797763  -7.9775035
#> [35,]  1  3.4 -4.10175791  -8.0743193
#> [36,]  1  3.5 -3.27146112  -9.3535745
#> [37,]  1  3.6 -1.63457083  -7.5351297
#> [38,]  1  3.7 -0.79818457  -8.2315563
#> [39,]  1  3.8 -2.32639988  -8.5229444
#> [40,]  1  3.9 -1.58563308  -9.2295258
#> [41,]  1  4.0 -2.25188821  -8.2865123
#> [42,]  1  4.1 -2.16836326  -8.2300217
#> [43,]  1  4.2 -2.43883193  -9.1315338
#> [44,]  1  4.3 -1.99134044 -10.8979781
#> [45,]  1  4.4  0.02339915  -7.1852146
#> [46,]  1  4.5 -1.04308397 -10.6124375
#> [47,]  1  4.6 -1.07100869  -9.0681720
#> [48,]  1  4.7 -2.74356884  -9.0765051
#> [49,]  1  4.8 -1.47551138  -8.9020332
#> [50,]  1  4.9 -1.40984089  -8.8462819

# Type 1
SimSSMOU(
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>       id time        y1          y2          x1          x2
#>  [1,]  1  0.0 -4.486980   0.4548382  1.37095845  0.32192527
#>  [2,]  1  0.1 -4.348837  -0.3987406 -0.56469817 -0.78383894
#>  [3,]  1  0.2 -4.181234  -0.2924930  0.36312841  1.57572752
#>  [4,]  1  0.3 -4.156087  -2.1269253  0.63286260  0.64289931
#>  [5,]  1  0.4 -5.524872  -2.0928129  0.40426832  0.08976065
#>  [6,]  1  0.5 -3.627320  -2.6056291 -0.10612452  0.27655075
#>  [7,]  1  0.6 -2.711617  -2.5153563  1.51152200  0.67928882
#>  [8,]  1  0.7 -2.768170  -3.6103075 -0.09465904  0.08983289
#>  [9,]  1  0.8 -3.332572  -2.4707466  2.01842371 -2.99309008
#> [10,]  1  0.9 -2.955686  -4.1819015 -0.06271410  0.28488295
#> [11,]  1  1.0 -4.090082  -4.2775912  1.30486965 -0.36723464
#> [12,]  1  1.1 -3.488130  -5.2105153  2.28664539  0.18523056
#> [13,]  1  1.2 -3.303554  -3.2590996 -1.38886070  0.58182373
#> [14,]  1  1.3 -3.758702  -5.2920549 -0.27878877  1.39973683
#> [15,]  1  1.4 -3.591844  -5.3068083 -0.13332134 -0.72729206
#> [16,]  1  1.5 -3.356298  -5.5543881  0.63595040  1.30254263
#> [17,]  1  1.6 -3.146979  -4.0759746 -0.28425292  0.33584812
#> [18,]  1  1.7 -4.083799  -6.5783457 -2.65645542  1.03850610
#> [19,]  1  1.8 -2.996370  -6.1212842 -2.44046693  0.92072857
#> [20,]  1  1.9 -2.491677  -6.0942245  1.32011335  0.72087816
#> [21,]  1  2.0 -3.020759  -6.8570972 -0.30663859 -1.04311894
#> [22,]  1  2.1 -3.294469  -6.0960848 -1.78130843 -0.09018639
#> [23,]  1  2.2 -2.903113  -3.9215612 -0.17191736  0.62351816
#> [24,]  1  2.3 -3.217207  -3.9194355  1.21467470 -0.95352336
#> [25,]  1  2.4 -4.997895  -3.5465326  1.89519346 -0.54282881
#> [26,]  1  2.5 -3.962170  -4.2129417 -0.43046913  0.58099650
#> [27,]  1  2.6 -5.140759  -5.2620900 -0.25726938  0.76817874
#> [28,]  1  2.7 -4.121975  -6.6708021 -1.76316309  0.46376759
#> [29,]  1  2.8 -3.408035  -8.2051041  0.46009735 -0.88577630
#> [30,]  1  2.9 -6.760769  -5.4998764 -0.63999488 -1.09978090
#> [31,]  1  3.0 -4.539720  -7.7017336  0.45545012  1.51270701
#> [32,]  1  3.1 -4.965861  -6.8867760  0.70483734  0.25792144
#> [33,]  1  3.2 -5.275197  -7.8024249  1.03510352  0.08844023
#> [34,]  1  3.3 -4.307836  -8.4899596 -0.60892638 -0.12089654
#> [35,]  1  3.4 -6.420489  -9.5533907  0.50495512 -1.19432890
#> [36,]  1  3.5 -6.402629  -7.9046091 -1.71700868  0.61199690
#> [37,]  1  3.6 -5.581378  -8.1330329 -0.78445901 -0.21713985
#> [38,]  1  3.7 -5.070015  -5.5184394 -0.85090759 -0.18275671
#> [39,]  1  3.8 -4.925354  -7.7167758 -2.41420765  0.93334633
#> [40,]  1  3.9 -4.458966  -9.5558055  0.03612261  0.82177311
#> [41,]  1  4.0 -4.404884 -10.1667055  0.20599860  1.39211638
#> [42,]  1  4.1 -5.330683  -9.2484117 -0.36105730 -0.47617392
#> [43,]  1  4.2 -3.330383 -11.0770605  0.75816324  0.65034856
#> [44,]  1  4.3 -4.543964  -9.7508490 -0.72670483  1.39111046
#> [45,]  1  4.4 -5.644550 -10.6881168 -1.36828104 -1.11078888
#> [46,]  1  4.5 -4.898390  -8.7241605  0.43281803 -0.86079259
#> [47,]  1  4.6 -3.488804  -9.9098248 -0.81139318 -1.13173868
#> [48,]  1  4.7 -3.524775  -8.7936072  1.44410126 -1.45921400
#> [49,]  1  4.8 -2.007279  -9.1920326 -0.43144620  0.07998255
#> [50,]  1  4.9 -2.809219  -8.1389422  0.65564788  0.65320434

# Type 2
SimSSMOU(
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>       id time        y1          y2          x1          x2
#>  [1,]  1  0.0 -4.938989  1.48951814  1.37095845  0.32192527
#>  [2,]  1  0.1 -3.554163  2.05051456 -0.56469817 -0.78383894
#>  [3,]  1  0.2 -4.730767  1.21447777  0.36312841  1.57572752
#>  [4,]  1  0.3 -5.577692  1.10355470  0.63286260  0.64289931
#>  [5,]  1  0.4 -6.501135  1.17575763  0.40426832  0.08976065
#>  [6,]  1  0.5 -5.733021  3.38258822 -0.10612452  0.27655075
#>  [7,]  1  0.6 -6.520811  3.97195396  1.51152200  0.67928882
#>  [8,]  1  0.7 -5.117550  3.60591629 -0.09465904  0.08983289
#>  [9,]  1  0.8 -4.350999  3.13495703  2.01842371 -2.99309008
#> [10,]  1  0.9 -4.888976  3.27978073 -0.06271410  0.28488295
#> [11,]  1  1.0 -5.051612  2.39348096  1.30486965 -0.36723464
#> [12,]  1  1.1 -3.313074  2.86146115  2.28664539  0.18523056
#> [13,]  1  1.2 -5.840677  2.86029658 -1.38886070  0.58182373
#> [14,]  1  1.3 -4.437699  0.89788175 -0.27878877  1.39973683
#> [15,]  1  1.4 -3.482911  1.86867178 -0.13332134 -0.72729206
#> [16,]  1  1.5 -3.956748  2.83470796  0.63595040  1.30254263
#> [17,]  1  1.6 -2.399963  1.51989923 -0.28425292  0.33584812
#> [18,]  1  1.7 -3.421212  0.26562544 -2.65645542  1.03850610
#> [19,]  1  1.8 -4.215894  1.04764813 -2.44046693  0.92072857
#> [20,]  1  1.9 -2.565554 -0.79366803  1.32011335  0.72087816
#> [21,]  1  2.0 -3.746046  1.74952216 -0.30663859 -1.04311894
#> [22,]  1  2.1 -3.562824  0.54042643 -1.78130843 -0.09018639
#> [23,]  1  2.2 -3.772702  0.70613279 -0.17191736  0.62351816
#> [24,]  1  2.3 -3.091660  1.75806629  1.21467470 -0.95352336
#> [25,]  1  2.4 -3.966512  1.40235989  1.89519346 -0.54282881
#> [26,]  1  2.5 -4.872187  0.40890446 -0.43046913  0.58099650
#> [27,]  1  2.6 -5.279395  0.46701754 -0.25726938  0.76817874
#> [28,]  1  2.7 -6.957168  1.32650994 -1.76316309  0.46376759
#> [29,]  1  2.8 -5.256231  1.83870188  0.46009735 -0.88577630
#> [30,]  1  2.9 -5.927106  0.89074323 -0.63999488 -1.09978090
#> [31,]  1  3.0 -5.572817  0.95661986  0.45545012  1.51270701
#> [32,]  1  3.1 -5.786286  3.65170621  0.70483734  0.25792144
#> [33,]  1  3.2 -4.889870  2.49461697  1.03510352  0.08844023
#> [34,]  1  3.3 -5.011970  3.29929570 -0.60892638 -0.12089654
#> [35,]  1  3.4 -6.102349  2.15845889  0.50495512 -1.19432890
#> [36,]  1  3.5 -4.605171  2.35873996 -1.71700868  0.61199690
#> [37,]  1  3.6 -3.677634  2.05742771 -0.78445901 -0.21713985
#> [38,]  1  3.7 -4.412068  1.69498669 -0.85090759 -0.18275671
#> [39,]  1  3.8 -5.129336  0.02120983 -2.41420765  0.93334633
#> [40,]  1  3.9 -4.908096  0.50267557  0.03612261  0.82177311
#> [41,]  1  4.0 -5.287605  0.61979903  0.20599860  1.39211638
#> [42,]  1  4.1 -6.886223 -0.66064965 -0.36105730 -0.47617392
#> [43,]  1  4.2 -5.659392 -0.14702425  0.75816324  0.65034856
#> [44,]  1  4.3 -6.905464  3.22346992 -0.72670483  1.39111046
#> [45,]  1  4.4 -5.667024 -0.28801892 -1.36828104 -1.11078888
#> [46,]  1  4.5 -7.541485 -1.06576067  0.43281803 -0.86079259
#> [47,]  1  4.6 -7.864306  1.22952295 -0.81139318 -1.13173868
#> [48,]  1  4.7 -5.299626  1.56116980  1.44410126 -1.45921400
#> [49,]  1  4.8 -6.266284  1.29960619 -0.43144620  0.07998255
#> [50,]  1  4.9 -6.241770  2.13971921  0.65564788  0.65320434