Simulate Data from an Ornstein–Uhlenbeck Model using a State Space Model Parameterization for n > 1 Individuals (Fixed Parameters)

This function simulates data from an Ornstein–Uhlenbeck model using a state space model parameterization for n > 1 individuals. In this model, the parameters are invariant across individuals. See details for more information.

Usage

SimSSMOUFixed(
  n,
  mu0,
  sigma0,
  mu,
  phi,
  sigma,
  nu,
  lambda,
  theta,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  delta_t,
  time,
  burn_in = 0
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0

Numeric matrix. The covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

mu

Numeric vector. The long-term mean or equilibrium level (\(\boldsymbol{\mu}\)).

phi

Numeric matrix. The rate of mean reversion, determining how quickly the variable returns to its mean (\(\boldsymbol{\Phi}\)).

sigma

Numeric matrix. The matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

nu

Numeric vector. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta

Numeric matrix. The measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the observed variables at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

Numeric matrix. The matrix of observed covariates in type = 1 or type = 2. The number of rows should be equal to time + burn_in.

type

Integer. State space model type.

delta_t

Numeric. Time interval (\(\delta_t\)).

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

time: A vector of continuous time points of length t starting from 0 with delta_t increments.

id: A vector of ID numbers of length t.

n: Number of individuals.

Details

Type 0

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\mathbf{y}_{i, t}\), \(\boldsymbol{\eta}_{i, t}\), and \(\boldsymbol{\varepsilon}_{i, t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{i, t}\) is a vector of observed random variables, \(\boldsymbol{\eta}_{i, t}\) is a vector of latent random variables, and \(\boldsymbol{\varepsilon}_{i, t}\) is a vector of random measurement errors, at time \(t\) and individual \(i\). \(\boldsymbol{\nu}\) is a vector of intercepts, \(\boldsymbol{\Lambda}\) is a matrix of factor loadings, and \(\boldsymbol{\Theta}\) is the covariance matrix of \(\boldsymbol{\varepsilon}\).

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$ where \(\boldsymbol{\mu}\) is the long-term mean or equilibrium level, \(\boldsymbol{\Phi}\) is the rate of mean reversion, determining how quickly the variable returns to its mean, \(\boldsymbol{\Sigma}\) is the matrix of volatility or randomness in the process, and \(\mathrm{d}\boldsymbol{W}\) is a Wiener process or Brownian motion, which represents random fluctuations.

Type 1

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) . $$

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{i, t} + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$ where \(\mathbf{x}_{i, t}\) is a vector of covariates at time \(t\) and individual \(i\), and \(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\) is the coefficient matrix linking the covariates to the latent variables.

Type 2

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\Gamma}_{\mathbf{y}} \mathbf{x}_{i, t} + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\boldsymbol{\Gamma}_{\mathbf{y}}\) is the coefficient matrix linking the covariates to the observed variables.

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{i, t} + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} . $$

References

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Uhlenbeck, G. E., & Ornstein, L. S. (1930). On the theory of the brownian motion. Physical Review, 36(5), 823–841. doi:10.1103/physrev.36.823

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
p <- k <- 2
iden <- diag(p)
n <- 5
mu0 <- c(-3.0, 1.5)
sigma0 <- iden
mu <- c(5.76, 5.18)
phi <- matrix(data = c(0.10, -0.05, -0.05, 0.10), nrow = p)
sigma <- matrix(
  data = c(2.79, 0.06, 0.06, 3.27),
  nrow = p
)
nu <- rep(x = 0, times = k)
lambda <- diag(k)
theta <- diag(x = 0.50, nrow = k)
delta_t <- 0.10
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- 0.10 * diag(k)
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = k * (time + burn_in)),
        ncol = k
      )
    )
  }
)

# Type 0
SimSSMOUFixed(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  type = 0,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>        id time          y1            y2
#>   [1,]  1  0.0 -1.68501472  1.3603337626
#>   [2,]  1  0.1 -1.79019531  3.1274097312
#>   [3,]  1  0.2 -2.61516106  3.2892949408
#>   [4,]  1  0.3 -2.80210821  3.4871149208
#>   [5,]  1  0.4 -2.61537907  2.4206470101
#>   [6,]  1  0.5 -2.31875433  2.6897275062
#>   [7,]  1  0.6 -2.91564950  3.6116164162
#>   [8,]  1  0.7 -2.93411510  2.8532245654
#>   [9,]  1  0.8 -3.48512401  3.1383512926
#>  [10,]  1  0.9 -3.28826544  3.9945384452
#>  [11,]  1  1.0 -4.34348305  5.0663617932
#>  [12,]  1  1.1 -3.40357262  5.7014280458
#>  [13,]  1  1.2 -3.40957112  5.6233424442
#>  [14,]  1  1.3 -2.32033809  4.6185965738
#>  [15,]  1  1.4 -2.85349222  4.8587084562
#>  [16,]  1  1.5 -2.66155123  5.3476856931
#>  [17,]  1  1.6 -1.92846943  5.2670784006
#>  [18,]  1  1.7 -3.78580999  4.5323085819
#>  [19,]  1  1.8 -3.61240778  5.2887924016
#>  [20,]  1  1.9 -3.73907177  6.1559798542
#>  [21,]  1  2.0 -2.16588345  4.7880623232
#>  [22,]  1  2.1 -1.57243733  3.6217927730
#>  [23,]  1  2.2 -1.10782594  4.1958805322
#>  [24,]  1  2.3 -0.63670794  2.9447581447
#>  [25,]  1  2.4 -1.33257775  3.4239461836
#>  [26,]  1  2.5 -1.20676250  4.6987671364
#>  [27,]  1  2.6 -2.65579868  6.3792802009
#>  [28,]  1  2.7 -2.69435067  4.1277364756
#>  [29,]  1  2.8 -4.65265286  5.9166962184
#>  [30,]  1  2.9 -1.60655132  3.0529886292
#>  [31,]  1  3.0 -3.72981038  4.3690874571
#>  [32,]  1  3.1 -2.46252961  4.1712763129
#>  [33,]  1  3.2 -2.08010958  4.3722788046
#>  [34,]  1  3.3 -1.77696435  2.5025614017
#>  [35,]  1  3.4 -3.34150964  2.9884369783
#>  [36,]  1  3.5 -1.18290161  5.8063626608
#>  [37,]  1  3.6 -1.36769767  4.2436737619
#>  [38,]  1  3.7 -2.40543659  3.0253130463
#>  [39,]  1  3.8  0.32971309  3.3900902169
#>  [40,]  1  3.9 -1.62662290  3.7279864061
#>  [41,]  1  4.0 -1.61327044  3.3531396345
#>  [42,]  1  4.1  0.07929040  2.7840669906
#>  [43,]  1  4.2  0.40636977  3.5060764495
#>  [44,]  1  4.3  0.22185549  3.8927196726
#>  [45,]  1  4.4  0.68440042  3.0714623948
#>  [46,]  1  4.5  0.80898300  4.0339468057
#>  [47,]  1  4.6  1.03738244  4.5463357080
#>  [48,]  1  4.7  1.91310309  4.3191533298
#>  [49,]  1  4.8  2.80596663  3.5033196895
#>  [50,]  1  4.9  2.74259472  5.6913964559
#>  [51,]  2  0.0 -3.59775671  1.6109025823
#>  [52,]  2  0.1 -1.22235325  0.5381468957
#>  [53,]  2  0.2 -3.16945262  0.9899326348
#>  [54,]  2  0.3 -1.98673865  2.2080035600
#>  [55,]  2  0.4 -1.24354409  0.7460726212
#>  [56,]  2  0.5 -2.60369078  0.6184088455
#>  [57,]  2  0.6 -2.22167630  1.0329189725
#>  [58,]  2  0.7 -1.69883622  0.6074484564
#>  [59,]  2  0.8 -0.24470186 -0.8452994461
#>  [60,]  2  0.9 -0.97485469 -0.0585576735
#>  [61,]  2  1.0 -0.93516648 -0.4313354775
#>  [62,]  2  1.1 -2.07331708 -2.1348382909
#>  [63,]  2  1.2 -0.58853424 -0.8251156837
#>  [64,]  2  1.3 -2.50646378  0.0015357796
#>  [65,]  2  1.4 -1.82021734 -0.8311198485
#>  [66,]  2  1.5 -1.32255328 -0.2278525284
#>  [67,]  2  1.6 -0.28065437 -0.2334332424
#>  [68,]  2  1.7 -0.03713787  1.4311481306
#>  [69,]  2  1.8  0.94031502  0.7142014596
#>  [70,]  2  1.9 -1.10685148  0.5932099749
#>  [71,]  2  2.0 -0.85032715 -0.1914124818
#>  [72,]  2  2.1 -1.70420754  1.4274301024
#>  [73,]  2  2.2 -0.55557061  1.2669522847
#>  [74,]  2  2.3 -2.22638146  1.6322229883
#>  [75,]  2  2.4 -1.19791470  0.6926967144
#>  [76,]  2  2.5 -1.25572391  1.1798974935
#>  [77,]  2  2.6 -1.57359679  1.0320443730
#>  [78,]  2  2.7 -1.24588918  1.2037357211
#>  [79,]  2  2.8 -1.97542304  0.8209940795
#>  [80,]  2  2.9 -0.51484190 -0.4949307945
#>  [81,]  2  3.0 -0.35580495 -1.1586490737
#>  [82,]  2  3.1 -1.30802184 -0.2025813502
#>  [83,]  2  3.2 -1.61665707 -0.6638205092
#>  [84,]  2  3.3 -1.94326117  0.3080963074
#>  [85,]  2  3.4 -2.86318065 -0.7551282501
#>  [86,]  2  3.5 -2.86683294 -1.1584389377
#>  [87,]  2  3.6 -3.98136315 -0.1867099843
#>  [88,]  2  3.7 -3.06106840 -2.2682272680
#>  [89,]  2  3.8 -4.23453898 -0.3364820997
#>  [90,]  2  3.9 -2.65363036 -0.0008997621
#>  [91,]  2  4.0 -1.40555425 -1.0291851942
#>  [92,]  2  4.1 -2.30261610  0.3972149114
#>  [93,]  2  4.2 -1.70605384 -0.6290539578
#>  [94,]  2  4.3 -3.84118380 -2.9589937468
#>  [95,]  2  4.4 -2.78997383 -1.5172659821
#>  [96,]  2  4.5 -2.84036675 -2.8521276463
#>  [97,]  2  4.6 -2.99905006 -1.5736614085
#>  [98,]  2  4.7 -4.05293501 -1.4474621979
#>  [99,]  2  4.8 -4.80651258 -2.2233390059
#> [100,]  2  4.9 -3.52111114  0.0893284311
#> [101,]  3  0.0 -4.10679982  1.6580502425
#> [102,]  3  0.1 -3.23962562 -0.3364508204
#> [103,]  3  0.2 -2.58049501 -1.8624981705
#> [104,]  3  0.3 -1.65918925 -0.8667337890
#> [105,]  3  0.4 -2.92196834 -0.5331705673
#> [106,]  3  0.5 -2.05972773  2.2001595680
#> [107,]  3  0.6 -3.30458305  2.9397474910
#> [108,]  3  0.7 -2.14288344  2.9161666561
#> [109,]  3  0.8 -3.22191509  2.4668983333
#> [110,]  3  0.9 -2.73857102  2.2523379091
#> [111,]  3  1.0 -1.72432572  1.6016344264
#> [112,]  3  1.1 -3.77222895  2.8979617815
#> [113,]  3  1.2 -3.33445763  3.8442412261
#> [114,]  3  1.3 -2.93104412  4.1743601136
#> [115,]  3  1.4 -3.64910418  5.5413491253
#> [116,]  3  1.5 -3.82179604  2.6509131245
#> [117,]  3  1.6 -3.23912711  2.4943370501
#> [118,]  3  1.7 -1.46863150  1.4344838386
#> [119,]  3  1.8 -4.85998035  2.3431955525
#> [120,]  3  1.9 -4.87436056  2.1665451410
#> [121,]  3  2.0 -3.04106067  2.0636791002
#> [122,]  3  2.1 -2.88430555  3.4743803908
#> [123,]  3  2.2 -4.32659124  2.3945712565
#> [124,]  3  2.3 -2.28902743  4.0374149484
#> [125,]  3  2.4 -2.17046099  4.3213856511
#> [126,]  3  2.5 -2.56522337  2.6609828610
#> [127,]  3  2.6 -1.64649592  3.3311378737
#> [128,]  3  2.7 -2.20602523  2.4962834041
#> [129,]  3  2.8 -2.45015674  2.2726275600
#> [130,]  3  2.9 -2.19000418 -0.4298554287
#> [131,]  3  3.0 -2.68305573  1.6610036942
#> [132,]  3  3.1 -2.02374513  1.3391881848
#> [133,]  3  3.2 -1.50975161  1.4910904097
#> [134,]  3  3.3 -1.10628584  0.5890923849
#> [135,]  3  3.4 -1.31940716  1.1093188814
#> [136,]  3  3.5 -2.56997252  0.2008466703
#> [137,]  3  3.6 -1.89358529  0.9999627763
#> [138,]  3  3.7 -2.52738550 -0.1230331771
#> [139,]  3  3.8 -0.58868604 -1.0210606468
#> [140,]  3  3.9 -2.10431131  0.9752098187
#> [141,]  3  4.0 -2.54239546  1.1155323698
#> [142,]  3  4.1 -2.31467610  0.1489044754
#> [143,]  3  4.2 -2.33021867  2.1788209724
#> [144,]  3  4.3 -0.88006056  0.0179958033
#> [145,]  3  4.4 -2.36125749  0.3705400487
#> [146,]  3  4.5 -3.04820779  2.8112681860
#> [147,]  3  4.6 -3.03702597  1.4192044216
#> [148,]  3  4.7 -3.94872829  2.7883512883
#> [149,]  3  4.8 -2.85412748  0.9788086668
#> [150,]  3  4.9 -2.32690719  0.4792198667
#> [151,]  4  0.0 -4.81711952  1.1933375842
#> [152,]  4  0.1 -4.40663209  2.3899028315
#> [153,]  4  0.2 -2.36187372  0.4695140778
#> [154,]  4  0.3 -4.31738332 -0.0681994070
#> [155,]  4  0.4 -4.92313145 -0.8876880556
#> [156,]  4  0.5 -3.60262032  0.5028087470
#> [157,]  4  0.6 -2.26512527  1.9050469756
#> [158,]  4  0.7 -1.85647024  1.1297609500
#> [159,]  4  0.8 -2.12697279  3.2438583924
#> [160,]  4  0.9 -1.66420654  3.0375281785
#> [161,]  4  1.0 -1.51108001  2.0236438292
#> [162,]  4  1.1 -0.76864732  1.8795053006
#> [163,]  4  1.2 -1.07665636  3.1397079578
#> [164,]  4  1.3 -1.62337149  2.0790404054
#> [165,]  4  1.4 -0.21450154  5.3572524036
#> [166,]  4  1.5 -1.05068097  4.3946259179
#> [167,]  4  1.6  0.13532568  5.4271258744
#> [168,]  4  1.7  0.15852453  5.8554074768
#> [169,]  4  1.8  1.39625028  5.6689860573
#> [170,]  4  1.9  2.46739816  5.7244556922
#> [171,]  4  2.0  1.60860082  6.2809191279
#> [172,]  4  2.1  3.06721999  5.5964381810
#> [173,]  4  2.2  2.64084841  5.8941450647
#> [174,]  4  2.3  2.73329312  5.2697542209
#> [175,]  4  2.4  1.93838801  3.5348562456
#> [176,]  4  2.5  2.07063262  3.5847482565
#> [177,]  4  2.6  1.87218307  2.8605075948
#> [178,]  4  2.7  2.74176219  3.3705999480
#> [179,]  4  2.8  2.49581811  2.2652897458
#> [180,]  4  2.9  3.09218688  1.7048203652
#> [181,]  4  3.0  1.75051493  1.9509886926
#> [182,]  4  3.1  2.35404700  1.4380131044
#> [183,]  4  3.2  3.64913584  1.2736221504
#> [184,]  4  3.3  4.06656201  2.2289220166
#> [185,]  4  3.4  4.22779598  2.5178815461
#> [186,]  4  3.5  3.82563075  1.8974138115
#> [187,]  4  3.6  3.94786022  1.4330345433
#> [188,]  4  3.7  4.62094770  0.4993369593
#> [189,]  4  3.8  5.20415914  1.7940384204
#> [190,]  4  3.9  5.14924545  1.3784766159
#> [191,]  4  4.0  3.81816728  0.3157422342
#> [192,]  4  4.1  5.20410134  1.8376571291
#> [193,]  4  4.2  4.28788786  1.7488897396
#> [194,]  4  4.3  4.49497557  2.0385803286
#> [195,]  4  4.4  5.04338571  2.8180688503
#> [196,]  4  4.5  4.02760180  2.7464213534
#> [197,]  4  4.6  5.05096669  4.0166186996
#> [198,]  4  4.7  2.35174795  3.4233218109
#> [199,]  4  4.8  4.53023873  2.9285730053
#> [200,]  4  4.9  3.27481205  2.6400129262
#> [201,]  5  0.0 -3.64434131  1.1225746368
#> [202,]  5  0.1 -3.45697729  0.0627392162
#> [203,]  5  0.2 -3.17715278  1.7456022053
#> [204,]  5  0.3 -3.42585475  2.2793435708
#> [205,]  5  0.4 -3.55250672  2.1440420843
#> [206,]  5  0.5 -5.55594624  2.0481285008
#> [207,]  5  0.6 -3.63455089  2.9556610261
#> [208,]  5  0.7 -4.09823255  0.7183703125
#> [209,]  5  0.8 -4.71777411  3.6105991865
#> [210,]  5  0.9 -6.66445034  1.3146661500
#> [211,]  5  1.0 -5.00482647  2.1459662056
#> [212,]  5  1.1 -5.46142231  2.4823068667
#> [213,]  5  1.2 -4.98557247  1.8681560289
#> [214,]  5  1.3 -5.02131003  3.0065067446
#> [215,]  5  1.4 -1.92391061  2.3644037096
#> [216,]  5  1.5 -2.40757538  2.6783458173
#> [217,]  5  1.6 -0.59967449  4.4452333360
#> [218,]  5  1.7 -1.63596223  4.1949879972
#> [219,]  5  1.8 -2.71758949  3.9016435227
#> [220,]  5  1.9 -3.31795513  2.9642715183
#> [221,]  5  2.0 -2.82823421  3.0935684479
#> [222,]  5  2.1 -2.23650584  4.9784962272
#> [223,]  5  2.2 -2.16643290  4.2936172267
#> [224,]  5  2.3 -2.47636517  3.6281504671
#> [225,]  5  2.4 -1.64101906  2.7863250853
#> [226,]  5  2.5 -1.45392716  3.1893847011
#> [227,]  5  2.6 -2.69947502  4.5010743640
#> [228,]  5  2.7 -1.57606177  3.9344260965
#> [229,]  5  2.8 -1.68015120  5.3433425029
#> [230,]  5  2.9 -2.85023511  6.9546404866
#> [231,]  5  3.0 -2.36998488  6.6360958012
#> [232,]  5  3.1 -1.28914043  4.9500640891
#> [233,]  5  3.2 -1.68942947  4.2486169658
#> [234,]  5  3.3 -2.86380817  4.3816008554
#> [235,]  5  3.4 -2.78805973  5.0062791994
#> [236,]  5  3.5 -1.10335022  3.8445138654
#> [237,]  5  3.6 -3.11631251  5.5469011982
#> [238,]  5  3.7 -2.85357754  6.3950340450
#> [239,]  5  3.8 -1.35583927  5.2988176795
#> [240,]  5  3.9  1.03569426  5.1190072937
#> [241,]  5  4.0  0.89386836  6.4276282771
#> [242,]  5  4.1 -1.38659607  7.4351330419
#> [243,]  5  4.2 -0.66042872  7.3018309537
#> [244,]  5  4.3 -0.96704864  8.3037580610
#> [245,]  5  4.4 -0.63682353 10.2473472407
#> [246,]  5  4.5 -2.27333058  9.1517690678
#> [247,]  5  4.6  0.90613340 11.1089633957
#> [248,]  5  4.7  0.02799021 10.5539293219
#> [249,]  5  4.8 -0.39979343 10.9697379629
#> [250,]  5  4.9  0.06494513  9.7227629533

# Type 1
SimSSMOUFixed(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>        id time          y1          y2            x1           x2
#>   [1,]  1  0.0 -4.70711009  1.59655419  1.370958e+00  0.321925265
#>   [2,]  1  0.1 -2.56864611  2.39116735 -5.646982e-01 -0.783838941
#>   [3,]  1  0.2 -2.54028078  3.72231030  3.631284e-01  1.575727520
#>   [4,]  1  0.3 -2.79840579  2.35602895  6.328626e-01  0.642899306
#>   [5,]  1  0.4 -2.12068156  3.16225678  4.042683e-01  0.089760647
#>   [6,]  1  0.5 -1.66334538  3.78814615 -1.061245e-01  0.276550747
#>   [7,]  1  0.6 -2.02011692  2.67739831  1.511522e+00  0.679288816
#>   [8,]  1  0.7 -3.12712457  1.63838282 -9.465904e-02  0.089832887
#>   [9,]  1  0.8 -1.24299767  1.96485984  2.018424e+00 -2.993090083
#>  [10,]  1  0.9 -0.75589221  2.28927997 -6.271410e-02  0.284882954
#>  [11,]  1  1.0 -1.14411998  0.79598102  1.304870e+00 -0.367234643
#>  [12,]  1  1.1 -0.31977703  1.24816901  2.286645e+00  0.185230565
#>  [13,]  1  1.2 -0.53157478  2.27881357 -1.388861e+00  0.581823727
#>  [14,]  1  1.3 -0.37281299  1.26152724 -2.787888e-01  1.399736827
#>  [15,]  1  1.4 -0.85768443  1.48884136 -1.333213e-01 -0.727292059
#>  [16,]  1  1.5 -1.76297948  0.52714792  6.359504e-01  1.302542632
#>  [17,]  1  1.6 -2.07671884  1.98307566 -2.842529e-01  0.335848120
#>  [18,]  1  1.7 -1.24307381  1.51789796 -2.656455e+00  1.038506099
#>  [19,]  1  1.8 -2.68804677  3.06442232 -2.440467e+00  0.920728568
#>  [20,]  1  1.9 -2.22181161  3.53143043  1.320113e+00  0.720878163
#>  [21,]  1  2.0 -2.42673747  2.92290765 -3.066386e-01 -1.043118939
#>  [22,]  1  2.1 -2.11485127  4.57640504 -1.781308e+00 -0.090186387
#>  [23,]  1  2.2 -1.17184140  4.04877397 -1.719174e-01  0.623518162
#>  [24,]  1  2.3 -0.93058016  3.64033257  1.214675e+00 -0.953523358
#>  [25,]  1  2.4 -1.04684794  3.41476170  1.895193e+00 -0.542828815
#>  [26,]  1  2.5 -0.32263614  4.68783805 -4.304691e-01  0.580996498
#>  [27,]  1  2.6 -0.67818152  2.94774330 -2.572694e-01  0.768178738
#>  [28,]  1  2.7 -0.54806284  4.12967155 -1.763163e+00  0.463767589
#>  [29,]  1  2.8  0.93159741  4.75315337  4.600974e-01 -0.885776297
#>  [30,]  1  2.9  0.42022119  3.70473266 -6.399949e-01 -1.099780899
#>  [31,]  1  3.0 -0.14986239  3.46655376  4.554501e-01  1.512707010
#>  [32,]  1  3.1 -2.30870600  2.43114865  7.048373e-01  0.257921438
#>  [33,]  1  3.2 -0.36916344  4.56001532  1.035104e+00  0.088440229
#>  [34,]  1  3.3 -0.70107826  3.91648903 -6.089264e-01 -0.120896538
#>  [35,]  1  3.4  1.13040855  1.40286659  5.049551e-01 -1.194328895
#>  [36,]  1  3.5 -0.18811997  2.92213542 -1.717009e+00  0.611996898
#>  [37,]  1  3.6 -0.26861827  1.64575844 -7.844590e-01 -0.217139846
#>  [38,]  1  3.7 -0.63089436  2.48595841 -8.509076e-01 -0.182756706
#>  [39,]  1  3.8 -0.53473911  2.03601623 -2.414208e+00  0.933346329
#>  [40,]  1  3.9 -1.12401522  2.14932950  3.612261e-02  0.821773111
#>  [41,]  1  4.0  0.31943008  1.59165783  2.059986e-01  1.392116376
#>  [42,]  1  4.1 -0.50705448  1.57505776 -3.610573e-01 -0.476173923
#>  [43,]  1  4.2 -0.99735128 -0.04179658  7.581632e-01  0.650348561
#>  [44,]  1  4.3 -2.11673186  1.73491423 -7.267048e-01  1.391110456
#>  [45,]  1  4.4 -2.14880365  0.38712437 -1.368281e+00 -1.110788879
#>  [46,]  1  4.5 -2.03763850  0.60992787  4.328180e-01 -0.860792587
#>  [47,]  1  4.6 -2.63176678  0.81966452 -8.113932e-01 -1.131738681
#>  [48,]  1  4.7 -4.70610588  0.85029248  1.444101e+00 -1.459214000
#>  [49,]  1  4.8 -5.15343907  3.39374885 -4.314462e-01  0.079982553
#>  [50,]  1  4.9 -4.69607490  0.97854692  6.556479e-01  0.653204340
#>  [51,]  2  0.0 -3.80041653  3.28459489  1.200965e+00 -0.040698475
#>  [52,]  2  0.1 -2.83529702  3.32198812  1.044751e+00 -1.551544822
#>  [53,]  2  0.2 -0.84925990  5.14577459 -1.003209e+00  1.167169549
#>  [54,]  2  0.3 -1.69481713  3.61054550  1.848482e+00 -0.273645701
#>  [55,]  2  0.4 -1.70856945  5.84433335 -6.667734e-01 -0.467845325
#>  [56,]  2  0.5 -2.89369594  5.01736414  1.055138e-01 -1.238252328
#>  [57,]  2  0.6 -2.81552921  4.14705754 -4.222559e-01 -0.007762034
#>  [58,]  2  0.7 -2.93790445  4.01632426 -1.223502e-01 -0.800282178
#>  [59,]  2  0.8 -2.00537470  5.46205266  1.881930e-01 -0.533492330
#>  [60,]  2  0.9 -1.57774323  6.04157012  1.191610e-01  1.287675246
#>  [61,]  2  1.0 -1.56130787  4.87398902 -2.509255e-02 -0.175525870
#>  [62,]  2  1.1 -1.74124644  5.03039635  1.080727e-01 -1.071782384
#>  [63,]  2  1.2 -1.48368967  4.27957476 -4.854352e-01  0.163206882
#>  [64,]  2  1.3 -1.77359291  4.73321541 -5.042171e-01 -0.362738416
#>  [65,]  2  1.4 -2.62843297  3.87935236 -1.661099e+00  0.590013548
#>  [66,]  2  1.5 -1.74259813  2.10397855 -3.823337e-01  1.432421928
#>  [67,]  2  1.6 -0.78959786  1.27261233 -5.126503e-01 -0.992692511
#>  [68,]  2  1.7 -1.43370275  0.52721814  2.701891e+00  0.454650298
#>  [69,]  2  1.8 -2.08685975  2.12821274 -1.362116e+00  0.084898059
#>  [70,]  2  1.9 -1.79347515  0.55297887  1.372562e-01  0.895565582
#>  [71,]  2  2.0 -2.10362517  0.57424035 -1.493625e+00 -0.229778139
#>  [72,]  2  2.1 -2.26350041  0.63842947 -1.470436e+00  0.836619068
#>  [73,]  2  2.2 -3.34908027  0.16410751  1.247024e-01 -1.745055861
#>  [74,]  2  2.3 -3.45638026  0.16716103 -9.966391e-01  1.689458921
#>  [75,]  2  2.4 -3.66354708 -0.30828014 -1.822614e-03  0.864777979
#>  [76,]  2  2.5 -3.13325217  1.11733291 -4.282589e-01 -0.150775989
#>  [77,]  2  2.6 -4.22686669  1.69226649 -6.136716e-01 -1.449007130
#>  [78,]  2  2.7 -4.10694378  1.07001286 -2.024678e+00  0.643008700
#>  [79,]  2  2.8 -4.14858967  1.05222391 -1.224748e+00  0.483193864
#>  [80,]  2  2.9 -4.18191401  1.60052371  1.795164e-01 -0.006355626
#>  [81,]  2  3.0 -3.29035999  0.07426085  5.676206e-01  0.151455893
#>  [82,]  2  3.1 -3.59379315 -0.32529445 -4.928774e-01 -0.584108970
#>  [83,]  2  3.2 -3.32214750 -0.82295418  6.288407e-05  0.368806733
#>  [84,]  2  3.3 -2.82286335 -2.20901193  1.122890e+00  0.294654340
#>  [85,]  2  3.4 -4.79229145 -0.74656528  1.439856e+00 -0.279259373
#>  [86,]  2  3.5 -4.80092813 -0.21893704 -1.097114e+00 -1.336236655
#>  [87,]  2  3.6 -4.38516313  0.50500520 -1.173196e-01  0.700748818
#>  [88,]  2  3.7 -3.62007355  1.99428640  1.201498e+00  0.554196622
#>  [89,]  2  3.8 -3.67275245 -0.78820605 -4.697296e-01 -0.836306593
#>  [90,]  2  3.9 -4.35789134 -0.55568678 -5.246948e-02 -1.594588162
#>  [91,]  2  4.0 -2.16025317  1.93270324 -8.610730e-02  0.204958581
#>  [92,]  2  4.1 -3.09619998  0.86518075 -8.876790e-01 -0.345087978
#>  [93,]  2  4.2 -2.46702448  0.34502147 -4.446840e-01  0.252611703
#>  [94,]  2  4.3 -2.19287302  0.48811528 -2.944488e-02 -1.294002465
#>  [95,]  2  4.4 -3.42615225  0.13819405 -4.138688e-01 -0.959170444
#>  [96,]  2  4.5 -4.06106271  1.11826629  1.113386e+00  1.085774854
#>  [97,]  2  4.6 -1.93909215  0.80345053 -4.809928e-01  0.403774905
#>  [98,]  2  4.7 -2.96441646  1.10278975 -4.331690e-01  0.586487537
#>  [99,]  2  4.8 -3.32076547 -0.93357252  6.968626e-01  1.815228446
#> [100,]  2  4.9 -5.25482172  0.05198079 -1.056368e+00  0.128821429
#> [101,]  3  0.0 -1.86109624  0.45101866 -2.000929e+00 -1.096156242
#> [102,]  3  0.1 -1.77271507  1.03066483  3.337772e-01  0.049050451
#> [103,]  3  0.2 -2.27519890 -0.95035823  1.171325e+00 -1.198495857
#> [104,]  3  0.3 -1.01759011 -0.64240225  2.059539e+00  0.190018999
#> [105,]  3  0.4 -3.04360230 -0.59911435 -1.376862e+00  1.297705900
#> [106,]  3  0.5 -2.92311625  1.13511898 -1.150856e+00 -1.033873723
#> [107,]  3  0.6 -4.26528435 -1.42311371 -7.058214e-01 -0.738440754
#> [108,]  3  0.7 -4.15688022 -1.75497929 -1.054056e+00  0.046563939
#> [109,]  3  0.8 -1.78568563  1.02997626 -6.457437e-01 -1.017596120
#> [110,]  3  0.9 -0.91069803  0.40791704 -1.853780e-01 -0.383283960
#> [111,]  3  1.0 -0.99632255  1.33869288 -1.201222e+00  0.872755412
#> [112,]  3  1.1 -1.00666441  3.59692297  2.036972e+00  0.969545014
#> [113,]  3  1.2 -0.83470519  1.83425520  1.077747e-01  0.383846665
#> [114,]  3  1.3 -1.68536076  2.25551706 -8.410810e-02 -1.851555663
#> [115,]  3  1.4 -0.41959603  3.48357463  4.956196e-01 -0.053996737
#> [116,]  3  1.5  1.05935521  3.26528053  3.741519e-02  1.064773214
#> [117,]  3  1.6  1.43164516  1.49783324 -1.320880e-01  0.813195037
#> [118,]  3  1.7  1.70776082  0.67301878  1.476787e+00 -0.190816474
#> [119,]  3  1.8  0.59497073 -0.60994382 -2.170302e-01 -2.699929809
#> [120,]  3  1.9 -0.97520084  2.93320299 -1.283602e+00  0.060966639
#> [121,]  3  2.0  1.48897078  1.24638738  3.856679e-01  0.573751697
#> [122,]  3  2.1  0.39697081  2.00373167 -3.515129e-01  0.045803580
#> [123,]  3  2.2  1.16799458  3.19054719 -5.217961e-01  0.157412540
#> [124,]  3  2.3  0.85116385  3.01226068 -1.068131e+00  0.431565373
#> [125,]  3  2.4  1.25766360  2.36268482  4.283659e-01 -0.396549736
#> [126,]  3  2.5  1.38286038  0.69167840 -1.740182e-01  1.309978226
#> [127,]  3  2.6 -0.47884589  3.22285068  5.156677e-01  0.470393400
#> [128,]  3  2.7 -0.49463111  0.67449177 -2.343653e-01 -1.242670271
#> [129,]  3  2.8  1.18878756  1.67189066 -6.585034e-01  1.381575456
#> [130,]  3  2.9 -1.77194643  1.17199023  1.250237e+00  1.204458937
#> [131,]  3  3.0 -0.01289777  1.33283744 -2.717637e-01  0.824073964
#> [132,]  3  3.1  0.18961016  0.97094205  9.479520e-01 -1.662629402
#> [133,]  3  3.2  0.51989235  1.22114138 -1.201582e+00 -0.569306344
#> [134,]  3  3.3  1.14382416 -0.16444732 -4.661161e-01  0.635513817
#> [135,]  3  3.4  0.33214697 -0.30170292 -2.693514e-01  0.043722008
#> [136,]  3  3.5  1.46615994 -0.24006356 -3.909654e-01  0.348012304
#> [137,]  3  3.6  1.37166231  1.15784842  1.348707e+00  2.459593549
#> [138,]  3  3.7  2.99678180 -0.10930787 -2.276470e-02 -0.818380324
#> [139,]  3  3.8  1.13736596  1.18338714  2.442259e-01 -2.113200115
#> [140,]  3  3.9  1.25949122  1.10509935 -9.423717e-01  0.273695272
#> [141,]  3  4.0 -0.51679931  2.31845509 -7.292173e-01 -0.687596841
#> [142,]  3  4.1  0.17169491  1.92608290  9.980689e-01  0.446041053
#> [143,]  3  4.2  1.44604622  3.36702912  1.258482e+00 -0.812384724
#> [144,]  3  4.3  0.76665842  3.01563115  1.248864e+00  2.212055480
#> [145,]  3  4.4  2.31952407  3.23834737 -1.380637e+00 -0.123705972
#> [146,]  3  4.5  3.21664952  3.17691265  2.049961e+00 -0.477335506
#> [147,]  3  4.6  3.60176600  3.31741853  1.016873e+00 -0.166261491
#> [148,]  3  4.7  3.89655470  2.56855816 -2.671746e-02  0.862563384
#> [149,]  3  4.8  2.74335237  3.22015995  7.036078e-01  0.097340485
#> [150,]  3  4.9  2.42064170  4.54613269 -9.713852e-01 -1.625616739
#> [151,]  4  0.0 -4.98958069  2.06541016 -4.620768e-03  0.724173801
#> [152,]  4  0.1 -5.23300625  2.56772344  7.602422e-01 -0.832552826
#> [153,]  4  0.2 -5.76994861  0.68869833  3.899091e-02  0.732528487
#> [154,]  4  0.3 -4.77353853  1.93291322  7.350721e-01 -0.871926870
#> [155,]  4  0.4 -4.74334182  2.25451758 -1.464726e-01 -0.453397511
#> [156,]  4  0.5 -5.45112848  3.98986408 -5.788734e-02  1.187534279
#> [157,]  4  0.6 -4.13726757  4.68380136  4.823695e-01 -0.290145312
#> [158,]  4  0.7 -5.46414497  4.75395990  9.929436e-01  0.828546145
#> [159,]  4  0.8 -3.34352159  4.47768982 -1.246395e+00 -0.291227709
#> [160,]  4  0.9 -5.51393554  5.07701083 -3.348752e-02 -1.576362405
#> [161,]  4  1.0 -4.56755872  4.17836989 -7.096218e-02 -0.848815697
#> [162,]  4  1.1 -3.73125116  5.32105750 -7.589207e-01 -1.088519862
#> [163,]  4  1.2 -3.97515005  4.45067247 -1.034359e+00 -0.484290571
#> [164,]  4  1.3 -3.65850150  5.77941605 -6.307320e-01 -0.336311209
#> [165,]  4  1.4 -3.12884196  3.89424762  5.868077e-01 -0.153357891
#> [166,]  4  1.5 -3.06278548  3.72445600 -4.163227e-01 -0.243247229
#> [167,]  4  1.6 -3.74532517  6.20278480 -7.848878e-01  1.892202042
#> [168,]  4  1.7 -3.69188269  4.27350773  1.634163e-01 -1.385998337
#> [169,]  4  1.8 -2.15405307  4.46937202 -1.236714e+00 -0.414824301
#> [170,]  4  1.9 -2.96753475  4.41055843  1.045874e+00  0.349081528
#> [171,]  4  2.0 -2.75447097  3.65207915 -4.845954e-01  1.628442266
#> [172,]  4  2.1 -3.82993651  3.31163429  1.891288e-01  0.088521896
#> [173,]  4  2.2 -4.45552946  4.54006004  5.100633e-02  1.239150708
#> [174,]  4  2.3 -5.31675775  3.22460185 -2.406689e-04 -1.644555536
#> [175,]  4  2.4 -6.14608582  3.21720150  1.809382e+00  1.446356525
#> [176,]  4  2.5 -4.38170705  3.76467358 -8.253280e-01 -0.690560172
#> [177,]  4  2.6 -4.90260468  4.46358008  1.145470e+00 -0.276431085
#> [178,]  4  2.7 -4.84279329  1.40504207  3.157319e-02 -1.109418760
#> [179,]  4  2.8 -4.79547733  2.25237535 -8.352058e-01  0.133869316
#> [180,]  4  2.9 -4.74100007  1.67554418 -6.876365e-02  1.785339052
#> [181,]  4  3.0 -2.80807668  2.09074968  7.467717e-01  2.422163355
#> [182,]  4  3.1 -5.18967906  2.33076414 -4.255187e-01 -1.076828902
#> [183,]  4  3.2 -4.23295998  2.27981892 -7.720822e-01  0.485941110
#> [184,]  4  3.3 -3.78329876  3.65771263  1.527641e-01  1.388521739
#> [185,]  4  3.4 -4.15260857  4.57276955  9.885968e-01 -0.195656817
#> [186,]  4  3.5 -4.85584835  4.66753702 -7.345833e-02 -0.218174798
#> [187,]  4  3.6 -3.81887611  5.04599975 -1.387027e+00 -0.304777955
#> [188,]  4  3.7 -4.61665749  3.19199951 -1.306676e+00  0.597832724
#> [189,]  4  3.8 -3.95767876  4.16181590 -7.683953e-01  1.397429411
#> [190,]  4  3.9 -5.13587865  4.63748094 -5.271081e-01  0.687619761
#> [191,]  4  4.0 -4.43245270  2.13785867 -2.142707e-02  0.320188014
#> [192,]  4  4.1 -4.28726430  4.79339369  6.704981e-01 -0.301869926
#> [193,]  4  4.2 -3.45277721  4.31120174 -4.346170e-01  0.498348686
#> [194,]  4  4.3 -4.82095577  5.36001360 -1.113880e+00 -0.549536918
#> [195,]  4  4.4 -4.05619534  5.16467092  6.071060e-01 -0.279256504
#> [196,]  4  4.5 -5.57635483  6.15995844  2.754570e-01  1.096513443
#> [197,]  4  4.6 -3.64458171  4.40198702  1.157347e+00  0.442013088
#> [198,]  4  4.7 -5.68562167  6.28258152 -1.682481e+00  0.241016294
#> [199,]  4  4.8 -4.85885555  3.32998720  8.731909e-02 -0.255607655
#> [200,]  4  4.9 -6.43969238  3.86146292  1.353362e+00  0.931032901
#> [201,]  5  0.0 -4.03128423  0.88651362  1.334913e+00 -1.303821154
#> [202,]  5  0.1 -0.69950443  1.58666189 -8.692718e-01 -0.250914465
#> [203,]  5  0.2 -2.98061330  2.21765252  5.548695e-02  0.171007374
#> [204,]  5  0.3 -2.18804137  1.02683523  4.906691e-02 -0.403467479
#> [205,]  5  0.4 -2.04953630  2.47697532 -5.783557e-01  0.104659441
#> [206,]  5  0.5 -3.72006759  1.89168940 -9.987387e-01 -0.318880787
#> [207,]  5  0.6 -2.63525375  1.89760080 -2.432780e-03  1.618343936
#> [208,]  5  0.7 -1.78287858  1.48579749  6.555119e-01  0.714188601
#> [209,]  5  0.8 -3.38381830  0.09739957  1.476842e+00  2.965865370
#> [210,]  5  0.9 -1.83646943  1.12846867 -1.909153e+00 -0.795077605
#> [211,]  5  1.0 -1.96504107  1.01147554 -7.024395e-01  0.814365915
#> [212,]  5  1.1 -1.21817432  3.40180058 -3.114302e-01  2.098030810
#> [213,]  5  1.2  0.14201131  0.89430764 -1.663157e+00  0.300980055
#> [214,]  5  1.3 -1.54050303  1.08794720 -7.505334e-01 -1.083075142
#> [215,]  5  1.4 -1.78810743  1.80277142 -7.773518e-01 -1.006322502
#> [216,]  5  1.5 -0.20416140  0.78079581 -7.225697e-01 -0.035414565
#> [217,]  5  1.6 -0.97328175  1.60848509 -2.188835e+00  1.309124361
#> [218,]  5  1.7 -0.97137025  0.92248827  2.134186e-01  0.750400487
#> [219,]  5  1.8 -1.10588514  1.73510859 -6.319229e-01 -2.138368328
#> [220,]  5  1.9 -2.00815352 -0.07322412  1.520491e+00 -0.700354109
#> [221,]  5  2.0 -1.82930914  0.53323593  7.959559e-01 -0.009056475
#> [222,]  5  2.1 -1.72178039  1.11179907 -1.453530e+00 -1.458133487
#> [223,]  5  2.2 -2.23677603  1.73004202  9.839542e-02  0.694529646
#> [224,]  5  2.3 -1.80728789  2.18926079 -5.937710e-01 -2.461335475
#> [225,]  5  2.4 -1.32582896  1.08978873  8.882812e-01  0.143289764
#> [226,]  5  2.5 -2.12051517  0.49141376  5.307042e-02 -0.391222119
#> [227,]  5  2.6 -0.83976290  0.24627586 -5.570236e-01 -0.491164086
#> [228,]  5  2.7 -1.80634997  1.47380274  4.383970e-01 -0.283647452
#> [229,]  5  2.8 -1.60489403  2.66202325  1.526082e-01  0.314794805
#> [230,]  5  2.9 -2.21234864  1.63763438 -1.646176e-01  0.396326578
#> [231,]  5  3.0 -2.09999546  2.36282160  2.019891e+00 -0.225603711
#> [232,]  5  3.1 -0.32743420  2.15105690 -5.293859e-01 -1.924950430
#> [233,]  5  3.2 -2.35909714  2.27317638 -4.707870e-01 -1.439229298
#> [234,]  5  3.3 -2.71087438  0.94312290 -1.545937e+00 -1.469657774
#> [235,]  5  3.4 -1.04941923  1.73428980 -4.052672e-02  0.761863447
#> [236,]  5  3.5 -2.87972278  0.68872574  8.903563e-01 -0.243614982
#> [237,]  5  3.6 -1.93895097  0.04542038 -2.071388e+00  0.269676607
#> [238,]  5  3.7 -2.67900770  0.46272512 -2.500651e-01 -1.558927521
#> [239,]  5  3.8 -3.52173677  0.28437948 -1.181650e+00 -0.535588007
#> [240,]  5  3.9 -3.30285077  1.46486688  1.441937e+00  0.562451973
#> [241,]  5  4.0 -2.99525570  0.90523251  1.357896e+00 -0.178326123
#> [242,]  5  4.1 -0.88064491  0.63115588  3.345028e-01 -0.115135986
#> [243,]  5  4.2 -1.48801990 -0.06421253  1.429338e+00 -0.072061472
#> [244,]  5  4.3 -1.61476412  1.40662338 -8.673179e-01  1.210909807
#> [245,]  5  4.4 -2.68626135  1.42252322  9.506517e-01 -0.614896901
#> [246,]  5  4.5 -2.98457084  1.43002996 -5.850115e-01  0.676126458
#> [247,]  5  4.6 -1.78461975  2.96869072  3.209575e-01  0.898599606
#> [248,]  5  4.7 -1.70860905  2.99487585 -2.993960e-01 -1.189317904
#> [249,]  5  4.8 -3.05160843  1.77376068 -2.785431e-01  0.121258850
#> [250,]  5  4.9 -2.54716353  2.02076922  5.461152e-01 -0.011221686

# Type 2
SimSSMOUFixed(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>        id time          y1           y2            x1           x2
#>   [1,]  1  0.0 -4.15861370  3.081004900  1.370958e+00  0.321925265
#>   [2,]  1  0.1 -3.19500261  2.965761872 -5.646982e-01 -0.783838941
#>   [3,]  1  0.2 -4.02381046  1.397749634  3.631284e-01  1.575727520
#>   [4,]  1  0.3 -4.41289457  2.177723497  6.328626e-01  0.642899306
#>   [5,]  1  0.4 -4.61514987  1.860471897  4.042683e-01  0.089760647
#>   [6,]  1  0.5 -4.51205243  2.906905436 -1.061245e-01  0.276550747
#>   [7,]  1  0.6 -3.49823426  4.349933949  1.511522e+00  0.679288816
#>   [8,]  1  0.7 -3.79835683  4.700522714 -9.465904e-02  0.089832887
#>   [9,]  1  0.8 -3.14784481  2.157266543  2.018424e+00 -2.993090083
#>  [10,]  1  0.9 -2.62680383  3.696889744 -6.271410e-02  0.284882954
#>  [11,]  1  1.0 -2.66374287  2.513645821  1.304870e+00 -0.367234643
#>  [12,]  1  1.1 -1.78694379  2.788894671  2.286645e+00  0.185230565
#>  [13,]  1  1.2 -3.85578965  3.292496265 -1.388861e+00  0.581823727
#>  [14,]  1  1.3 -3.48047060  3.877803887 -2.787888e-01  1.399736827
#>  [15,]  1  1.4 -2.88915979  4.911762787 -1.333213e-01 -0.727292059
#>  [16,]  1  1.5 -1.58952555  7.142140480  6.359504e-01  1.302542632
#>  [17,]  1  1.6 -2.43657028  5.228555387 -2.842529e-01  0.335848120
#>  [18,]  1  1.7 -3.00356030  6.075010672 -2.656455e+00  1.038506099
#>  [19,]  1  1.8 -2.85055441  8.235491662 -2.440467e+00  0.920728568
#>  [20,]  1  1.9 -2.04332764  7.447304864  1.320113e+00  0.720878163
#>  [21,]  1  2.0 -3.58660301  6.023776071 -3.066386e-01 -1.043118939
#>  [22,]  1  2.1 -2.73709879  8.090766244 -1.781308e+00 -0.090186387
#>  [23,]  1  2.2  0.11579452  8.345510693 -1.719174e-01  0.623518162
#>  [24,]  1  2.3 -0.62738268  7.413256875  1.214675e+00 -0.953523358
#>  [25,]  1  2.4  1.47500936  7.395888099  1.895193e+00 -0.542828815
#>  [26,]  1  2.5  0.24033027  7.927882778 -4.304691e-01  0.580996498
#>  [27,]  1  2.6 -0.30626914  8.082455121 -2.572694e-01  0.768178738
#>  [28,]  1  2.7  0.40147210  8.515848088 -1.763163e+00  0.463767589
#>  [29,]  1  2.8  1.08450806  7.124231547  4.600974e-01 -0.885776297
#>  [30,]  1  2.9  0.74957513  8.337379733 -6.399949e-01 -1.099780899
#>  [31,]  1  3.0  1.23128412  8.369634571  4.554501e-01  1.512707010
#>  [32,]  1  3.1  0.88565290  8.567437885  7.048373e-01  0.257921438
#>  [33,]  1  3.2  1.25268933  7.737581841  1.035104e+00  0.088440229
#>  [34,]  1  3.3 -0.42834842  6.752787823 -6.089264e-01 -0.120896538
#>  [35,]  1  3.4  0.90099172  6.411914808  5.049551e-01 -1.194328895
#>  [36,]  1  3.5 -1.49083850  5.761253432 -1.717009e+00  0.611996898
#>  [37,]  1  3.6  0.54554075  7.231525619 -7.844590e-01 -0.217139846
#>  [38,]  1  3.7  0.16004705  6.827060505 -8.509076e-01 -0.182756706
#>  [39,]  1  3.8 -1.16993915  8.456634063 -2.414208e+00  0.933346329
#>  [40,]  1  3.9  0.77781377  7.518569739  3.612261e-02  0.821773111
#>  [41,]  1  4.0 -1.51388597  7.591537753  2.059986e-01  1.392116376
#>  [42,]  1  4.1 -0.73229816  5.423052145 -3.610573e-01 -0.476173923
#>  [43,]  1  4.2 -2.05659766  6.344370804  7.581632e-01  0.650348561
#>  [44,]  1  4.3 -1.86559396  6.175139787 -7.267048e-01  1.391110456
#>  [45,]  1  4.4 -1.35979163  5.504529888 -1.368281e+00 -1.110788879
#>  [46,]  1  4.5  0.07322013  5.891083045  4.328180e-01 -0.860792587
#>  [47,]  1  4.6 -0.27164713  4.961591820 -8.113932e-01 -1.131738681
#>  [48,]  1  4.7  0.18432097  4.712492043  1.444101e+00 -1.459214000
#>  [49,]  1  4.8  0.23937761  4.875819151 -4.314462e-01  0.079982553
#>  [50,]  1  4.9  0.94585893  3.930630210  6.556479e-01  0.653204340
#>  [51,]  2  0.0 -3.26298505  1.775969424  1.200965e+00 -0.040698475
#>  [52,]  2  0.1 -2.01410241 -0.826294713  1.044751e+00 -1.551544822
#>  [53,]  2  0.2 -1.91078829  1.126818088 -1.003209e+00  1.167169549
#>  [54,]  2  0.3 -1.70420116  1.319037041  1.848482e+00 -0.273645701
#>  [55,]  2  0.4 -1.60161944  1.235719804 -6.667734e-01 -0.467845325
#>  [56,]  2  0.5 -1.63195064  0.419904656  1.055138e-01 -1.238252328
#>  [57,]  2  0.6 -0.96707150  2.693727499 -4.222559e-01 -0.007762034
#>  [58,]  2  0.7 -1.02602254  2.348949111 -1.223502e-01 -0.800282178
#>  [59,]  2  0.8  0.66105154  0.458359266  1.881930e-01 -0.533492330
#>  [60,]  2  0.9 -0.05659913  3.651571405  1.191610e-01  1.287675246
#>  [61,]  2  1.0 -1.02856809  2.233792487 -2.509255e-02 -0.175525870
#>  [62,]  2  1.1 -0.95791933  1.088390440  1.080727e-01 -1.071782384
#>  [63,]  2  1.2 -1.17096748  1.233820026 -4.854352e-01  0.163206882
#>  [64,]  2  1.3 -1.97585331  2.366060556 -5.042171e-01 -0.362738416
#>  [65,]  2  1.4 -0.67056987  0.884848404 -1.661099e+00  0.590013548
#>  [66,]  2  1.5 -1.08362592  4.249186586 -3.823337e-01  1.432421928
#>  [67,]  2  1.6 -2.55533194  1.954310115 -5.126503e-01 -0.992692511
#>  [68,]  2  1.7 -1.46813969  1.177462668  2.701891e+00  0.454650298
#>  [69,]  2  1.8 -2.12738913  3.660404393 -1.362116e+00  0.084898059
#>  [70,]  2  1.9 -2.31689460  3.651817821  1.372562e-01  0.895565582
#>  [71,]  2  2.0 -3.38298249  2.842416594 -1.493625e+00 -0.229778139
#>  [72,]  2  2.1 -3.02748190  2.614308838 -1.470436e+00  0.836619068
#>  [73,]  2  2.2 -1.60113582  0.842573570  1.247024e-01 -1.745055861
#>  [74,]  2  2.3 -1.00620844  2.542509042 -9.966391e-01  1.689458921
#>  [75,]  2  2.4 -0.28780582  1.803448964 -1.822614e-03  0.864777979
#>  [76,]  2  2.5 -0.70979200  3.207306126 -4.282589e-01 -0.150775989
#>  [77,]  2  2.6 -0.93225559  2.939486050 -6.136716e-01 -1.449007130
#>  [78,]  2  2.7 -2.14747914  1.904274973 -2.024678e+00  0.643008700
#>  [79,]  2  2.8 -0.46012372  1.486542931 -1.224748e+00  0.483193864
#>  [80,]  2  2.9  0.06344073  3.179574765  1.795164e-01 -0.006355626
#>  [81,]  2  3.0 -0.89036617  2.479384432  5.676206e-01  0.151455893
#>  [82,]  2  3.1 -0.68821425  2.433071784 -4.928774e-01 -0.584108970
#>  [83,]  2  3.2 -1.68453475  1.121127845  6.288407e-05  0.368806733
#>  [84,]  2  3.3 -0.30366995  1.500320977  1.122890e+00  0.294654340
#>  [85,]  2  3.4 -1.73099976  2.019470641  1.439856e+00 -0.279259373
#>  [86,]  2  3.5 -1.23240203  0.381160267 -1.097114e+00 -1.336236655
#>  [87,]  2  3.6  0.38452351  0.386259166 -1.173196e-01  0.700748818
#>  [88,]  2  3.7 -0.14379238  2.764661748  1.201498e+00  0.554196622
#>  [89,]  2  3.8 -0.09450590  1.067556215 -4.697296e-01 -0.836306593
#>  [90,]  2  3.9 -1.76497151  0.874343947 -5.246948e-02 -1.594588162
#>  [91,]  2  4.0  0.08239535  1.027633493 -8.610730e-02  0.204958581
#>  [92,]  2  4.1 -0.68478614  0.482499461 -8.876790e-01 -0.345087978
#>  [93,]  2  4.2 -3.20870047 -0.287624989 -4.446840e-01  0.252611703
#>  [94,]  2  4.3 -1.37233094  0.706847603 -2.944488e-02 -1.294002465
#>  [95,]  2  4.4 -1.88008245  0.537240246 -4.138688e-01 -0.959170444
#>  [96,]  2  4.5 -3.73461745  0.888363825  1.113386e+00  1.085774854
#>  [97,]  2  4.6 -1.37906294  1.128276657 -4.809928e-01  0.403774905
#>  [98,]  2  4.7 -1.77834462  0.605929736 -4.331690e-01  0.586487537
#>  [99,]  2  4.8 -0.79556543  0.316057376  6.968626e-01  1.815228446
#> [100,]  2  4.9 -1.60164333  2.280704733 -1.056368e+00  0.128821429
#> [101,]  3  0.0 -3.03719035  1.205615007 -2.000929e+00 -1.096156242
#> [102,]  3  0.1 -3.55859779  0.896268090  3.337772e-01  0.049050451
#> [103,]  3  0.2 -2.48152610  1.103148137  1.171325e+00 -1.198495857
#> [104,]  3  0.3 -3.36090080  2.149514607  2.059539e+00  0.190018999
#> [105,]  3  0.4 -4.22820788  1.345877442 -1.376862e+00  1.297705900
#> [106,]  3  0.5 -4.55250996  0.613710847 -1.150856e+00 -1.033873723
#> [107,]  3  0.6 -5.34031563  1.152957795 -7.058214e-01 -0.738440754
#> [108,]  3  0.7 -2.48585424  2.288752332 -1.054056e+00  0.046563939
#> [109,]  3  0.8 -2.83231528  1.069878033 -6.457437e-01 -1.017596120
#> [110,]  3  0.9 -4.02833654  1.046121590 -1.853780e-01 -0.383283960
#> [111,]  3  1.0 -4.44251343  0.702906664 -1.201222e+00  0.872755412
#> [112,]  3  1.1 -3.26800894  1.286604534  2.036972e+00  0.969545014
#> [113,]  3  1.2 -4.28233367  1.027559947  1.077747e-01  0.383846665
#> [114,]  3  1.3 -4.45607680  0.470118528 -8.410810e-02 -1.851555663
#> [115,]  3  1.4 -4.25009508  0.753191351  4.956196e-01 -0.053996737
#> [116,]  3  1.5 -4.98705086 -0.580563064  3.741519e-02  1.064773214
#> [117,]  3  1.6 -4.13009159 -0.053519499 -1.320880e-01  0.813195037
#> [118,]  3  1.7 -3.16610330  1.008513669  1.476787e+00 -0.190816474
#> [119,]  3  1.8 -4.22517591 -0.052220941 -2.170302e-01 -2.699929809
#> [120,]  3  1.9 -3.78302131 -0.099637577 -1.283602e+00  0.060966639
#> [121,]  3  2.0 -3.48348355  1.398092186  3.856679e-01  0.573751697
#> [122,]  3  2.1 -5.10637291  1.713673831 -3.515129e-01  0.045803580
#> [123,]  3  2.2 -4.77818520  4.581017387 -5.217961e-01  0.157412540
#> [124,]  3  2.3 -3.15578976  2.837840105 -1.068131e+00  0.431565373
#> [125,]  3  2.4 -2.77230420  2.810253138  4.283659e-01 -0.396549736
#> [126,]  3  2.5 -2.27579898  3.488487225 -1.740182e-01  1.309978226
#> [127,]  3  2.6 -2.67734901  3.468503420  5.156677e-01  0.470393400
#> [128,]  3  2.7 -3.66801318  3.120465596 -2.343653e-01 -1.242670271
#> [129,]  3  2.8 -2.80128613  3.675967035 -6.585034e-01  1.381575456
#> [130,]  3  2.9 -2.73922635  5.250395636  1.250237e+00  1.204458937
#> [131,]  3  3.0 -1.56199099  5.147669606 -2.717637e-01  0.824073964
#> [132,]  3  3.1 -3.51466696  3.720943789  9.479520e-01 -1.662629402
#> [133,]  3  3.2 -2.82072185  4.940246499 -1.201582e+00 -0.569306344
#> [134,]  3  3.3 -3.32355695  4.046322081 -4.661161e-01  0.635513817
#> [135,]  3  3.4 -3.68962006  3.561264288 -2.693514e-01  0.043722008
#> [136,]  3  3.5 -2.76709079  3.909425408 -3.909654e-01  0.348012304
#> [137,]  3  3.6 -2.73348065  3.763954265  1.348707e+00  2.459593549
#> [138,]  3  3.7 -3.56145451  4.781339046 -2.276470e-02 -0.818380324
#> [139,]  3  3.8 -3.16601247  2.431943714  2.442259e-01 -2.113200115
#> [140,]  3  3.9 -1.77274852  2.959362547 -9.423717e-01  0.273695272
#> [141,]  3  4.0 -1.27755170  2.471549494 -7.292173e-01 -0.687596841
#> [142,]  3  4.1 -2.61309152  5.114754909  9.980689e-01  0.446041053
#> [143,]  3  4.2 -2.22616404  5.061461696  1.258482e+00 -0.812384724
#> [144,]  3  4.3 -1.63341481  5.937465532  1.248864e+00  2.212055480
#> [145,]  3  4.4 -2.66196280  5.608734323 -1.380637e+00 -0.123705972
#> [146,]  3  4.5 -0.98760995  6.682512237  2.049961e+00 -0.477335506
#> [147,]  3  4.6  0.03586551  5.781818853  1.016873e+00 -0.166261491
#> [148,]  3  4.7  0.42714001  6.280599135 -2.671746e-02  0.862563384
#> [149,]  3  4.8 -0.84761479  6.014800385  7.036078e-01  0.097340485
#> [150,]  3  4.9  1.02715839  6.255902367 -9.713852e-01 -1.625616739
#> [151,]  4  0.0 -4.08387551  0.270481051 -4.620768e-03  0.724173801
#> [152,]  4  0.1 -1.77817274  1.838488310  7.602422e-01 -0.832552826
#> [153,]  4  0.2 -4.18874128 -1.048662090  3.899091e-02  0.732528487
#> [154,]  4  0.3 -3.94006841 -0.494480212  7.350721e-01 -0.871926870
#> [155,]  4  0.4 -3.67137739  0.256855992 -1.464726e-01 -0.453397511
#> [156,]  4  0.5 -5.53225894  1.287579699 -5.788734e-02  1.187534279
#> [157,]  4  0.6 -3.72971131  1.309432579  4.823695e-01 -0.290145312
#> [158,]  4  0.7 -3.14566732  1.665491278  9.929436e-01  0.828546145
#> [159,]  4  0.8 -4.00229352  0.008746648 -1.246395e+00 -0.291227709
#> [160,]  4  0.9 -5.66090938 -2.192308260 -3.348752e-02 -1.576362405
#> [161,]  4  1.0 -4.83016996 -1.090170186 -7.096218e-02 -0.848815697
#> [162,]  4  1.1 -6.34033294 -1.705853370 -7.589207e-01 -1.088519862
#> [163,]  4  1.2 -6.84718391 -2.637203357 -1.034359e+00 -0.484290571
#> [164,]  4  1.3 -5.82553870 -1.689876573 -6.307320e-01 -0.336311209
#> [165,]  4  1.4 -7.96162161 -2.162816258  5.868077e-01 -0.153357891
#> [166,]  4  1.5 -7.38586087 -1.383326029 -4.163227e-01 -0.243247229
#> [167,]  4  1.6 -6.40138005 -0.387348382 -7.848878e-01  1.892202042
#> [168,]  4  1.7 -6.30563598 -0.683693252  1.634163e-01 -1.385998337
#> [169,]  4  1.8 -5.75705175 -0.480196478 -1.236714e+00 -0.414824301
#> [170,]  4  1.9 -5.07504288 -0.657695495  1.045874e+00  0.349081528
#> [171,]  4  2.0 -4.81487021  0.570498825 -4.845954e-01  1.628442266
#> [172,]  4  2.1 -3.60355277 -2.421747671  1.891288e-01  0.088521896
#> [173,]  4  2.2 -4.55171807  1.532795954  5.100633e-02  1.239150708
#> [174,]  4  2.3 -5.27421198  0.892544195 -2.406689e-04 -1.644555536
#> [175,]  4  2.4 -4.83040773  1.490983992  1.809382e+00  1.446356525
#> [176,]  4  2.5 -5.65917594  1.796518242 -8.253280e-01 -0.690560172
#> [177,]  4  2.6 -4.12345391  1.495594927  1.145470e+00 -0.276431085
#> [178,]  4  2.7 -5.37810708  1.071369548  3.157319e-02 -1.109418760
#> [179,]  4  2.8 -5.51174159  1.343760653 -8.352058e-01  0.133869316
#> [180,]  4  2.9 -5.69455604  2.428243970 -6.876365e-02  1.785339052
#> [181,]  4  3.0 -4.69725078  1.594988822  7.467717e-01  2.422163355
#> [182,]  4  3.1 -3.70009719  2.002396299 -4.255187e-01 -1.076828902
#> [183,]  4  3.2 -5.02242490  2.737561590 -7.720822e-01  0.485941110
#> [184,]  4  3.3 -3.96775593  3.967047578  1.527641e-01  1.388521739
#> [185,]  4  3.4 -1.68491489  1.822671083  9.885968e-01 -0.195656817
#> [186,]  4  3.5 -2.00330147  2.424729935 -7.345833e-02 -0.218174798
#> [187,]  4  3.6 -3.42918360  2.390752980 -1.387027e+00 -0.304777955
#> [188,]  4  3.7 -3.07002853  2.802512409 -1.306676e+00  0.597832724
#> [189,]  4  3.8 -4.27318323  0.303268964 -7.683953e-01  1.397429411
#> [190,]  4  3.9 -4.39501688  1.316817954 -5.271081e-01  0.687619761
#> [191,]  4  4.0 -3.75306880  1.811653492 -2.142707e-02  0.320188014
#> [192,]  4  4.1 -2.45912634  2.091186859  6.704981e-01 -0.301869926
#> [193,]  4  4.2 -1.99759959  0.773234305 -4.346170e-01  0.498348686
#> [194,]  4  4.3 -1.02499879  1.640989839 -1.113880e+00 -0.549536918
#> [195,]  4  4.4 -1.40470393  1.098920947  6.071060e-01 -0.279256504
#> [196,]  4  4.5 -0.95186468  0.865666518  2.754570e-01  1.096513443
#> [197,]  4  4.6  0.98903560  2.753107299  1.157347e+00  0.442013088
#> [198,]  4  4.7 -1.12180331  2.504090065 -1.682481e+00  0.241016294
#> [199,]  4  4.8 -1.10551004  2.431151797  8.731909e-02 -0.255607655
#> [200,]  4  4.9  0.70910588  1.645870600  1.353362e+00  0.931032901
#> [201,]  5  0.0 -4.49808098  0.398703072  1.334913e+00 -1.303821154
#> [202,]  5  0.1 -4.88148703 -1.893179308 -8.692718e-01 -0.250914465
#> [203,]  5  0.2 -3.49806238 -0.980364209  5.548695e-02  0.171007374
#> [204,]  5  0.3 -4.51810116 -0.588145118  4.906691e-02 -0.403467479
#> [205,]  5  0.4 -2.81332282  0.334495788 -5.783557e-01  0.104659441
#> [206,]  5  0.5 -3.48607230 -0.826075652 -9.987387e-01 -0.318880787
#> [207,]  5  0.6 -1.61422922 -0.765646445 -2.432780e-03  1.618343936
#> [208,]  5  0.7 -1.44948851  1.586998488  6.555119e-01  0.714188601
#> [209,]  5  0.8 -1.38909286  2.122778754  1.476842e+00  2.965865370
#> [210,]  5  0.9 -0.35075638  1.700025860 -1.909153e+00 -0.795077605
#> [211,]  5  1.0 -1.68765375  1.202344797 -7.024395e-01  0.814365915
#> [212,]  5  1.1 -1.74529362  1.951628324 -3.114302e-01  2.098030810
#> [213,]  5  1.2 -2.66251147  1.713762663 -1.663157e+00  0.300980055
#> [214,]  5  1.3 -4.65553335  1.200641633 -7.505334e-01 -1.083075142
#> [215,]  5  1.4 -3.00591959  2.346078428 -7.773518e-01 -1.006322502
#> [216,]  5  1.5 -2.24305967  2.712361995 -7.225697e-01 -0.035414565
#> [217,]  5  1.6 -2.16557552  2.821322663 -2.188835e+00  1.309124361
#> [218,]  5  1.7 -2.16000410  3.014385928  2.134186e-01  0.750400487
#> [219,]  5  1.8 -1.93193673  0.718318428 -6.319229e-01 -2.138368328
#> [220,]  5  1.9 -3.08973482  2.246648569  1.520491e+00 -0.700354109
#> [221,]  5  2.0 -0.65444973  3.750469200  7.959559e-01 -0.009056475
#> [222,]  5  2.1 -2.27914807  3.541322296 -1.453530e+00 -1.458133487
#> [223,]  5  2.2 -1.79863087  2.568675000  9.839542e-02  0.694529646
#> [224,]  5  2.3 -1.52770036  1.950377072 -5.937710e-01 -2.461335475
#> [225,]  5  2.4 -2.62233614  2.834388064  8.882812e-01  0.143289764
#> [226,]  5  2.5 -1.84314971  3.907217667  5.307042e-02 -0.391222119
#> [227,]  5  2.6 -3.91116444  2.250365836 -5.570236e-01 -0.491164086
#> [228,]  5  2.7 -3.77879982  3.381289600  4.383970e-01 -0.283647452
#> [229,]  5  2.8 -5.66485639  0.837065296  1.526082e-01  0.314794805
#> [230,]  5  2.9 -6.10637108  2.672682036 -1.646176e-01  0.396326578
#> [231,]  5  3.0 -6.79166810  2.677760416  2.019891e+00 -0.225603711
#> [232,]  5  3.1 -3.69947558  3.768060235 -5.293859e-01 -1.924950430
#> [233,]  5  3.2 -5.96955044  1.202901164 -4.707870e-01 -1.439229298
#> [234,]  5  3.3 -5.18274848  1.452513663 -1.545937e+00 -1.469657774
#> [235,]  5  3.4 -4.71287994  1.877386104 -4.052672e-02  0.761863447
#> [236,]  5  3.5 -3.22961923  2.601131640  8.903563e-01 -0.243614982
#> [237,]  5  3.6 -5.27320635  1.662948750 -2.071388e+00  0.269676607
#> [238,]  5  3.7 -3.75794796  1.593522281 -2.500651e-01 -1.558927521
#> [239,]  5  3.8 -5.37616320  2.431225081 -1.181650e+00 -0.535588007
#> [240,]  5  3.9 -4.21558256  1.755342523  1.441937e+00  0.562451973
#> [241,]  5  4.0 -4.10978026  0.649135800  1.357896e+00 -0.178326123
#> [242,]  5  4.1 -4.30353100  2.015614895  3.345028e-01 -0.115135986
#> [243,]  5  4.2 -3.90529295  2.822965360  1.429338e+00 -0.072061472
#> [244,]  5  4.3 -3.11466994  2.746407722 -8.673179e-01  1.210909807
#> [245,]  5  4.4 -3.51318597  0.367976632  9.506517e-01 -0.614896901
#> [246,]  5  4.5 -5.63697088  3.396869101 -5.850115e-01  0.676126458
#> [247,]  5  4.6 -5.06439142  1.624182738  3.209575e-01  0.898599606
#> [248,]  5  4.7 -6.73369622  3.315213862 -2.993960e-01 -1.189317904
#> [249,]  5  4.8 -5.96976168  2.868095590 -2.785431e-01  0.121258850
#> [250,]  5  4.9 -4.59314928  3.906299718  5.461152e-01 -0.011221686