Simulate Data from an Ornstein–Uhlenbeck Model using a State Space Model Parameterization for n > 1 Individuals (Individual-Varying Parameters)

This function simulates data from an Ornstein–Uhlenbeck model using a state space model parameterization for n > 1 individuals. In this model, the parameters can vary across individuals.

Usage

SimSSMOUIVary(
  n,
  mu0,
  sigma0,
  mu,
  phi,
  sigma,
  nu,
  lambda,
  theta,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  delta_t,
  time,
  burn_in = 0
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0

Numeric matrix. The covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

mu

List of numeric vectors. Each element of the list is the long-term mean or equilibrium level (\(\boldsymbol{\mu}\)).

phi

List of numeric matrices. Each element of the list is the rate of mean reversion, determining how quickly the variable returns to its mean (\(\boldsymbol{\Phi}\)).

sigma

List of numeric matrices. Each element of the list is the matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

nu

Numeric vector. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta

Numeric matrix. The measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the observed variables at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

Numeric matrix. The matrix of observed covariates in type = 1 or type = 2. The number of rows should be equal to time + burn_in.

type

Integer. State space model type.

delta_t

Numeric. Time interval (\(\delta_t\)).

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list with the following elements:

call: Function call.

args: Function arguments.

data: Generated data which is a list of length n.

fun: Function used.

Each element of data is a list with the following elements:

id: A vector of ID numbers of length t.

time: A vector of discrete time points from 0 to t - 1.

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

Details

Parameters can vary across individuals by providing a list of parameter values. If the length of any of the parameters (mu0, sigma0, mu, phi, sigma, nu, lambda, theta, gamma_y, or gamma_eta) is less the n, the function will cycle through the available values.

References

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Uhlenbeck, G. E., & Ornstein, L. S. (1930). On the theory of the brownian motion. Physical Review, 36(5), 823–841. doi:10.1103/physrev.36.823

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
# In this example, phi varies across individuals
set.seed(42)
p <- k <- 2
iden <- diag(p)
n <- 5
mu0 <- list(c(-3.0, 1.5))
sigma0 <- list(iden)
mu <- list(c(5.76, 5.18))
phi <- list(
  as.matrix(Matrix::expm(diag(x = -0.1, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.2, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.3, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.4, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.5, nrow = k)))
)
sigma <- list(
  matrix(data = c(2.79, 0.06, 0.06, 3.27), nrow = p)
)
nu <- list(rep(x = 0, times = k))
lambda <- list(diag(k))
theta <- list(diag(x = 0.50, nrow = k))
delta_t <- 0.10
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- list(0.10 * diag(k))
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = k * (time + burn_in)),
        ncol = k
      )
    )
  }
)

# Type 0
SimSSMOUIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  type = 0,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>        id time          y1          y2
#>   [1,]  1  0.0 -1.68501472  1.36033376
#>   [2,]  1  0.1 -1.21709640  3.39556396
#>   [3,]  1  0.2 -1.49792726  3.72842716
#>   [4,]  1  0.3 -1.14894736  4.04585609
#>   [5,]  1  0.4 -0.48895787  3.09783182
#>   [6,]  1  0.5  0.23616039  3.49511678
#>   [7,]  1  0.6  0.08539169  4.52975673
#>   [8,]  1  0.7  0.51416724  3.88555327
#>   [9,]  1  0.8  0.39209156  4.26485529
#>  [10,]  1  0.9  0.99329687  5.15712618
#>  [11,]  1  1.0  0.30743901  6.23121268
#>  [12,]  1  1.1  1.54601995  6.85605147
#>  [13,]  1  1.2  1.75476333  6.76277298
#>  [14,]  1  1.3  3.02597817  5.76359170
#>  [15,]  1  1.4  2.67245872  6.00566758
#>  [16,]  1  1.5  3.05710202  6.45440819
#>  [17,]  1  1.6  3.97473549  6.31587448
#>  [18,]  1  1.7  2.31344220  5.53405990
#>  [19,]  1  1.8  2.67679161  6.25139805
#>  [20,]  1  1.9  2.69838130  7.09540698
#>  [21,]  1  2.0  4.37328206  5.73224650
#>  [22,]  1  2.1  5.00822100  4.61277143
#>  [23,]  1  2.2  5.46754297  5.25024826
#>  [24,]  1  2.3  5.91427802  4.09210289
#>  [25,]  1  2.4  5.17440321  4.66838493
#>  [26,]  1  2.5  5.25866800  5.94576978
#>  [27,]  1  2.6  3.82521353  7.56669613
#>  [28,]  1  2.7  3.85641801  5.26839054
#>  [29,]  1  2.8  1.96722644  7.01410726
#>  [30,]  1  2.9  5.06455323  4.10426101
#>  [31,]  1  3.0  3.01779851  5.40038733
#>  [32,]  1  3.1  4.35330151  5.22429540
#>  [33,]  1  3.2  4.75549599  5.45290280
#>  [34,]  1  3.3  5.07611404  3.62500444
#>  [35,]  1  3.4  3.50587391  4.15444368
#>  [36,]  1  3.5  5.64113678  6.97405991
#>  [37,]  1  3.6  5.44138999  5.41160480
#>  [38,]  1  3.7  4.36734198  4.21453501
#>  [39,]  1  3.8  7.03360050  4.62397046
#>  [40,]  1  3.9  5.02154079  5.00698938
#>  [41,]  1  4.0  4.97436640  4.65501519
#>  [42,]  1  4.1  6.58404599  4.09837299
#>  [43,]  1  4.2  6.83281230  4.85426283
#>  [44,]  1  4.3  6.58089490  5.27537978
#>  [45,]  1  4.4  6.94804692  4.47433291
#>  [46,]  1  4.5  6.91464696  5.46188094
#>  [47,]  1  4.6  6.95781633  5.97706278
#>  [48,]  1  4.7  7.62944567  5.70738379
#>  [49,]  1  4.8  8.31489139  4.83770105
#>  [50,]  1  4.9  8.05901128  6.96702552
#>  [51,]  2  0.0 -3.59775671  1.61090258
#>  [52,]  2  0.1 -0.66790729  0.89039524
#>  [53,]  2  0.2 -2.11004011  1.64192363
#>  [54,]  2  0.3 -0.47408168  3.10028123
#>  [55,]  2  0.4  0.68911964  1.85832908
#>  [56,]  2  0.5 -0.24370355  1.97836917
#>  [57,]  2  0.6  0.51739545  2.62791738
#>  [58,]  2  0.7  1.36949416  2.40563578
#>  [59,]  2  0.8  3.11458344  1.17850906
#>  [60,]  2  0.9  2.63176615  2.21847149
#>  [61,]  2  1.0  2.88062459  2.09469318
#>  [62,]  2  1.1  1.94058963  0.63533963
#>  [63,]  2  1.2  3.64240104  2.18524843
#>  [64,]  2  1.3  1.96955858  3.23827405
#>  [65,]  2  1.4  2.89823446  2.61293077
#>  [66,]  2  1.5  3.60190244  3.41903918
#>  [67,]  2  1.6  4.77722053  3.56144136
#>  [68,]  2  1.7  5.06965245  5.28313810
#>  [69,]  2  1.8  6.06998541  4.60831450
#>  [70,]  2  1.9  4.07006096  4.52232105
#>  [71,]  2  2.0  4.40193273  3.76437182
#>  [72,]  2  2.1  3.62674123  5.38703871
#>  [73,]  2  2.2  4.82604852  5.20700148
#>  [74,]  2  2.3  3.19831284  5.53180384
#>  [75,]  2  2.4  4.29014423  4.56707615
#>  [76,]  2  2.5  4.29275127  5.05714703
#>  [77,]  2  2.6  4.01718579  4.92479115
#>  [78,]  2  2.7  4.38026137  5.12688850
#>  [79,]  2  2.8  3.72227905  4.77231334
#>  [80,]  2  2.9  5.23876572  3.51568557
#>  [81,]  2  3.0  5.39266623  2.95847608
#>  [82,]  2  3.1  4.42863295  4.00449039
#>  [83,]  2  3.2  4.15169772  3.61403873
#>  [84,]  2  3.3  3.89527111  4.67113797
#>  [85,]  2  3.4  3.10094676  3.70899175
#>  [86,]  2  3.5  3.26371537  3.43935409
#>  [87,]  2  3.6  2.31210446  4.54026506
#>  [88,]  2  3.7  3.39802639  2.57279096
#>  [89,]  2  3.8  2.37352265  4.61397890
#>  [90,]  2  3.9  4.06711531  5.02547573
#>  [91,]  2  4.0  5.39536459  4.04413564
#>  [92,]  2  4.1  4.57049840  5.52244529
#>  [93,]  2  4.2  5.24713378  4.57901527
#>  [94,]  2  4.3  3.19956828  2.36033380
#>  [95,]  2  4.4  4.33118142  3.93536073
#>  [96,]  2  4.5  4.36982909  2.73875607
#>  [97,]  2  4.6  4.31386925  4.12080460
#>  [98,]  2  4.7  3.36737688  4.30638873
#>  [99,]  2  4.8  2.71816779  3.59796323
#> [100,]  2  4.9  4.10030244  5.95266173
#> [101,]  3  0.0 -4.10679982  1.65805024
#> [102,]  3  0.1 -2.65586956  0.01739116
#> [103,]  3  0.2 -1.46967335 -1.13861066
#> [104,]  3  0.3 -0.07113598  0.20244030
#> [105,]  3  0.4 -0.90131138  0.80877116
#> [106,]  3  0.5  0.34872532  3.72896179
#> [107,]  3  0.6 -0.54316275  4.59847049
#> [108,]  3  0.7  0.95426273  4.69008706
#> [109,]  3  0.8  0.20358331  4.36470220
#> [110,]  3  0.9  1.01422361  4.26881726
#> [111,]  3  1.0  2.34442161  3.70775560
#> [112,]  3  1.1  0.57276579  5.04953981
#> [113,]  3  1.2  1.26895508  5.97345367
#> [114,]  3  1.3  1.93942654  6.24333604
#> [115,]  3  1.4  1.47887013  7.56175280
#> [116,]  3  1.5  1.55467295  4.65629613
#> [117,]  3  1.6  2.36910856  4.52626022
#> [118,]  3  1.7  4.35282864  3.53607144
#> [119,]  3  1.8  1.16507457  4.52736454
#> [120,]  3  1.9  1.34470343  4.42920345
#> [121,]  3  2.0  3.33057183  4.39326612
#> [122,]  3  2.1  3.58800636  5.82121501
#> [123,]  3  2.2  2.23638225  4.72585069
#> [124,]  3  2.3  4.35265064  6.34149050
#> [125,]  3  2.4  4.52445516  6.57990820
#> [126,]  3  2.5  4.16506454  4.90239831
#> [127,]  3  2.6  5.10172835  5.58954300
#> [128,]  3  2.7  4.55952503  4.77876361
#> [129,]  3  2.8  4.34128287  4.60301305
#> [130,]  3  2.9  4.64779735  1.99322485
#> [131,]  3  3.0  4.20830801  4.19325443
#> [132,]  3  3.1  4.89237888  3.95306471
#> [133,]  3  3.2  5.39013167  4.17468627
#> [134,]  3  3.3  5.76068397  3.34887232
#> [135,]  3  3.4  5.53130753  3.96513640
#> [136,]  3  3.5  4.27978500  3.17746024
#> [137,]  3  3.6  4.96678940  4.11504071
#> [138,]  3  3.7  4.36062247  3.14269797
#> [139,]  3  3.8  6.32274223  2.40419562
#> [140,]  3  3.9  4.82936657  4.52965418
#> [141,]  3  4.0  4.43543007  4.74225773
#> [142,]  3  4.1  4.71361315  3.83756028
#> [143,]  3  4.2  4.75051990  5.92374523
#> [144,]  3  4.3  6.24748647  3.82162526
#> [145,]  3  4.4  4.78642239  4.22547012
#> [146,]  3  4.5  4.11531429  6.64851155
#> [147,]  3  4.6  4.15076859  5.21110421
#> [148,]  3  4.7  3.25697516  6.56080859
#> [149,]  3  4.8  4.36355429  4.77462881
#> [150,]  3  4.9  4.91521963  4.31769099
#> [151,]  4  0.0 -4.81711952  1.19333758
#> [152,]  4  0.1 -3.81299002  2.64708838
#> [153,]  4  0.2 -1.20003743  0.99898314
#> [154,]  4  0.3 -2.64371654  0.73549542
#> [155,]  4  0.4 -2.79316402  0.17498307
#> [156,]  4  0.5 -1.08364150  1.79695098
#> [157,]  4  0.6  0.57905787  3.37459743
#> [158,]  4  0.7  1.28379881  2.74390813
#> [159,]  4  0.8  1.28739928  4.99231782
#> [160,]  4  0.9  2.01066811  4.91636748
#> [161,]  4  1.0  2.37420048  4.00265019
#> [162,]  4  1.1  3.27364349  3.89478962
#> [163,]  4  1.2  3.10315951  5.17385041
#> [164,]  4  1.3  2.68510721  4.12895221
#> [165,]  4  1.4  4.20075061  7.38908638
#> [166,]  4  1.5  3.44321558  6.37759739
#> [167,]  4  1.6  4.67985981  7.34268635
#> [168,]  4  1.7  4.71866008  7.69069083
#> [169,]  4  1.8  5.93053433  7.42086294
#> [170,]  4  1.9  6.93806724  7.38830592
#> [171,]  4  2.0  6.00999039  7.84526246
#> [172,]  4  2.1  7.38767800  7.04626973
#> [173,]  4  2.2  6.86553352  7.23053450
#> [174,]  4  2.3  6.87833845  6.53772967
#> [175,]  4  2.4  6.02720095  4.79424520
#> [176,]  4  2.5  6.12168381  4.86939161
#> [177,]  4  2.6  5.90205529  4.17069367
#> [178,]  4  2.7  6.74868034  4.71305693
#> [179,]  4  2.8  6.45045580  3.67461314
#> [180,]  4  2.9  6.97608520  3.21989735
#> [181,]  4  3.0  5.56743956  3.57905144
#> [182,]  4  3.1  6.10172291  3.17041325
#> [183,]  4  3.2  7.31493266  3.11046454
#> [184,]  4  3.3  7.63071799  4.15633241
#> [185,]  4  3.4  7.67359587  4.52518380
#> [186,]  4  3.5  7.13727599  4.00278992
#> [187,]  4  3.6  7.13247834  3.62296223
#> [188,]  4  3.7  7.69688632  2.75868109
#> [189,]  4  3.8  8.17301971  4.11955724
#> [190,]  4  3.9  8.00383253  3.75711096
#> [191,]  4  4.0  6.56982408  2.74807613
#> [192,]  4  4.1  7.86166656  4.32851899
#> [193,]  4  4.2  6.85530986  4.27658817
#> [194,]  4  4.3  6.98764935  4.58750685
#> [195,]  4  4.4  7.46100378  5.36989932
#> [196,]  4  4.5  6.36056953  5.27299977
#> [197,]  4  4.6  7.30571019  6.51174540
#> [198,]  4  4.7  4.55266027  5.88957374
#> [199,]  4  4.8  6.70310036  5.37584203
#> [200,]  4  4.9  5.45228851  5.09505681
#> [201,]  5  0.0 -3.64434131  1.12257464
#> [202,]  5  0.1 -2.98749343  0.31126104
#> [203,]  5  0.2 -2.27362467  2.21782889
#> [204,]  5  0.3 -2.09071236  2.94779280
#> [205,]  5  0.4 -1.78545851  3.00219950
#> [206,]  5  0.5 -3.36916077  3.07285640
#> [207,]  5  0.6 -1.04704365  4.14893080
#> [208,]  5  0.7 -1.14962003  2.06891406
#> [209,]  5  0.8 -1.43436328  5.08160687
#> [210,]  5  0.9 -3.07257770  2.88797442
#> [211,]  5  1.0 -1.11677939  3.81102339
#> [212,]  5  1.1 -1.26879892  4.22836097
#> [213,]  5  1.2 -0.51777612  3.67688314
#> [214,]  5  1.3 -0.30941647  4.85136753
#> [215,]  5  1.4  2.99485958  4.23633138
#> [216,]  5  1.5  2.65090768  4.55592503
#> [217,]  5  1.6  4.55842266  6.29925922
#> [218,]  5  1.7  3.61057920  6.02046613
#> [219,]  5  1.8  2.63132144  5.72633729
#> [220,]  5  1.9  2.13475247  4.80050237
#> [221,]  5  2.0  2.72009441  4.92965669
#> [222,]  5  2.1  3.39538404  6.79626997
#> [223,]  5  2.2  3.54643191  6.08473401
#> [224,]  5  2.3  3.30850429  5.42807144
#> [225,]  5  2.4  4.19326346  4.60315967
#> [226,]  5  2.5  4.40616106  5.00296558
#> [227,]  5  2.6  3.16995160  6.28571060
#> [228,]  5  2.7  4.31222260  5.66820020
#> [229,]  5  2.8  4.24825972  7.00682320
#> [230,]  5  2.9  3.12569572  8.52248141
#> [231,]  5  3.0  3.65292471  8.10309741
#> [232,]  5  3.1  4.75845181  6.34378563
#> [233,]  5  3.2  4.37017897  5.61143066
#> [234,]  5  3.3  3.22441252  5.73503104
#> [235,]  5  3.4  3.34746307  6.32773871
#> [236,]  5  3.5  5.07488538  5.13874784
#> [237,]  5  3.6  3.09108869  6.79434147
#> [238,]  5  3.7  3.37401214  7.56666653
#> [239,]  5  3.8  4.86846172  6.39957656
#> [240,]  5  3.9  7.23706199  6.16426208
#> [241,]  5  4.0  7.04017099  7.41499209
#> [242,]  5  4.1  4.70404400  8.33475782
#> [243,]  5  4.2  5.36853710  8.09792841
#> [244,]  5  4.3  4.99837426  8.97376232
#> [245,]  5  4.4  5.29199796 10.75249712
#> [246,]  5  4.5  3.62939166  9.46102175
#> [247,]  5  4.6  6.76011112 11.21084374
#> [248,]  5  4.7  5.79405669 10.43275108
#> [249,]  5  4.8  5.25784511 10.62052916
#> [250,]  5  4.9  5.61394656  9.15516446

# Type 1
SimSSMOUIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>        id time          y1         y2            x1           x2
#>   [1,]  1  0.0 -4.70711009  1.5965542  1.370958e+00  0.321925265
#>   [2,]  1  0.1 -1.85993623  2.6583445 -5.646982e-01 -0.783838941
#>   [3,]  1  0.2 -1.20703367  4.1972215  3.631284e-01  1.575727520
#>   [4,]  1  0.3 -0.93219191  3.0152736  6.328626e-01  0.642899306
#>   [5,]  1  0.4  0.19888747  3.9949348  4.042683e-01  0.089760647
#>   [6,]  1  0.5  1.04530756  4.7391192 -1.061245e-01  0.276550747
#>   [7,]  1  0.6  1.05595481  3.7452485  1.511522e+00  0.679288816
#>   [8,]  1  0.7  0.30439974  2.8550045 -9.465904e-02  0.089832887
#>   [9,]  1  0.8  2.50046134  3.3449705  2.018424e+00 -2.993090083
#>  [10,]  1  0.9  3.22765547  3.8779714 -6.271410e-02  0.284882954
#>  [11,]  1  1.0  3.05600908  2.5940199  1.304870e+00 -0.367234643
#>  [12,]  1  1.1  4.05396100  3.2041844  2.286645e+00  0.185230565
#>  [13,]  1  1.2  3.97791949  4.3417702 -1.388861e+00  0.581823727
#>  [14,]  1  1.3  4.24294705  3.4181426 -2.787888e-01  1.399736827
#>  [15,]  1  1.4  3.83635092  3.7400799 -1.333213e-01 -0.727292059
#>  [16,]  1  1.5  3.05264938  2.9039469  6.359504e-01  1.302542632
#>  [17,]  1  1.6  2.92523256  4.4862314 -2.842529e-01  0.335848120
#>  [18,]  1  1.7  3.95598274  4.1272324 -2.656455e+00  1.038506099
#>  [19,]  1  1.8  2.67918940  5.7166435 -2.440467e+00  0.920728568
#>  [20,]  1  1.9  3.32096225  6.1811165  1.320113e+00  0.720878163
#>  [21,]  1  2.0  3.26536465  5.5410486 -3.066386e-01 -1.043118939
#>  [22,]  1  2.1  3.69197471  7.1100127 -1.781308e+00 -0.090186387
#>  [23,]  1  2.2  4.72876925  6.4938713 -1.719174e-01  0.623518162
#>  [24,]  1  2.3  5.00635162  6.0233849  1.214675e+00 -0.953523358
#>  [25,]  1  2.4  4.87116407  5.7205080  1.895193e+00 -0.542828815
#>  [26,]  1  2.5  5.56917796  6.9000144 -4.304691e-01  0.580996498
#>  [27,]  1  2.6  5.20577625  5.0865549 -2.572694e-01  0.768178738
#>  [28,]  1  2.7  5.32124668  6.2064376 -1.763163e+00  0.463767589
#>  [29,]  1  2.8  6.79160104  6.7512381  4.600974e-01 -0.885776297
#>  [30,]  1  2.9  6.25626785  5.6187348 -6.399949e-01 -1.099780899
#>  [31,]  1  3.0  5.68540509  5.3411843  4.554501e-01  1.512707010
#>  [32,]  1  3.1  3.55523542  4.2812897  7.048373e-01  0.257921438
#>  [33,]  1  3.2  5.53243489  6.3884204  1.035104e+00  0.088440229
#>  [34,]  1  3.3  5.18598026  5.7294149 -6.089264e-01 -0.120896538
#>  [35,]  1  3.4  6.97735516  3.2597638  5.049551e-01 -1.194328895
#>  [36,]  1  3.5  5.60797168  4.8990909 -1.717009e+00  0.611996898
#>  [37,]  1  3.6  5.48881151  3.7082991 -7.844590e-01 -0.217139846
#>  [38,]  1  3.7  5.09924824  4.6330248 -8.509076e-01 -0.182756706
#>  [39,]  1  3.8  5.15821257  4.2737667 -2.414208e+00  0.933346329
#>  [40,]  1  3.9  4.54467733  4.4790319  3.612261e-02  0.821773111
#>  [41,]  1  4.0  5.97167614  4.0418653  2.059986e-01  1.392116376
#>  [42,]  1  4.1  5.15001414  4.1708115 -3.610573e-01 -0.476173923
#>  [43,]  1  4.2  4.71752153  2.7465370  7.581632e-01  0.650348561
#>  [44,]  1  4.3  3.70332727  4.6933374 -7.267048e-01  1.391110456
#>  [45,]  1  4.4  3.78045915  3.4731323 -1.368281e+00 -1.110788879
#>  [46,]  1  4.5  4.00720699  3.7888039  4.328180e-01 -0.860792587
#>  [47,]  1  4.6  3.56577579  4.0781019 -8.113932e-01 -1.131738681
#>  [48,]  1  4.7  1.69159146  4.1979314  1.444101e+00 -1.459214000
#>  [49,]  1  4.8  1.49120526  6.7968324 -4.314462e-01  0.079982553
#>  [50,]  1  4.9  2.21588541  4.4077939  6.556479e-01  0.653204340
#>  [51,]  2  0.0 -3.80041653  3.2845949  1.200965e+00 -0.040698475
#>  [52,]  2  0.1 -2.20886623  3.4373004  1.044751e+00 -1.551544822
#>  [53,]  2  0.2  0.30931083  5.3394731 -1.003209e+00  1.167169549
#>  [54,]  2  0.3 -0.07537028  3.8527932  1.848482e+00 -0.273645701
#>  [55,]  2  0.4  0.33098483  6.1248308 -6.667734e-01 -0.467845325
#>  [56,]  2  0.5 -0.46181721  5.3202219  1.055138e-01 -1.238252328
#>  [57,]  2  0.6 -0.03475194  4.4906278 -4.222559e-01 -0.007762034
#>  [58,]  2  0.7  0.17083252  4.4536790 -1.223502e-01 -0.800282178
#>  [59,]  2  0.8  1.40615188  5.9642817  1.881930e-01 -0.533492330
#>  [60,]  2  0.9  2.11256568  6.5489033  1.191610e-01  1.287675246
#>  [61,]  2  1.0  2.38339074  5.3722678 -2.509255e-02 -0.175525870
#>  [62,]  2  1.1  2.43464049  5.5439644  1.080727e-01 -1.071782384
#>  [63,]  2  1.2  2.90882078  4.8329766 -4.854352e-01  0.163206882
#>  [64,]  2  1.3  2.80790798  5.3533244 -5.042171e-01 -0.362738416
#>  [65,]  2  1.4  2.12221084  4.6234641 -1.661099e+00  0.590013548
#>  [66,]  2  1.5  3.18329035  3.0223538 -3.823337e-01  1.432421928
#>  [67,]  2  1.6  4.28391099  2.3769136 -5.126503e-01 -0.992692511
#>  [68,]  2  1.7  3.76757598  1.8204428  2.701891e+00  0.454650298
#>  [69,]  2  1.8  3.21893716  3.6009218 -1.362116e+00  0.084898059
#>  [70,]  2  1.9  3.64739003  2.2367135  1.372562e-01  0.895565582
#>  [71,]  2  2.0  3.46839584  2.4705766 -1.493625e+00 -0.229778139
#>  [72,]  2  2.1  3.43056650  2.7362268 -1.470436e+00  0.836619068
#>  [73,]  2  2.2  2.48482568  2.4571300  1.247024e-01 -1.745055861
#>  [74,]  2  2.3  2.52049480  2.6550086 -9.966391e-01  1.689458921
#>  [75,]  2  2.4  2.47442092  2.3601567 -1.822614e-03  0.864777979
#>  [76,]  2  2.5  3.20287199  3.9347932 -4.282589e-01 -0.150775989
#>  [77,]  2  2.6  2.32148566  4.6238655 -6.136716e-01 -1.449007130
#>  [78,]  2  2.7  2.61750966  4.0981098 -2.024678e+00  0.643008700
#>  [79,]  2  2.8  2.73445291  4.1621778 -1.224748e+00  0.483193864
#>  [80,]  2  2.9  2.83837204  4.8101881  1.795164e-01 -0.006355626
#>  [81,]  2  3.0  3.85458606  3.4259639  5.676206e-01  0.151455893
#>  [82,]  2  3.1  3.69329595  3.2191504 -4.928774e-01 -0.584108970
#>  [83,]  2  3.2  4.08122872  2.9358836  6.288407e-05  0.368806733
#>  [84,]  2  3.3  4.68707963  1.7568569  1.122890e+00  0.294654340
#>  [85,]  2  3.4  2.85449309  3.4069449  1.439856e+00 -0.279259373
#>  [86,]  2  3.5  2.97616192  4.0768169 -1.097114e+00 -1.336236655
#>  [87,]  2  3.6  3.49813090  4.8775316 -1.173196e-01  0.700748818
#>  [88,]  2  3.7  4.34301981  6.4108118  1.201498e+00  0.554196622
#>  [89,]  2  3.8  4.35124329  3.6830383 -4.697296e-01 -0.836306593
#>  [90,]  2  3.9  3.71842321  3.9411310 -5.246948e-02 -1.594588162
#>  [91,]  2  4.0  5.95214728  6.3915707 -8.610730e-02  0.204958581
#>  [92,]  2  4.1  5.02453849  5.2724725 -8.876790e-01 -0.345087978
#>  [93,]  2  4.2  5.61898919  4.7305027 -4.446840e-01  0.252611703
#>  [94,]  2  4.3  5.85026840  4.8768479 -2.944488e-02 -1.294002465
#>  [95,]  2  4.4  4.58898261  4.5531428 -4.138688e-01 -0.959170444
#>  [96,]  2  4.5  3.96263085  5.5398266  1.113386e+00  1.085774854
#>  [97,]  2  4.6  6.11197085  5.2133744 -4.809928e-01  0.403774905
#>  [98,]  2  4.7  5.12348848  5.5188755 -4.331690e-01  0.586487537
#>  [99,]  2  4.8  4.80816725  3.5404537  6.968626e-01  1.815228446
#> [100,]  2  4.9  2.92696685  4.5821195 -1.056368e+00  0.128821429
#> [101,]  3  0.0 -1.86109624  0.4510187 -2.000929e+00 -1.096156242
#> [102,]  3  0.1 -1.26582217  1.3442108  3.337772e-01  0.049050451
#> [103,]  3  0.2 -1.29303980 -0.3173724  1.171325e+00 -1.198495857
#> [104,]  3  0.3  0.40249339  0.3350212  2.059539e+00  0.190018999
#> [105,]  3  0.4 -1.22331008  0.7262334 -1.376862e+00  1.297705900
#> [106,]  3  0.5 -0.69892859  2.7430542 -1.150856e+00 -1.033873723
#> [107,]  3  0.6 -1.61205616  0.4277799 -7.058214e-01 -0.738440754
#> [108,]  3  0.7 -1.10479654  0.3506862 -1.054056e+00  0.046563939
#> [109,]  3  0.8  1.58618689  3.3479136 -6.457437e-01 -1.017596120
#> [110,]  3  0.9  2.69256968  2.8893381 -1.853780e-01 -0.383283960
#> [111,]  3  1.0  2.78090428  3.9335809 -1.201222e+00  0.872755412
#> [112,]  3  1.1  2.92182664  6.2330644  2.036972e+00  0.969545014
#> [113,]  3  1.2  3.22343985  4.4985315  1.077747e-01  0.383846665
#> [114,]  3  1.3  2.50020371  4.9529084 -8.410810e-02 -1.851555663
#> [115,]  3  1.4  3.89368766  6.1855624  4.956196e-01 -0.053996737
#> [116,]  3  1.5  5.45313126  5.9687181  3.741519e-02  1.064773214
#> [117,]  3  1.6  5.84396552  4.2420414 -1.320880e-01  0.813195037
#> [118,]  3  1.7  6.11414506  3.4898771  1.476787e+00 -0.190816474
#> [119,]  3  1.8  5.00459116  2.2778808 -2.170302e-01 -2.699929809
#> [120,]  3  1.9  3.44771269  5.8753734 -1.283602e+00  0.060966639
#> [121,]  3  2.0  5.90566506  4.1928036  3.856679e-01  0.573751697
#> [122,]  3  2.1  4.80616726  4.9235738 -3.515129e-01  0.045803580
#> [123,]  3  2.2  5.56096258  6.0794653 -5.217961e-01  0.157412540
#> [124,]  3  2.3  5.21843981  5.8637818 -1.068131e+00  0.431565373
#> [125,]  3  2.4  5.61023204  5.1933714  4.283659e-01 -0.396549736
#> [126,]  3  2.5  5.70977542  3.5273970 -1.740182e-01  1.309978226
#> [127,]  3  2.6  3.86146370  6.0653036  5.156677e-01  0.470393400
#> [128,]  3  2.7  3.89186501  3.5459683 -2.343653e-01 -1.242670271
#> [129,]  3  2.8  5.63196046  4.5820979 -6.585034e-01  1.381575456
#> [130,]  3  2.9  2.75315764  4.1220578  1.250237e+00  1.204458937
#> [131,]  3  3.0  4.59265506  4.3437049 -2.717637e-01  0.824073964
#> [132,]  3  3.1  4.81437727  4.0661283  9.479520e-01 -1.662629402
#> [133,]  3  3.2  5.08245649  4.4057037 -1.201582e+00 -0.569306344
#> [134,]  3  3.3  5.67898317  3.1289063 -4.661161e-01  0.635513817
#> [135,]  3  3.4  4.88132489  3.1220021 -2.693514e-01  0.043722008
#> [136,]  3  3.5  6.03785613  3.2931313 -3.909654e-01  0.348012304
#> [137,]  3  3.6  5.95701893  4.7709530  1.348707e+00  2.459593549
#> [138,]  3  3.7  7.56229163  3.5489593 -2.276470e-02 -0.818380324
#> [139,]  3  3.8  5.65994435  4.8737467  2.442259e-01 -2.113200115
#> [140,]  3  3.9  5.74335831  4.8153618 -9.423717e-01  0.273695272
#> [141,]  3  4.0  3.97290634  6.0101494 -7.292173e-01 -0.687596841
#> [142,]  3  4.1  4.69166723  5.5685922  9.980689e-01  0.446041053
#> [143,]  3  4.2  5.95909106  6.9245661  1.258482e+00 -0.812384724
#> [144,]  3  4.3  5.26373374  6.4802469  1.248864e+00  2.212055480
#> [145,]  3  4.4  6.75755010  6.6020885 -1.380637e+00 -0.123705972
#> [146,]  3  4.5  7.55917131  6.4483831  2.049961e+00 -0.477335506
#> [147,]  3  4.6  7.84570837  6.5332028  1.016873e+00 -0.166261491
#> [148,]  3  4.7  8.03479899  5.7427735 -2.671746e-02  0.862563384
#> [149,]  3  4.8  6.77848765  6.3216172  7.036078e-01  0.097340485
#> [150,]  3  4.9  6.36540044  7.5559464 -9.713852e-01 -1.625616739
#> [151,]  4  0.0 -4.98958069  2.0654102 -4.620768e-03  0.724173801
#> [152,]  4  0.1 -4.61175935  2.7687399  7.602422e-01 -0.832552826
#> [153,]  4  0.2 -4.55354130  1.0986109  3.899091e-02  0.732528487
#> [154,]  4  0.3 -2.98676171  2.5604115  7.350721e-01 -0.871926870
#> [155,]  4  0.4 -2.45099057  3.0637164 -1.464726e-01 -0.453397511
#> [156,]  4  0.5 -2.69946923  4.9076152 -5.788734e-02  1.187534279
#> [157,]  4  0.6 -0.98437136  5.6521645  4.823695e-01 -0.290145312
#> [158,]  4  0.7 -1.96997335  5.7399336  9.929436e-01  0.828546145
#> [159,]  4  0.8  0.45362582  5.4619840 -1.246395e+00 -0.291227709
#> [160,]  4  0.9 -1.43351254  6.0794781 -3.348752e-02 -1.576362405
#> [161,]  4  1.0 -0.21907091  5.2098767 -7.096218e-02 -0.848815697
#> [162,]  4  1.1  0.86545969  6.3551439 -7.589207e-01 -1.088519862
#> [163,]  4  1.2  0.85829537  5.4805248 -1.034359e+00 -0.484290571
#> [164,]  4  1.3  1.38845829  6.8159379 -6.307320e-01 -0.336311209
#> [165,]  4  1.4  2.09898797  4.9568605  5.868077e-01 -0.153357891
#> [166,]  4  1.5  2.31980091  4.8163219 -4.163227e-01 -0.243247229
#> [167,]  4  1.6  1.76873818  7.3147094 -7.848878e-01  1.892202042
#> [168,]  4  1.7  1.91978393  5.3902387  1.634163e-01 -1.385998337
#> [169,]  4  1.8  3.53781318  5.5958188 -1.236714e+00 -0.414824301
#> [170,]  4  1.9  2.84386152  5.5551234  1.045874e+00  0.349081528
#> [171,]  4  2.0  3.20240400  4.8368017 -4.845954e-01  1.628442266
#> [172,]  4  2.1  2.27728718  4.5409714  1.891288e-01  0.088521896
#> [173,]  4  2.2  1.81900830  5.8022116  5.100633e-02  1.239150708
#> [174,]  4  2.3  1.14099887  4.5223207 -2.406689e-04 -1.644555536
#> [175,]  4  2.4  0.49808422  4.5639025  1.809382e+00  1.446356525
#> [176,]  4  2.5  2.42058418  5.1587096 -8.253280e-01 -0.690560172
#> [177,]  4  2.6  2.01908172  5.9070679  1.145470e+00 -0.276431085
#> [178,]  4  2.7  2.18419843  2.9128975  3.157319e-02 -1.109418760
#> [179,]  4  2.8  2.35253349  3.8527253 -8.352058e-01  0.133869316
#> [180,]  4  2.9  2.53153158  3.3669801 -6.876365e-02  1.785339052
#> [181,]  4  3.0  4.55774940  3.8499679  7.467717e-01  2.422163355
#> [182,]  4  3.1  2.26572354  4.1390655 -4.255187e-01 -1.076828902
#> [183,]  4  3.2  3.31325518  4.1487919 -7.720822e-01  0.485941110
#> [184,]  4  3.3  3.83740141  5.5747752  1.527641e-01  1.388521739
#> [185,]  4  3.4  3.52718810  6.4845445  9.885968e-01 -0.195656817
#> [186,]  4  3.5  2.88438779  6.5252267 -7.345833e-02 -0.218174798
#> [187,]  4  3.6  3.96735494  6.8638348 -1.387027e+00 -0.304777955
#> [188,]  4  3.7  3.21266441  5.0151405 -1.306676e+00  0.597832724
#> [189,]  4  3.8  3.93993649  6.0362349 -7.683953e-01  1.397429411
#> [190,]  4  3.9  2.85079294  6.5537988 -5.271081e-01  0.687619761
#> [191,]  4  4.0  3.61859474  4.0769999 -2.142707e-02  0.320188014
#> [192,]  4  4.1  3.78833334  6.7312932  6.704981e-01 -0.301869926
#> [193,]  4  4.2  4.64607549  6.2103965 -4.346170e-01  0.498348686
#> [194,]  4  4.3  3.29574322  7.1912876 -1.113880e+00 -0.549536918
#> [195,]  4  4.4  4.09432061  6.9295221  6.071060e-01 -0.279256504
#> [196,]  4  4.5  2.63615299  7.8542804  2.754570e-01  1.096513443
#> [197,]  4  4.6  4.62817471  6.0133369  1.157347e+00  0.442013088
#> [198,]  4  4.7  2.61758930  7.7886639 -1.682481e+00  0.241016294
#> [199,]  4  4.8  3.50110487  4.7569858  8.731909e-02 -0.255607655
#> [200,]  4  4.9  1.98966731  5.2795923  1.353362e+00  0.931032901
#> [201,]  5  0.0 -4.03128423  0.8865136  1.334913e+00 -1.303821154
#> [202,]  5  0.1 -0.29115117  1.8355479 -8.692718e-01 -0.250914465
#> [203,]  5  0.2 -2.18045300  2.6952521  5.548695e-02  0.171007374
#> [204,]  5  0.3 -0.99444932  1.7224662  4.906691e-02 -0.403467479
#> [205,]  5  0.4 -0.49810888  3.3589190 -5.783557e-01  0.104659441
#> [206,]  5  0.5 -1.83153334  2.9555017 -9.987387e-01 -0.318880787
#> [207,]  5  0.6 -0.43967883  3.1619701 -2.432780e-03  1.618343936
#> [208,]  5  0.7  0.68124279  2.9166893  6.555119e-01  0.714188601
#> [209,]  5  0.8 -0.66964412  1.6735519  1.476842e+00  2.965865370
#> [210,]  5  0.9  1.09692352  2.8344580 -1.909153e+00 -0.795077605
#> [211,]  5  1.0  1.16624295  2.8486237 -7.024395e-01  0.814365915
#> [212,]  5  1.1  2.10177124  5.3606242 -3.114302e-01  2.098030810
#> [213,]  5  1.2  3.61776131  2.9667171 -1.663157e+00  0.300980055
#> [214,]  5  1.3  2.09494522  3.2460014 -7.505334e-01 -1.083075142
#> [215,]  5  1.4  2.03252086  4.0342169 -7.773518e-01 -1.006322502
#> [216,]  5  1.5  3.78209428  3.1043603 -7.225697e-01 -0.035414565
#> [217,]  5  1.6  3.15819500  4.0290077 -2.188835e+00  1.309124361
#> [218,]  5  1.7  3.29683399  3.4128072  2.134186e-01  0.750400487
#> [219,]  5  1.8  3.28324647  4.2784645 -6.319229e-01 -2.138368328
#> [220,]  5  1.9  2.50686218  2.5377312  1.520491e+00 -0.700354109
#> [221,]  5  2.0  2.81402075  3.2166954  7.959559e-01 -0.009056475
#> [222,]  5  2.1  3.04026262  3.8599688 -1.453530e+00 -1.458133487
#> [223,]  5  2.2  2.63909261  4.5286998  9.839542e-02  0.694529646
#> [224,]  5  2.3  3.17382100  5.0124309 -5.937710e-01 -2.461335475
#> [225,]  5  2.4  3.74996732  3.9530605  8.882812e-01  0.143289764
#> [226,]  5  2.5  3.03050626  3.4204777  5.307042e-02 -0.391222119
#> [227,]  5  2.6  4.36851242  3.2524588 -5.570236e-01 -0.491164086
#> [228,]  5  2.7  3.46262481  4.5405343  4.383970e-01 -0.283647452
#> [229,]  5  2.8  3.73544015  5.7695999  1.526082e-01  0.314794805
#> [230,]  5  2.9  3.20876833  4.7611626 -1.646176e-01  0.396326578
#> [231,]  5  3.0  3.40711552  5.4957989  2.019891e+00 -0.225603711
#> [232,]  5  3.1  5.23065851  5.2860264 -5.293859e-01 -1.924950430
#> [233,]  5  3.2  3.23216272  5.3951894 -4.707870e-01 -1.439229298
#> [234,]  5  3.3  2.90885213  4.0779472 -1.545937e+00 -1.469657774
#> [235,]  5  3.4  4.60788329  4.9245081 -4.052672e-02  0.761863447
#> [236,]  5  3.5  2.83867121  3.9479610  8.903563e-01 -0.243614982
#> [237,]  5  3.6  3.84364058  3.3897607 -2.071388e+00  0.269676607
#> [238,]  5  3.7  3.18769905  3.8910033 -2.500651e-01 -1.558927521
#> [239,]  5  3.8  2.44587048  3.7871219 -1.181650e+00 -0.535588007
#> [240,]  5  3.9  2.75673871  5.0264509  1.441937e+00  0.562451973
#> [241,]  5  4.0  3.11937332  4.5121017  1.357896e+00 -0.178326123
#> [242,]  5  4.1  5.26403921  4.2888617  3.345028e-01 -0.115135986
#> [243,]  5  4.2  4.67898944  3.6424031  1.429338e+00 -0.072061472
#> [244,]  5  4.3  4.56384627  5.1478054 -8.673179e-01  1.210909807
#> [245,]  5  4.4  3.49387250  5.1762342  9.506517e-01 -0.614896901
#> [246,]  5  4.5  3.19712385  5.1839928 -5.850115e-01  0.676126458
#> [247,]  5  4.6  4.41832694  6.6946050  3.209575e-01  0.898599606
#> [248,]  5  4.7  4.51684781  6.6725602 -2.993960e-01 -1.189317904
#> [249,]  5  4.8  3.20498216  5.4089019 -2.785431e-01  0.121258850
#> [250,]  5  4.9  3.76565787  5.6292252  5.461152e-01 -0.011221686

# Type 2
SimSSMOUIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  mu = mu,
  phi = phi,
  sigma = sigma,
  nu = nu,
  lambda = lambda,
  theta = theta,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)
#>        id time           y1          y2            x1           x2
#>   [1,]  1  0.0 -4.158613696  3.08100490  1.370958e+00  0.321925265
#>   [2,]  1  0.1 -2.439543803  3.19547900 -5.646982e-01 -0.783838941
#>   [3,]  1  0.2 -2.559666951  1.88250831  3.631284e-01  1.575727520
#>   [4,]  1  0.3 -2.312735847  2.92240581  6.328626e-01  0.642899306
#>   [5,]  1  0.4 -1.939724698  2.81768446  4.042683e-01  0.089760647
#>   [6,]  1  0.5 -1.295355008  4.02245105 -1.061245e-01  0.276550747
#>   [7,]  1  0.6  0.235955237  5.54995823  1.511522e+00  0.679288816
#>   [8,]  1  0.7  0.392228771  5.94097916 -9.465904e-02  0.089832887
#>   [9,]  1  0.8  1.455145560  3.45276951  2.018424e+00 -2.993090083
#>  [10,]  1  0.9  2.327367511  5.06921396 -6.271410e-02  0.284882954
#>  [11,]  1  1.0  2.580911537  3.94960055  1.304870e+00 -0.367234643
#>  [12,]  1  1.1  3.704398094  4.29417116  2.286645e+00  0.185230565
#>  [13,]  1  1.2  1.852907953  4.84722415 -1.388861e+00  0.581823727
#>  [14,]  1  1.3  2.429430187  5.42331595 -2.787888e-01  1.399736827
#>  [15,]  1  1.4  3.186500471  6.38099914 -1.333213e-01 -0.727292059
#>  [16,]  1  1.5  4.643685108  8.49999093  6.359504e-01  1.302542632
#>  [17,]  1  1.6  3.930255169  6.46174287 -2.842529e-01  0.335848120
#>  [18,]  1  1.7  3.440905645  7.19569765 -2.656455e+00  1.038506099
#>  [19,]  1  1.8  3.637551475  9.20173365 -2.440467e+00  0.920728568
#>  [20,]  1  1.9  4.487222599  8.22475950  1.320113e+00  0.720878163
#>  [21,]  1  2.0  2.964441612  6.61050186 -3.066386e-01 -1.043118939
#>  [22,]  1  2.1  3.804424863  8.49400937 -1.781308e+00 -0.090186387
#>  [23,]  1  2.2  6.584687878  8.53646376 -1.719174e-01  0.623518162
#>  [24,]  1  2.3  5.753649973  7.40001446  1.214675e+00 -0.953523358
#>  [25,]  1  2.4  7.776109206  7.21664220  1.895193e+00 -0.542828815
#>  [26,]  1  2.5  6.437738061  7.58439262 -4.304691e-01  0.580996498
#>  [27,]  1  2.6  5.772984609  7.59219501 -2.572694e-01  0.768178738
#>  [28,]  1  2.7  6.344098255  7.88588255 -1.763163e+00  0.463767589
#>  [29,]  1  2.8  6.903923340  6.35617974  4.600974e-01 -0.885776297
#>  [30,]  1  2.9  6.439746930  7.44217093 -6.399949e-01 -1.099780899
#>  [31,]  1  3.0  6.807553624  7.35111164  4.554501e-01  1.512707010
#>  [32,]  1  3.1  6.374122604  7.42254466  7.048373e-01  0.257921438
#>  [33,]  1  3.2  6.655331375  6.52840894  1.035104e+00  0.088440229
#>  [34,]  1  3.3  4.883582522  5.50517129 -6.089264e-01 -0.120896538
#>  [35,]  1  3.4  6.149318389  5.13808590  5.049551e-01 -1.194328895
#>  [36,]  1  3.5  3.735500854  4.48966423 -1.717009e+00  0.611996898
#>  [37,]  1  3.6  5.790517067  5.94768200 -7.844590e-01 -0.217139846
#>  [38,]  1  3.7  5.410072945  5.49773860 -8.509076e-01 -0.182756706
#>  [39,]  1  3.8  4.107794629  7.04374344 -2.414208e+00  0.933346329
#>  [40,]  1  3.9  6.101059913  6.02741013  3.612261e-02  0.821773111
#>  [41,]  1  4.0  3.829426316  6.08624465  2.059986e-01  1.392116376
#>  [42,]  1  4.1  4.626544005  3.96201286 -3.610573e-01 -0.476173923
#>  [43,]  1  4.2  3.359397807  4.96920030  7.581632e-01  0.650348561
#>  [44,]  1  4.3  3.638759357  4.87940065 -7.267048e-01  1.391110456
#>  [45,]  1  4.4  4.217214601  4.25891959 -1.368281e+00 -1.110788879
#>  [46,]  1  4.5  5.686550342  4.70320783  4.328180e-01 -0.860792587
#>  [47,]  1  4.6  5.322936378  3.84712674 -8.113932e-01 -1.131738681
#>  [48,]  1  4.7  5.739626452  3.71940808  1.444101e+00 -1.459214000
#>  [49,]  1  4.8  5.746016580  4.05281302 -4.314462e-01  0.079982553
#>  [50,]  1  4.9  6.387048448  3.26581478  6.556479e-01  0.653204340
#>  [51,]  2  0.0 -3.262985055  1.77596942  1.200965e+00 -0.040698475
#>  [52,]  2  0.1 -1.389347704 -0.51402407  1.044751e+00 -1.551544822
#>  [53,]  2  0.2 -0.773959672  1.76102049 -1.003209e+00  1.167169549
#>  [54,]  2  0.3 -0.115251131  2.21939809  1.848482e+00 -0.273645701
#>  [55,]  2  0.4  0.379236019  2.36439951 -6.667734e-01 -0.467845325
#>  [56,]  2  0.5  0.688968717  1.73317872  1.055138e-01 -1.238252328
#>  [57,]  2  0.6  1.622484797  4.15059514 -4.222559e-01 -0.007762034
#>  [58,]  2  0.7  1.794964439  3.95155843 -1.223502e-01 -0.800282178
#>  [59,]  2  0.8  3.707858944  2.24230416  1.881930e-01 -0.533492330
#>  [60,]  2  0.9  3.202409686  5.59140253  1.191610e-01  1.287675246
#>  [61,]  2  1.0  2.432027170  4.29865199 -2.509255e-02 -0.175525870
#>  [62,]  2  1.1  2.693837687  3.29330843  1.080727e-01 -1.071782384
#>  [63,]  2  1.2  2.658910749  3.58945581 -4.854352e-01  0.163206882
#>  [64,]  2  1.3  2.036373001  4.83580994 -5.042171e-01 -0.362738416
#>  [65,]  2  1.4  3.517080618  3.43910049 -1.661099e+00  0.590013548
#>  [66,]  2  1.5  3.293969064  6.85402828 -3.823337e-01  1.432421928
#>  [67,]  2  1.6  2.056235525  4.57731419 -5.126503e-01 -0.992692511
#>  [68,]  2  1.7  3.396576257  3.82121045  2.701891e+00  0.454650298
#>  [69,]  2  1.8  2.948803135  6.28438676 -1.362116e+00  0.084898059
#>  [70,]  2  1.9  2.969848588  6.25323290  1.372562e-01  0.895565582
#>  [71,]  2  2.0  2.076522227  5.42175444 -1.493625e+00 -0.229778139
#>  [72,]  2  2.1  2.593278613  5.19158203 -1.470436e+00  0.836619068
#>  [73,]  2  2.2  4.163442216  3.45503696  1.247024e-01 -1.745055861
#>  [74,]  2  2.3  4.857417346  5.22363229 -9.966391e-01  1.689458921
#>  [75,]  2  2.4  5.635693759  4.53727741 -1.822614e-03  0.864777979
#>  [76,]  2  2.5  5.244601266  5.95751666 -4.282589e-01 -0.150775989
#>  [77,]  2  2.6  5.044847724  5.71470837 -6.136716e-01 -1.449007130
#>  [78,]  2  2.7  3.860380920  4.72280659 -2.024678e+00  0.643008700
#>  [79,]  2  2.8  5.581597199  4.34273311 -1.224748e+00  0.483193864
#>  [80,]  2  2.9  6.108363147  6.08495472  1.795164e-01 -0.006355626
#>  [81,]  2  3.0  5.133549054  5.42650991  5.676206e-01  0.151455893
#>  [82,]  2  3.1  5.330672272  5.39203621 -4.928774e-01 -0.584108970
#>  [83,]  2  3.2  4.350905415  4.08099525  6.288407e-05  0.368806733
#>  [84,]  2  3.3  5.743749986  4.46096306  1.122890e+00  0.294654340
#>  [85,]  2  3.4  4.347231955  5.00472228  1.439856e+00 -0.279259373
#>  [86,]  2  3.5  4.868607503  3.43693753 -1.097114e+00 -1.336236655
#>  [87,]  2  3.6  6.456492003  3.50982194 -1.173196e-01  0.700748818
#>  [88,]  2  3.7  5.877803427  5.90505430  1.201498e+00  0.554196622
#>  [89,]  2  3.8  5.877668287  4.21222247 -4.697296e-01 -0.836306593
#>  [90,]  2  3.9  4.202820279  4.07228875 -5.246948e-02 -1.594588162
#>  [91,]  2  4.0  6.075798220  4.31740220 -8.610730e-02  0.204958581
#>  [92,]  2  4.1  5.347131607  3.84177268 -8.876790e-01 -0.345087978
#>  [93,]  2  4.2  2.883835944  3.12407886 -4.446840e-01  0.252611703
#>  [94,]  2  4.3  4.775505403  4.15351307 -2.944488e-02 -1.294002465
#>  [95,]  2  4.4  4.348016339  4.01107790 -4.138688e-01 -0.959170444
#>  [96,]  2  4.5  2.628183358  4.41851775  1.113386e+00  1.085774854
#>  [97,]  2  4.6  5.081761323  4.73687851 -4.809928e-01  0.403774905
#>  [98,]  2  4.7  4.732582769  4.29025215 -4.331690e-01  0.586487537
#>  [99,]  2  4.8  5.734398507  4.07194632  6.968626e-01  1.815228446
#> [100,]  2  4.9  4.921726766  6.05121635 -1.056368e+00  0.128821429
#> [101,]  3  0.0 -3.037190349  1.20561501 -2.000929e+00 -1.096156242
#> [102,]  3  0.1 -2.976346442  1.20534625  3.337772e-01  0.049050451
#> [103,]  3  0.2 -1.381139781  1.67836817  1.171325e+00 -1.198495857
#> [104,]  3  0.3 -1.784096174  2.97018514  2.059539e+00  0.190018999
#> [105,]  3  0.4 -2.201272906  2.37142855 -1.376862e+00  1.297705900
#> [106,]  3  0.5 -2.078380619  1.81904755 -1.150856e+00 -1.033873723
#> [107,]  3  0.6 -2.430477302  2.55607476 -7.058214e-01 -0.738440754
#> [108,]  3  0.7  0.834937026  3.87750187 -1.054056e+00  0.046563939
#> [109,]  3  0.8  0.858409604  2.83842364 -6.457437e-01 -1.017596120
#> [110,]  3  0.9  0.008717101  3.01229362 -1.853780e-01 -0.383283960
#> [111,]  3  1.0 -0.047467738  2.84964942 -1.201222e+00  0.872755412
#> [112,]  3  1.1  1.492509320  3.59067790  2.036972e+00  0.969545014
#> [113,]  3  1.2  0.815303228  3.48975684  1.077747e-01  0.383846665
#> [114,]  3  1.3  0.940488479  3.08089986 -8.410810e-02 -1.851555663
#> [115,]  3  1.4  1.405049771  3.53116314  4.956196e-01 -0.053996737
#> [116,]  3  1.5  0.919526640  2.36747810  3.741519e-02  1.064773214
#> [117,]  3  1.6  2.037038866  3.06720237 -1.320880e-01  0.813195037
#> [118,]  3  1.7  3.222213478  4.28083874  1.476787e+00 -0.190816474
#> [119,]  3  1.8  2.356276280  3.32178036 -2.170302e-01 -2.699929809
#> [120,]  3  1.9  2.980578528  3.37609081 -1.283602e+00  0.060966639
#> [121,]  3  2.0  3.444226254  4.93120910  3.856679e-01  0.573751697
#> [122,]  3  2.1  1.982333542  5.25353595 -3.515129e-01  0.045803580
#> [123,]  3  2.2  2.458170156  8.09093019 -5.217961e-01  0.157412540
#> [124,]  3  2.3  4.183140603  6.31129164 -1.068131e+00  0.431565373
#> [125,]  3  2.4  4.618201836  6.25153206  4.283659e-01 -0.396549736
#> [126,]  3  2.5  5.118255522  6.88694215 -1.740182e-01  1.309978226
#> [127,]  3  2.6  4.694955170  6.78275010  5.156677e-01  0.470393400
#> [128,]  3  2.7  3.704045166  6.31673308 -2.343653e-01 -1.242670271
#> [129,]  3  2.8  4.579951848  6.75586291 -6.585034e-01  1.381575456
#> [130,]  3  2.9  4.641616811  8.19466985  1.250237e+00  1.204458937
#> [131,]  3  3.0  5.792569449  7.91931503 -2.717637e-01  0.824073964
#> [132,]  3  3.1  3.825843514  6.31718691  9.479520e-01 -1.662629402
#> [133,]  3  3.2  4.516389948  7.42181408 -1.201582e+00 -0.569306344
#> [134,]  3  3.3  4.020269355  6.46774858 -4.661161e-01  0.635513817
#> [135,]  3  3.4  3.666992667  5.94138325 -2.693514e-01  0.043722008
#> [136,]  3  3.5  4.600641112  6.22891639 -3.909654e-01  0.348012304
#> [137,]  3  3.6  4.660425689  6.03673717  1.348707e+00  2.459593549
#> [138,]  3  3.7  3.850290354  7.00702970 -2.276470e-02 -0.818380324
#> [139,]  3  3.8  4.288519846  4.62530788  2.442259e-01 -2.113200115
#> [140,]  3  3.9  5.721968726  5.15804391 -9.423717e-01  0.273695272
#> [141,]  3  4.0  6.205863390  4.64419721 -7.292173e-01 -0.687596841
#> [142,]  3  4.1  4.837528739  7.22842257  9.980689e-01  0.446041053
#> [143,]  3  4.2  5.180718321  7.08981800  1.258482e+00 -0.812384724
#> [144,]  3  4.3  5.722255860  7.85450270  1.248864e+00  2.212055480
#> [145,]  3  4.4  4.634365450  7.38856606 -1.380637e+00 -0.123705972
#> [146,]  3  4.5  6.238227521  8.33042632  2.049961e+00 -0.477335506
#> [147,]  3  4.6  7.143130464  7.30362300  1.016873e+00 -0.166261491
#> [148,]  3  4.7  7.385021299  7.67524955 -2.671746e-02  0.862563384
#> [149,]  3  4.8  5.969922511  7.28179930  7.036078e-01  0.097340485
#> [150,]  3  4.9  7.693119228  7.40627483 -9.713852e-01 -1.625616739
#> [151,]  4  0.0 -4.083875510  0.27048105 -4.620768e-03  0.724173801
#> [152,]  4  0.1 -1.202135112  2.16148754  7.602422e-01 -0.832552826
#> [153,]  4  0.2 -3.096701806 -0.40967242  3.899091e-02  0.732528487
#> [154,]  4  0.3 -2.356750662  0.43279518  7.350721e-01 -0.871926870
#> [155,]  4  0.4 -1.640922439  1.44490365 -1.464726e-01 -0.453397511
#> [156,]  4  0.5 -3.075579416  2.69698448 -5.788734e-02  1.187534279
#> [157,]  4  0.6 -0.873902220  2.92973148  4.823695e-01 -0.290145312
#> [158,]  4  0.7  0.091883462  3.49626349  9.929436e-01  0.828546145
#> [159,]  4  0.8 -0.376602631  2.05724312 -1.246395e+00 -0.291227709
#> [160,]  4  0.9 -1.639010018  0.07452298 -3.348752e-02 -1.576362405
#> [161,]  4  1.0 -0.399696694  1.41494209 -7.096218e-02 -0.848815697
#> [162,]  4  1.1 -1.520831790  1.06678116 -7.589207e-01 -1.088519862
#> [163,]  4  1.2 -1.651948700  0.40929116 -1.034359e+00 -0.484290571
#> [164,]  4  1.3 -0.254702034  1.61416564 -6.307320e-01 -0.336311209
#> [165,]  4  1.4 -2.028173858  1.39851153  5.868077e-01 -0.153357891
#> [166,]  4  1.5 -1.129300738  2.40571037 -4.163227e-01 -0.243247229
#> [167,]  4  1.6  0.149287579  3.57207721 -7.848878e-01  1.892202042
#> [168,]  4  1.7  0.513830463  3.38870245  1.634163e-01 -1.385998337
#> [169,]  4  1.8  1.303327764  3.69029623 -1.236714e+00 -0.414824301
#> [170,]  4  1.9  2.176776843  3.61461495  1.045874e+00  0.349081528
#> [171,]  4  2.0  2.563566224  4.91288317 -4.845954e-01  1.628442266
#> [172,]  4  2.1  3.888201975  1.97375114  1.891288e-01  0.088521896
#> [173,]  4  2.2  3.050161411  5.97025632  5.100633e-02  1.239150708
#> [174,]  4  2.3  2.457949010  5.32395341 -2.406689e-04 -1.644555536
#> [175,]  4  2.4  3.047697394  5.91440807  1.809382e+00  1.446356525
#> [176,]  4  2.5  2.315550611  6.20893702 -8.253280e-01 -0.690560172
#> [177,]  4  2.6  3.915035586  5.90166169  1.145470e+00 -0.276431085
#> [178,]  4  2.7  2.735896474  5.47009749  3.157319e-02 -1.109418760
#> [179,]  4  2.8  2.696096279  5.74563588 -8.352058e-01  0.133869316
#> [180,]  4  2.9  2.594018000  6.82471473 -6.876365e-02  1.785339052
#> [181,]  4  3.0  3.655980565  5.98466699  7.467717e-01  2.422163355
#> [182,]  4  3.1  4.690638713  6.38392300 -4.255187e-01 -1.076828902
#> [183,]  4  3.2  3.388854968  7.07403944 -7.720822e-01  0.485941110
#> [184,]  4  3.3  4.435165292  8.21659962  1.527641e-01  1.388521739
#> [185,]  4  3.4  6.659595687  5.96917368  9.885968e-01 -0.195656817
#> [186,]  4  3.5  6.266146899  6.47769305 -7.345833e-02 -0.218174798
#> [187,]  4  3.6  4.772641518  6.36793538 -1.387027e+00 -0.304777955
#> [188,]  4  3.7  5.074074415  6.73172570 -1.306676e+00  0.597832724
#> [189,]  4  3.8  3.861424828  4.22825324 -7.683953e-01  1.397429411
#> [190,]  4  3.9  3.758979615  5.25188581 -5.271081e-01  0.687619761
#> [191,]  4  4.0  4.374330474  5.72453030 -2.142707e-02  0.320188014
#> [192,]  4  4.1  5.614890709  5.98021919  6.704981e-01 -0.301869926
#> [193,]  4  4.2  6.012394608  4.67065739 -4.346170e-01  0.498348686
#> [194,]  4  4.3  6.908343566  5.55619747 -1.113880e+00 -0.549536918
#> [195,]  4  4.4  6.429500597  5.01980094  6.071060e-01 -0.279256504
#> [196,]  4  4.5  6.781508541  4.76720059  2.754570e-01  1.096513443
#> [197,]  4  4.6  8.606629278  6.62019848  1.157347e+00  0.442013088
#> [198,]  4  4.7  6.351961136  6.31503227 -1.682481e+00  0.241016294
#> [199,]  4  4.8  6.239655669  6.19174702  8.731909e-02 -0.255607655
#> [200,]  4  4.9  7.928651450  5.37623869  1.353362e+00  0.931032901
#> [201,]  5  0.0 -4.498080983  0.39870307  1.334913e+00 -1.303821154
#> [202,]  5  0.1 -4.380035768 -1.55669780 -8.692718e-01 -0.250914465
#> [203,]  5  0.2 -2.513341866 -0.30370572  5.548695e-02  0.171007374
#> [204,]  5  0.3 -3.095389475  0.39585582  4.906691e-02 -0.403467479
#> [205,]  5  0.4 -1.024388874  1.57635238 -5.783557e-01  0.104659441
#> [206,]  5  0.5 -1.370803847  0.65766213 -9.987387e-01 -0.318880787
#> [207,]  5  0.6  0.783943965  0.95496584 -2.432780e-03  1.618343936
#> [208,]  5  0.7  1.180449835  3.48181445  6.555119e-01  0.714188601
#> [209,]  5  0.8  1.449872222  4.15520809  1.476842e+00  2.965865370
#> [210,]  5  0.9  2.683723361  3.83424559 -1.909153e+00 -0.795077605
#> [211,]  5  1.0  1.550857901  3.41440177 -7.024395e-01  0.814365915
#> [212,]  5  1.1  1.695056651  4.24229796 -3.114302e-01  2.098030810
#> [213,]  5  1.2  0.991901316  4.08119112 -1.663157e+00  0.300980055
#> [214,]  5  1.3 -0.754654605  3.64458965 -7.505334e-01 -1.083075142
#> [215,]  5  1.4  1.124686029  4.85339129 -7.773518e-01 -1.006322502
#> [216,]  5  1.5  2.070872841  5.25786775 -7.225697e-01 -0.035414565
#> [217,]  5  1.6  2.281253856  5.36379039 -2.188835e+00  1.309124361
#> [218,]  5  1.7  2.430905962  5.53786004  2.134186e-01  0.750400487
#> [219,]  5  1.8  2.808431339  3.26625702 -6.319229e-01 -2.138368328
#> [220,]  5  1.9  1.779981393  4.84589335  1.520491e+00 -0.700354109
#> [221,]  5  2.0  4.327665766  6.37626338  7.959559e-01 -0.009056475
#> [222,]  5  2.1  2.820163181  6.16004979 -1.453530e+00 -1.458133487
#> [223,]  5  2.2  3.419657402  5.16735028  9.839542e-02  0.694529646
#> [224,]  5  2.3  3.803517429  4.51757331 -5.937710e-01 -2.461335475
#> [225,]  5  2.4  2.822122572  5.36223797  8.882812e-01  0.143289764
#> [226,]  5  2.5  3.694675437  6.38441698  5.307042e-02 -0.391222119
#> [227,]  5  2.6  1.730722101  4.70828976 -5.570236e-01 -0.491164086
#> [228,]  5  2.7  2.018968955  5.84833442  4.383970e-01 -0.283647452
#> [229,]  5  2.8  0.330765101  3.35221964  1.526082e-01  0.314794805
#> [230,]  5  2.9  0.084925398  5.23918442 -1.646176e-01  0.396326578
#> [231,]  5  3.0 -0.390094882  5.28168623  2.019891e+00 -0.225603711
#> [232,]  5  3.1  2.892399106  6.42481807 -5.293859e-01 -1.924950430
#> [233,]  5  3.2  0.783861630  3.89994501 -4.707870e-01 -1.439229298
#> [234,]  5  3.3  1.742099247  4.18517607 -1.545937e+00 -1.469657774
#> [235,]  5  3.4  2.347701753  4.66349137 -4.052672e-02  0.761863447
#> [236,]  5  3.5  3.913658166  5.41560249  8.903563e-01 -0.243614982
#> [237,]  5  3.6  1.959708033  4.49852312 -2.071388e+00  0.269676607
#> [238,]  5  3.7  3.553942822  4.45981960 -2.500651e-01 -1.558927521
#> [239,]  5  3.8  2.007120768  5.32414765 -1.181650e+00 -0.535588007
#> [240,]  5  3.9  3.260320081  4.67100552  1.441937e+00  0.562451973
#> [241,]  5  4.0  3.439459999  3.57723546  1.357896e+00 -0.178326123
#> [242,]  5  4.1  3.308312790  4.94107605  3.345028e-01 -0.115135986
#> [243,]  5  4.2  3.767832323  5.75979508  1.429338e+00 -0.072061472
#> [244,]  5  4.3  4.604689188  5.70262879 -8.673179e-01  1.210909807
#> [245,]  5  4.4  4.280120216  3.33512239  9.506517e-01 -0.614896901
#> [246,]  5  4.5  2.248149536  6.39837807 -5.850115e-01  0.676126458
#> [247,]  5  4.6  2.899048254  4.64364797  3.209575e-01  0.898599606
#> [248,]  5  4.7  1.319421932  6.32478919 -2.993960e-01 -1.189317904
#> [249,]  5  4.8  2.186892603  5.85859107 -2.785431e-01  0.121258850
#> [250,]  5  4.9  3.666542134  6.83959985  5.461152e-01 -0.011221686