Simulate Data from an Ornstein–Uhlenbeck Model using a State Space Model Parameterization for n > 1 Individuals (Varying Parameters)

This function simulates data from an Ornstein–Uhlenbeck model using a state space model parameterization for n > 1 individuals. In this model, the parameters can vary across individuals.

Usage

SimSSMOUVary(
  n,
  mu0,
  sigma0_sqrt,
  mu,
  phi,
  sigma_sqrt,
  nu,
  lambda,
  theta_sqrt,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  delta_t,
  time,
  burn_in
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_sqrt

Numeric matrix. Cholesky decomposition of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

mu

List of numeric vectors. The long-term mean or equilibrium level (\(\boldsymbol{\mu}\)).

phi

List of numeric matrices. The rate of mean reversion, determining how quickly the variable returns to its mean (\(\boldsymbol{\Phi}\)).

sigma_sqrt

List of numeric matrices. Cholesky decomposition of the matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

nu

Numeric vector. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta_sqrt

Numeric matrix. Cholesky decomposition of the measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the observed variables at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

Numeric matrix. The matrix of observed covariates in type = 1 or type = 2. The number of rows should be equal to time + burn_in.

type

Integer. State space model type.

delta_t

Numeric. Time interval (\(\delta_t\)).

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

time: A vector of discrete time points from 1 to t.

id: A vector of ID numbers of length t.

n: Number of individuals.

Details

Parameters can vary across individuals by providing a list of parameter values. If the length of any of the parameters (mu0, sigma0_sqrt, mu, phi, sigma_sqrt, nu, lambda, theta_sqrt, gamma_y, or gamma_eta) is less the n, the function will cycle through the available values.

References

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Uhlenbeck, G. E., & Ornstein, L. S. (1930). On the theory of the brownian motion. Physical Review, 36(5), 823–841. doi:10.1103/physrev.36.823

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
# In this example, phi varies across individuals
set.seed(42)
p <- k <- 2
iden <- diag(p)
iden_sqrt <- chol(iden)
n <- 5
mu0 <- list(c(-3.0, 1.5))
sigma0_sqrt <- list(iden_sqrt)
mu <- list(c(5.76, 5.18))
phi <- list(
  as.matrix(Matrix::expm(diag(x = -0.1, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.2, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.3, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.4, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.5, nrow = k)))
)
sigma_sqrt <- list(
  chol(
    matrix(data = c(2.79, 0.06, 0.06, 3.27), nrow = p)
  )
)
nu <- list(rep(x = 0, times = k))
lambda <- list(diag(k))
theta_sqrt <- list(chol(diag(x = 0.50, nrow = k)))
delta_t <- 0.10
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- list(0.10 * diag(k))
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = k * (time + burn_in)),
        ncol = k
      )
    )
  }
)

# Type 0
ssm <- SimSSMOUVary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  mu = mu,
  phi = phi,
  sigma_sqrt = sigma_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  type = 0,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -1.685 -1.216 -1.495 -1.146 -0.487 ...
#>   ..$ eta : num [1:50, 1:2] -1.684 -0.805 -1.201 -0.764 0.274 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -3.598 -0.668 -2.109 -0.472 0.69 ...
#>   ..$ eta : num [1:50, 1:2] -2.2483 -1.1278 -1.0766 0.0847 -0.1682 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.1068 -2.6573 -1.4717 -0.0725 -0.9007 ...
#>   ..$ eta : num [1:50, 1:2] -3.429 -2.516 -1.765 -1.169 -0.696 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.82 -3.81 -1.2 -2.65 -2.79 ...
#>   ..$ eta : num [1:50, 1:2] -4.37 -4.34 -3.57 -2.55 -1.77 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -3.64 -2.99 -2.27 -2.09 -1.79 ...
#>   ..$ eta : num [1:50, 1:2] -3.41 -2.84 -2.37 -2.84 -2.62 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...

# Type 1
ssm <- SimSSMOUVary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  mu = mu,
  phi = phi,
  sigma_sqrt = sigma_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.844 -1.984 -1.32 -1.036 0.105 ...
#>   ..$ eta : num [1:50, 1:2] -3.398 -2.789 -1.411 -0.531 0.378 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -3.921 -2.318 0.209 -0.167 0.246 ...
#>   ..$ eta : num [1:50, 1:2] -3.703 -2.364 -1.407 -0.28 -0.589 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -1.661 -1.08 -1.122 0.561 -1.077 ...
#>   ..$ eta : num [1:50, 1:2] -1.8667 -1.3993 -0.8418 -0.2251 -0.0834 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.99 -4.61 -4.55 -2.99 -2.45 ...
#>   ..$ eta : num [1:50, 1:2] -5.31 -4.53 -4.56 -3.76 -2.77 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.165 -0.417 -2.299 -1.106 -0.601 ...
#>   ..$ eta : num [1:50, 1:2] -2.82 -1.38 -2.21 -1.6 -1.16 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...

# Type 2
ssm <- SimSSMOUVary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  mu = mu,
  phi = phi,
  sigma_sqrt = sigma_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.43 -2.56 -2.68 -2.42 -2.04 ...
#>   ..$ eta : num [1:50, 1:2] -3.87 -3.51 -2.64 -1.85 -1.35 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -3.503 -1.502 -0.877 -0.21 0.293 ...
#>   ..$ eta : num [1:50, 1:2] -3.431 -2.038 -0.68 -0.22 0.581 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -2.64 -2.79 -1.21 -1.62 -2.05 ...
#>   ..$ eta : num [1:50, 1:2] -3.27 -2.43 -1.41 -1.2 -1.22 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.08 -1.2 -3.1 -2.36 -1.64 ...
#>   ..$ eta : num [1:50, 1:2] -4.5 -3.29 -2.68 -2.52 -1.49 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.77 -4.51 -2.63 -3.21 -1.13 ...
#>   ..$ eta : num [1:50, 1:2] -3.873 -3.579 -3.288 -2.024 -0.835 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...