Simulate Data from a Vector Autoregressive Model using a State Space Model Parameterization for n > 1 Individuals (Fixed Parameters)

This function simulates data from a vector autoregressive model using a state space model parameterization for n > 1 individuals. In this model, the parameters are invariant across individuals.

Usage

SimSSMVARFixed(
  n,
  mu0,
  sigma0,
  alpha,
  beta,
  psi,
  gamma_eta = NULL,
  x = NULL,
  time = 0,
  burn_in = 0
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0

Numeric matrix. The covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

alpha

Numeric vector. Vector of intercepts for the dynamic model (\(\boldsymbol{\alpha}\)).

beta

Numeric matrix. Transition matrix relating the values of the latent variables at time t - 1 to those at time t (\(\boldsymbol{\beta}\)).

psi

Numeric matrix. The process noise covariance matrix (\(\boldsymbol{\Psi}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

A list of length n of numeric matrices. Each element of the list is a matrix of observed covariates in type = 1 or type = 2. The number of rows in each matrix should be equal to time + burn_in.

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list with the following elements:

call: Function call.

args: Function arguments.

data: Generated data which is a list of length n.

fun: Function used.

Each element of data is a list with the following elements:

id: A vector of ID numbers of length t.

time: A vector of discrete time points from 0 to t - 1.

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
k <- 3
iden <- diag(k)
null_vec <- rep(x = 0, times = k)
n <- 5
mu0 <- null_vec
sigma0 <- iden
alpha <- null_vec
beta <- diag(x = 0.5, nrow = k)
psi <- iden
time <- 50
burn_in <- 0
gamma_eta <- 0.10 * diag(k)
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = k * (time + burn_in)),
        ncol = k
      )
    )
  }
)

# No covariates
SimSSMVARFixed(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  alpha = alpha,
  beta = beta,
  psi = psi,
  time = time,
  burn_in = burn_in
)
#>        id time           y1          y2           y3
#>   [1,]  1    0 -0.958385183  1.42600303  0.543567535
#>   [2,]  1    1 -0.678815770  0.68087912  0.603930524
#>   [3,]  1    2  0.075801673 -1.23729067  0.404291185
#>   [4,]  1    3  1.588710900 -1.26978257  0.110670581
#>   [5,]  1    4  0.505490141 -2.49221521  0.386058132
#>   [6,]  1    5  0.815411515 -0.59482163 -0.122497316
#>   [7,]  1    6 -0.661787961  0.81385810 -0.431349477
#>   [8,]  1    7  0.429399600  1.79702039 -1.006854371
#>   [9,]  1    8 -0.048360036  2.08576128 -1.422322937
#>  [10,]  1    9  0.971849380  1.58653911 -0.995782343
#>  [11,]  1   10  3.004151943 -0.27213196  0.260164320
#>  [12,]  1   11  0.423405932 -0.42116508 -1.493051354
#>  [13,]  1   12  0.804616339 -0.94866161 -1.230189301
#>  [14,]  1   13  1.791212367 -0.93285221 -0.958792359
#>  [15,]  1   14  1.403961875  1.49432080 -0.839948555
#>  [16,]  1   15  1.124262563 -0.51130985 -0.213035843
#>  [17,]  1   16  1.549958698 -0.77525638 -0.212335605
#>  [18,]  1   17  4.252736626 -1.26589331 -0.696244114
#>  [19,]  1   18  1.124026241 -0.53771825 -0.311190923
#>  [20,]  1   19 -0.599287621 -0.08153910  1.046378788
#>  [21,]  1   20  0.107144793 -0.58737636  1.209517815
#>  [22,]  1   21  0.030184949 -0.15001158 -0.242055775
#>  [23,]  1   22 -1.529153926 -1.50423087  0.595081656
#>  [24,]  1   23 -0.089826816 -0.47411683  0.432762409
#>  [25,]  1   24  0.580258570  0.23044796  0.977115747
#>  [26,]  1   25 -0.284232383 -0.37411182  0.475128849
#>  [27,]  1   26  0.501532837  0.49068327  0.003915195
#>  [28,]  1   27  0.186366799 -0.08410848 -0.376345697
#>  [29,]  1   28 -0.105594846  0.29157467 -1.106252179
#>  [30,]  1   29  0.736747785 -2.14831887 -0.553548895
#>  [31,]  1   30  0.363584809 -0.63735454  0.836126953
#>  [32,]  1   31  0.173782234 -0.99798316  1.588675052
#>  [33,]  1   32 -0.137347010 -0.44137460  0.574632801
#>  [34,]  1   33  0.370954153 -1.28854399 -0.322651359
#>  [35,]  1   34  0.727785088  0.01284842 -0.492074516
#>  [36,]  1   35 -0.679451164 -0.22273063 -0.616662881
#>  [37,]  1   36 -0.204360897  1.56469838 -1.145922527
#>  [38,]  1   37 -0.306295546  1.07427167  0.345000310
#>  [39,]  1   38  1.641842888  0.16530107 -1.253189152
#>  [40,]  1   39  1.433957160  1.18532868 -1.139408439
#>  [41,]  1   40  0.985717045  0.90811963 -0.735494993
#>  [42,]  1   41  1.134890260  0.09942634 -0.940197784
#>  [43,]  1   42  1.508038040  1.81139834  0.183849609
#>  [44,]  1   43  3.190650325  0.60297223  0.419992887
#>  [45,]  1   44  1.584825793 -0.54492202 -0.653341077
#>  [46,]  1   45  0.320667093  0.44311657 -0.255368727
#>  [47,]  1   46 -0.436246339 -0.55696857 -0.017532629
#>  [48,]  1   47 -2.266957831  1.46364183  0.261255260
#>  [49,]  1   48 -1.892756244  1.24836219 -1.212493381
#>  [50,]  1   49 -0.051140440  0.10861977 -1.976472351
#>  [51,]  2    0 -0.632003364 -0.55150889 -1.217198582
#>  [52,]  2    1  0.432265979  1.37496531 -0.346619282
#>  [53,]  2    2  3.562600720  0.59968482  0.641451017
#>  [54,]  2    3  1.643613431 -0.46620796  0.800937957
#>  [55,]  2    4 -0.632603431 -1.98556776  2.062776302
#>  [56,]  2    5 -1.256604363 -1.41695474  1.428629230
#>  [57,]  2    6 -0.094496527 -0.37756288  0.665792861
#>  [58,]  2    7  1.014196533 -0.65636447  0.244209643
#>  [59,]  2    8  1.151924103  1.86086125 -0.177736982
#>  [60,]  2    9  2.019447268  3.40781955  1.790400215
#>  [61,]  2   10  1.138772148  1.03614477  1.216347190
#>  [62,]  2   11  1.400605055 -1.32056717  0.735545772
#>  [63,]  2   12  0.920923962 -0.46652326 -0.063562836
#>  [64,]  2   13  0.137258899 -1.85250997  0.703048580
#>  [65,]  2   14  1.069284084  0.68533031  0.766987804
#>  [66,]  2   15 -0.047089784  0.47119613  1.061316518
#>  [67,]  2   16  0.460696796  0.90576838  1.819105921
#>  [68,]  2   17 -0.290155512  1.86589859  2.104757355
#>  [69,]  2   18  0.114842712  1.70109253  2.390425682
#>  [70,]  2   19  0.774131839  1.73179951  0.198870163
#>  [71,]  2   20  0.558270968  1.80609740  1.870213108
#>  [72,]  2   21  1.120710371  0.38556964  0.506063286
#>  [73,]  2   22  1.075360954  0.49946545  0.264194164
#>  [74,]  2   23  1.158602660  1.14890223 -0.841784060
#>  [75,]  2   24  0.788709991  0.38905480 -0.495262680
#>  [76,]  2   25  0.806130741  0.29750520 -1.286031142
#>  [77,]  2   26  1.258686718 -0.89286905  0.572698221
#>  [78,]  2   27  1.769262499 -0.34514417  1.524262338
#>  [79,]  2   28  0.756248118 -0.01515493  1.213288510
#>  [80,]  2   29  0.753786911  0.11793036  0.199799142
#>  [81,]  2   30 -0.857387648  0.55642352  0.869919614
#>  [82,]  2   31 -1.403643467 -0.09089632  0.962301943
#>  [83,]  2   32 -0.235520024 -0.05314072  1.434596091
#>  [84,]  2   33  0.202124430  0.43294083  1.536762085
#>  [85,]  2   34  0.055488429  1.84386053  1.744059226
#>  [86,]  2   35 -1.291635462  2.21471528 -0.139909257
#>  [87,]  2   36 -1.042982338  0.96180928  0.276410849
#>  [88,]  2   37 -0.215447654 -0.83299047  0.181415544
#>  [89,]  2   38  0.998197165  0.05722900  0.928423175
#>  [90,]  2   39  0.903027092 -0.48569446 -0.422386411
#>  [91,]  2   40 -0.337759523 -1.54894896  0.456678748
#>  [92,]  2   41  0.504721314 -0.32902474 -0.380584722
#>  [93,]  2   42  0.084371899 -0.27446301 -0.323940047
#>  [94,]  2   43  0.263653811 -0.16503282  0.172231318
#>  [95,]  2   44  0.869306576  0.43484454  0.512436302
#>  [96,]  2   45 -0.599506532 -0.19669826 -0.119690931
#>  [97,]  2   46  0.406227381  1.76527925 -1.127162375
#>  [98,]  2   47 -2.126228224  1.40634718 -0.860483973
#>  [99,]  2   48 -0.081178746  1.29749991 -2.111683127
#> [100,]  2   49  0.420432412  1.33426595 -1.462355339
#> [101,]  3    0 -0.336338925 -0.23968802  0.077441303
#> [102,]  3    1 -0.380202827 -1.10469357 -0.066523133
#> [103,]  3    2 -0.057154577 -0.46283132 -1.868707619
#> [104,]  3    3  1.026269297  1.09869794 -1.507073389
#> [105,]  3    4  1.690482090  1.24486782 -1.622381621
#> [106,]  3    5 -0.290882482  0.87693092 -0.467555036
#> [107,]  3    6  1.076838270  2.85319088  0.103662847
#> [108,]  3    7  0.715225406 -0.20075154 -0.422056802
#> [109,]  3    8 -0.122909825  1.73088278  0.406834713
#> [110,]  3    9 -3.878712031 -0.60071760 -1.205479650
#> [111,]  3   10 -2.482243211 -0.68630922 -1.648701240
#> [112,]  3   11 -1.789446556 -0.64715789  0.568306411
#> [113,]  3   12 -1.953908096 -2.00784896 -0.479057236
#> [114,]  3   13 -1.660197812 -1.64865705  1.924460237
#> [115,]  3   14  1.126717506 -2.29209336  2.464627796
#> [116,]  3   15  0.581560142 -3.05032646  1.715957926
#> [117,]  3   16  2.385017607 -1.46614112  0.564889006
#> [118,]  3   17  1.939693136 -0.76783946 -0.809958340
#> [119,]  3   18  0.923891217  0.38067579 -0.339411916
#> [120,]  3   19 -0.604656202 -1.01710389 -0.370467854
#> [121,]  3   20 -0.634436303 -1.48830372 -0.108510071
#> [122,]  3   21 -0.005376414 -0.15353799 -0.465039733
#> [123,]  3   22  0.615796205  0.13086305  0.608825086
#> [124,]  3   23 -0.215866188  0.65324271  0.449177897
#> [125,]  3   24  0.327686490 -0.95601271  1.274019606
#> [126,]  3   25 -0.024236367 -1.57121315  0.208719228
#> [127,]  3   26 -1.758295844 -0.95610699 -0.672282279
#> [128,]  3   27 -0.569622231 -1.91820895 -1.851261018
#> [129,]  3   28  0.875221600 -1.31896577 -0.809750102
#> [130,]  3   29 -0.288213619  0.27942705 -0.931177431
#> [131,]  3   30  0.086977484  1.02586458  1.470101284
#> [132,]  3   31  0.259843771 -0.19467526 -0.078581047
#> [133,]  3   32  0.322771407 -0.09759109 -0.984179218
#> [134,]  3   33 -0.721838232 -0.52522295 -1.664793439
#> [135,]  3   34 -0.395184621 -0.63723765  0.360158468
#> [136,]  3   35  1.162061221 -1.54081656 -0.964795425
#> [137,]  3   36 -0.207961564 -1.39568678  0.783165182
#> [138,]  3   37 -1.590408297 -0.28356326  0.653304096
#> [139,]  3   38 -0.462603188 -0.88255540  1.399189252
#> [140,]  3   39  2.584546784 -1.12818501  1.813578800
#> [141,]  3   40  2.969533105  0.14535183 -0.541373851
#> [142,]  3   41  0.890182609  0.93483156  1.520649825
#> [143,]  3   42 -0.564967266  1.01657275 -1.220143302
#> [144,]  3   43 -0.378748441  1.23697093 -1.264862871
#> [145,]  3   44  0.512317425  2.41282738 -1.222944506
#> [146,]  3   45 -1.061717979  0.72181826  1.261958318
#> [147,]  3   46  1.106172326  1.99200345  1.464045578
#> [148,]  3   47  0.192207688  1.19242277  1.475289206
#> [149,]  3   48 -1.544722202  1.13618935  0.249090914
#> [150,]  3   49 -1.499742895 -0.49710869 -0.273068500
#> [151,]  4    0 -2.045789988 -0.49811231 -0.278856036
#> [152,]  4    1  0.115937957 -0.40388173  1.031705580
#> [153,]  4    2  0.187688241  0.61609942  0.933939910
#> [154,]  4    3 -0.620055771  0.15390139  1.091250238
#> [155,]  4    4 -0.696575760  0.43710720  0.586646257
#> [156,]  4    5 -0.213921986  0.66727584 -0.341125472
#> [157,]  4    6 -0.284815232  0.27772629 -1.202403504
#> [158,]  4    7 -1.182457633 -0.84512347  0.919107030
#> [159,]  4    8 -0.467381414 -0.12216587  0.484320522
#> [160,]  4    9  0.521444497  1.76968700  0.147457940
#> [161,]  4   10  0.275099214  0.33261903  0.480359606
#> [162,]  4   11  0.598812543 -0.74769105 -0.171537441
#> [163,]  4   12  0.890086634 -0.17443799  1.545235668
#> [164,]  4   13  0.019733254 -1.30246664  0.127589754
#> [165,]  4   14 -0.646977817 -0.78451525 -1.725981144
#> [166,]  4   15 -1.068369592 -1.55047518 -3.196684371
#> [167,]  4   16 -0.015939845  0.21081086 -0.854183861
#> [168,]  4   17  1.409144883 -0.29712083 -0.068236278
#> [169,]  4   18  0.040817381  0.43278018 -0.881369372
#> [170,]  4   19  0.359150895  1.15538681  0.007311057
#> [171,]  4   20 -0.178728084  0.09511156  0.163251153
#> [172,]  4   21 -0.011349514  0.33808255  1.035531682
#> [173,]  4   22  0.642178672  0.74513168  1.789236232
#> [174,]  4   23  0.095696751  0.35660707  1.381132421
#> [175,]  4   24 -1.071339870 -0.80439655 -0.324121554
#> [176,]  4   25  0.102100115  0.14731883  0.108408835
#> [177,]  4   26 -0.047768592 -1.24749156 -0.018930428
#> [178,]  4   27  0.277405176 -0.60237833  0.198306258
#> [179,]  4   28  2.217153002  0.28753564  0.216829909
#> [180,]  4   29  2.377933695 -0.81670425 -0.744973910
#> [181,]  4   30  2.171153860 -0.63229433 -1.458067384
#> [182,]  4   31 -0.372765782 -2.53066942 -0.645348726
#> [183,]  4   32  0.810699425  0.02193931  1.324567100
#> [184,]  4   33 -0.298158659  0.12832662  0.872397738
#> [185,]  4   34  1.497444013 -0.69249917  0.695404817
#> [186,]  4   35  0.525397445  1.09144049  0.025834198
#> [187,]  4   36  0.196584355 -0.62824280  0.236090045
#> [188,]  4   37 -0.566988810  0.68934079  0.590918285
#> [189,]  4   38 -0.894505723  0.08049606  0.285666231
#> [190,]  4   39 -1.923841071  0.91765171 -0.184390285
#> [191,]  4   40 -0.227268736  0.83547539 -0.750769884
#> [192,]  4   41 -0.034171485  2.02494338 -2.145391136
#> [193,]  4   42  0.542456810  1.05017472 -2.529242925
#> [194,]  4   43  0.345801124  1.90627033 -0.421628913
#> [195,]  4   44 -0.306321007  0.78614584 -1.730429060
#> [196,]  4   45  0.487203629 -0.64923342 -1.937254779
#> [197,]  4   46  0.884596370 -0.16828396 -2.884514111
#> [198,]  4   47 -0.737039658 -0.22817916 -3.687286323
#> [199,]  4   48 -0.580403412  2.37290174 -1.893482266
#> [200,]  4   49 -0.041178650  0.03936605 -1.649618365
#> [201,]  5    0 -0.307783557 -0.53522234  0.924444955
#> [202,]  5    1 -0.087959130 -1.91692996  1.284065670
#> [203,]  5    2  2.242131316 -0.25833955  1.372973195
#> [204,]  5    3  1.281271475 -2.05287832 -0.720513057
#> [205,]  5    4  1.822178029  0.54770254  0.476718669
#> [206,]  5    5 -0.361895919 -0.21555906  0.119976977
#> [207,]  5    6 -1.266683883 -0.19188446 -0.106454642
#> [208,]  5    7 -1.838543405  0.41220196 -0.494740718
#> [209,]  5    8 -0.648462749  1.06813016 -0.343271184
#> [210,]  5    9  0.478716531  1.38006423 -0.242209273
#> [211,]  5   10  1.018515718  0.36710281 -0.045844458
#> [212,]  5   11  0.862913524  0.70119341 -0.513486192
#> [213,]  5   12  1.483831173  0.09210980  0.907822081
#> [214,]  5   13  0.504395897  1.39754191 -0.091321963
#> [215,]  5   14 -0.607243342  2.41995190  0.041275423
#> [216,]  5   15  0.009044375  1.10647357  0.712610844
#> [217,]  5   16  1.149399918 -0.01672563  0.132277485
#> [218,]  5   17  0.512831734 -1.74959677 -0.134897604
#> [219,]  5   18 -0.452696596  0.07741287 -1.357297853
#> [220,]  5   19  0.376730058  0.04610770 -0.132605236
#> [221,]  5   20  0.065627768 -0.43173480 -0.168850687
#> [222,]  5   21 -0.093084942  0.02239690 -1.159426826
#> [223,]  5   22 -1.007417335 -0.29100065 -0.488501725
#> [224,]  5   23 -1.624417770 -0.54534447 -1.743744782
#> [225,]  5   24 -1.186553743 -1.08238878 -2.582848914
#> [226,]  5   25  0.995948111 -0.30136276 -1.412780578
#> [227,]  5   26  0.601767445  0.93007797  0.310354697
#> [228,]  5   27 -0.021821224 -0.01916144 -0.086546342
#> [229,]  5   28 -0.135088318 -0.02473672  0.971916595
#> [230,]  5   29 -1.115040025  1.28280109 -0.378685539
#> [231,]  5   30  0.161048825  1.39161918 -1.059093800
#> [232,]  5   31  1.013439262  1.84140441  1.185043775
#> [233,]  5   32  0.462669534  1.68933696 -1.223850079
#> [234,]  5   33  2.011604821 -0.08532907 -1.700137487
#> [235,]  5   34  0.503216157  0.63737072 -1.014903074
#> [236,]  5   35 -0.105992038  1.02944713  0.184969657
#> [237,]  5   36 -0.466662273  0.43482930  0.285593733
#> [238,]  5   37  0.037480049  3.05702707 -0.087701310
#> [239,]  5   38  0.352488385  0.45825982  0.064679208
#> [240,]  5   39 -0.660659073 -1.46105483 -0.045511722
#> [241,]  5   40  1.864653592 -0.69680227  1.349501976
#> [242,]  5   41  0.826736287 -0.63849903  1.744052130
#> [243,]  5   42  0.369914517 -1.41956008  0.593216702
#> [244,]  5   43  0.683032738 -0.25011579 -0.329758543
#> [245,]  5   44 -0.476617116 -0.62453507 -1.590736541
#> [246,]  5   45 -1.151139173 -0.02971460 -1.191714797
#> [247,]  5   46  1.773877508 -0.31696763 -0.756932251
#> [248,]  5   47  1.885823608  0.72661652 -0.383780859
#> [249,]  5   48  1.279527758 -0.44458365 -1.699240710
#> [250,]  5   49 -0.600286223 -0.43805301  0.283648689

# With covariates
SimSSMVARFixed(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  alpha = alpha,
  beta = beta,
  psi = psi,
  gamma_eta = gamma_eta,
  x = x,
  time = time,
  burn_in = burn_in
)
#>        id time           y1           y2           y3            x1
#>   [1,]  1    0  0.428038658 -0.367711921 -0.094993026  1.3709584471
#>   [2,]  1    1  0.608995767  0.211685545 -0.046057819 -0.5646981714
#>   [3,]  1    2 -0.055433457 -0.658345258 -0.223342932  0.3631284113
#>   [4,]  1    3  1.146705978 -0.140230002 -0.194859099  0.6328626050
#>   [5,]  1    4 -0.792027541  0.684481826 -2.087929710  0.4042683231
#>   [6,]  1    5 -0.170464788  1.444989237 -2.054218032 -0.1061245161
#>   [7,]  1    6 -0.807084718 -0.805685641  0.015543786  1.5115219974
#>   [8,]  1    7 -1.879571431 -2.300938127  1.905257265 -0.0946590384
#>   [9,]  1    8 -0.293033437 -0.416669902  3.007822169  2.0184237139
#>  [10,]  1    9 -0.045447846 -0.437124934  1.617249771 -0.0627140991
#>  [11,]  1   10  0.097620604 -0.470793912  1.175565996  1.3048696542
#>  [12,]  1   11 -0.417098985  1.088049858  0.209870002  2.2866453927
#>  [13,]  1   12 -0.560336931  0.793780415 -0.427869125 -1.3888607011
#>  [14,]  1   13 -1.443814972  0.556149977 -0.674134298 -0.2787887668
#>  [15,]  1   14  0.062338085  1.232734717  1.951624103 -0.1333213364
#>  [16,]  1   15  1.089764296  1.877938596  2.196085072  0.6359503981
#>  [17,]  1   16  1.109751629  1.320153603  1.218917581 -0.2842529214
#>  [18,]  1   17  0.901121819  0.116706843 -0.461990348 -2.6564554209
#>  [19,]  1   18  0.239096131 -1.909184033  0.028783440 -2.4404669286
#>  [20,]  1   19 -2.127504216  0.600189117  0.938500306  1.3201133457
#>  [21,]  1   20 -0.782798052 -1.401844467 -0.893514109 -0.3066385941
#>  [22,]  1   21 -0.906267600 -1.741539157  1.167269578 -1.7813084340
#>  [23,]  1   22 -1.004777986 -0.100452849 -0.475097987 -0.1719173558
#>  [24,]  1   23 -0.463630387  0.014472779  0.618386004  1.2146746992
#>  [25,]  1   24 -0.352188754  0.098188338 -0.071963139  1.8951934613
#>  [26,]  1   25 -0.076608545 -2.012312628 -2.015504980 -0.4304691316
#>  [27,]  1   26 -1.225964393  0.603440682 -1.360282358 -0.2572693828
#>  [28,]  1   27 -1.758150837 -0.720614225 -1.589341618 -1.7631630852
#>  [29,]  1   28  1.129034383 -0.967296183 -1.701751896  0.4600973548
#>  [30,]  1   29 -1.351439800 -0.849823491 -0.904093963 -0.6399948760
#>  [31,]  1   30  0.042355687 -0.182130003  3.293849589  0.4554501232
#>  [32,]  1   31 -1.854480070  1.010664178  2.300684295  0.7048373372
#>  [33,]  1   32 -2.697561957  1.261551741 -0.565710228  1.0351035220
#>  [34,]  1   33 -1.036809830  1.148058319 -0.868563669 -0.6089263754
#>  [35,]  1   34 -0.726057571  0.274671540 -0.207026127  0.5049551233
#>  [36,]  1   35  0.766740049 -0.033431709 -0.006000009 -1.7170086791
#>  [37,]  1   36  1.072594341  0.561127698  1.116760540 -0.7844590084
#>  [38,]  1   37  2.647305036 -0.593805859  0.832878482 -0.8509075942
#>  [39,]  1   38  0.453495365  0.304426323 -0.042933931 -2.4142076499
#>  [40,]  1   39  0.180251747  0.332786374 -2.258710764  0.0361226069
#>  [41,]  1   40 -0.757501796  0.801145428 -0.458500393  0.2059986002
#>  [42,]  1   41 -1.025226848 -0.278033074  0.346666051 -0.3610572985
#>  [43,]  1   42  0.004122769  0.396939903 -0.493390841  0.7581632357
#>  [44,]  1   43 -0.347420116  0.287907161  2.308974997 -0.7267048271
#>  [45,]  1   44 -0.159597948 -0.050282929  1.638901559 -1.3682810444
#>  [46,]  1   45  0.682298757  0.450818802  0.364217721  0.4328180259
#>  [47,]  1   46  1.791999222  1.579049329  1.058067884 -0.8113931762
#>  [48,]  1   47  2.273880781  0.627328534 -0.422822027  1.4441012617
#>  [49,]  1   48  1.229933991  0.309326072 -0.004382130 -0.4314462026
#>  [50,]  1   49  0.359224198  2.055446683 -2.412833777  0.6556478834
#>  [51,]  2    0  0.442662192 -0.879137511 -0.138436015 -0.0406984752
#>  [52,]  2    1 -0.052458455 -0.134114688 -1.424006158 -1.5515448223
#>  [53,]  2    2  0.101714151 -1.689206201 -0.701328879  1.1671695492
#>  [54,]  2    3  1.113718077 -0.051501790  0.376283435 -0.2736457014
#>  [55,]  2    4  0.960293872 -0.797297635 -0.182136096 -0.4678453247
#>  [56,]  2    5  0.014469359 -0.277950328  1.551011764 -1.2382523280
#>  [57,]  2    6 -0.011350530  0.152946059  0.612575720 -0.0077620338
#>  [58,]  2    7 -2.200361679 -0.128593295  0.669059397 -0.8002821780
#>  [59,]  2    8 -0.500242641 -0.792377342  0.313638329 -0.5334923300
#>  [60,]  2    9 -2.697945127  1.170923300 -0.679729991  1.2876752456
#>  [61,]  2   10 -2.150980869 -0.069971167  0.212359348 -0.1755258702
#>  [62,]  2   11 -1.172340369  0.792776203 -0.615647424 -1.0717823842
#>  [63,]  2   12 -0.281818011 -0.104194664  1.190987017  0.1632068825
#>  [64,]  2   13 -0.569671838  1.626101483  0.393988270 -0.3627384156
#>  [65,]  2   14 -0.044769680  0.750772615  0.514383533  0.5900135480
#>  [66,]  2   15  0.079840390 -0.542236571  0.836491590  1.4324219277
#>  [67,]  2   16 -1.420006621  2.244137538  1.607163398 -0.9926925111
#>  [68,]  2   17 -2.920129107  0.937294255  0.229887400  0.4546502976
#>  [69,]  2   18 -1.042832453  0.820740237 -2.480887411  0.0848980587
#>  [70,]  2   19  0.295138791  0.414969727 -1.671917777  0.8955655823
#>  [71,]  2   20  1.449793042  0.211747022 -1.588498943 -0.2297781389
#>  [72,]  2   21  1.055754918 -0.816276499 -2.285346663  0.8366190685
#>  [73,]  2   22  0.686832962  0.063650325 -1.758680457 -1.7450558613
#>  [74,]  2   23 -0.029382660 -0.599506048 -1.627409037  1.6894589213
#>  [75,]  2   24 -1.458390168 -1.291106729  0.111551644  0.8647779785
#>  [76,]  2   25 -0.506174195 -0.079626128  1.028417866 -0.1507759889
#>  [77,]  2   26 -1.830916388  1.079676676 -0.202921232 -1.4490071301
#>  [78,]  2   27 -1.723920284 -1.036116661 -1.347955543  0.6430087000
#>  [79,]  2   28 -0.796595459 -1.238426278 -0.138226435  0.4831938638
#>  [80,]  2   29 -0.712754234 -1.808929174  0.685111748 -0.0063556264
#>  [81,]  2   30  1.356591214 -2.720867201 -0.868046758  0.1514558929
#>  [82,]  2   31 -0.135010117 -1.772558302 -0.062191749 -0.5841089703
#>  [83,]  2   32  0.134562468 -2.146446908  0.081114507  0.3688067326
#>  [84,]  2   33  0.628504447 -0.248755042 -0.413780343  0.2946543397
#>  [85,]  2   34  0.356228505 -0.343160748 -0.900601090 -0.2792593733
#>  [86,]  2   35 -0.494334704 -0.447089008  0.811282723 -1.3362366549
#>  [87,]  2   36 -0.092655523  0.879965260  0.062331955  0.7007488184
#>  [88,]  2   37 -0.502909193  0.312656470 -0.980749549  0.5541966223
#>  [89,]  2   38  0.777253139  2.070717009 -1.820862858 -0.8363065928
#>  [90,]  2   39  0.427940809  2.536575805  1.311733541 -1.5945881620
#>  [91,]  2   40 -0.342872387 -0.698964993  0.280917237  0.2049585806
#>  [92,]  2   41 -0.592891035  0.809256853 -0.460139874 -0.3450879780
#>  [93,]  2   42  0.903474700 -0.324213189  0.230312888  0.2526117034
#>  [94,]  2   43 -0.833726530  0.275071290 -1.272587870 -1.2940024655
#>  [95,]  2   44  0.388070007  0.231357500 -1.964409651 -0.9591704444
#>  [96,]  2   45 -0.175422932  0.972331835 -0.435759725  1.0857748537
#>  [97,]  2   46  1.444211257 -1.177128892 -0.256786215  0.4037749047
#>  [98,]  2   47 -0.543940922 -1.017599460 -1.162974678  0.5864875367
#>  [99,]  2   48  0.472731675 -2.684019150 -0.733388669  1.8152284462
#> [100,]  2   49 -1.628771087 -2.239765631 -0.352060516  0.1288214286
#> [101,]  3    0 -1.898295636 -0.029902147  2.126111030 -0.0046207678
#> [102,]  3    1  0.494317528  0.858965584 -1.477783347  0.7602421677
#> [103,]  3    2  0.131010716  2.190078238 -0.469943221  0.0389909129
#> [104,]  3    3  0.844594441  1.072394160 -0.120273914  0.7350721416
#> [105,]  3    4  1.191534791  1.448148137  0.146162183 -0.1464726270
#> [106,]  3    5 -1.099030182  2.579608960  0.720190789 -0.0578873354
#> [107,]  3    6 -1.294099538  2.842559123  0.691713244  0.4823694661
#> [108,]  3    7 -0.572516410  1.643671128 -0.104225781  0.9929436368
#> [109,]  3    8 -2.594963450 -1.067219291  0.983566948 -1.2463954980
#> [110,]  3    9 -1.766115680 -0.940173743  0.721078320 -0.0334875248
#> [111,]  3   10 -1.816699557 -0.742886094 -0.303564358 -0.0709621812
#> [112,]  3   11 -0.443902549  2.259229618  2.089711119 -0.7589206537
#> [113,]  3   12  0.622282459  0.845028788 -2.335000998 -1.0343593609
#> [114,]  3   13  1.215976307 -0.866484671 -0.611598494 -0.6307319540
#> [115,]  3   14  0.841447817 -0.302897976  0.030281603  0.5868077200
#> [116,]  3   15  2.526770836 -0.591842211  0.372181030 -0.4163226562
#> [117,]  3   16  2.153005439  0.220636618 -0.075055023 -0.7848878095
#> [118,]  3   17  1.972302657 -1.380222933  0.312157153  0.1634163195
#> [119,]  3   18  1.319210927 -0.529011045 -0.937698801 -1.2367142351
#> [120,]  3   19  0.428856460 -2.130749206 -0.841048916  1.0458737762
#> [121,]  3   20  0.278897055 -2.079734941 -0.217906199 -0.4845954162
#> [122,]  3   21  0.453681227 -1.292753205 -0.423100809  0.1891288117
#> [123,]  3   22 -0.192911280 -0.914056722 -0.273942329  0.0510063316
#> [124,]  3   23  0.131937833 -0.240660655  0.436732746 -0.0002406689
#> [125,]  3   24  0.486372447 -0.672535663  0.593632610  1.8093820424
#> [126,]  3   25 -0.988963830 -0.607103047  0.767344407 -0.8253279571
#> [127,]  3   26 -0.064018368 -1.274374979 -0.705886916  1.1454704522
#> [128,]  3   27 -0.446578756 -0.214869399 -0.983992513  0.0315731876
#> [129,]  3   28 -0.053686001  1.243214514 -1.297625376 -0.8352058053
#> [130,]  3   29 -0.099274606  0.182457964 -0.780453329 -0.0687636490
#> [131,]  3   30 -0.168277004  0.907299060  1.080751505  0.7467716870
#> [132,]  3   31  1.218834335  0.356182924  0.212997719 -0.4255187344
#> [133,]  3   32 -0.819413091 -0.057711721  1.031164393 -0.7720822352
#> [134,]  3   33 -2.787977790 -0.187362794 -0.770124321  0.1527641067
#> [135,]  3   34 -0.546221475  1.128890883 -1.284680896  0.9885968452
#> [136,]  3   35 -1.640343420  0.843448851 -0.547186182 -0.0734583347
#> [137,]  3   36 -0.762319376  0.237562345 -1.788535380 -1.3870265536
#> [138,]  3   37 -0.434886836  0.597444902 -1.561564480 -1.3066759044
#> [139,]  3   38 -0.864626629  0.036383718  0.092280438 -0.7683953251
#> [140,]  3   39 -1.785483076  1.012808394  0.841305969 -0.5271081254
#> [141,]  3   40 -2.518918216  0.741566428  1.026252071 -0.0214270650
#> [142,]  3   41 -0.285127439  0.698629409  0.125401093  0.6704980710
#> [143,]  3   42 -0.105140584 -0.620216062  0.522586230 -0.4346170386
#> [144,]  3   43 -0.452978401  0.241731797  0.522386237 -1.1138797833
#> [145,]  3   44 -2.446372980  0.321533005 -0.609959048  0.6071059949
#> [146,]  3   45 -3.171992547  0.156552423 -1.333510968  0.2754569687
#> [147,]  3   46 -1.184311788  0.739618605 -0.237147398  1.1573470694
#> [148,]  3   47 -0.705891096  1.416069131 -1.464611161 -1.6824808595
#> [149,]  3   48 -1.264314270 -0.343463131 -1.411853028  0.0873190889
#> [150,]  3   49 -0.369852882  0.167317250 -1.525571823  1.3533618939
#> [151,]  4    0 -0.920289065  0.561872762 -0.982216865 -1.3038211536
#> [152,]  4    1  0.934133659  0.772718701 -0.349346663 -0.2509144650
#> [153,]  4    2  0.382315705 -0.467503981 -0.190330209  0.1710073737
#> [154,]  4    3 -0.744285672  0.023737629 -0.415764314 -0.4034674787
#> [155,]  4    4 -1.393276730 -0.856945649 -1.117188679  0.1046594408
#> [156,]  4    5 -0.929069472 -0.478502060 -1.316714459 -0.3188807867
#> [157,]  4    6  0.907993300 -0.039283610  0.144687559  1.6183439359
#> [158,]  4    7  0.937270207  1.111088146 -0.730276365  0.7141886015
#> [159,]  4    8  1.876052585 -0.669471088  0.539098928  2.9658653698
#> [160,]  4    9  2.271405605  0.036149087  0.519357974 -0.7950776046
#> [161,]  4   10  1.781121603 -0.902209054  0.079280816  0.8143659147
#> [162,]  4   11  2.510747792 -1.066929261 -0.601994977  2.0980308101
#> [163,]  4   12  0.845031936 -1.240187384  0.227309158  0.3009800552
#> [164,]  4   13 -0.185464043 -2.013688066  0.140357491 -1.0830751418
#> [165,]  4   14 -0.013636377 -1.044743036 -0.886939561 -1.0063225018
#> [166,]  4   15  2.254538666  1.125624424  0.087278643 -0.0354145649
#> [167,]  4   16  1.828709812 -0.332687141  1.278960745  1.3091243609
#> [168,]  4   17  0.459860261 -0.196678826  1.190968010  0.7504004866
#> [169,]  4   18 -0.551504516  1.704230892  0.354268874 -2.1383683276
#> [170,]  4   19 -0.484344902  1.134392917  0.310949653 -0.7003541089
#> [171,]  4   20 -2.546196808 -0.873076939  1.039103363 -0.0090564753
#> [172,]  4   21 -2.899993953  0.392718717  2.788379568 -1.4581334872
#> [173,]  4   22 -0.880011481  0.040370870  0.070367956  0.6945296465
#> [174,]  4   23 -0.767444456  0.123485666  0.175407398 -2.4613354752
#> [175,]  4   24  1.777232493 -0.381584990  0.280942465  0.1432897640
#> [176,]  4   25  1.381292429 -0.056618775  1.064439088 -0.3912221191
#> [177,]  4   26 -0.299384755 -0.009751115  1.151715776 -0.4911640856
#> [178,]  4   27 -0.071241824  0.744037949  0.413981944 -0.2836474525
#> [179,]  4   28  0.707176908 -0.743624317  0.764463652  0.3147948047
#> [180,]  4   29  0.361951847  0.212560247 -0.366933757  0.3963265779
#> [181,]  4   30  1.113936600 -0.380935262 -0.540960049 -0.2256037110
#> [182,]  4   31  0.877376479  0.214864245  0.103442768 -1.9249504303
#> [183,]  4   32  0.763932345  0.959900802 -0.648716502 -1.4392292983
#> [184,]  4   33 -1.064478541 -0.537789819 -0.718941886 -1.4696577740
#> [185,]  4   34  0.026023332 -0.680670687 -2.252273731  0.7618634469
#> [186,]  4   35 -1.011541144 -2.263825796 -0.956491309 -0.2436149819
#> [187,]  4   36  1.034414070 -1.106535752 -0.287131505  0.2696766074
#> [188,]  4   37  1.265942991 -2.174483902 -1.595713219 -1.5589275210
#> [189,]  4   38  1.023222560 -1.045888645  0.164600741 -0.5355880067
#> [190,]  4   39  2.576060609 -1.558643574  0.213150716  0.5624519732
#> [191,]  4   40 -0.179909709 -0.061794993 -0.084626306 -0.1783261234
#> [192,]  4   41 -0.258818882 -1.374775366 -1.876221249 -0.1151359862
#> [193,]  4   42 -1.136272451 -0.056077548 -1.534872990 -0.0720614721
#> [194,]  4   43 -0.501744713 -0.581478914  0.226888541  1.2109098069
#> [195,]  4   44 -0.226543636 -1.134792772  0.813888469 -0.6148969008
#> [196,]  4   45  1.102445585 -0.738635755  1.961848708  0.6761264583
#> [197,]  4   46  0.363828489 -1.072261918  0.853135062  0.8985996055
#> [198,]  4   47 -0.086082126 -0.313214890  0.728364391 -1.1893179035
#> [199,]  4   48  0.053634783  0.765712744  0.918712026  0.1212588498
#> [200,]  4   49  0.205728509 -0.554782780  0.130827255 -0.0112216861
#> [201,]  5    0 -0.126599418 -0.084897319  1.194652900 -0.2484829332
#> [202,]  5    1  0.589741223 -1.730263465  2.285264743  0.4223203858
#> [203,]  5    2  0.257482174 -1.053831904  0.582324023  0.9876532936
#> [204,]  5    3 -0.034144433 -0.184579300  1.024138102  0.8355681720
#> [205,]  5    4 -0.395590124 -0.318044457  0.743584212 -0.6605218590
#> [206,]  5    5 -0.854046324 -2.027847487  2.009938204  1.5640694933
#> [207,]  5    6 -1.686437859 -0.642893734  0.395859141 -1.6229759353
#> [208,]  5    7 -1.623217004  1.424663707 -0.152153347  0.8638963735
#> [209,]  5    8  0.706597540 -0.221679149 -0.363615963 -0.5116027734
#> [210,]  5    9  0.894244631  2.555790917 -0.407415107 -1.9173650254
#> [211,]  5   10 -0.268435072  2.288972346 -0.325278893 -1.8658138486
#> [212,]  5   11 -0.548322040  1.706169620 -0.184602584  0.2451790634
#> [213,]  5   12 -0.556961764  1.184276750 -1.080789491  2.2235343019
#> [214,]  5   13 -1.173770296  1.850918525 -0.085231929  0.2733760641
#> [215,]  5   14  0.391694605 -0.230500388 -1.764758671  1.1307847945
#> [216,]  5   15  1.679091985  2.303721671 -2.571186162  0.8386693741
#> [217,]  5   16  1.306341430  0.677886976 -1.652868557 -0.6546152779
#> [218,]  5   17  2.251945621 -1.660163451 -0.787681158  0.9539609785
#> [219,]  5   18  2.306171944 -0.820396036 -0.941032820  0.3529514655
#> [220,]  5   19  2.100270466  0.700693067  0.499881067  0.2065992926
#> [221,]  5   20  0.070577522 -0.445965753 -0.249562064  1.0011132343
#> [222,]  5   21  0.073479979  0.158360901  1.099039769  0.7474519916
#> [223,]  5   22  0.806931783 -1.599254282  1.290496038 -0.6265748367
#> [224,]  5   23  1.778265528  0.308132726  1.527085312  0.3952236859
#> [225,]  5   24  2.266467847 -0.686229613  1.033876805 -0.8921679540
#> [226,]  5   25  1.933365651  0.916102068  0.762458502  0.6308184101
#> [227,]  5   26  1.311684982  1.977459863 -0.082117268 -0.4327051833
#> [228,]  5   27  0.179709155  0.987802102  0.584201517  0.4521386442
#> [229,]  5   28  1.423493855 -0.464982255  1.472049944  0.3679990452
#> [230,]  5   29  1.515497787  2.346779010  0.319916958 -0.2703874897
#> [231,]  5   30  0.320216420  1.770542896 -0.725092485  0.4655126177
#> [232,]  5   31  0.836419626  1.329294562 -0.746192144  0.5743562256
#> [233,]  5   32 -0.267385622 -1.207026743  0.116356642 -0.2306999902
#> [234,]  5   33  0.555016843 -1.555058520 -2.023281417  1.1722425083
#> [235,]  5   34  0.162227614 -0.146837420  0.770355560  1.3927027067
#> [236,]  5   35 -0.424716051 -0.429747177  1.434595492 -0.6619912557
#> [237,]  5   36  0.618499211 -2.196185702  0.792290719 -0.7773692067
#> [238,]  5   37  0.034529940 -0.050251065 -0.209825792  0.5135385789
#> [239,]  5   38  0.316992570 -1.028211247 -1.322179610 -0.9133122380
#> [240,]  5   39 -0.953250847  0.194826524 -1.142831563 -0.4494238053
#> [241,]  5   40  1.451781549  0.214445766 -1.193871207  0.8029326991
#> [242,]  5   41  1.994375478 -0.584606944 -1.235767450 -0.5734768510
#> [243,]  5   42 -0.921925858 -2.513330075  0.017347695 -1.9281251678
#> [244,]  5   43 -0.219863185 -1.867523645 -2.765233797  0.6643908338
#> [245,]  5   44  1.144195055 -2.179376226 -2.181930275 -1.6025402238
#> [246,]  5   45 -0.784207275 -1.030750227  0.350398055 -1.3546002575
#> [247,]  5   46  0.600739642  0.101932196  0.959299591 -3.0179326794
#> [248,]  5   47  0.470032098 -0.127997112  1.092422995  0.8312378222
#> [249,]  5   48  0.932364356 -0.755553721  0.414299956  0.2510970888
#> [250,]  5   49  0.453736665  0.347472948 -0.114359287  0.4622934659
#>                  x2            x3
#>   [1,]  0.321925265  1.200965e+00
#>   [2,] -0.783838941  1.044751e+00
#>   [3,]  1.575727520 -1.003209e+00
#>   [4,]  0.642899306  1.848482e+00
#>   [5,]  0.089760647 -6.667734e-01
#>   [6,]  0.276550747  1.055138e-01
#>   [7,]  0.679288816 -4.222559e-01
#>   [8,]  0.089832887 -1.223502e-01
#>   [9,] -2.993090083  1.881930e-01
#>  [10,]  0.284882954  1.191610e-01
#>  [11,] -0.367234643 -2.509255e-02
#>  [12,]  0.185230565  1.080727e-01
#>  [13,]  0.581823727 -4.854352e-01
#>  [14,]  1.399736827 -5.042171e-01
#>  [15,] -0.727292059 -1.661099e+00
#>  [16,]  1.302542632 -3.823337e-01
#>  [17,]  0.335848120 -5.126503e-01
#>  [18,]  1.038506099  2.701891e+00
#>  [19,]  0.920728568 -1.362116e+00
#>  [20,]  0.720878163  1.372562e-01
#>  [21,] -1.043118939 -1.493625e+00
#>  [22,] -0.090186387 -1.470436e+00
#>  [23,]  0.623518162  1.247024e-01
#>  [24,] -0.953523358 -9.966391e-01
#>  [25,] -0.542828815 -1.822614e-03
#>  [26,]  0.580996498 -4.282589e-01
#>  [27,]  0.768178738 -6.136716e-01
#>  [28,]  0.463767589 -2.024678e+00
#>  [29,] -0.885776297 -1.224748e+00
#>  [30,] -1.099780899  1.795164e-01
#>  [31,]  1.512707010  5.676206e-01
#>  [32,]  0.257921438 -4.928774e-01
#>  [33,]  0.088440229  6.288407e-05
#>  [34,] -0.120896538  1.122890e+00
#>  [35,] -1.194328895  1.439856e+00
#>  [36,]  0.611996898 -1.097114e+00
#>  [37,] -0.217139846 -1.173196e-01
#>  [38,] -0.182756706  1.201498e+00
#>  [39,]  0.933346329 -4.697296e-01
#>  [40,]  0.821773111 -5.246948e-02
#>  [41,]  1.392116376 -8.610730e-02
#>  [42,] -0.476173923 -8.876790e-01
#>  [43,]  0.650348561 -4.446840e-01
#>  [44,]  1.391110456 -2.944488e-02
#>  [45,] -1.110788879 -4.138688e-01
#>  [46,] -0.860792587  1.113386e+00
#>  [47,] -1.131738681 -4.809928e-01
#>  [48,] -1.459214000 -4.331690e-01
#>  [49,]  0.079982553  6.968626e-01
#>  [50,]  0.653204340 -1.056368e+00
#>  [51,] -2.000929238 -1.096156e+00
#>  [52,]  0.333777197  4.905045e-02
#>  [53,]  1.171325127 -1.198496e+00
#>  [54,]  2.059539242  1.900190e-01
#>  [55,] -1.376861598  1.297706e+00
#>  [56,] -1.150855566 -1.033874e+00
#>  [57,] -0.705821395 -7.384408e-01
#>  [58,] -1.054055782  4.656394e-02
#>  [59,] -0.645743723 -1.017596e+00
#>  [60,] -0.185377968 -3.832840e-01
#>  [61,] -1.201222051  8.727554e-01
#>  [62,]  2.036972167  9.695450e-01
#>  [63,]  0.107774745  3.838467e-01
#>  [64,] -0.084108101 -1.851556e+00
#>  [65,]  0.495619642 -5.399674e-02
#>  [66,]  0.037415186  1.064773e+00
#>  [67,] -0.132088037  8.131950e-01
#>  [68,]  1.476787424 -1.908165e-01
#>  [69,] -0.217030210 -2.699930e+00
#>  [70,] -1.283602204  6.096664e-02
#>  [71,]  0.385667890  5.737517e-01
#>  [72,] -0.351512874  4.580358e-02
#>  [73,] -0.521796093  1.574125e-01
#>  [74,] -1.068131201  4.315654e-01
#>  [75,]  0.428365903 -3.965497e-01
#>  [76,] -0.174018234  1.309978e+00
#>  [77,]  0.515667729  4.703934e-01
#>  [78,] -0.234365277 -1.242670e+00
#>  [79,] -0.658503426  1.381575e+00
#>  [80,]  1.250236604  1.204459e+00
#>  [81,] -0.271763715  8.240740e-01
#>  [82,]  0.947951996 -1.662629e+00
#>  [83,] -1.201582430 -5.693063e-01
#>  [84,] -0.466116096  6.355138e-01
#>  [85,] -0.269351395  4.372201e-02
#>  [86,] -0.390965408  3.480123e-01
#>  [87,]  1.348707012  2.459594e+00
#>  [88,] -0.022764701 -8.183803e-01
#>  [89,]  0.244225851 -2.113200e+00
#>  [90,] -0.942371708  2.736953e-01
#>  [91,] -0.729217277 -6.875968e-01
#>  [92,]  0.998068909  4.460411e-01
#>  [93,]  1.258481665 -8.123847e-01
#>  [94,]  1.248863689  2.212055e+00
#>  [95,] -1.380637050 -1.237060e-01
#>  [96,]  2.049960694 -4.773355e-01
#>  [97,]  1.016872830 -1.662615e-01
#>  [98,] -0.026717464  8.625634e-01
#>  [99,]  0.703607779  9.734049e-02
#> [100,] -0.971385229 -1.625617e+00
#> [101,]  0.724173801  1.334913e+00
#> [102,] -0.832552826 -8.692718e-01
#> [103,]  0.732528487  5.548695e-02
#> [104,] -0.871926870  4.906691e-02
#> [105,] -0.453397511 -5.783557e-01
#> [106,]  1.187534279 -9.987387e-01
#> [107,] -0.290145312 -2.432780e-03
#> [108,]  0.828546145  6.555119e-01
#> [109,] -0.291227709  1.476842e+00
#> [110,] -1.576362405 -1.909153e+00
#> [111,] -0.848815697 -7.024395e-01
#> [112,] -1.088519862 -3.114302e-01
#> [113,] -0.484290571 -1.663157e+00
#> [114,] -0.336311209 -7.505334e-01
#> [115,] -0.153357891 -7.773518e-01
#> [116,] -0.243247229 -7.225697e-01
#> [117,]  1.892202042 -2.188835e+00
#> [118,] -1.385998337  2.134186e-01
#> [119,] -0.414824301 -6.319229e-01
#> [120,]  0.349081528  1.520491e+00
#> [121,]  1.628442266  7.959559e-01
#> [122,]  0.088521896 -1.453530e+00
#> [123,]  1.239150708  9.839542e-02
#> [124,] -1.644555536 -5.937710e-01
#> [125,]  1.446356525  8.882812e-01
#> [126,] -0.690560172  5.307042e-02
#> [127,] -0.276431085 -5.570236e-01
#> [128,] -1.109418760  4.383970e-01
#> [129,]  0.133869316  1.526082e-01
#> [130,]  1.785339052 -1.646176e-01
#> [131,]  2.422163355  2.019891e+00
#> [132,] -1.076828902 -5.293859e-01
#> [133,]  0.485941110 -4.707870e-01
#> [134,]  1.388521739 -1.545937e+00
#> [135,] -0.195656817 -4.052672e-02
#> [136,] -0.218174798  8.903563e-01
#> [137,] -0.304777955 -2.071388e+00
#> [138,]  0.597832724 -2.500651e-01
#> [139,]  1.397429411 -1.181650e+00
#> [140,]  0.687619761  1.441937e+00
#> [141,]  0.320188014  1.357896e+00
#> [142,] -0.301869926  3.345028e-01
#> [143,]  0.498348686  1.429338e+00
#> [144,] -0.549536918 -8.673179e-01
#> [145,] -0.279256504  9.506517e-01
#> [146,]  1.096513443 -5.850115e-01
#> [147,]  0.442013088  3.209575e-01
#> [148,]  0.241016294 -2.993960e-01
#> [149,] -0.255607655 -2.785431e-01
#> [150,]  0.931032901  5.461152e-01
#> [151,]  1.029140719 -1.561266e+00
#> [152,]  0.914774868  3.247707e-01
#> [153,] -0.002456267 -1.567903e-01
#> [154,]  0.136009552  8.777839e-01
#> [155,] -0.720153545  7.501666e-01
#> [156,] -0.198124330  3.010534e-01
#> [157,] -1.029208806  1.492811e+00
#> [158,] -0.966955896 -1.525494e+00
#> [159,] -1.220813089  9.107176e-01
#> [160,]  0.836207704 -1.579532e+00
#> [161,]  1.114971798  5.877163e-01
#> [162,] -0.412944853  8.964230e-02
#> [163,] -1.128977398  9.672617e-01
#> [164,] -0.087931332  7.881235e-02
#> [165,]  2.241903693 -1.568700e+00
#> [166,]  2.041313171 -2.007823e+00
#> [167,] -1.719800337  5.409735e-01
#> [168,] -0.356906656 -7.337660e-02
#> [169,]  1.533280492 -5.710184e-01
#> [170,] -0.038244097 -3.110685e-01
#> [171,]  1.597413242 -6.714151e-01
#> [172,] -0.333585393 -1.573405e-01
#> [173,]  0.604985924 -9.313051e-01
#> [174,]  0.224241239 -1.983009e+00
#> [175,]  3.229069495 -2.195998e-01
#> [176,]  0.920452567  1.045276e+00
#> [177,] -1.206539069  1.877330e+00
#> [178,] -0.603881169  2.606196e-03
#> [179,]  0.370235285 -8.066993e-02
#> [180,] -1.901000647  9.629823e-01
#> [181,] -1.804409945  5.357102e-02
#> [182,] -1.121822606 -4.348984e-01
#> [183,] -0.347929607 -1.737297e+00
#> [184,]  1.238902149 -1.263696e+00
#> [185,] -0.274197565  4.063085e-01
#> [186,]  0.162380716 -1.459654e+00
#> [187,] -0.064606910  1.048457e+00
#> [188,] -0.705237097 -1.346431e+00
#> [189,]  1.362197269 -1.935706e-01
#> [190,] -1.096513188 -2.335957e-03
#> [191,] -0.228433519 -1.282973e-02
#> [192,] -0.347828081  1.519473e-01
#> [193,]  0.532128620  5.985111e-01
#> [194,]  1.607234562 -1.262124e-01
#> [195,]  0.513814526 -2.485356e-01
#> [196,]  1.382373161  1.603274e-01
#> [197,]  0.763097049 -4.336419e-01
#> [198,] -0.624584545  1.537412e+00
#> [199,]  0.081543800 -2.170247e+00
#> [200,]  1.376079111  1.027005e+00
#> [201,]  0.844792223  2.946924e-01
#> [202,] -0.041971524  3.927413e-01
#> [203,] -1.105575906 -1.000844e+00
#> [204,]  0.563775652 -3.257271e-01
#> [205,]  1.303364756 -1.008349e+00
#> [206,] -1.500220938 -6.354315e-01
#> [207,] -0.606989235 -1.209841e+00
#> [208,] -0.292245066 -1.116464e+00
#> [209,] -1.289683343  6.298812e-01
#> [210,]  0.694105849 -2.725216e-01
#> [211,] -0.599181903 -2.588412e-01
#> [212,]  1.256906644  1.729558e+00
#> [213,]  0.053508014 -5.839217e-02
#> [214,]  0.728092518 -5.370638e-01
#> [215,]  1.561098068  7.472867e-01
#> [216,]  0.265624754 -4.872578e-01
#> [217,]  1.076726259  1.372908e+00
#> [218,]  0.210697932 -3.776724e-01
#> [219,] -1.511673526 -6.161526e-01
#> [220,]  0.022402260 -1.168125e+00
#> [221,]  0.718136206  3.286404e-01
#> [222,]  0.489457018  1.466511e+00
#> [223,] -0.173888346 -3.560095e-01
#> [224,] -1.217699489  2.614676e-01
#> [225,]  0.646398369  3.333289e-01
#> [226,] -0.916456028  1.422193e+00
#> [227,] -1.251823442  6.638766e-01
#> [228,]  0.594927718 -1.073655e+00
#> [229,] -1.232810665 -6.969018e-01
#> [230,]  0.244364415 -7.461305e-01
#> [231,]  0.002772185  1.415729e-01
#> [232,] -1.328209686 -3.947221e-03
#> [233,]  1.179696412  3.679375e-01
#> [234,] -0.592804999 -6.573429e-01
#> [235,]  1.199978316 -3.763467e-01
#> [236,] -0.475033680  7.413601e-01
#> [237,] -0.575057096 -9.960676e-02
#> [238,] -0.031226025 -6.542899e-01
#> [239,] -0.358056996  9.711644e-01
#> [240,] -0.356601078  1.349631e-02
#> [241,] -0.877664383 -9.165347e-01
#> [242,] -1.212897125  1.709689e+00
#> [243,]  0.613286619 -1.168101e+00
#> [244,] -0.806203341 -1.781036e+00
#> [245,] -1.376456930 -2.253132e+00
#> [246,] -0.507847899  6.511260e-01
#> [247,] -0.800935487 -5.328326e-01
#> [248,] -2.192785686 -2.755594e-01
#> [249,] -0.290937149  2.896270e-01
#> [250,]  0.167174121 -4.664841e-01