Test the Drift Matrix

Both have to be true for the function to return TRUE.

Test that the real part of all eigenvalues of \(\boldsymbol{\Phi}\) are less than zero.

Test that the diagonal values of \(\boldsymbol{\Phi}\) are between 0 to negative inifinity.

Usage

TestPhi(phi, a_target = 0, auto_ubound = 0)

Arguments

phi

Numeric matrix. The drift matrix (\(\boldsymbol{\Phi}\)).

a_target

Numeric scalar specifying the stability threshold for the real part of the eigenvalues. The default 0.0 corresponds to the imaginary axis; values less than 0.0 enforce a stricter stability margin.

auto_ubound

Numeric scalar specifying the upper bound for the diagonal elements of \(\boldsymbol{\Phi}\). Default is 0.0, requiring all diagonal values to be \(\leq 0\).

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

phi <- matrix(
  data = c(
    -0.357, 0.771, -0.450,
    0.0, -0.511, 0.729,
    0, 0, -0.693
  ),
  nrow = 3
)
TestPhi(phi = phi)
#> [1] TRUE