Coerce an Object of Class simstatespace to a Data Frame

Coerce an Object of Class simstatespace to a Data Frame

Usage

# S3 method for class 'simstatespace'
as.data.frame(
  x,
  row.names = NULL,
  optional = FALSE,
  eta = FALSE,
  long = TRUE,
  ...
)

Arguments

x

Object of class simstatespace.

row.names

NULL or character vector giving the row names for the data frame. Missing values are not allowed.

optional

Logical. If TRUE, setting row names and converting column names is optional.

eta

Logical. If eta = TRUE, include eta. If eta = FALSE, exclude eta.

long

Logical. If long = TRUE, use long format. If long = FALSE, use wide format.

...

Additional arguments.

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
## number of individuals
n <- 5
## time points
time <- 50
## dynamic structure
p <- 3
mu0 <- rep(x = 0, times = p)
sigma0 <- diag(p)
sigma0_l <- t(chol(sigma0))
alpha <- rep(x = 0, times = p)
beta <- 0.50 * diag(p)
psi <- diag(p)
psi_l <- t(chol(psi))
## measurement model
k <- 3
nu <- rep(x = 0, times = k)
lambda <- diag(k)
theta <- 0.50 * diag(k)
theta_l <- t(chol(theta))
## covariates
j <- 2
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    matrix(
      data = stats::rnorm(n = time * j),
      nrow = j,
      ncol = time
    )
  }
)
gamma <- diag(x = 0.10, nrow = p, ncol = j)
kappa <- diag(x = 0.10, nrow = k, ncol = j)

# Type 0
ssm <- SimSSMFixed(
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  nu = nu,
  lambda = lambda,
  theta_l = theta_l,
  type = 0
)

head(as.data.frame(ssm))
#>   id time         y1          y2         y3
#> 1  1    0  1.6162534  1.38408950 -0.4111695
#> 2  1    1 -1.5232952  1.92627469 -0.7992414
#> 3  1    2 -0.3544375 -0.02220621 -1.3175714
#> 4  1    3 -1.9809839  0.16221765  0.2537055
#> 5  1    4 -2.2357532  2.58289080  0.9846965
#> 6  1    5 -0.3865249  1.52753010 -0.9550317
head(as.data.frame(ssm, long = FALSE))
#>   id        y1_0       y2_0       y3_0       y1_1       y2_1       y3_1
#> 1  1  1.61625337  1.3840895 -0.4111695 -1.5232952  1.9262747 -0.7992414
#> 2  2 -0.04475887 -2.0473263 -1.0995401  0.1899425 -0.8022084  1.0326053
#> 3  3 -0.35175450 -0.8620198 -0.4862692 -1.6063570  1.1059695  0.7106665
#> 4  4 -0.04919594  3.4244821 -0.2611621  0.9872575  1.7031256  2.2455017
#> 5  5 -2.32538496  0.9005817  0.3628444 -2.4199019 -0.7584027  1.0054983
#>         y1_2        y2_2        y3_2       y1_3       y2_3       y3_3
#> 1 -0.3544375 -0.02220621 -1.31757139 -1.9809839  0.1622177  0.2537055
#> 2  0.2011749  2.65001795  0.07700843 -0.5935679  1.2088950 -0.6923134
#> 3 -1.3413361  1.61786709  0.35637548  0.2187572  0.3953878  1.9271685
#> 4 -0.5799665  1.32218398  1.34751556 -0.8219741 -2.3408688 -0.8141161
#> 5 -0.9017397  3.27872010  0.32578948  1.4083978  2.6259266 -1.2293404
#>          y1_4      y2_4       y3_4        y1_5       y2_5       y3_5
#> 1 -2.23575323 2.5828908  0.9846965 -0.38652493  1.5275301 -0.9550317
#> 2 -1.60687060 0.5230821 -0.1737218 -0.22715129  1.9330535  1.3164819
#> 3 -0.04691762 0.6900648 -2.2786571 -2.39040982  0.8390154 -0.7203359
#> 4 -0.86146604 1.9597340  0.8576239 -0.41418366 -0.1887423  1.1741567
#> 5  1.15099319 0.4568205 -1.0356634  0.07745305  2.5743280 -0.3840216
#>          y1_6        y2_6       y3_6       y1_7       y2_7       y3_7
#> 1 -0.47791533 -0.05186306 -1.5771627 -1.9745206  1.9076700 -0.3836298
#> 2 -2.34672407  2.97237423  2.6146254 -1.6010079  0.1317900  1.6232148
#> 3  0.02758404  1.45522180 -1.7086169  3.5099721  1.2209718  1.3569414
#> 4  0.09920765 -1.00373952  1.3464648 -0.8458918 -2.9079858  0.1188553
#> 5  0.08823787  0.85463074 -0.1851625 -0.2603230  0.1072301  0.1982461
#>          y1_8       y2_8      y3_8       y1_9       y2_9       y3_9      y1_10
#> 1 -1.84909940  0.3162936  1.143882  0.6510890 -0.1876209 -0.5045624  2.3831996
#> 2 -0.54091538 -1.6836697  4.109281 -0.7970075 -0.2145923 -0.9984378  1.2757208
#> 3  1.50020639  0.7332345  3.294612 -0.1923136 -0.3978935  0.5829798 -0.2658483
#> 4  0.07987889 -1.7158004  1.050014 -0.9939521 -2.6680120  0.2642090  0.4325857
#> 5  1.01172797 -0.1018807 -2.841976 -2.0849596 -0.4872781  0.2992914 -1.2304391
#>        y2_10      y3_10       y1_11       y2_11      y3_11      y1_12
#> 1 -1.3758294  0.4845554  1.64979582 -1.11965602  0.8464700  0.2986962
#> 2  1.9613728  0.3016141  0.05475791  1.07159722 -0.9080502  0.4473381
#> 3  0.4920791  0.5569878  0.02408590 -2.04233004  0.4163683  0.6014075
#> 4 -0.7117496 -0.2928771 -0.75795600 -0.03295955 -1.8552625 -3.1205632
#> 5 -1.3019437  1.3926087 -2.27126241 -1.03470055  0.5592967 -1.0212046
#>        y2_12      y3_12      y1_13       y2_13      y3_13       y1_14
#> 1  3.4085424 -0.5178566 -2.3503508  1.02991334 -0.9227273 -0.70897783
#> 2 -0.6233495 -0.4735484 -0.2239352 -0.09317512  0.5644049 -1.05684353
#> 3  0.7830307  0.6959900 -1.0477854  1.85814585 -1.1128188 -1.68248836
#> 4  0.4237490  0.2180062 -0.3577689  0.54474564 -0.5772663 -0.06223421
#> 5 -0.4610074 -0.8906314 -0.5504551 -0.51539350 -2.5795614 -0.84613947
#>        y2_14       y3_14       y1_15       y2_15       y3_15      y1_16
#> 1  1.2279481 -2.88552243 -1.10910481 -0.49726091 -2.87605245 -1.2853619
#> 2 -2.2102637  0.66448570  0.58268562  0.28523090  0.04264334  0.9131296
#> 3  2.8175307 -0.01904589 -0.03977694 -0.08811798  0.63116514 -2.1860271
#> 4 -1.1343966 -1.03940884 -1.86969158 -3.15045449  1.22888055 -0.1971206
#> 5  0.4678982 -1.18607203 -0.84415745  0.67704173  0.68194793  0.6927647
#>        y2_16      y3_16       y1_17      y2_17       y3_17      y1_18
#> 1 -0.2842451 -1.0167008 -0.08604294  1.7148485 -2.69775499  0.5812529
#> 2 -0.9245131  0.1107793 -0.30960035 -1.9897506 -0.29548962 -1.0007548
#> 3 -1.1531789 -0.1925946 -1.08838493 -0.4985157  1.14312425 -1.2158110
#> 4 -1.9700271  1.4993960 -0.66114376  0.5260738 -0.63517024  0.5566757
#> 5  1.1351382  1.1490660 -0.72492632  3.9159707  0.07542342  0.6944526
#>        y2_18       y3_18       y1_19      y2_19      y3_19      y1_20
#> 1 -0.2369862 -2.49272996 -0.43617985 -0.5114359  0.5075110  1.4113725
#> 2 -1.9941789  0.21216641 -0.29439721 -0.5181764  2.4067212 -0.2731870
#> 3  2.5857987 -1.74920314 -0.03522969  0.5649820  1.9746364 -0.3068359
#> 4  1.2124295 -0.41332821  0.05630755  2.1318854 -0.5604419  0.8822642
#> 5 -0.2262339 -0.04938728 -1.16588456  0.3797271  1.4092183  1.7910299
#>        y2_20      y3_20      y1_21      y2_21      y3_21        y1_22     y2_22
#> 1 -0.2775884 -0.5805856  0.5153738 -0.5774842  1.3461776  3.596343553 -1.015069
#> 2 -0.2414759  1.7093432 -0.1880165  0.2443441  3.2912305 -0.607617672  3.498036
#> 3  2.1865472  1.2485908  0.7495193  1.8971623 -0.2436155  0.002613899  3.302662
#> 4  1.7698072  2.1359884  0.8281188  0.3965071  1.4140650 -1.121966830  1.835665
#> 5  0.5259533 -0.2166186  0.7415509  1.7855709  0.3856954 -2.360038750  1.374296
#>        y3_22      y1_23       y2_23      y3_23      y1_24      y2_24      y3_24
#> 1  0.0226050  2.5672517  1.35666661 -0.6335853  0.9607637  0.6095592  1.4023794
#> 2  0.4400620  0.1506179 -0.14861547 -1.6585065 -0.6433440 -1.1929799 -0.2311295
#> 3 -0.3179609  2.6725823  2.90101024  1.6888110  1.4395645  1.5184101 -1.1831673
#> 4 -0.9326660 -0.5671104  0.09194007  2.9121667 -1.2977790 -0.7019866  0.5626923
#> 5  1.1644061 -0.1364091  1.43208504  1.3060583 -1.2954587 -3.3133986  0.3565990
#>        y1_25      y2_25       y3_25      y1_26      y2_26      y3_26     y1_27
#> 1  1.9757187 -0.7188623 -1.08740006  0.5135102  1.1672849 -2.7907014 -2.147634
#> 2 -2.0777220 -1.0833064 -0.53497728  0.2295865 -1.5825029  2.7170672  1.482373
#> 3  0.2977171  1.9956090 -1.57492885  1.7142149  0.5874467 -1.5651869  1.362578
#> 4  0.8652171 -1.4067314  1.32993361  0.3349119 -2.8006450  0.8948044 -1.425575
#> 5  0.9956078 -1.7670528 -0.08862718 -2.4743535 -0.7087822 -0.1489025 -1.033740
#>        y2_27       y3_27      y1_28      y2_28      y3_28       y1_29
#> 1 -1.6352704  0.01636576  0.2899833  0.0791213 -0.4072794  1.45918868
#> 2 -0.4078197 -0.71209126  0.5164032  0.7878999  1.2607620 -0.02535035
#> 3  2.2117553 -0.82172855 -0.6972844 -1.1158735 -1.3716504  1.55399030
#> 4 -0.1066037 -1.18425988 -0.2862932 -2.2443239 -2.5856469  2.82294540
#> 5 -0.2955502 -0.73433686  0.3713890  1.1065388 -0.5803437  0.18063800
#>        y2_29      y3_29      y1_30      y2_30      y3_30       y1_31
#> 1 -2.1495539  1.3943432  0.8209480 -1.4734985 -0.4914223 -1.24810796
#> 2  0.4410402  1.9206276  2.1996889  3.4661757  1.1667738  0.80867357
#> 3 -1.7740351  0.3396966  0.9610677  0.8411052  0.2434068  1.07834520
#> 4  0.7514304 -0.2535041  1.0541691  2.9329244 -0.6659829 -0.33468009
#> 5 -0.2898897 -2.7944895 -0.2233469 -0.4254074 -1.7772095  0.09860742
#>         y2_31      y3_31      y1_32       y2_32      y3_32     y1_33
#> 1 -0.53574403  1.8486074  0.8113517 -0.46419208  2.2016191  3.415350
#> 2  1.30923202  0.2815890  1.5854441  2.42009251  0.8443721  2.234726
#> 3  0.12788082  0.2485631 -1.9480654 -1.12356865 -0.9089033 -2.148746
#> 4 -0.07606522 -1.1034329  1.3299005 -0.05077669  1.3509872  1.177275
#> 5  0.21793047  0.5524406  2.1090294 -1.58307664  1.6274305  1.307622
#>         y2_33     y3_33      y1_34      y2_34      y3_34       y1_35     y2_35
#> 1  3.79006196 1.5123009 -0.3304899  1.7468472  2.0719680 -0.74759469 1.4146564
#> 2  2.23747372 0.7440492  2.8031746  1.6726827  1.3227618  1.18430270 2.2366559
#> 3  0.01415462 1.3291999 -1.5777140  1.2882232 -0.2606539  0.07153844 2.8709768
#> 4 -2.20632144 1.3496753 -0.5338542  0.1748380 -1.1911688 -0.35455243 0.6370667
#> 5  0.41535689 1.6443757  0.2683620 -0.4900236  0.7308648  0.06128940 0.1717690
#>        y3_35      y1_36      y2_36      y3_36       y1_37     y2_37       y3_37
#> 1  1.0198628 -0.1944395  0.8854330 -0.4322774  0.07233919 -1.388633  0.05290312
#> 2  1.1742738 -0.4550634 -0.8409145  1.2436593 -1.25794738 -1.652569  0.90287860
#> 3 -0.5532294  1.8323151 -0.4501924  0.1842798  0.23564658  0.814098 -0.49638849
#> 4  0.3190500  1.2553414  0.2026020 -0.3143567  1.69260785 -1.309939 -0.33704338
#> 5  0.8057427 -0.2327205  1.3692313  0.0351231  0.74480078 -1.156616 -0.63663308
#>          y1_38      y2_38      y3_38      y1_39     y2_39       y3_39
#> 1  0.443147072 -1.7593746 -1.5113774  0.5297982 -1.917194 -0.87312929
#> 2 -0.781440030 -0.9407196  1.9675205  0.7363899 -2.988072  0.71798859
#> 3  0.695325057 -0.2591509 -0.3594548  0.6645011  0.551236  2.68977216
#> 4 -1.161582117 -1.4425935 -2.0659299 -1.7393826 -0.550202 -0.03785555
#> 5  0.005303157  0.7985164  0.1086735 -0.8871705  1.698271  0.40879492
#>        y1_40      y2_40      y3_40      y1_41      y2_41      y3_41      y1_42
#> 1 -0.1055503 -0.6508614 -1.7886765 -1.7211278 -1.7478217 -1.3823884 -3.1341750
#> 2  0.3463280 -1.1180259 -1.6974972  1.1024505 -0.6403250  0.1884423  1.7235042
#> 3 -1.3303646 -0.9439727  0.8470835  0.9670889 -1.0168700  1.4387706  0.2945571
#> 4  1.9451513 -1.1986014 -0.0677613 -0.8421198 -0.7832004 -0.9559934 -0.4534582
#> 5  0.6287751  0.9767175  0.9856875  0.2784196 -0.3577045 -3.0269869 -1.9687273
#>        y2_42       y3_42      y1_43      y2_43      y3_43       y1_44
#> 1 -2.4218606 -0.34383038 -2.1621986 -0.5874906 -0.9592735  0.07684553
#> 2 -2.1913185 -0.99739117 -0.2502237 -0.1004548 -0.2129836  0.38799381
#> 3 -3.8564551  2.45672208 -0.1084257 -1.6772485  0.7707791  0.33283223
#> 4  1.0494567  0.05230591  0.7511409  1.2285816  1.8865604 -0.16173517
#> 5 -0.5729698 -0.87992403 -1.5863713 -0.4946458  0.2917907 -0.27219374
#>        y2_44      y3_44      y1_45      y2_45      y3_45      y1_46      y2_46
#> 1  0.6498703 -2.0858223  0.1913725 -0.3984163  0.7564219 -0.1914062 -1.5603954
#> 2  0.3487298  1.1768471 -0.5165214 -0.6068654  0.1326481 -1.1978164  0.2747516
#> 3 -0.7693597 -1.0345149 -0.6230276 -1.5716891  0.7956358 -2.8786132 -0.4584111
#> 4  1.8282077 -0.4973262 -1.6866041  2.9826468 -1.5307716 -0.8588464  2.3200099
#> 5  0.6703838 -1.3712142  0.4966287 -0.3086405 -1.2083143 -0.1148448  0.8002702
#>        y3_46       y1_47      y2_47      y3_47      y1_48      y2_48      y3_48
#> 1  0.8761995 -0.26632228 -1.4031044  1.3346467 -0.2329856  1.0748771  0.7753996
#> 2 -0.1832172  0.08399896  1.3640486 -0.1636911 -1.2393437  0.1201421 -0.6573276
#> 3  0.9819970 -1.15587658  0.2381519  0.4382173 -2.4674501 -0.8868146 -0.2002175
#> 4 -0.3044024  2.72098966  1.1599923  0.6048861  1.2614324  0.2585708  2.4524881
#> 5 -1.0484130 -0.50487573  1.0964549 -4.1001085 -0.5997751  1.9156734 -1.0268122
#>         y1_49      y2_49      y3_49
#> 1  0.05097052  0.8516830 -0.8218398
#> 2 -1.72817957 -1.5878168  1.8028099
#> 3 -2.83662766  0.6451701  0.2810525
#> 4 -0.50741093  1.3275485  0.6990856
#> 5 -2.21734387  0.6382063 -1.3630000

# Type 1
ssm <- SimSSMFixed(
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  nu = nu,
  lambda = lambda,
  theta_l = theta_l,
  type = 1,
  x = x,
  gamma = gamma
)

head(as.data.frame(ssm))
#>   id time         y1         y2         y3        x1          x2
#> 1  1    0 -1.3212950  1.3069112  0.2137389 1.3709584 -0.56469817
#> 2  1    1 -0.2229673 -1.1548953 -1.1297343 0.3631284  0.63286260
#> 3  1    2 -1.1754638  0.8212813 -1.4218078 0.4042683 -0.10612452
#> 4  1    3 -0.4145136  1.8191834  0.8619176 1.5115220 -0.09465904
#> 5  1    4 -0.3116308  0.8777874  2.3952392 2.0184237 -0.06271410
#> 6  1    5 -0.0741827  0.1952049  2.2593628 1.3048697  2.28664539
head(as.data.frame(ssm, long = FALSE))
#>   id        y1_0        y2_0       y3_0         x1_0       x2_0        y1_1
#> 1  1 -1.32129502  1.30691121  0.2137389  1.370958447 -0.5646982 -0.22296728
#> 2  2  0.06803976 -0.04990546  0.5051611  1.200965376  1.0447511 -0.27313571
#> 3  3 -1.29748492 -2.91033474 -2.1032602 -2.000929238  0.3337772  0.05172567
#> 4  4 -0.25584749 -0.25219544  1.3222454 -0.004620768  0.7602422 -0.96322916
#> 5  5  1.36620158  1.35333546 -0.6372020  1.334912585 -0.8692718  0.78256512
#>          y2_1        y3_1        x1_1       x2_1        y1_2       y2_2
#> 1 -1.15489529 -1.12973434  0.36312841 0.63286260 -1.17546383 0.82128133
#> 2  0.23540119 -0.05306924 -1.00320865 1.84848190 -1.09640628 0.66881405
#> 3  0.19611625 -0.90129537  1.17132513 2.05953924  1.12632688 0.08719612
#> 4 -0.02352759 -0.19357488  0.03899091 0.73507214  0.07984138 0.52840906
#> 5 -0.94363718  0.41208380  0.05548695 0.04906691 -1.21504746 0.43538045
#>         y3_2       x1_2        x2_2       y1_3      y2_3       y3_3        x1_3
#> 1 -1.4218078  0.4042683 -0.10612452 -0.4145136 1.8191834  0.8619176  1.51152200
#> 2 -1.3358199 -0.6667734  0.10551381 -0.7999442 0.7045620 -0.4957978 -0.42225588
#> 3 -1.1610320 -1.3768616 -1.15085557  2.2700874 2.7845463 -0.5357065 -0.70582139
#> 4  1.8357221 -0.1464726 -0.05788734  0.4556301 1.8346636  0.7831771  0.48236947
#> 5 -0.8136944 -0.5783557 -0.99873866 -0.7081730 0.8893509  3.6320074 -0.00243278
#>          x2_3          y1_4       y2_4       y3_4       x1_4        x2_4
#> 1 -0.09465904 -0.3116308457  0.8777874  2.3952392  2.0184237 -0.06271410
#> 2 -0.12235017  0.0003275326 -1.1810319 -0.0725149  0.1881930  0.11916096
#> 3 -1.05405578  0.0523547052  1.5919880  0.7848450 -0.6457437 -0.18537797
#> 4  0.99294364  1.0515491779  2.4426596  0.7943206 -1.2463955 -0.03348752
#> 5  0.65551188 -2.6143792827  1.2718223  2.8713631  1.4768423 -1.90915279
#>         y1_5       y2_5      y3_5        x1_5       x2_5       y1_6       y2_6
#> 1 -0.0741827  0.1952049 2.2593628  1.30486965  2.2866454 -0.4368443  2.0929767
#> 2 -1.3228055  0.3247527 1.9619418 -0.02509255  0.1080727 -0.4160145  0.0470476
#> 3 -0.4094134  0.6089027 0.2004189 -1.20122205  2.0369722 -3.1257716  0.9779773
#> 4  2.2540317 -0.3995432 1.6048742 -0.07096218 -0.7589207  1.3679994 -2.4941578
#> 5 -1.5389980  1.0286935 1.2371762 -0.70243947 -0.3114302  0.5103027  0.0342447
#>         y3_6       x1_6       x2_6       y1_7      y2_7       y3_7       x1_7
#> 1 -0.1627264 -1.3888607 -0.2787888 -0.9313803  2.423123  1.8765506 -0.1333213
#> 2  1.1472796 -0.4854352 -0.5042171 -1.4850458 -1.194179  0.1440756 -1.6610991
#> 3 -0.1317061  0.1077747 -0.0841081  0.6404262  2.360186  0.8461550  0.4956196
#> 4 -0.7912684 -1.0343594 -0.6307320 -0.6828575 -1.932106 -1.2325272  0.5868077
#> 5  2.1412550 -1.6631570 -0.7505334 -1.5170361 -1.277911  0.2377416 -0.7773518
#>          x2_7       y1_8       y2_8       y3_8       x1_8       x2_8
#> 1  0.63595040  1.3555973  0.2776942  0.3336033 -0.2842529 -2.6564554
#> 2 -0.38233373  0.1689121 -0.8308893  2.0655163 -0.5126503  2.7018910
#> 3  0.03741519 -0.6297808  2.3492915  0.8100695 -0.1320880  1.4767874
#> 4 -0.41632266  1.0278824 -3.0135325 -0.0690037 -0.7848878  0.1634163
#> 5 -0.72256970  0.4862914 -1.8894371 -0.2718721 -2.1888346  0.2134186
#>          y1_9       y2_9       y3_9       x1_9       x2_9      y1_10     y2_10
#> 1  0.44476374  2.0663970 -0.3115522 -2.4404669  1.3201133  1.0940299  1.551743
#> 2 -0.96189170 -0.4509793  1.9482290 -1.3621162  0.1372562 -1.8715290  1.768182
#> 3  0.80833110  0.2288769 -0.3313643 -0.2170302 -1.2836022 -1.3556844  2.128277
#> 4 -0.03444984 -0.7515914  0.7226250 -1.2367142  1.0458738  0.6706872 -1.726603
#> 5 -1.76555996 -2.2453781 -0.7889341 -0.6319229  1.5204912 -1.7029262 -1.546746
#>       y3_10      x1_10      x2_10      y1_11     y2_11      y3_11       x1_11
#> 1 -3.457191 -0.3066386 -1.7813084  1.6080070  1.341208 -0.7619126 -0.17191736
#> 2  1.680022 -1.4936251 -1.4704357  0.9153270  1.200346  1.3312963  0.12470239
#> 3  3.247484  0.3856679 -0.3515129 -0.7769676  1.198546  3.6750765 -0.52179609
#> 4  2.009070 -0.4845954  0.1891288  0.1231655 -1.289442  2.6975231  0.05100633
#> 5 -1.579077  0.7959559 -1.4535296 -1.9696088 -4.089331 -1.2921864  0.09839542
#>           x2_11     y1_12      y2_12       y3_12        x1_12       x2_12
#> 1  1.2146746992 1.9193308  2.0938548  2.17045341  1.895193461 -0.43046913
#> 2 -0.9966391349 1.6719831  1.5324441  0.69583843 -0.001822614 -0.42825888
#> 3 -1.0681312007 0.1247800  1.7248144  1.68928420  0.428365903 -0.17401823
#> 4 -0.0002406689 1.9714391 -2.2886164 -0.02502037  1.809382042 -0.82532796
#> 5 -0.5937709842 0.9537555  0.1167135 -0.55523419  0.888281169  0.05307042
#>       y1_13      y2_13       y3_13      x1_13       x2_13      y1_14      y2_14
#> 1  1.982597  1.0425821  0.03182284 -0.2572694 -1.76316309  1.1224173  0.9823196
#> 2 -1.064327  2.4528216 -0.86275293 -0.6136716 -2.02467785  0.3482144  1.9676994
#> 3 -3.914508 -0.3014104  0.56301961  0.5156677 -0.23436528 -2.6245766 -0.8698393
#> 4  1.957702 -0.7604525  0.34113488  1.1454705  0.03157319 -0.3332116 -0.9814938
#> 5 -1.773687 -1.1221734 -1.32431578 -0.5570236  0.43839704 -0.4898259 -1.7693607
#>        y3_14      x1_14       x2_14      y1_15      y2_15       y3_15
#> 1  0.6673541  0.4600974 -0.63999488 -0.6952325  1.0229887  1.66293082
#> 2 -0.4420744 -1.2247480  0.17951644  0.1102308  1.6923343 -0.07650175
#> 3 -1.7154935 -0.6585034  1.25023660 -2.2388038 -2.0003179 -0.47657878
#> 4  1.7911445 -0.8352058 -0.06876365  1.1053596 -0.2092331  0.93481928
#> 5  0.1158383  0.1526082 -0.16461758  0.6003556 -1.4797100 -2.04300966
#>        x1_15      x2_15       y1_16      y2_16      y3_16         x1_16
#> 1  0.4554501  0.7048373 -0.07084522 -0.5595938  0.2004234  1.035104e+00
#> 2  0.5676206 -0.4928774  0.59248600 -0.4661674 -0.1810490  6.288407e-05
#> 3 -0.2717637  0.9479520 -2.85135034 -0.4035567 -1.0487320 -1.201582e+00
#> 4  0.7467717 -0.4255187  1.99492931 -2.6126058  1.1768557 -7.720822e-01
#> 5  2.0198906 -0.5293859  1.16386902 -1.6207066  0.0318715 -4.707870e-01
#>        x2_16       y1_17       y2_17      y3_17       x1_17       x2_17
#> 1 -0.6089264 -1.43406576 -1.52341239  0.3344811  0.50495512 -1.71700868
#> 2  1.1228896  0.32089263  1.44632469 -0.8328417  1.43985574 -1.09711377
#> 3 -0.4661161  2.24942801  0.09748498 -1.2315448 -0.26935140 -0.39096541
#> 4  0.1527641  1.84507091 -1.04563662  1.9960705  0.98859685 -0.07345833
#> 5 -1.5459369  0.09640684 -0.78013019 -0.3272448 -0.04052672  0.89035631
#>        y1_18       y2_18      y3_18      x1_18      x2_18       y1_19
#> 1  0.2647052  0.37468039  2.6737476 -0.7844590 -0.8509076 -0.76282546
#> 2 -0.5119163 -0.03934419 -0.8031779 -0.1173196  1.2014984 -1.24289238
#> 3  2.0495729 -0.94087452 -1.5905452  1.3487070 -0.0227647 -1.23929468
#> 4  1.3493736 -0.03620265  1.3392054 -1.3870266 -1.3066759  0.01149444
#> 5  1.1458171  0.14204054 -2.5427818 -2.0713879 -0.2500651  0.27888846
#>         y2_19      y3_19      x1_19       x2_19       y1_20      y2_20
#> 1  0.06189121  2.8808639 -2.4142076  0.03612261 -0.71114553 -2.5089723
#> 2 -1.18786430  2.1575269 -0.4697296 -0.05246948  0.02807218  1.2439153
#> 3  0.19919403 -1.4852461  0.2442259 -0.94237171 -0.81764122  1.0367502
#> 4 -2.85293147  2.3386243 -0.7683953 -0.52710813  0.55518284  0.2245032
#> 5 -2.47708708  0.5582733 -1.1816504  1.44193727  0.09094230 -1.1840430
#>        y3_20       x1_20      x2_20      y1_21      y2_21      y3_21      x1_21
#> 1 -0.8309464  0.20599860 -0.3610573 -2.5260087 -1.3048137 -1.3981766  0.7581632
#> 2  1.4833495 -0.08610730 -0.8876790 -0.3199901  2.2064632  0.6489734 -0.4446840
#> 3 -0.8555348 -0.72921728  0.9980689  0.0996544 -0.4360866  0.6828378  1.2584817
#> 4  0.6364247 -0.02142707  0.6704981  1.2731180  0.4635977 -0.2848650 -0.4346170
#> 5  0.4087034  1.35789554  0.3345028  3.8716091  0.2527161  0.8965945  1.4293381
#>         x2_21      y1_22      y2_22       y3_22      x1_22      x2_22
#> 1 -0.72670483  0.3521247 -1.0971151  0.21790193 -1.3682810  0.4328180
#> 2 -0.02944488  0.6911039  1.1117158 -0.72756334 -0.4138688  1.1133860
#> 3  1.24886369 -1.9390432 -1.1699622  1.36893560 -1.3806370  2.0499607
#> 4 -1.11387978 -1.1151190  0.1135364  0.05691346  0.6071060  0.2754570
#> 5 -0.86731785  2.0364593 -1.9058403 -0.03871032  0.9506517 -0.5850115
#>        y1_23      y2_23      y3_23      x1_23       x2_23      y1_24
#> 1  1.3176657 -2.1642990 -0.6796242 -0.8113932  1.44410126  1.3391641
#> 2  1.4954691  0.7122494  0.3633060 -0.4809928 -0.43316903  1.0337595
#> 3 -1.7492765  1.3050955 -0.3266634  1.0168728 -0.02671746 -0.6841845
#> 4 -0.2646706  0.5000499  2.1818351  1.1573471 -1.68248086 -1.0163380
#> 5  1.2545266  0.3112625 -2.1339148  0.3209575 -0.29939602  0.2409194
#>         y2_24      y3_24       x1_24      x2_24      y1_25      y2_25
#> 1 -0.82130880 -0.0752571 -0.43144620  0.6556479  0.8216689 -1.8147205
#> 2 -0.06397518 -0.5531983  0.69686258 -1.0563684 -0.7541386 -2.1541924
#> 3  3.16868302  1.5155881  0.70360778 -0.9713852  0.1454184  0.3460727
#> 4 -0.42079016  3.0327701  0.08731909  1.3533619 -1.9542314  0.7789802
#> 5  0.18425062 -2.2777921 -0.27854308  0.5461152 -0.2883920 -1.8790973
#>        y3_25       x1_25       x2_25     y1_26      y2_26      y3_26      x1_26
#> 1  0.2288761  0.32192527 -0.78383894  1.479689  0.7835702 -0.4119303  1.5757275
#> 2  2.0231696 -0.04069848 -1.55154482 -1.204166 -0.4808715 -0.5025113  1.1671695
#> 3  0.2923802 -1.09615624  0.04905045  2.080418 -1.9194933  0.3897390 -1.1984959
#> 4  1.6673886  0.72417380 -0.83255283 -1.172532  1.1331902  2.3754289  0.7325285
#> 5 -1.3474831 -1.30382115 -0.25091447  1.368135 -2.2429527 -2.5437840  0.1710074
#>        x2_26       y1_27      y2_27     y3_27       x1_27      x2_27      y1_28
#> 1  0.6428993  1.91209020  1.5502746 -1.005937  0.08976065  0.2765507  2.2630026
#> 2 -0.2736457 -0.63264824  1.0368186 -1.620818 -0.46784532 -1.2382523  0.1246215
#> 3  0.1900190  0.51148147  0.0728170 -3.060578  1.29770590 -1.0338737  2.4021004
#> 4 -0.8719269 -0.37614524  2.4562279  3.266426 -0.45339751  1.1875343 -0.4664616
#> 5 -0.4034675  0.09799243 -0.4597594 -2.391649  0.10465944 -0.3188808  1.4656842
#>        y2_28      y3_28        x1_28       x2_28      y1_29      y2_29
#> 1  0.2672299 -1.9253959  0.679288816  0.08983289  0.5184740  1.3615360
#> 2 -0.5721696  0.6460846 -0.007762034 -0.80028218 -0.6565600 -0.8537611
#> 3 -0.5177981 -0.3705207 -0.738440754  0.04656394  0.5111742 -2.4207984
#> 4  1.4204688  3.5127039 -0.290145312  0.82854614  0.0876693 -0.3946798
#> 5 -0.6122942 -1.5410249  1.618343936  0.71418860  1.0471282 -0.7406853
#>        y3_29      x1_29      x2_29       y1_30       y2_30      y3_30
#> 1  1.4310673 -2.9930901  0.2848830  0.08042765 -0.53075827 -0.3826375
#> 2  0.8903193 -0.5334923  1.2876752 -0.21823923 -2.74552124  1.6616192
#> 3  1.4869472 -1.0175961 -0.3832840  1.01685499 -1.50281241  1.4140044
#> 4  2.4726918 -0.2912277 -1.5763624 -1.11657467  0.62752893  0.6372705
#> 5 -2.5462526  2.9658654 -0.7950776 -0.04829977 -0.05633085 -2.6124751
#>        x1_30      x2_30      y1_31      y2_31      y3_31      x1_31      x2_31
#> 1 -0.3672346  0.1852306 -0.7575300  1.0660180 -1.1937923  0.5818237  1.3997368
#> 2 -0.1755259 -1.0717824 -1.5893939 -0.9641858  1.4969122  0.1632069 -0.3627384
#> 3  0.8727554  0.9695450  2.8954413 -0.8172052  1.0713061  0.3838467 -1.8515557
#> 4 -0.8488157 -1.0885199 -0.5761493  0.6555097 -0.8429513 -0.4842906 -0.3363112
#> 5  0.8143659  2.0980308  2.2346414 -1.3267072  0.2265222  0.3009801 -1.0830751
#>        y1_32      y2_32      y3_32       x1_32       x2_32      y1_33     y2_33
#> 1 -0.8657861  1.1594111  1.5843125 -0.72729206  1.30254263 -1.7608815 0.7269482
#> 2 -0.3281431 -0.2782823 -2.7009377  0.59001355  1.43242193 -0.2476019 1.0308517
#> 3  1.9802173 -1.1335253 -1.4082145 -0.05399674  1.06477321 -2.0353089 1.5750707
#> 4  0.6702878  1.6342009  0.1738516 -0.15335789 -0.24324723  0.3690930 0.8654226
#> 5  0.6742349 -1.6456944  0.7162619 -1.00632250 -0.03541456 -1.2679077 0.5067678
#>         y3_33      x1_33      x2_33      y1_34      y2_34      y3_34
#> 1  1.89270783  0.3358481  1.0385061 -0.8768112  1.1036715  2.2148850
#> 2 -0.84104312 -0.9926925  0.4546503  1.8621741 -0.4604311 -0.1387278
#> 3 -0.03280635  0.8131950 -0.1908165 -0.4522141  0.2369424 -0.1114640
#> 4 -0.33863586  1.8922020 -1.3859983  0.6567835  1.9270995  0.6349295
#> 5  2.24073481  1.3091244  0.7504005 -2.1886602 -1.7674955  1.4045945
#>         x1_34       x2_34      y1_35      y2_35      y3_35        x1_35
#> 1  0.92072857  0.72087816  0.3397298  0.8155806 -0.3256500 -1.043118939
#> 2  0.08489806  0.89556558  0.7237191  0.2834713  0.5157692 -0.229778139
#> 3 -2.69992981  0.06096664 -0.3752383 -0.7829329 -0.8206978  0.573751697
#> 4 -0.41482430  0.34908153 -0.2851125  0.6157689 -3.8247503  1.628442266
#> 5 -2.13836833 -0.70035411 -1.8370658  1.0129193 -1.8720094 -0.009056475
#>         x2_35      y1_36      y2_36      y3_36      x1_36      x2_36      y1_37
#> 1 -0.09018639  2.0467294  0.1813455  1.6271317  0.6235182 -0.9535234  1.7177961
#> 2  0.83661907 -1.6404856  2.9320693  0.4389722 -1.7450559  1.6894589  1.1006471
#> 3  0.04580358  0.1458506  0.2831584 -0.0647262  0.1574125  0.4315654  2.8766446
#> 4  0.08852190  0.1730077  0.9797463  0.5413674  1.2391507 -1.6445555 -0.3845699
#> 5 -1.45813349 -1.8592593 -1.4851963 -2.2134587  0.6945296 -2.4613355 -1.1164536
#>         y2_37      y3_37      x1_37      x2_37      y1_38     y2_38     y3_38
#> 1  1.93999530  2.3291989 -0.5428288  0.5809965  0.5806455 0.5890687 0.9236375
#> 2  0.91649600 -0.5028793  0.8647780 -0.1507760 -1.5437966 0.7768215 0.0760361
#> 3  0.02653138  0.5382699 -0.3965497  1.3099782  0.3231083 1.0383649 0.2899639
#> 4 -0.40591425 -1.1825058  1.4463565 -0.6905602 -2.5216510 1.4791234 0.8107824
#> 5 -1.15210573 -0.6631066  0.1432898 -0.3912221  0.6531909 0.1385496 0.4259709
#>        x1_38      x2_38      y1_39      y2_39      y3_39      x1_39
#> 1  0.7681787  0.4637676  2.0052886 -0.3835412  2.5304392 -0.8857763
#> 2 -1.4490071  0.6430087  0.1418797  0.7309179 -1.0435802  0.4831939
#> 3  0.4703934 -1.2426703  1.3222493  0.1552634  1.0564194  1.3815755
#> 4 -0.2764311 -1.1094188 -2.8024327  1.1789045  0.5664178  0.1338693
#> 5 -0.4911641 -0.2836475 -2.9328006  1.4429639  0.5536245  0.3147948
#>          x2_39      y1_40       y2_40      y3_40      x1_40      x2_40
#> 1 -1.099780899 -0.2652305  1.50027725  0.9250186  1.5127070  0.2579214
#> 2 -0.006355626  0.5980214 -0.76641709  0.2154274  0.1514559 -0.5841090
#> 3  1.204458937  2.1905223  0.05355728 -0.7026008  0.8240740 -1.6626294
#> 4  1.785339052 -0.3865157  0.97305800  1.2350377  2.4221634 -1.0768289
#> 5  0.396326578  0.1574557 -0.98809480  0.8396418 -0.2256037 -1.9249504
#>        y1_41      y2_41       y3_41       x1_41      x2_41      y1_42
#> 1  0.1309740  0.8219069 -0.01472055  0.08844023 -0.1208965 -0.9430371
#> 2  1.2234728  2.1725154 -1.85474634  0.36880673  0.2946543  0.8238318
#> 3 -0.2257734 -0.4322122 -0.30076709 -0.56930634  0.6355138 -0.5528709
#> 4 -0.2699938  0.1163131  0.17479170  0.48594111  1.3885217 -0.7994950
#> 5 -2.4896525 -0.9838587 -0.19082061 -1.43922930 -1.4696578 -2.4903470
#>         y2_42      y3_42       x1_42      x2_42       y1_43      y2_43
#> 1 -0.96621247 -0.7748278 -1.19432890  0.6119969 -0.10121621  1.8060507
#> 2  0.65315701  0.8901553 -0.27925937 -1.3362367  0.61060692 -0.3885625
#> 3  0.40038204  1.3984393  0.04372201  0.3480123  0.45115331 -2.0828687
#> 4 -0.07212676 -0.3533231 -0.19565682 -0.2181748 -0.04885723 -1.0599949
#> 5 -0.71147283 -1.1043962  0.76186345 -0.2436150 -0.16237658 -0.2331553
#>        y3_43      x1_43      x2_43     y1_44     y2_44       y3_44      x1_44
#> 1  0.4460444 -0.2171398 -0.1827567 -1.457261 -0.702341 -0.02497946  0.9333463
#> 2 -1.0955178  0.7007488  0.5541966 -2.175534 -1.107481 -1.26217429 -0.8363066
#> 3  1.6863428  2.4595935 -0.8183803  1.952765 -2.492334  0.00432441 -2.1132001
#> 4  0.6701214 -0.3047780  0.5978327  2.512454 -2.018226  0.16243943  1.3974294
#> 5 -0.6697140  0.2696766 -1.5589275 -3.896427  2.019988 -0.04529128 -0.5355880
#>        x2_44      y1_45      y2_45      y3_45      x1_45      x2_45      y1_46
#> 1  0.8217731 -0.9831391  0.9994024  3.1503015  1.3921164 -0.4761739 -0.5168949
#> 2 -1.5945882  0.7340985 -0.7533449 -2.1924199  0.2049586 -0.3450880  1.5346657
#> 3  0.2736953  2.2299542  0.4874205  0.4514499 -0.6875968  0.4460411  1.4045247
#> 4  0.6876198  2.1679166 -1.5166678 -1.3601942  0.3201880 -0.3018699  3.2596240
#> 5  0.5624520  0.6531392  1.3666939  1.7850713 -0.1783261 -0.1151360  0.4616186
#>         y2_46       y3_46       x1_46      x2_46      y1_47       y2_47
#> 1  1.24877752 -0.49639914  0.65034856  1.3911105 -3.6370966  0.02088656
#> 2 -0.94852167 -0.02162973  0.25261170 -1.2940025  1.0756810  1.03974512
#> 3  1.45055467  0.35575850 -0.81238472  2.2120555 -0.0574027  0.91049398
#> 4  0.12775727 -0.38482344  0.49834869 -0.5495369  1.8684240 -0.15782705
#> 5 -0.06109306  1.49734976 -0.07206147  1.2109098  0.7463014 -1.45892392
#>        y3_47      x1_47      x2_47      y1_48       y2_48     y3_48      x1_48
#> 1  0.2411616 -1.1107889 -0.8607926  0.2395879 -0.06391602  1.908043 -1.1317387
#> 2 -0.5862474 -0.9591704  1.0857749  3.0474133  1.33874980  0.559258  0.4037749
#> 3  0.8101820 -0.1237060 -0.4773355  1.5218235 -0.50759186  1.018314 -0.1662615
#> 4 -1.6212787 -0.2792565  1.0965134  2.4411195  0.78024158 -2.333490  0.4420131
#> 5  0.3547960 -0.6148969  0.6761265 -0.6336531 -0.83164252 -2.093679  0.8985996
#>        x2_48      y1_49      y2_49       y3_49       x1_49       x2_49
#> 1 -1.4592140  0.3583062  1.3518190  1.55417479  0.07998255  0.65320434
#> 2  0.5864875  2.4212966 -0.3200761 -0.06399690  1.81522845  0.12882143
#> 3  0.8625634 -0.1229318 -3.3831398  0.51101047  0.09734049 -1.62561674
#> 4  0.2410163  0.9136470 -0.8633060 -0.22048714 -0.25560766  0.93103290
#> 5 -1.1893179 -0.1293048  0.4391199  0.06054169  0.12125885 -0.01122169

# Type 2
ssm <- SimSSMFixed(
  n = n,
  time = time,
  mu0 = mu0,
  sigma0_l = sigma0_l,
  alpha = alpha,
  beta = beta,
  psi_l = psi_l,
  nu = nu,
  lambda = lambda,
  theta_l = theta_l,
  type = 2,
  x = x,
  gamma = gamma,
  kappa = kappa
)

head(as.data.frame(ssm))
#>   id time         y1         y2        y3        x1          x2
#> 1  1    0 -2.2617108 -0.7667176 0.5340921 1.3709584 -0.56469817
#> 2  1    1  0.9759577  0.5206830 1.1328299 0.3631284  0.63286260
#> 3  1    2  1.4864513  1.5139406 0.7636724 0.4042683 -0.10612452
#> 4  1    3  0.6895916 -1.4110927 0.5326012 1.5115220 -0.09465904
#> 5  1    4 -2.0426250 -1.6092847 2.9555453 2.0184237 -0.06271410
#> 6  1    5  0.1554637 -0.2390945 0.7660269 1.3048697  2.28664539
head(as.data.frame(ssm, long = FALSE))
#>   id        y1_0       y2_0       y3_0         x1_0       x2_0        y1_1
#> 1  1 -2.26171084 -0.7667176  0.5340921  1.370958447 -0.5646982  0.97595771
#> 2  2 -0.01083935  1.2735290  2.1028202  1.200965376  1.0447511 -1.22065483
#> 3  3 -1.17628203  1.5504647 -0.4657054 -2.000929238  0.3337772 -0.16068750
#> 4  4 -1.09443990  0.7378863 -1.5560017 -0.004620768  0.7602422 -0.33062582
#> 5  5  2.91767302  0.2292802  1.1298055  1.334912585 -0.8692718  0.06649816
#>         y2_1       y3_1        x1_1       x2_1       y1_2       y2_2
#> 1  0.5206830  1.1328299  0.36312841 0.63286260  1.4864513  1.5139406
#> 2  2.3612218  1.1376440 -1.00320865 1.84848190 -0.4653553  0.5947964
#> 3  0.7164697 -0.5730460  1.17132513 2.05953924 -1.6635127  0.2287985
#> 4  0.5212429  0.9009251  0.03899091 0.73507214 -0.7938779 -0.3201251
#> 5 -1.0677271  2.2473308  0.05548695 0.04906691 -0.3308651 -0.5287427
#>          y3_2       x1_2        x2_2       y1_3       y2_3       y3_3
#> 1  0.76367243  0.4042683 -0.10612452  0.6895916 -1.4110927  0.5326012
#> 2 -0.44834655 -0.6667734  0.10551381 -0.7405792 -0.5751752  1.2894237
#> 3  0.10846422 -1.3768616 -1.15085557 -0.3987294 -2.2224797 -0.2573785
#> 4  0.09072877 -0.1464726 -0.05788734  1.2413697  2.0108480  1.0675637
#> 5 -0.39058089 -0.5783557 -0.99873866  0.4627781  0.7531802 -1.4845421
#>          x1_3        x2_3        y1_4       y2_4        y3_4       x1_4
#> 1  1.51152200 -0.09465904 -2.04262498 -1.6092847  2.95554532  2.0184237
#> 2 -0.42225588 -0.12235017  0.08938988 -2.1971766 -0.29360369  0.1881930
#> 3 -0.70582139 -1.05405578 -1.04567361 -2.2322137  0.94546923 -0.6457437
#> 4  0.48236947  0.99294364 -0.41209094 -0.6080460  2.11912321 -1.2463955
#> 5 -0.00243278  0.65551188 -0.16190873  0.9545628  0.09236095  1.4768423
#>          x2_4       y1_5       y2_5       y3_5        x1_5       x2_5
#> 1 -0.06271410  0.1554637 -0.2390945  0.7660269  1.30486965  2.2866454
#> 2  0.11916096 -0.3856124  1.0283889  0.4224272 -0.02509255  0.1080727
#> 3 -0.18537797 -1.1221677 -1.7405010  1.4310039 -1.20122205  2.0369722
#> 4 -0.03348752  1.6593916 -2.2213285  0.5968630 -0.07096218 -0.7589207
#> 5 -1.90915279 -1.1067060 -0.3530877 -0.1913458 -0.70243947 -0.3114302
#>         y1_6        y2_6      y3_6       x1_6       x2_6       y1_7       y2_7
#> 1 -0.3577100 -1.11826529 0.3574445 -1.3888607 -0.2787888 -0.9035040  1.6431612
#> 2 -0.2227387 -0.44385541 0.2499583 -0.4854352 -0.5042171  0.8001131 -0.3852358
#> 3 -0.7254209  0.34005317 1.5168391  0.1077747 -0.0841081 -1.8839166 -1.3832898
#> 4 -2.7752740 -1.45383703 0.8223615 -1.0343594 -0.6307320 -3.4695359 -0.2338029
#> 5  0.1157016  0.01079781 1.8961065 -1.6631570 -0.7505334  2.1013598 -0.3598088
#>         y3_7       x1_7        x2_7       y1_8       y2_8       y3_8       x1_8
#> 1 -0.5654410 -0.1333213  0.63595040 -0.3055039  0.5140589  1.3956298 -0.2842529
#> 2 -0.6260134 -1.6610991 -0.38233373  0.6097453  0.3018880 -0.2750748 -0.5126503
#> 3  0.8759856  0.4956196  0.03741519 -0.9261098  0.8598035  0.1280309 -0.1320880
#> 4 -2.2144320  0.5868077 -0.41632266 -3.0403706 -1.7676560 -0.5040090 -0.7848878
#> 5  0.8637659 -0.7773518 -0.72256970 -0.6761900 -0.3791144  1.4061523 -2.1888346
#>         x2_8       y1_9       y2_9       y3_9       x1_9       x2_9       y1_10
#> 1 -2.6564554  0.7170630  0.9319849  1.7523587 -2.4404669  1.3201133  0.61528870
#> 2  2.7018910  1.0110175  1.1011928 -2.1915604 -1.3621162  0.1372562  1.84601432
#> 3  1.4767874  0.1551347 -0.3676341  0.3566809 -0.2170302 -1.2836022 -0.85339502
#> 4  0.1634163 -1.6478691 -1.7046013 -0.8163977 -1.2367142  1.0458738  0.07623835
#> 5  0.2134186 -2.1447521  0.9616863  1.7582653 -0.6319229  1.5204912  0.44929767
#>         y2_10      y3_10      x1_10      x2_10      y1_11      y2_11      y3_11
#> 1  0.06595301  0.3283583 -0.3066386 -1.7813084 -0.6584546  1.1780206  2.2128578
#> 2  1.07596242 -1.1666866 -1.4936251 -1.4704357  1.3268751  1.4581645  1.6710793
#> 3 -0.02230328  1.9582845  0.3856679 -0.3515129 -2.3708827 -1.0934615 -2.3682302
#> 4 -2.27620204  0.2104809 -0.4845954  0.1891288 -1.4402695 -0.3698054 -1.1972837
#> 5  0.23189896  0.1892874  0.7959559 -1.4535296  0.5914779  2.3065489 -0.6066982
#>         x1_11         x2_11      y1_12      y2_12      y3_12        x1_12
#> 1 -0.17191736  1.2146746992 -1.6574764  1.8291666  0.2199046  1.895193461
#> 2  0.12470239 -0.9966391349  1.1932733 -0.1249057 -0.6931150 -0.001822614
#> 3 -0.52179609 -1.0681312007 -0.3379368 -0.1078337  1.3526176  0.428365903
#> 4  0.05100633 -0.0002406689  0.1015011 -0.7048022 -1.1787278  1.809382042
#> 5  0.09839542 -0.5937709842 -1.2065887  0.2941488 -1.7351641  0.888281169
#>         x2_12      y1_13       y2_13      y3_13      x1_13       x2_13
#> 1 -0.43046913 -2.2151404 -0.42539203 -1.8083666 -0.2572694 -1.76316309
#> 2 -0.42825888 -0.1552379 -0.06603291  0.7551181 -0.6136716 -2.02467785
#> 3 -0.17401823  1.1126262 -1.11051383 -0.3005570  0.5156677 -0.23436528
#> 4 -0.82532796  2.0640866 -1.93662451  1.5293285  1.1454705  0.03157319
#> 5  0.05307042 -0.4365304 -1.07341658  0.2023655 -0.5570236  0.43839704
#>         y1_14      y2_14      y3_14      x1_14       x2_14      y1_15
#> 1 -1.42026784 -0.5793268 -0.6061088  0.4600974 -0.63999488 -0.9428814
#> 2  1.33106571  0.2816104 -2.1650570 -1.2247480  0.17951644  1.0131716
#> 3 -1.18306966 -2.7370652 -0.1639112 -0.6585034  1.25023660 -0.2079178
#> 4  0.08559636  1.0091553  2.6040149 -0.8352058 -0.06876365 -0.3314998
#> 5  0.08712358 -1.4457050  0.3418526  0.1526082 -0.16461758 -0.7151986
#>         y2_15      y3_15      x1_15      x2_15      y1_16      y2_16
#> 1  0.59601820  2.1629074  0.4554501  0.7048373 -0.3427984  0.7387568
#> 2  0.03519945 -1.1499958  0.5676206 -0.4928774  0.6484758  0.7295049
#> 3 -0.26248612 -0.1580101 -0.2717637  0.9479520 -2.2845259  2.6062737
#> 4  0.24641057  3.5398440  0.7467717 -0.4255187  0.3390863 -0.4424833
#> 5  0.08833419  1.0854586  2.0198906 -0.5293859 -0.3580190 -0.8553143
#>         y3_16         x1_16      x2_16      y1_17      y2_17      y3_17
#> 1  1.10667527  1.035104e+00 -0.6089264  0.7831213  2.5436538 -1.9466145
#> 2 -0.91152816  6.288407e-05  1.1228896 -1.3577298 -1.2937834  0.9420950
#> 3 -0.16398670 -1.201582e+00 -0.4661161 -0.6930692  1.6987587 -1.3298241
#> 4  3.00894938 -7.720822e-01  0.1527641  2.0548933 -0.4498286 -1.3480820
#> 5 -0.06110253 -4.707870e-01 -1.5459369 -0.7032934 -0.3767284  0.4782274
#>         x1_17       x2_17      y1_18        y2_18      y3_18      x1_18
#> 1  0.50495512 -1.71700868 -0.1979541  0.934651116 -2.4605171 -0.7844590
#> 2  1.43985574 -1.09711377  1.0538112  0.482245291 -0.2161476 -0.1173196
#> 3 -0.26935140 -0.39096541 -0.1742509  0.002553713 -0.5221625  1.3487070
#> 4  0.98859685 -0.07345833 -0.2050521 -0.620461603 -0.2960910 -1.3870266
#> 5 -0.04052672  0.89035631  0.5126820 -0.990310060 -1.6786754 -2.0713879
#>        x2_18       y1_19      y2_19       y3_19      x1_19       x2_19
#> 1 -0.8509076  0.09501072 0.16565982 -2.67127807 -2.4142076  0.03612261
#> 2  1.2014984  0.55752311 1.63942582 -0.06732495 -0.4697296 -0.05246948
#> 3 -0.0227647 -2.30534138 1.66243142 -1.05916120  0.2442259 -0.94237171
#> 4 -1.3066759 -1.54067494 0.01575045  0.76033397 -0.7683953 -0.52710813
#> 5 -0.2500651 -1.24765403 0.38733170  0.73859327 -1.1816504  1.44193727
#>         y1_20      y2_20     y3_20       x1_20      x2_20      y1_21      y2_21
#> 1  0.06524747  0.1468387 -1.078771  0.20599860 -0.3610573 -1.2301772  0.5091490
#> 2  0.57675232 -0.2908006  1.491129 -0.08610730 -0.8876790 -0.2472754 -0.4192988
#> 3 -3.28260298  1.3880891 -1.116361 -0.72921728  0.9980689 -0.7938003 -0.1749302
#> 4 -0.62321824  0.9650449  2.117989 -0.02142707  0.6704981 -1.2653429  2.3447854
#> 5 -1.96697562  0.1014138 -1.252068  1.35789554  0.3345028  0.0408507 -2.1493766
#>         y3_21      x1_21       x2_21      y1_22      y2_22     y3_22      x1_22
#> 1  0.02231443  0.7581632 -0.72670483 -3.0469130  0.3743534 1.0232007 -1.3682810
#> 2 -1.41122249 -0.4446840 -0.02944488 -0.5645269 -0.9572478 0.7276238 -0.4138688
#> 3  1.51059438  1.2584817  1.24886369 -1.5603645 -0.1838495 0.7381851 -1.3806370
#> 4 -1.05513299 -0.4346170 -1.11387978 -0.4357499  1.1624015 0.2351577  0.6071060
#> 5  0.93601973  1.4293381 -0.86731785 -0.7831636 -1.6882271 1.8769630  0.9506517
#>        x2_22      y1_23      y2_23      y3_23      x1_23       x2_23      y1_24
#> 1  0.4328180 -1.8262502 -0.7376244 -0.3671736 -0.8113932  1.44410126 -0.7958582
#> 2  1.1133860 -0.1703043  0.2797799  1.5287615 -0.4809928 -0.43316903 -1.5288576
#> 3  2.0499607 -0.9550793  0.2786485 -1.3416689  1.0168728 -0.02671746 -0.4028647
#> 4  0.2754570 -0.3413122 -0.4190110 -1.2849335  1.1573471 -1.68248086 -0.7521130
#> 5 -0.5850115 -0.1491781 -0.7253244 -1.3660979  0.3209575 -0.29939602 -0.3675176
#>         y2_24       y3_24       x1_24      x2_24      y1_25      y2_25
#> 1  1.22671231 -0.46755175 -0.43144620  0.6556479  0.4792682 -0.6222010
#> 2 -1.18580336 -0.09326903  0.69686258 -1.0563684 -0.8426745  2.6837173
#> 3 -1.36256873  1.44578030  0.70360778 -0.9713852  1.5342963 -0.4870606
#> 4  1.65155850  0.37153081  0.08731909  1.3533619  1.6027442 -1.0151345
#> 5 -0.02479883 -0.90361727 -0.27854308  0.5461152 -0.4166378  2.3206512
#>        y3_25       x1_25       x2_25       y1_26      y2_26      y3_26
#> 1 -0.8141507  0.32192527 -0.78383894  1.49478044 -0.7011373  0.9654151
#> 2  0.3922965 -0.04069848 -1.55154482  0.06169944  1.4954911 -2.1184110
#> 3 -0.5943964 -1.09615624  0.04905045  0.72900810 -2.7314058  0.3839384
#> 4  1.7426810  0.72417380 -0.83255283 -0.06770521 -0.3768668 -0.2867897
#> 5 -0.1405577 -1.30382115 -0.25091447  0.10216523  0.5191222  0.3139478
#>        x1_26      x2_26      y1_27      y2_27      y3_27       x1_27      x2_27
#> 1  1.5757275  0.6428993 -1.1647060 -2.2184199  3.4363355  0.08976065  0.2765507
#> 2  1.1671695 -0.2736457  2.6494635  0.1662511  0.1836773 -0.46784532 -1.2382523
#> 3 -1.1984959  0.1900190 -0.4266472 -1.1554926 -1.1296515  1.29770590 -1.0338737
#> 4  0.7325285 -0.8719269  0.7947885 -2.3842952 -2.1643112 -0.45339751  1.1875343
#> 5  0.1710074 -0.4034675  1.9508978 -0.3848814 -3.0563611  0.10465944 -0.3188808
#>        y1_28      y2_28      y3_28        x1_28       x2_28      y1_29
#> 1 -1.1232442 -0.3386327  2.4856565  0.679288816  0.08983289 -0.2116782
#> 2  0.4372636  1.8478779 -1.3167820 -0.007762034 -0.80028218  0.6312305
#> 3 -0.9236442 -3.0363753  1.1332017 -0.738440754  0.04656394 -0.8130017
#> 4  1.6876901 -1.3855848 -1.1868768 -0.290145312  0.82854614  0.4826578
#> 5  1.5762711  2.9005160 -0.6647209  1.618343936  0.71418860  3.3908647
#>        y2_29      y3_29      x1_29      x2_29      y1_30      y2_30       y3_30
#> 1 -0.2639721  0.7380767 -2.9930901  0.2848830 -1.2035309 -0.3571279 -1.64509425
#> 2  2.1808281 -0.6859749 -0.5334923  1.2876752  2.0804276  0.2775435  0.06526954
#> 3 -1.9517001 -0.8381478 -1.0175961 -0.3832840 -1.0822334 -0.9186276  1.39868427
#> 4 -3.0943325 -0.9111990 -0.2912277 -1.5763624 -1.7304351 -2.5725873 -0.55611115
#> 5  2.3261089 -1.4026520  2.9658654 -0.7950776 -0.2029005  1.1005079 -1.85546589
#>        x1_30      x2_30       y1_31       y2_31       y3_31      x1_31
#> 1 -0.3672346  0.1852306  0.07109925 -2.24613575  1.99010583  0.5818237
#> 2 -0.1755259 -1.0717824  0.55629458  0.69940331 -0.97774634  0.1632069
#> 3  0.8727554  0.9695450 -0.46278148 -1.93447356  0.98093559  0.3838467
#> 4 -0.8488157 -1.0885199  1.36947093 -0.06717144 -0.09082801 -0.4842906
#> 5  0.8143659  2.0980308 -0.01129053  0.20335470 -3.45852622  0.3009801
#>        x2_31      y1_32       y2_32       y3_32       x1_32       x2_32
#> 1  1.3997368  3.7186580 -0.05208359 -0.56197007 -0.72729206  1.30254263
#> 2 -0.3627384 -0.5323864  2.47158612 -0.97034857  0.59001355  1.43242193
#> 3 -1.8515557 -0.9562222  1.44806587 -2.37649172 -0.05399674  1.06477321
#> 4 -0.3363112 -1.0024854 -0.90709147  0.06740686 -0.15335789 -0.24324723
#> 5 -1.0830751 -0.3644994 -1.24486220 -3.45887740 -1.00632250 -0.03541456
#>        y1_33      y2_33       y3_33      x1_33      x2_33      y1_34      y2_34
#> 1 -0.1160670  0.3101314  1.53358498  0.3358481  1.0385061 -0.4771736  1.9733541
#> 2  0.1011716 -0.1749360 -0.26398660 -0.9926925  0.4546503  3.5078962 -0.6752161
#> 3 -2.4240088  1.2592813 -0.08246484  0.8131950 -0.1908165  0.7883838  1.6639357
#> 4  1.2153747  0.1826516 -0.99893422  1.8922020 -1.3859983  0.4488331  1.4923626
#> 5  0.7289556 -1.5118393 -0.91846250  1.3091244  0.7504005 -0.3396360 -2.6796190
#>         y3_34       x1_34       x2_34      y1_35      y2_35      y3_35
#> 1 -0.06159458  0.92072857  0.72087816  0.1334822  0.6098394  0.9771480
#> 2 -0.09798790  0.08489806  0.89556558  0.7529208  0.2315859 -0.2440604
#> 3  1.78529327 -2.69992981  0.06096664  2.7238966 -0.4761221 -1.4859338
#> 4 -0.27997541 -0.41482430  0.34908153  1.2359666  0.7233830  1.4193523
#> 5  0.09732401 -2.13836833 -0.70035411 -1.3900866 -1.2572970  1.5649886
#>          x1_35       x2_35      y1_36      y2_36      y3_36      x1_36
#> 1 -1.043118939 -0.09018639 -2.1793162  0.7935969 -0.5925976  0.6235182
#> 2 -0.229778139  0.83661907  0.6036193  0.2024711  1.3342233 -1.7450559
#> 3  0.573751697  0.04580358  0.9637950 -0.8749543 -0.1194499  0.1574125
#> 4  1.628442266  0.08852190  1.6951209 -1.2556609  0.2163004  1.2391507
#> 5 -0.009056475 -1.45813349 -1.8757981 -0.6122068  1.5551625  0.6945296
#>        x2_36       y1_37      y2_37      y3_37      x1_37      x2_37      y1_38
#> 1 -0.9535234 -1.69080557 -0.3025021  1.7319803 -0.5428288  0.5809965 -2.6354214
#> 2  1.6894589 -0.09277337 -1.1021267  1.6043880  0.8647780 -0.1507760 -0.6308415
#> 3  0.4315654  0.94630702 -0.5591950 -0.8697912 -0.3965497  1.3099782  1.5346053
#> 4 -1.6445555  1.13743144 -2.4450389  1.0365605  1.4463565 -0.6905602  2.2584497
#> 5 -2.4613355 -0.87174442 -0.9280160 -0.8723213  0.1432898 -0.3912221  0.4400004
#>        y2_38      y3_38      x1_38      x2_38      y1_39      y2_39
#> 1 -0.7183577 -0.4737886  0.7681787  0.4637676 -2.8071935 -0.5614764
#> 2 -1.1112348  1.6758752 -1.4490071  0.6430087  0.3740888 -0.1523084
#> 3 -1.9888093 -0.2844570  0.4703934 -1.2426703  0.2343761  2.1184055
#> 4 -2.7826021 -1.3484993 -0.2764311 -1.1094188  0.4663693 -1.6985818
#> 5 -1.7603171 -0.9955418 -0.4911641 -0.2836475  1.4811656  0.8301803
#>          y3_39      x1_39        x2_39      y1_40      y2_40      y3_40
#> 1  0.881604859 -0.8857763 -1.099780899 -2.2040033  0.1882097 -0.1269184
#> 2 -0.006343342  0.4831939 -0.006355626  0.6789519  2.3019134  0.6939580
#> 3  0.795769292  1.3815755  1.204458937  0.4976630  1.2046355 -1.1505839
#> 4  0.626488740  0.1338693  1.785339052  1.3594825 -0.6741237  2.0404639
#> 5 -0.163071934  0.3147948  0.396326578  1.4834493  2.9148698 -0.2685219
#>        x1_40      x2_40       y1_41      y2_41      y3_41       x1_41
#> 1  1.5127070  0.2579214  0.57684619 -0.6899581  0.5429887  0.08844023
#> 2  0.1514559 -0.5841090 -0.08175649  0.6560388  2.2044339  0.36880673
#> 3  0.8240740 -1.6626294 -0.90625055 -0.2168597 -1.9086710 -0.56930634
#> 4  2.4221634 -1.0768289  0.92517233  0.0159189  0.5964618  0.48594111
#> 5 -0.2256037 -1.9249504 -1.51243420  0.4976952 -1.4491377 -1.43922930
#>        x2_41      y1_42       y2_42       y3_42       x1_42      x2_42
#> 1 -0.1208965 -0.9841455  1.95659308  0.10631121 -1.19432890  0.6119969
#> 2  0.2946543  1.6252713  1.40303796 -0.64688033 -0.27925937 -1.3362367
#> 3  0.6355138  0.5611223  0.09548179  0.02792649  0.04372201  0.3480123
#> 4  1.3885217 -1.8603537 -1.05576936 -1.43747523 -0.19565682 -0.2181748
#> 5 -1.4696578  0.7375841 -1.27621056 -0.28031657  0.76186345 -0.2436150
#>        y1_43       y2_43      y3_43      x1_43      x2_43      y1_44      y2_44
#> 1 -0.4172575 -0.03975722 -1.0649549 -0.2171398 -0.1827567 -0.6356891 -0.8823777
#> 2 -0.4323630  0.69400066 -0.4854508  0.7007488  0.5541966 -1.5815027  1.2093094
#> 3  0.6717540  0.59886248 -1.4107830  2.4595935 -0.8183803 -1.5925834  0.2028067
#> 4 -0.1999509 -0.27958137 -0.1115583 -0.3047780  0.5978327 -1.3860432  0.8088848
#> 5  1.5568158 -0.60905958 -2.4224375  0.2696766 -1.5589275 -2.1225482  0.3167143
#>        y3_44      x1_44      x2_44      y1_45      y2_45      y3_45      x1_45
#> 1  1.0255378  0.9333463  0.8217731  1.4168399 -1.2529312 -0.0138464  1.3921164
#> 2  0.1015089 -0.8363066 -1.5945882  1.0726562 -0.3726381 -3.8113788  0.2049586
#> 3  2.1139917 -2.1132001  0.2736953  0.7580356 -1.1751165 -0.1074872 -0.6875968
#> 4  1.2056015  1.3974294  0.6876198 -1.8356432  1.1946067  1.3876294  0.3201880
#> 5 -0.5508063 -0.5355880  0.5624520 -2.3342729 -0.4466525 -0.9093240 -0.1783261
#>        x2_45       y1_46      y2_46       y3_46       x1_46      x2_46
#> 1 -0.4761739  1.12795639 -0.6834568 -2.15905811  0.65034856  1.3911105
#> 2 -0.3450880 -3.52737494 -0.7430851 -0.08598658  0.25261170 -1.2940025
#> 3  0.4460411  0.05897484  2.5111017  0.04442733 -0.81238472  2.2120555
#> 4 -0.3018699 -0.97038509  1.3974840  0.24494246  0.49834869 -0.5495369
#> 5 -0.1151360 -2.60616376 -0.7653270 -4.59670290 -0.07206147  1.2109098
#>        y1_47      y2_47      y3_47      x1_47      x2_47      y1_48      y2_48
#> 1 -1.3150620 -0.4756001 -1.8329691 -1.1107889 -0.8607926 -1.3449987  1.7682712
#> 2 -0.7472187 -2.0455025 -2.0144549 -0.9591704  1.0857749 -1.4057986 -1.7109581
#> 3  2.8324142  1.4446795 -0.1309829 -0.1237060 -0.4773355  0.1231669 -1.3976359
#> 4 -0.6650360 -1.8388994  1.8045724 -0.2792565  1.0965134 -1.9632644 -1.0236716
#> 5 -1.3209728  1.4065665 -0.7600015 -0.6148969  0.6761265 -0.8411154  0.7324089
#>        y3_48      x1_48      x2_48      y1_49       y2_49      y3_49
#> 1  0.5772184 -1.1317387 -1.4592140 -2.4462217  1.19775527 -1.2621964
#> 2 -0.5982397  0.4037749  0.5864875 -0.2728646  1.19842157  2.2908512
#> 3  1.5982764 -0.1662615  0.8625634 -0.4690810 -3.23643735 -0.6867484
#> 4  1.2249418  0.4420131  0.2410163  2.2338761  0.32421845  2.2357651
#> 5  0.6988805  0.8985996 -1.1893179 -0.1271400  0.09189132  0.1123442
#>         x1_49       x2_49
#> 1  0.07998255  0.65320434
#> 2  1.81522845  0.12882143
#> 3  0.09734049 -1.62561674
#> 4 -0.25560766  0.93103290
#> 5  0.12125885 -0.01122169