Coerce an Object of Class simstatespace to a Matrix

Coerce an Object of Class simstatespace to a Matrix

Usage

# S3 method for simstatespace
as.matrix(x, eta = FALSE, long = TRUE, ...)

Arguments

x

Object of class simstatespace.

eta

Logical. If eta = TRUE, include eta. If eta = FALSE, exclude eta.

long

Logical. If long = TRUE, use long format. If long = FALSE, use wide format.

...

Additional arguments.

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
k <- p <- 3
iden <- diag(k)
null_vec <- rep(x = 0, times = k)
n <- 5
mu0 <- null_vec
sigma0 <- iden
alpha <- null_vec
beta <- diag(x = 0.50, nrow = k)
psi <- iden
nu <- null_vec
lambda <- iden
theta <- diag(x = 0.50, nrow = k)
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- 0.10 * diag(k)
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = k * (time + burn_in)),
        ncol = k
      )
    )
  }
)

# Type 0
ssm <- SimSSMFixed(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  alpha = alpha,
  beta = beta,
  psi = psi,
  nu = nu,
  lambda = lambda,
  theta = theta,
  type = 0,
  time = time,
  burn_in = burn_in
)

head(as.matrix(ssm))
#>      id time          y1         y2         y3
#> [1,]  1    0 -1.09954009  1.4032891  0.7784308
#> [2,]  1    1  1.03260528 -1.3251523  0.3094271
#> [3,]  1    2  0.07700843 -1.8291596  0.2946668
#> [4,]  1    3 -0.69231339  0.9493678 -0.6706605
#> [5,]  1    4 -0.17372176  1.5548881 -1.1757591
#> [6,]  1    5  1.31648190 -0.6817653 -0.8769241
head(as.matrix(ssm, long = FALSE))
#>      id       y1_0       y2_0         y3_0       y1_1       y2_1       y3_1
#> [1,]  1 -1.0995401  1.4032891  0.778430759  1.0326053 -1.3251523  0.3094271
#> [2,]  2 -0.4862692 -0.9360818  0.003022709  0.7106665  0.9102038 -2.2016993
#> [3,]  3 -0.2611621 -1.7014668  1.505575850  2.2455017 -0.9277531  0.1982638
#> [4,]  4  0.3628444 -0.4866576 -0.990268670  1.0054983 -1.6636278  0.6167070
#> [5,]  5  0.2137389  0.7420170 -0.748814331 -1.1297343 -0.4516830 -0.7017896
#>             y1_2       y2_2       y3_2       y1_3       y2_3        y3_3
#> [1,]  0.07700843 -1.8291596  0.2946668 -0.6923134  0.9493678 -0.67066046
#> [2,]  0.35637548  0.8603111 -1.5242204  1.9271685  1.6041946 -0.88125303
#> [3,]  1.34751556  1.4443335  1.3498476 -0.8141161  1.1704067  0.23813806
#> [4,]  0.32578948 -1.5065379  0.7922401 -1.2293404 -0.5445986 -0.05685624
#> [5,] -1.42180784 -1.1302348 -1.7481942  0.8619176  0.6163751 -0.40132311
#>            y1_4       y2_4       y3_4       y1_5       y2_5       y3_5
#> [1,] -0.1737218  1.5548881 -1.1757591  1.3164819 -0.6817653 -0.8769241
#> [2,] -2.2786571  2.1719789 -0.6514592 -0.7203359  1.0034417  0.1111865
#> [3,]  0.8576239  1.8657882  0.1293636  1.1741567  0.6768871 -0.1819875
#> [4,] -1.0356634  0.2206535 -0.6386000 -0.3840216 -0.5443154 -0.1511036
#> [5,]  2.3952392 -1.0255805  1.1306696  2.2593628 -2.0523884 -0.2780439
#>            y1_6       y2_6       y3_6      y1_7        y2_7        y3_7
#> [1,]  2.6146254 -1.3023876 -0.5912923 1.6232148  0.58179208 -0.38835546
#> [2,] -1.7086169 -1.5309625  0.3301754 1.3569414 -3.35182785  1.24438775
#> [3,]  1.3464648  0.8525882  0.8614740 0.1188553  1.59646498  1.19974128
#> [4,] -0.1851625  0.1884550  1.6627963 0.1982461 -1.26542216  1.49430371
#> [5,] -0.1627264 -2.3960944  0.5358891 1.8765506  0.05043105 -0.02168787
#>            y1_8       y2_8        y3_8       y1_9       y2_9         y3_9
#> [1,]  4.1092815 -0.4047962 -0.79040616 -0.9984378  0.3357992  0.700275676
#> [2,]  3.2946122 -0.9682126 -0.61433553  0.5829798 -0.5610084  0.002662121
#> [3,]  1.0500138 -0.9663746 -0.01096091  0.2642090  1.0898766 -0.838141228
#> [4,] -2.8419764  1.9864251  1.29541159  0.2992914  1.5785655 -1.791456386
#> [5,]  0.3336033 -1.2050002  1.30689412 -0.3115522  1.2471533  0.388376698
#>           y1_10       y2_10       y3_10      y1_11      y2_11      y3_11
#> [1,]  0.3016141 -0.34334016  0.01271578 -0.9080502 -1.4551178  1.1174777
#> [2,]  0.5569878 -0.41573375 -0.14143368  0.4163683  0.2065969  1.0182270
#> [3,] -0.2928771  0.25601120 -1.47481512 -1.8552625 -0.5411652 -1.3264278
#> [4,]  1.3926087  0.08076689 -2.60854254  0.5592967  0.5070627 -1.9666986
#> [5,] -3.4571907 -0.46762892  2.70207887 -0.7619126 -0.6980942 -0.4479936
#>           y1_12      y2_12      y3_12       y1_13      y2_13      y3_13
#> [1,] -0.4735484 -0.7043525  1.2621696  0.56440495  0.2656128  0.1346826
#> [2,]  0.6959900 -2.1601722  1.2045147 -1.11281875 -1.8824232 -1.0943125
#> [3,]  0.2180062 -0.7693463 -1.9298510 -0.57726626  0.2689115 -0.0761991
#> [4,] -0.8906314 -0.1827378  0.8564346 -2.57956143 -0.3274803 -1.4270533
#> [5,]  2.1704534 -1.1140866  0.4509473  0.03182284  0.1395369  0.2528732
#>            y1_14      y2_14      y3_14      y1_15       y2_15      y3_15
#> [1,]  0.66448570 -1.0392645 -0.7172866 0.04264334  0.24791986  1.5821550
#> [2,] -0.01904589 -0.1279915 -0.2677521 0.63116514 -0.01015358  1.4125655
#> [3,] -1.03940884  1.2063136  0.4511569 1.22888055  1.70552173  0.8739239
#> [4,] -1.18607203 -1.1990209  0.5289841 0.68194793 -2.28235253 -0.8721024
#> [5,]  0.66735414 -0.8213950  0.1514670 1.66293082 -0.62535149  0.3069551
#>           y1_16       y2_16      y3_16       y1_17      y2_17      y3_17
#> [1,]  0.1107793 -0.33334090 -0.2738133 -0.29548962  0.7059574 -0.5140712
#> [2,] -0.1925946 -0.09203825 -0.7801705  1.14312425 -1.1164271  0.4075347
#> [3,]  1.4993960  0.55875912 -2.7550918 -0.63517024  1.7908031 -0.6680233
#> [4,]  1.1490660 -1.48601594 -0.8233582  0.07542342 -1.4099979 -0.0591949
#> [5,]  0.2004234 -0.41563760 -0.7104642  0.33448108 -1.5601427 -1.9706671
#>            y1_18      y2_18      y3_18      y1_19      y2_19       y3_19
#> [1,]  0.21216641  2.3164812 -0.3144939  2.4067212  1.4629069 -2.27009883
#> [2,] -1.74920314 -0.4584396  2.0591696  1.9746364 -1.6021805  2.79729525
#> [3,] -0.41332821  2.5721242 -0.5486974 -0.5604419 -1.9832873 -0.13937464
#> [4,] -0.04938728  0.2588009 -1.7102633  1.4092183  3.1917351 -0.09351291
#> [5,]  2.67374765 -1.1589410 -1.9422668  2.8808639 -0.7808992  0.56917952
#>           y1_20      y2_20      y3_20      y1_21      y2_21       y3_21
#> [1,]  1.7093432  0.4062894 -1.5243166  3.2912305  1.8761515  0.32616111
#> [2,]  1.2485908  0.7608486  0.8232979 -0.2436155  1.1400210 -2.66515978
#> [3,]  2.1359884 -0.5654781  2.0862052  1.4140650 -0.9554528 -0.05667008
#> [4,] -0.2166186 -0.1137637 -1.5851508  0.3856954 -1.0651373 -1.16053990
#> [5,] -0.8309464 -0.2483420 -1.2591873 -1.3981766  0.4117713  0.07956801
#>           y1_22       y2_22      y3_22      y1_23      y2_23      y3_23
#> [1,]  0.4400620  0.70468749 -0.7658286 -1.6585065  0.5526912 -0.1070375
#> [2,] -0.3179609  1.76406281 -0.3668738  1.6888110  2.8233467  0.5128405
#> [3,] -0.9326660 -0.93992435 -1.5129975  2.9121667 -0.2461096 -0.7812020
#> [4,]  1.1644061  0.04969627 -0.9070176  1.3060583 -2.2720139 -1.4784150
#> [5,]  0.2179019 -0.65588721  1.4972191 -0.6796242 -0.7959906  0.2248023
#>           y1_24      y2_24      y3_24       y1_25      y2_25        y3_25
#> [1,] -0.2311295 -0.1876046 -2.4610026 -0.53497728  0.7042024 -1.778004131
#> [2,] -1.1831673  1.1289935 -0.7760733 -1.57492885  0.3335125  0.301801571
#> [3,]  0.5626923 -1.3964431 -1.0947300  1.32993361 -0.6867271 -1.235983436
#> [4,]  0.3565990 -3.7969670 -0.3792707 -0.08862718 -1.0409192  0.005158764
#> [5,] -0.0752571 -0.3450391  0.5266850  0.22887606 -1.1522363  2.746770307
#>           y1_26      y2_26      y3_26      y1_27      y2_27      y3_27
#> [1,]  2.7170672 -0.8609936  0.1726962 -0.7120913 -2.2120550  1.8597665
#> [2,] -1.5651869  0.9305363  0.5963625 -0.8217285  1.1972189 -0.3301369
#> [3,]  0.8948044 -1.3445395 -0.8710860 -1.1842599  0.7894854 -2.9864186
#> [4,] -0.1489025  0.8741137  1.2112019 -0.7343369  2.5996166  1.3591158
#> [5,] -0.4119303  0.1431916  0.8715951 -1.0059365 -1.6009747  1.7007612
#>           y1_28      y2_28      y3_28      y1_29       y2_29      y3_29
#> [1,]  1.2607620 -0.9557771  0.6782045  1.9206276  3.62824901  1.3994596
#> [2,] -1.3716504 -0.3117331  0.1024229  0.3396966  1.78697097 -0.5284965
#> [3,] -2.5856469 -2.9299775  1.3029542 -0.2535041  0.06045029  2.0843798
#> [4,] -0.5803437  2.3662000  0.3827125 -2.7944895 -0.90218442  1.5587893
#> [5,] -1.9253959  1.4942818 -0.4079834  1.4310673  1.15093429 -0.6104772
#>           y1_30      y2_30      y3_30      y1_31      y2_31       y3_31
#> [1,]  1.1667738 -1.0296442  0.4465208  0.2815890 -0.8159860  0.26649509
#> [2,]  0.2434068 -0.9523099 -0.1152634  0.2485631 -0.8129093  1.26283301
#> [3,] -0.6659829  0.8283992  1.1009152 -1.1034329  0.1580498 -0.08612643
#> [4,] -1.7772095  1.3861359  1.6439649  0.5524406 -1.3993364 -2.74890387
#> [5,] -0.3826375  1.0364998 -0.4686253 -1.1937923  1.5285524 -0.78822250
#>           y1_32      y2_32      y3_32     y1_33      y2_33      y3_33
#> [1,]  0.8443721  1.8669682  0.7682024 0.7440492  2.5573639  2.2780395
#> [2,] -0.9089033 -0.9290651 -2.5804784 1.3291999  0.6463786  0.5327538
#> [3,]  1.3509872  1.8995571  3.4465148 1.3496753  0.4397430  0.5017668
#> [4,]  1.6274305 -0.4043533 -0.8801534 1.6443757 -0.6837037 -0.4019448
#> [5,]  1.5843125  0.7245916 -3.5306075 1.8927078 -2.4936758 -1.6359057
#>           y1_34      y2_34      y3_34      y1_35      y2_35      y3_35
#> [1,]  1.3227618  2.1703893  1.3499775  1.1742738  1.3479093  2.4946477
#> [2,] -0.2606539  1.7108154 -0.3909682 -0.5532294  0.2462738  0.4802346
#> [3,] -1.1911688 -0.5624893  1.7649523  0.3190500 -3.1325332  0.5924701
#> [4,]  0.7308648 -1.0056029 -1.7183440  0.8057427 -1.1929240 -0.6701554
#> [5,]  2.2148850  0.4121900 -0.5618103 -0.3256500 -0.8189572 -0.4359845
#>           y1_36      y2_36      y3_36      y1_37      y2_37      y3_37
#> [1,]  1.2436593  1.7994006  0.7215381  0.9028786  0.4692159 -0.1035422
#> [2,]  0.1842798  0.5852260  1.5183485 -0.4963885 -0.2958775 -0.2362714
#> [3,] -0.3143567 -0.3804031 -0.7218898 -0.3370434 -1.6002708 -2.4492260
#> [4,]  0.0351231 -0.6262361  0.1000247 -0.6366331 -1.2872799  0.4415835
#> [5,]  1.6271317 -1.3186361  0.7828124  2.3291989 -1.0437543  0.3989199
#>           y1_38      y2_38      y3_38       y1_39      y2_39      y3_39
#> [1,]  1.9675205 -0.7717988  2.4064093  0.71798859 -0.1755463  0.9290106
#> [2,] -0.3594548 -1.6482643 -0.1399794  2.68977216 -0.9221182 -0.6292083
#> [3,] -2.0659299  0.7263094 -1.3307852 -0.03785555  0.0414309 -1.2504557
#> [4,]  0.1086735 -1.1384088 -1.4547639  0.40879492  0.1427540 -1.3911065
#> [5,]  0.9236375  0.6023882  0.1512480  2.53043915  1.2934940  0.8247677
#>           y1_40      y2_40      y3_40       y1_41       y2_41      y3_41
#> [1,] -1.6974972  0.1043126  1.7512540  0.18844232  0.49988664  2.2220771
#> [2,]  0.8470835 -1.9113268 -1.0392918  1.43877055  1.19754635  1.2465816
#> [3,] -0.0677613  0.9636083  3.5861984 -0.95599343  1.22310119 -0.6650830
#> [4,]  0.9856875  0.8708387  0.4200998 -3.02698688 -0.06285456  0.4790056
#> [5,]  0.9250186  0.7482434 -0.6370872 -0.01472055 -2.45039352 -1.1557180
#>            y1_42       y2_42      y3_42      y1_43      y2_43      y3_43
#> [1,] -0.99739117  2.62900875  1.0814436 -0.2129836 -0.2171767  1.2754159
#> [2,]  2.45672208  0.81721915  0.5806150  0.7707791 -0.5640988  0.9616499
#> [3,]  0.05230591  0.08043901  0.1440378  1.8865604  0.1164399  1.2393873
#> [4,] -0.87992403  1.52855549 -0.2518280  0.2917907  0.8004304 -1.8900486
#> [5,] -0.77482783 -0.61741188  2.6440668  0.4460444 -1.3801821  0.6799519
#>            y1_44      y2_44       y3_44      y1_45      y2_45      y3_45
#> [1,]  1.17684706  0.4659966  0.83322680  0.1326481 -0.5762872  0.9673707
#> [2,] -1.03451494 -0.2805831  0.07246422  0.7956358  1.0626150 -1.7582351
#> [3,] -0.49732616  0.9384659 -1.42553706 -1.5307716  0.4838092  1.2258299
#> [4,] -1.37121423  0.5047150  0.73359407 -1.2083143  0.3813296  0.3086373
#> [5,] -0.02497946 -1.5024369 -1.12489390  3.1503015 -1.2131054 -1.4950659
#>           y1_46      y2_46      y3_46      y1_47       y2_47        y3_47
#> [1,] -0.1832172 -0.5752430  0.8016409 -0.1636911 -0.05325874  0.443175209
#> [2,]  0.9819970  0.9774213 -1.1137875  0.4382173 -0.90362245 -3.132598135
#> [3,] -0.3044024  0.9007896  1.5648721  0.6048861  0.78222147  0.003003359
#> [4,] -1.0484130 -0.5262169  0.8661278 -4.1001085 -0.32956460 -1.342043098
#> [5,] -0.4963991 -3.0621266 -1.5446943  0.2411616 -2.37397885  0.544564179
#>           y1_48       y2_48      y3_48      y1_49      y2_49       y3_49
#> [1,] -0.6573276  2.35372989 -0.9227673  1.8028099  2.0707657 -2.32530247
#> [2,] -0.2002175 -0.02881721 -4.6906377  0.2810525  1.7123961 -2.09842952
#> [3,]  2.4524881  1.64770558  0.4834392  0.6990856  2.1151388 -0.92960600
#> [4,] -1.0268122  1.21539129 -0.8611734 -1.3630000 -0.6057190 -1.23516193
#> [5,]  1.9080431 -0.49987413  1.0758689  1.5541748 -0.4256441 -0.04368629