Coerce an Object of Class ssm to a Matrix

Coerce an Object of Class ssm to a Matrix

Usage

# S3 method for ssm
as.matrix(x, eta = FALSE, long = TRUE, ...)

Arguments

x

Object of class ssm.

eta

Logical. If eta = TRUE, include eta. If eta = FALSE, exclude eta.

long

Logical. If long = TRUE, use long format. If long = FALSE, use wide format.

...

Additional arguments.

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
k <- p <- 3
iden <- diag(k)
null_vec <- rep(x = 0, times = k)
n <- 5
mu0 <- null_vec
sigma0 <- iden
alpha <- null_vec
beta <- diag(x = 0.50, nrow = k)
psi <- iden
nu <- null_vec
lambda <- iden
theta <- diag(x = 0.50, nrow = k)
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- 0.10 * diag(k)
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = k * (time + burn_in)),
        ncol = k
      )
    )
  }
)

# Type 0
ssm <- SimSSMFixed(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  alpha = alpha,
  beta = beta,
  psi = psi,
  nu = nu,
  lambda = lambda,
  theta = theta,
  type = 0,
  time = time,
  burn_in = burn_in
)

as.matrix(ssm)
#>        id time          y1          y2           y3
#>   [1,]  1    0 -1.09954009  1.40328907  0.778430759
#>   [2,]  1    1  1.03260528 -1.32515227  0.309427089
#>   [3,]  1    2  0.07700843 -1.82915955  0.294666843
#>   [4,]  1    3 -0.69231339  0.94936778 -0.670660458
#>   [5,]  1    4 -0.17372176  1.55488807 -1.175759089
#>   [6,]  1    5  1.31648190 -0.68176528 -0.876924093
#>   [7,]  1    6  2.61462535 -1.30238755 -0.591292294
#>   [8,]  1    7  1.62321478  0.58179208 -0.388355462
#>   [9,]  1    8  4.10928146 -0.40479621 -0.790406161
#>  [10,]  1    9 -0.99843785  0.33579917  0.700275676
#>  [11,]  1   10  0.30161407 -0.34334016  0.012715784
#>  [12,]  1   11 -0.90805025 -1.45511781  1.117477735
#>  [13,]  1   12 -0.47354838 -0.70435255  1.262169607
#>  [14,]  1   13  0.56440495  0.26561283  0.134682627
#>  [15,]  1   14  0.66448570 -1.03926452 -0.717286573
#>  [16,]  1   15  0.04264334  0.24791986  1.582154981
#>  [17,]  1   16  0.11077926 -0.33334090 -0.273813265
#>  [18,]  1   17 -0.29548962  0.70595738 -0.514071191
#>  [19,]  1   18  0.21216641  2.31648122 -0.314493883
#>  [20,]  1   19  2.40672116  1.46290685 -2.270098835
#>  [21,]  1   20  1.70934319  0.40628938 -1.524316591
#>  [22,]  1   21  3.29123052  1.87615146  0.326161112
#>  [23,]  1   22  0.44006198  0.70468749 -0.765828554
#>  [24,]  1   23 -1.65850645  0.55269119 -0.107037454
#>  [25,]  1   24 -0.23112948 -0.18760463 -2.461002638
#>  [26,]  1   25 -0.53497728  0.70420238 -1.778004131
#>  [27,]  1   26  2.71706716 -0.86099362  0.172696150
#>  [28,]  1   27 -0.71209126 -2.21205498  1.859766520
#>  [29,]  1   28  1.26076202 -0.95577708  0.678204547
#>  [30,]  1   29  1.92062763  3.62824901  1.399459630
#>  [31,]  1   30  1.16677381 -1.02964425  0.446520805
#>  [32,]  1   31  0.28158896 -0.81598603  0.266495090
#>  [33,]  1   32  0.84437214  1.86696815  0.768202390
#>  [34,]  1   33  0.74404917  2.55736389  2.278039505
#>  [35,]  1   34  1.32276182  2.17038929  1.349977527
#>  [36,]  1   35  1.17427385  1.34790926  2.494647718
#>  [37,]  1   36  1.24365932  1.79940059  0.721538123
#>  [38,]  1   37  0.90287860  0.46921585 -0.103542151
#>  [39,]  1   38  1.96752049 -0.77179878  2.406409325
#>  [40,]  1   39  0.71798859 -0.17554635  0.929010646
#>  [41,]  1   40 -1.69749719  0.10431262  1.751254035
#>  [42,]  1   41  0.18844232  0.49988664  2.222077117
#>  [43,]  1   42 -0.99739117  2.62900875  1.081443615
#>  [44,]  1   43 -0.21298359 -0.21717666  1.275415872
#>  [45,]  1   44  1.17684706  0.46599663  0.833226798
#>  [46,]  1   45  0.13264806 -0.57628721  0.967370724
#>  [47,]  1   46 -0.18321725 -0.57524301  0.801640889
#>  [48,]  1   47 -0.16369109 -0.05325874  0.443175209
#>  [49,]  1   48 -0.65732765  2.35372989 -0.922767347
#>  [50,]  1   49  1.80280990  2.07076569 -2.325302475
#>  [51,]  2    0 -0.48626915 -0.93608182  0.003022709
#>  [52,]  2    1  0.71066655  0.91020380 -2.201699299
#>  [53,]  2    2  0.35637548  0.86031114 -1.524220440
#>  [54,]  2    3  1.92716849  1.60419463 -0.881253025
#>  [55,]  2    4 -2.27865713  2.17197894 -0.651459237
#>  [56,]  2    5 -0.72033586  1.00344175  0.111186511
#>  [57,]  2    6 -1.70861690 -1.53096252  0.330175392
#>  [58,]  2    7  1.35694144 -3.35182785  1.244387749
#>  [59,]  2    8  3.29461220 -0.96821257 -0.614335528
#>  [60,]  2    9  0.58297982 -0.56100837  0.002662121
#>  [61,]  2   10  0.55698780 -0.41573375 -0.141433679
#>  [62,]  2   11  0.41636833  0.20659691  1.018226998
#>  [63,]  2   12  0.69599002 -2.16017218  1.204514706
#>  [64,]  2   13 -1.11281875 -1.88242323 -1.094312496
#>  [65,]  2   14 -0.01904589 -0.12799149 -0.267752052
#>  [66,]  2   15  0.63116514 -0.01015358  1.412565479
#>  [67,]  2   16 -0.19259459 -0.09203825 -0.780170543
#>  [68,]  2   17  1.14312425 -1.11642708  0.407534695
#>  [69,]  2   18 -1.74920314 -0.45843960  2.059169610
#>  [70,]  2   19  1.97463636 -1.60218054  2.797295248
#>  [71,]  2   20  1.24859081  0.76084858  0.823297855
#>  [72,]  2   21 -0.24361549  1.14002099 -2.665159781
#>  [73,]  2   22 -0.31796090  1.76406281 -0.366873844
#>  [74,]  2   23  1.68881102  2.82334673  0.512840490
#>  [75,]  2   24 -1.18316733  1.12899348 -0.776073331
#>  [76,]  2   25 -1.57492885  0.33351248  0.301801571
#>  [77,]  2   26 -1.56518687  0.93053631  0.596362472
#>  [78,]  2   27 -0.82172855  1.19721886 -0.330136944
#>  [79,]  2   28 -1.37165045 -0.31173308  0.102422863
#>  [80,]  2   29  0.33969661  1.78697097 -0.528496539
#>  [81,]  2   30  0.24340681 -0.95230990 -0.115263362
#>  [82,]  2   31  0.24856306 -0.81290927  1.262833006
#>  [83,]  2   32 -0.90890328 -0.92906513 -2.580478364
#>  [84,]  2   33  1.32919989  0.64637858  0.532753763
#>  [85,]  2   34 -0.26065395  1.71081535 -0.390968157
#>  [86,]  2   35 -0.55322936  0.24627381  0.480234608
#>  [87,]  2   36  0.18427982  0.58522603  1.518348522
#>  [88,]  2   37 -0.49638849 -0.29587752 -0.236271390
#>  [89,]  2   38 -0.35945483 -1.64826430 -0.139979372
#>  [90,]  2   39  2.68977216 -0.92211819 -0.629208268
#>  [91,]  2   40  0.84708353 -1.91132685 -1.039291830
#>  [92,]  2   41  1.43877055  1.19754635  1.246581612
#>  [93,]  2   42  2.45672208  0.81721915  0.580614976
#>  [94,]  2   43  0.77077913 -0.56409875  0.961649922
#>  [95,]  2   44 -1.03451494 -0.28058308  0.072464218
#>  [96,]  2   45  0.79563584  1.06261499 -1.758235072
#>  [97,]  2   46  0.98199698  0.97742135 -1.113787454
#>  [98,]  2   47  0.43821730 -0.90362245 -3.132598135
#>  [99,]  2   48 -0.20021749 -0.02881721 -4.690637742
#> [100,]  2   49  0.28105246  1.71239610 -2.098429524
#> [101,]  3    0 -0.26116213 -1.70146684  1.505575850
#> [102,]  3    1  2.24550172 -0.92775310  0.198263841
#> [103,]  3    2  1.34751556  1.44433346  1.349847649
#> [104,]  3    3 -0.81411613  1.17040668  0.238138061
#> [105,]  3    4  0.85762386  1.86578815  0.129363635
#> [106,]  3    5  1.17415675  0.67688707 -0.181987545
#> [107,]  3    6  1.34646484  0.85258818  0.861474003
#> [108,]  3    7  0.11885530  1.59646498  1.199741277
#> [109,]  3    8  1.05001380 -0.96637464 -0.010960909
#> [110,]  3    9  0.26420895  1.08987658 -0.838141228
#> [111,]  3   10 -0.29287710  0.25601120 -1.474815119
#> [112,]  3   11 -1.85526246 -0.54116516 -1.326427771
#> [113,]  3   12  0.21800618 -0.76934630 -1.929850985
#> [114,]  3   13 -0.57726626  0.26891151 -0.076199099
#> [115,]  3   14 -1.03940884  1.20631356  0.451156926
#> [116,]  3   15  1.22888055  1.70552173  0.873923885
#> [117,]  3   16  1.49939600  0.55875912 -2.755091772
#> [118,]  3   17 -0.63517024  1.79080306 -0.668023294
#> [119,]  3   18 -0.41332821  2.57212423 -0.548697427
#> [120,]  3   19 -0.56044193 -1.98328729 -0.139374639
#> [121,]  3   20  2.13598838 -0.56547814  2.086205206
#> [122,]  3   21  1.41406500 -0.95545284 -0.056670083
#> [123,]  3   22 -0.93266598 -0.93992435 -1.512997488
#> [124,]  3   23  2.91216672 -0.24610961 -0.781202022
#> [125,]  3   24  0.56269235 -1.39644313 -1.094730024
#> [126,]  3   25  1.32993361 -0.68672705 -1.235983436
#> [127,]  3   26  0.89480439 -1.34453947 -0.871085950
#> [128,]  3   27 -1.18425988  0.78948537 -2.986418646
#> [129,]  3   28 -2.58564690 -2.92997753  1.302954218
#> [130,]  3   29 -0.25350412  0.06045029  2.084379825
#> [131,]  3   30 -0.66598288  0.82839921  1.100915239
#> [132,]  3   31 -1.10343287  0.15804982 -0.086126432
#> [133,]  3   32  1.35098716  1.89955709  3.446514757
#> [134,]  3   33  1.34967532  0.43974303  0.501766751
#> [135,]  3   34 -1.19116877 -0.56248927  1.764952329
#> [136,]  3   35  0.31905000 -3.13253316  0.592470111
#> [137,]  3   36 -0.31435668 -0.38040307 -0.721889832
#> [138,]  3   37 -0.33704338 -1.60027085 -2.449225982
#> [139,]  3   38 -2.06592986  0.72630936 -1.330785176
#> [140,]  3   39 -0.03785555  0.04143090 -1.250455703
#> [141,]  3   40 -0.06776130  0.96360829  3.586198402
#> [142,]  3   41 -0.95599343  1.22310119 -0.665082976
#> [143,]  3   42  0.05230591  0.08043901  0.144037835
#> [144,]  3   43  1.88656038  0.11643992  1.239387298
#> [145,]  3   44 -0.49732616  0.93846594 -1.425537063
#> [146,]  3   45 -1.53077161  0.48380919  1.225829910
#> [147,]  3   46 -0.30440242  0.90078956  1.564872105
#> [148,]  3   47  0.60488609  0.78222147  0.003003359
#> [149,]  3   48  2.45248814  1.64770558  0.483439185
#> [150,]  3   49  0.69908565  2.11513881 -0.929605996
#> [151,]  4    0  0.36284438 -0.48665756 -0.990268670
#> [152,]  4    1  1.00549834 -1.66362776  0.616706989
#> [153,]  4    2  0.32578948 -1.50653791  0.792240074
#> [154,]  4    3 -1.22934036 -0.54459863 -0.056856245
#> [155,]  4    4 -1.03566338  0.22065346 -0.638600015
#> [156,]  4    5 -0.38402159 -0.54431544 -0.151103570
#> [157,]  4    6 -0.18516254  0.18845500  1.662796324
#> [158,]  4    7  0.19824605 -1.26542216  1.494303713
#> [159,]  4    8 -2.84197642  1.98642510  1.295411591
#> [160,]  4    9  0.29929141  1.57856545 -1.791456386
#> [161,]  4   10  1.39260871  0.08076689 -2.608542538
#> [162,]  4   11  0.55929670  0.50706273 -1.966698587
#> [163,]  4   12 -0.89063144 -0.18273784  0.856434583
#> [164,]  4   13 -2.57956143 -0.32748026 -1.427053340
#> [165,]  4   14 -1.18607203 -1.19902092  0.528984079
#> [166,]  4   15  0.68194793 -2.28235253 -0.872102410
#> [167,]  4   16  1.14906598 -1.48601594 -0.823358157
#> [168,]  4   17  0.07542342 -1.40999795 -0.059194896
#> [169,]  4   18 -0.04938728  0.25880090 -1.710263262
#> [170,]  4   19  1.40921825  3.19173506 -0.093512910
#> [171,]  4   20 -0.21661858 -0.11376365 -1.585150807
#> [172,]  4   21  0.38569541 -1.06513732 -1.160539898
#> [173,]  4   22  1.16440615  0.04969627 -0.907017636
#> [174,]  4   23  1.30605832 -2.27201386 -1.478414984
#> [175,]  4   24  0.35659895 -3.79696704 -0.379270694
#> [176,]  4   25 -0.08862718 -1.04091923  0.005158764
#> [177,]  4   26 -0.14890245  0.87411373  1.211201900
#> [178,]  4   27 -0.73433686  2.59961656  1.359115824
#> [179,]  4   28 -0.58034370  2.36620000  0.382712459
#> [180,]  4   29 -2.79448953 -0.90218442  1.558789274
#> [181,]  4   30 -1.77720952  1.38613588  1.643964905
#> [182,]  4   31  0.55244062 -1.39933644 -2.748903872
#> [183,]  4   32  1.62743047 -0.40435333 -0.880153353
#> [184,]  4   33  1.64437571 -0.68370366 -0.401944807
#> [185,]  4   34  0.73086478 -1.00560286 -1.718343966
#> [186,]  4   35  0.80574272 -1.19292398 -0.670155352
#> [187,]  4   36  0.03512310 -0.62623614  0.100024689
#> [188,]  4   37 -0.63663308 -1.28727985  0.441583494
#> [189,]  4   38  0.10867352 -1.13840881 -1.454763911
#> [190,]  4   39  0.40879492  0.14275403 -1.391106451
#> [191,]  4   40  0.98568752  0.87083875  0.420099774
#> [192,]  4   41 -3.02698688 -0.06285456  0.479005565
#> [193,]  4   42 -0.87992403  1.52855549 -0.251827977
#> [194,]  4   43  0.29179069  0.80043043 -1.890048568
#> [195,]  4   44 -1.37121423  0.50471497  0.733594074
#> [196,]  4   45 -1.20831426  0.38132959  0.308637345
#> [197,]  4   46 -1.04841296 -0.52621687  0.866127827
#> [198,]  4   47 -4.10010854 -0.32956460 -1.342043098
#> [199,]  4   48 -1.02681220  1.21539129 -0.861173438
#> [200,]  4   49 -1.36299997 -0.60571904 -1.235161927
#> [201,]  5    0  0.21373895  0.74201695 -0.748814331
#> [202,]  5    1 -1.12973434 -0.45168303 -0.701789556
#> [203,]  5    2 -1.42180784 -1.13023479 -1.748194241
#> [204,]  5    3  0.86191764  0.61637508 -0.401323106
#> [205,]  5    4  2.39523917 -1.02558055  1.130669596
#> [206,]  5    5  2.25936280 -2.05238840 -0.278043934
#> [207,]  5    6 -0.16272639 -2.39609444  0.535889149
#> [208,]  5    7  1.87655056  0.05043105 -0.021687868
#> [209,]  5    8  0.33360327 -1.20500018  1.306894120
#> [210,]  5    9 -0.31155215  1.24715332  0.388376698
#> [211,]  5   10 -3.45719067 -0.46762892  2.702078872
#> [212,]  5   11 -0.76191261 -0.69809421 -0.447993625
#> [213,]  5   12  2.17045341 -1.11408658  0.450947266
#> [214,]  5   13  0.03182284  0.13953689  0.252873160
#> [215,]  5   14  0.66735414 -0.82139497  0.151466991
#> [216,]  5   15  1.66293082 -0.62535149  0.306955124
#> [217,]  5   16  0.20042340 -0.41563760 -0.710464241
#> [218,]  5   17  0.33448108 -1.56014270 -1.970667119
#> [219,]  5   18  2.67374765 -1.15894097 -1.942266775
#> [220,]  5   19  2.88086390 -0.78089915  0.569179516
#> [221,]  5   20 -0.83094639 -0.24834204 -1.259187254
#> [222,]  5   21 -1.39817664  0.41177133  0.079568012
#> [223,]  5   22  0.21790193 -0.65588721  1.497219129
#> [224,]  5   23 -0.67962420 -0.79599056  0.224802279
#> [225,]  5   24 -0.07525710 -0.34503909  0.526684956
#> [226,]  5   25  0.22887606 -1.15223630  2.746770307
#> [227,]  5   26 -0.41193034  0.14319161  0.871595120
#> [228,]  5   27 -1.00593650 -1.60097466  1.700761232
#> [229,]  5   28 -1.92539591  1.49428177 -0.407983371
#> [230,]  5   29  1.43106734  1.15093429 -0.610477188
#> [231,]  5   30 -0.38263755  1.03649977 -0.468625260
#> [232,]  5   31 -1.19379233  1.52855236 -0.788222497
#> [233,]  5   32  1.58431251  0.72459159 -3.530607481
#> [234,]  5   33  1.89270783 -2.49367581 -1.635905718
#> [235,]  5   34  2.21488503  0.41219001 -0.561810319
#> [236,]  5   35 -0.32564999 -0.81895724 -0.435984542
#> [237,]  5   36  1.62713168 -1.31863611  0.782812399
#> [238,]  5   37  2.32919888 -1.04375429  0.398919854
#> [239,]  5   38  0.92363749  0.60238820  0.151247986
#> [240,]  5   39  2.53043915  1.29349400  0.824767659
#> [241,]  5   40  0.92501863  0.74824342 -0.637087225
#> [242,]  5   41 -0.01472055 -2.45039352 -1.155718026
#> [243,]  5   42 -0.77482783 -0.61741188  2.644066796
#> [244,]  5   43  0.44604438 -1.38018214  0.679951863
#> [245,]  5   44 -0.02497946 -1.50243687 -1.124893902
#> [246,]  5   45  3.15030148 -1.21310540 -1.495065880
#> [247,]  5   46 -0.49639914 -3.06212665 -1.544694261
#> [248,]  5   47  0.24116161 -2.37397885  0.544564179
#> [249,]  5   48  1.90804312 -0.49987413  1.075868925
#> [250,]  5   49  1.55417479 -0.42564408 -0.043686291