Simulate Transition Matrices from the Multivariate Normal Distribution

This function simulates random transition matrices from the multivariate normal distribution. The function ensures that the generated transition matrices are stationary using TestStationarity() with a rejection sampling approach.

Usage

SimBetaN(n, beta, vcov_beta_vec_l)

Arguments

n

Positive integer. Number of replications.

beta

Numeric matrix. The transition matrix (\(\boldsymbol{\beta}\)).

vcov_beta_vec_l

Numeric matrix. Cholesky factorization (t(chol(vcov_beta_vec))) of the sampling variance-covariance matrix of \(\mathrm{vec} \left( \boldsymbol{\beta} \right)\).

Value

Returns a list of random transition matrices.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

n <- 10
beta <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0, 0, 0.5
  ),
  nrow = 3
)
vcov_beta_vec_l <- t(chol(0.001 * diag(9)))
SimBetaN(n = n, beta = beta, vcov_beta_vec_l = vcov_beta_vec_l)
#> [[1]]
#>            [,1]         [,2]        [,3]
#> [1,]  0.6725908 -0.001067837 -0.06472688
#> [2,]  0.4339222  0.598248607  0.03152077
#> [3,] -0.1148166  0.392356415  0.53225375
#> 
#> [[2]]
#>            [,1]        [,2]        [,3]
#> [1,]  0.7211563 -0.02811988 -0.01461553
#> [2,]  0.5762658  0.59875269 -0.04222542
#> [3,] -0.1319610  0.42629276  0.50533896
#> 
#> [[3]]
#>             [,1]        [,2]          [,3]
#> [1,]  0.70204951 -0.03019389 -0.0148599972
#> [2,]  0.53822759  0.60550486 -0.0009637857
#> [3,] -0.07452351  0.43917744  0.5160152581
#> 
#> [[4]]
#>            [,1]       [,2]        [,3]
#> [1,]  0.6852822 0.05638168 0.001401591
#> [2,]  0.4568725 0.60321232 0.039885923
#> [3,] -0.1272571 0.46733423 0.512104329
#> 
#> [[5]]
#>            [,1]       [,2]        [,3]
#> [1,]  0.7054159 0.03904114 0.034901343
#> [2,]  0.5000021 0.63616728 0.006509402
#> [3,] -0.1200285 0.45736071 0.494628159
#> 
#> [[6]]
#>            [,1]       [,2]        [,3]
#> [1,]  0.7177862 0.04439135 -0.03078154
#> [2,]  0.4878249 0.59226451  0.01497701
#> [3,] -0.1159894 0.39379345  0.54608043
#> 
#> [[7]]
#>            [,1]        [,2]        [,3]
#> [1,]  0.7252945 -0.03631255  0.02246809
#> [2,]  0.4655519  0.61618050 -0.03148214
#> [3,] -0.0875958  0.35450922  0.52242287
#> 
#> [[8]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.68349206 -0.04973111 0.027437274
#> [2,]  0.43693983  0.59772075 0.005540657
#> [3,] -0.08628195  0.42645314 0.522051822
#> 
#> [[9]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.66948154 -0.01187867  0.05434223
#> [2,]  0.53249946  0.58767516 -0.02980019
#> [3,] -0.08334942  0.44992734  0.47454086
#> 
#> [[10]]
#>            [,1]        [,2]         [,3]
#> [1,]  0.7078639 -0.05207038  0.009697993
#> [2,]  0.4801459  0.61570994 -0.003815280
#> [3,] -0.1031305  0.37167562  0.490623981
#>