Simulate Transition Matrices from the Multivariate Normal Distribution and Project to Stability

This function simulates random transition matrices from the multivariate normal distribution then projects each draw to the stability region using ProjectToStability().

Usage

SimBetaN2(n, beta, vcov_beta_vec_l, margin = 0.98, tol = 1e-12)

Arguments

n

Positive integer. Number of replications.

beta

Numeric matrix. The transition matrix (\(\boldsymbol{\beta}\)).

vcov_beta_vec_l

Numeric matrix. Cholesky factorization (t(chol(vcov_beta_vec))) of the sampling variance-covariance matrix of \(\mathrm{vec} \left( \boldsymbol{\beta} \right)\).

margin

Double in \((0, 1)\). Target upper bound for the spectral radius (default = 0.98).

tol

Small positive double added to the denominator in the scaling factor to avoid division by zero (default = 1e-12).

Value

Returns a list of random transition matrices.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

n <- 10
beta <- matrix(
  data = c(
    0.7, 0.5, -0.1,
    0.0, 0.6, 0.4,
    0, 0, 0.5
  ),
  nrow = 3
)
vcov_beta_vec_l <- t(chol(0.001 * diag(9)))
SimBetaN2(n = n, beta = beta, vcov_beta_vec_l = vcov_beta_vec_l)
#> [[1]]
#>            [,1]       [,2]        [,3]
#> [1,]  0.7077795 0.01438246 -0.02230688
#> [2,]  0.5052252 0.57989453  0.02458508
#> [3,] -0.1048454 0.37969557  0.48802190
#> 
#> [[2]]
#>             [,1]      [,2]         [,3]
#> [1,]  0.71605138 0.0155194 -0.019794236
#> [2,]  0.54630413 0.5923103  0.007366145
#> [3,] -0.08015487 0.3663594  0.534151590
#> 
#> [[3]]
#>            [,1]        [,2]        [,3]
#> [1,]  0.7122546 -0.03391019  0.05249404
#> [2,]  0.5212208  0.58444080 -0.04965351
#> [3,] -0.0494798  0.40400738  0.53639433
#> 
#> [[4]]
#>             [,1]        [,2]         [,3]
#> [1,]  0.74652801 0.007668595  0.008189777
#> [2,]  0.46546768 0.624127584 -0.049843932
#> [3,] -0.08198905 0.437405757  0.509051799
#> 
#> [[5]]
#>            [,1]       [,2]        [,3]
#> [1,]  0.7430272 0.01507708  0.01208606
#> [2,]  0.5261038 0.59688925 -0.04475027
#> [3,] -0.1033991 0.39305056  0.52310965
#> 
#> [[6]]
#>            [,1]         [,2]        [,3]
#> [1,]  0.7532249 -0.007071082 0.004734181
#> [2,]  0.5764983  0.571244635 0.012059007
#> [3,] -0.1255910  0.416738663 0.462439962
#> 
#> [[7]]
#>            [,1]      [,2]       [,3]
#> [1,]  0.7051078 0.0533864 0.03567517
#> [2,]  0.5501585 0.6082728 0.01117277
#> [3,] -0.1657389 0.4114660 0.55195525
#> 
#> [[8]]
#>            [,1]      [,2]         [,3]
#> [1,]  0.7192515 0.0401671  0.034624353
#> [2,]  0.4656795 0.5813792 -0.008413454
#> [3,] -0.0940404 0.3880602  0.530368234
#> 
#> [[9]]
#>            [,1]        [,2]        [,3]
#> [1,]  0.7004622 -0.03370255 -0.01332358
#> [2,]  0.5142915  0.61462578 -0.03574804
#> [3,] -0.1142357  0.38315191  0.51346426
#> 
#> [[10]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.67565413 -0.04383784 0.006704065
#> [2,]  0.51219682  0.57388074 0.014437401
#> [3,] -0.06396186  0.44816789 0.468903941
#>