Simulate Data from a State Space Model (n = 1)

This function simulates data from a state space model. See details for more information.

Usage

SimSSM(
  mu0,
  sigma0_sqrt,
  alpha,
  beta,
  psi_sqrt,
  nu,
  lambda,
  theta_sqrt,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  time,
  burn_in = 0
)

Arguments

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_sqrt

Numeric matrix. Cholesky decomposition of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

alpha

Numeric vector. Vector of intercepts for the dynamic model (\(\boldsymbol{\alpha}\)).

beta

Numeric matrix. Transition matrix relating the values of the latent variables at time t - 1 to those at time t (\(\boldsymbol{\beta}\)).

psi_sqrt

Numeric matrix. Cholesky decomposition of the process noise covariance matrix (\(\boldsymbol{\Psi}\)).

nu

Numeric vector. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta_sqrt

Numeric matrix. Cholesky decomposition of the measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the observed variables at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

Numeric matrix. The matrix of observed covariates in type = 1 or type = 2. The number of rows should be equal to time + burn_in.

type

Integer. State space model type.

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

time: A vector of discrete time points from 0 to t - 1.

id: A vector of ones.

Details

Type 0

The measurement model is given by $$ \mathbf{y}_{t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{t} + \boldsymbol{\varepsilon}_{t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\mathbf{y}_{t}\), \(\boldsymbol{\eta}_{t}\), and \(\boldsymbol{\varepsilon}_{t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{t}\) is a vector of observed random variables, \(\boldsymbol{\eta}_{t}\) is a vector of latent random variables, and \(\boldsymbol{\varepsilon}_{t}\) is a vector of random measurement errors, at time \(t\). \(\boldsymbol{\nu}\) is a vector of intercepts, \(\boldsymbol{\Lambda}\) is a matrix of factor loadings, and \(\boldsymbol{\Theta}\) is the covariance matrix of \(\boldsymbol{\varepsilon}\).

The dynamic structure is given by $$ \boldsymbol{\eta}_{t} = \boldsymbol{\alpha} + \boldsymbol{\beta} \boldsymbol{\eta}_{t - 1} + \boldsymbol{\zeta}_{t} \quad \mathrm{with} \quad \boldsymbol{\zeta}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Psi} \right) $$ where \(\boldsymbol{\eta}_{t}\), \(\boldsymbol{\eta}_{t - 1}\), and \(\boldsymbol{\zeta}_{t}\) are random variables, and \(\boldsymbol{\alpha}\), \(\boldsymbol{\beta}\), and \(\boldsymbol{\Psi}\) are model parameters. \(\boldsymbol{\eta}_{t}\) is a vector of latent variables at time \(t\), \(\boldsymbol{\eta}_{t - 1}\) is a vector of latent variables at time \(t - 1\), and \(\boldsymbol{\zeta}_{t}\) is a vector of dynamic noise at time \(t\). \(\boldsymbol{\alpha}\) is a vector of intercepts, \(\boldsymbol{\beta}\) is a matrix of autoregression and cross regression coefficients, and \(\boldsymbol{\Psi}\) is the covariance matrix of \(\boldsymbol{\zeta}_{t}\).

Type 1

The measurement model is given by $$ \mathbf{y}_{t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{t} + \boldsymbol{\varepsilon}_{t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) . $$

The dynamic structure is given by $$ \boldsymbol{\eta}_{t} = \boldsymbol{\alpha} + \boldsymbol{\beta} \boldsymbol{\eta}_{t - 1} + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{t} + \boldsymbol{\zeta}_{t} \quad \mathrm{with} \quad \boldsymbol{\zeta}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Psi} \right) $$ where \(\mathbf{x}_{t}\) is a vector of covariates at time \(t\), and \(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\) is the coefficient matrix linking the covariates to the latent variables.

Type 2

The measurement model is given by $$ \mathbf{y}_{t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{t} + \boldsymbol{\Gamma}_{\mathbf{y}} \mathbf{x}_{t} + \boldsymbol{\varepsilon}_{t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\boldsymbol{\Gamma}_{\mathbf{y}}\) is the coefficient matrix linking the covariates to the observed variables.

The dynamic structure is given by $$ \boldsymbol{\eta}_{t} = \boldsymbol{\alpha} + \boldsymbol{\beta} \boldsymbol{\eta}_{t - 1} + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{t} + \boldsymbol{\zeta}_{t} \quad \mathrm{with} \quad \boldsymbol{\zeta}_{t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Psi} \right) . $$

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Examples

# prepare parameters
set.seed(42)
k <- p <- 3
iden <- diag(k)
iden_sqrt <- chol(iden)
null_vec <- rep(x = 0, times = k)
mu0 <- null_vec
sigma0_sqrt <- iden_sqrt
alpha <- null_vec
beta <- diag(x = 0.50, nrow = k)
psi_sqrt <- iden_sqrt
nu <- null_vec
lambda <- iden
theta_sqrt <- chol(diag(x = 0.50, nrow = k))
time <- 50
burn_in <- 0
gamma_y <- gamma_eta <- 0.10 * diag(k)
x <- matrix(
  data = rnorm(n = k * (time + burn_in)),
  ncol = k
)

# Type 0
ssm <- SimSSM(
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  type = 0,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ y   : num [1:50, 1:3] -1.072 -1.328 0.489 2.603 1.038 ...
#>  $ eta : num [1:50, 1:3] -0.826 -0.42 -0.104 1.976 2.156 ...
#>  $ x   : int 0
#>  $ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>  $ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...

# Type 1
ssm <- SimSSM(
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ y   : num [1:50, 1:3] -0.2107 0.8796 0.6269 -0.4614 0.0537 ...
#>  $ eta : num [1:50, 1:3] 0.117 -0.129 0.541 0.765 0.525 ...
#>  $ x   : num [1:50, 1:3] 1.371 -0.565 0.363 0.633 0.404 ...
#>  $ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>  $ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...

# Type 2
ssm <- SimSSM(
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ y   : num [1:50, 1:3] 1.8463 1.3891 -0.0868 -1.0306 -2.2401 ...
#>  $ eta : num [1:50, 1:3] 1.079 0.274 -0.412 -1.037 -2.143 ...
#>  $ x   : num [1:50, 1:3] 1.371 -0.565 0.363 0.633 0.404 ...
#>  $ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>  $ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...