Simulate Data from a Linear Growth Curve Model

This function simulates data from a linear growth curve model.

Usage

SimSSMLinGrowth(
  n,
  mu0,
  sigma0_sqrt,
  theta_sqrt,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  time
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. A vector of length two. The first element is the mean of the intercept, and the second element is the mean of the slope.

sigma0_sqrt

Numeric matrix. Cholesky decomposition of the covariance matrix of the intercept and the slope.

theta_sqrt

Numeric. Square root of the common measurement error variance.

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to y at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables (intercept and slope) at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

A list of length n of numeric matrices. Each element of the list is a matrix of observed covariates in type = 1 or type = 2. The number of rows in each matrix should be equal to time.

type

Integer. State space model type.

time

Positive integer. Number of time points to simulate.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

time: A vector of discrete time points from 1 to t.

id: A vector of ID numbers of length t.

Details

Type 0

The measurement model is given by $$ y_{i, t} = \left( \begin{array}{cc} 1 & 0 \\ \end{array} \right) \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( 0, \theta^{2} \right) $$ where \(y_{i, t}\), \(\eta_{0_{i, t}}\), \(\eta_{1_{i, t}}\), and \(\boldsymbol{\varepsilon}_{i, t}\) are random variables and and \(\theta^{2}\) is a model parameter. \(y_{i, t}\) is a vector of observed random variables at time \(t\) and individual \(i\), \(\eta_{0_{i, t}}\), \(\eta_{1_{i, t}}\) is forms a vector of latent random variables at time \(t\) and individual \(i\), and \(\boldsymbol{\varepsilon}_{i, t}\) is a vector of random measurement errors at time \(t\) and individual \(i\), and \(\theta^{2}\) is the variance of \(\boldsymbol{\varepsilon}\).

The dynamic structure is given by $$ \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) = \left( \begin{array}{cc} 1 & 1 \\ 0 & 1 \\ \end{array} \right) \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) . $$

The mean vector and covariance matrix of the intercept and slope are captured in the mean vector and covariance matrix of the initial condition given by $$ \boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0} = \left( \begin{array}{c} \mu_{\eta_{0}} \\ \mu_{\eta_{1}} \\ \end{array} \right) \quad \mathrm{and,} $$

$$ \boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0} = \left( \begin{array}{cc} \sigma^{2}_{\eta_{0}} & \sigma_{\eta_{0}, \eta_{1}} \\ \sigma_{\eta_{1}, \eta_{0}} & \sigma^{2}_{\eta_{1}} \\ \end{array} \right) . $$

Type 1

The measurement model is given by $$ y_{i, t} = \left( \begin{array}{cc} 1 & 0 \\ \end{array} \right) \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( 0, \theta^{2} \right) . $$

The dynamic structure is given by $$ \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) = \left( \begin{array}{cc} 1 & 1 \\ 0 & 1 \\ \end{array} \right) \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{i, t} $$ where \(\mathbf{x}_{i, t}\) is a vector of covariates at time \(t\) and individual \(i\), and \(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\) is the coefficient matrix linking the covariates to the latent variables.

Type 2

The measurement model is given by $$ y_{i, t} = \left( \begin{array}{cc} 1 & 0 \\ \end{array} \right) \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) + \boldsymbol{\Gamma}_{\mathbf{y}} \mathbf{x}_{i, t} + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( 0, \theta^{2} \right) $$ where \(\boldsymbol{\Gamma}_{\mathbf{y}}\) is the coefficient matrix linking the covariates to the observed variables.

The dynamic structure is given by $$ \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) = \left( \begin{array}{cc} 1 & 1 \\ 0 & 1 \\ \end{array} \right) \left( \begin{array}{c} \eta_{0_{i, t}} \\ \eta_{1_{i, t}} \\ \end{array} \right) + \boldsymbol{\Gamma}_{\boldsymbol{\eta}} \mathbf{x}_{i, t} . $$

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
n <- 5
mu0 <- c(0.615, 1.006)
sigma0 <- matrix(
  data = c(
    1.932,
    0.618,
    0.618,
    0.587
  ),
  nrow = 2
)
sigma0_sqrt <- chol(sigma0)
theta <- 0.6
theta_sqrt <- sqrt(theta)
time <- 10
gamma_y <- matrix(data = 0.10, nrow = 1, ncol = 2)
gamma_eta <- matrix(data = 0.10, nrow = 2, ncol = 2)
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = 2 * time),
        ncol = 2
      )
    )
  }
)

# Type 0
ssm <- SimSSMLinGrowth(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  theta_sqrt = theta_sqrt,
  type = 0,
  time = time
)

str(ssm)
#> List of 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.94 2.16 4.52 4.48 6.64 ...
#>   ..$ eta : num [1:10, 1:2] 2.05 3.09 4.12 5.15 6.19 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.504 1.832 3.289 6.111 5.759 ...
#>   ..$ eta : num [1:10, 1:2] 0.93 2.28 3.63 4.97 6.32 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.195 0.651 1.322 2.499 3.855 ...
#>   ..$ eta : num [1:10, 1:2] -0.707 0.362 1.43 2.499 3.568 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.127 0.865 1.441 2.581 1.49 ...
#>   ..$ eta : num [1:10, 1:2] 0.264 0.872 1.48 2.088 2.697 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -1.426 -2.16 -1.861 -1.14 -0.232 ...
#>   ..$ eta : num [1:10, 1:2] -1.42 -1.144 -0.867 -0.59 -0.314 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5

# Type 1
ssm <- SimSSMLinGrowth(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  theta_sqrt = theta_sqrt,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  time = time
)

str(ssm)
#> List of 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.45 3.3 3.47 2.63 5.95 ...
#>   ..$ eta : num [1:10, 1:2] 1.59 2.4 3.11 3.86 4.63 ...
#>   ..$ x   : num [1:10, 1:2] 1.371 -0.565 0.363 0.633 0.404 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -1.642 -1.015 -1.301 0.834 -1.417 ...
#>   ..$ eta : num [1:10, 1:2] -1.163 -1.217 -1.184 -1.091 -0.758 ...
#>   ..$ x   : num [1:10, 1:2] -0.307 -1.781 -0.172 1.215 1.895 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.452 0.347 1.277 2.452 4.074 ...
#>   ..$ eta : num [1:10, 1:2] 0.0238 0.9227 2.0551 3.179 4.1751 ...
#>   ..$ x   : num [1:10, 1:2] 0.206 -0.361 0.758 -0.727 -1.368 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 4.14 4.43 5.39 6.28 8.33 ...
#>   ..$ eta : num [1:10, 1:2] 3.4 4.51 5.74 7.01 8.15 ...
#>   ..$ x   : num [1:10, 1:2] -0.367 0.185 0.582 1.4 -0.727 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.24 3.58 6.05 7.38 9.18 ...
#>   ..$ eta : num [1:10, 1:2] 2.86 4.49 6.19 8.03 9.63 ...
#>   ..$ x   : num [1:10, 1:2] 1.5127 0.2579 0.0884 -0.1209 -1.1943 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5

# Type 2
ssm <- SimSSMLinGrowth(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  theta_sqrt = theta_sqrt,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  time = time
)

str(ssm)
#> List of 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.21 3.46 4.76 5.02 6.92 ...
#>   ..$ eta : num [1:10, 1:2] 1.87 2.98 3.98 5.01 6.08 ...
#>   ..$ x   : num [1:10, 1:2] 1.371 -0.565 0.363 0.633 0.404 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.975 -1.377 -1.548 1.25 1.311 ...
#>   ..$ eta : num [1:10, 1:2] -1.156 -0.723 -0.205 0.375 1.194 ...
#>   ..$ x   : num [1:10, 1:2] -0.307 -1.781 -0.172 1.215 1.895 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.424 0.632 -0.114 1.133 1.196 ...
#>   ..$ eta : num [1:10, 1:2] 0.136 0.213 0.523 0.824 0.998 ...
#>   ..$ x   : num [1:10, 1:2] 0.206 -0.361 0.758 -0.727 -1.368 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.0868 2.6387 1.523 3.0533 1.0046 ...
#>   ..$ eta : num [1:10, 1:2] 0.846 1.12 1.514 1.952 2.264 ...
#>   ..$ x   : num [1:10, 1:2] -0.367 0.185 0.582 1.4 -0.727 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.56935 1.13029 -0.00616 -2.35272 -2.56286 ...
#>   ..$ eta : num [1:10, 1:2] -0.218 -0.374 -0.455 -0.409 -0.594 ...
#>   ..$ x   : num [1:10, 1:2] 1.5127 0.2579 0.0884 -0.1209 -1.1943 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5