Simulate Data from a Linear Growth Curve Model (Individual-Varying Parameters)

This function simulates data from a linear growth curve model for n > 1 individuals. In this model, the parameters can vary across individuals.

Usage

SimSSMLinGrowthIVary(
  n,
  mu0,
  sigma0,
  theta,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  time
)

Arguments

n

Positive integer. Number of individuals.

mu0

A list of numeric vectors. Each element of the list is a vector of length two. The first element is the mean of the intercept, and the second element is the mean of the slope.

sigma0

A list of numeric matrices. Each element of the list is the covariance matrix of the intercept and the slope.

theta

A list numeric values. Each element of the list is the common measurement error variance.

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to y at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables (intercept and slope) at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

A list of length n of numeric matrices. Each element of the list is a matrix of observed covariates in type = 1 or type = 2. The number of rows in each matrix should be equal to time.

type

Integer. State space model type.

time

Positive integer. Number of time points to simulate.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

time: A vector of discrete time points from 1 to t.

id: A vector of ID numbers of length t.

Details

Parameters can vary across individuals by providing a list of parameter values. If the length of any of the parameters (mu0, sigma0, mu, theta, gamma_y, or gamma_eta) is less the n, the function will cycle through the available values.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
# In this example, the mean vector of the intercept and slope vary.
# Specifically,
# there are two sets of values representing two latent classes.
set.seed(42)
n <- 10
mu0_1 <- c(0.615, 1.006) # lower starting point, higher growth
mu0_2 <- c(1.000, 0.500) # higher starting point, lower growth
mu0 <- list(mu0_1, mu0_2)
sigma0 <- list(
  matrix(
    data = c(
      1.932,
      0.618,
      0.618,
      0.587
    ),
    nrow = 2
  )
)
theta <- list(0.6)
time <- 10
gamma_y <- list(matrix(data = 0.10, nrow = 1, ncol = 2))
gamma_eta <- list(matrix(data = 0.10, nrow = 2, ncol = 2))
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = 2 * time),
        ncol = 2
      )
    )
  }
)

# Type 0
ssm <- SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  theta = theta,
  type = 0,
  time = time
)

str(ssm)
#> List of 10
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.447 0.266 0.885 1.759 3.208 ...
#>   ..$ eta : num [1:10, 1:2] 0.0185 0.8411 1.6637 2.4864 3.309 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 4.46 5.1 6.3 7.38 9.75 ...
#>   ..$ eta : num [1:10, 1:2] 3.72 5.19 6.65 8.11 9.57 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.78 3.71 6.7 8.42 10.85 ...
#>   ..$ eta : num [1:10, 1:2] 2.4 4.62 6.84 9.07 11.29 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.65 3.47 4.88 4.81 6.5 ...
#>   ..$ eta : num [1:10, 1:2] 2.31 3.16 4 4.84 5.69 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.643 -1.291 -2.094 0.23 -0.628 ...
#>   ..$ eta : num [1:10, 1:2] -0.824 -0.744 -0.664 -0.584 -0.505 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.39 1.353 -0.333 0.572 0.298 ...
#>   ..$ eta : num [1:10, 1:2] 1.1019 0.8196 0.5373 0.255 -0.0273 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 6 6 6 6 6 6 6 6 6 6
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.442 3.391 2.278 3.952 2.271 ...
#>   ..$ eta : num [1:10, 1:2] 1.37 1.88 2.39 2.9 3.4 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 7 7 7 7 7 7 7 7 7 7
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.7662 1.8579 0.0608 -3.0309 -3.3449 ...
#>   ..$ eta : num [1:10, 1:2] 0.978 0.332 -0.314 -0.96 -1.606 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 8 8 8 8 8 8 8 8 8 8
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -1.76 1.796 0.769 1.684 3.613 ...
#>   ..$ eta : num [1:10, 1:2] -0.519 0.412 1.343 2.274 3.205 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 9 9 9 9 9 9 9 9 9 9
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.56 1.6 -0.82 -1.88 -2.39 ...
#>   ..$ eta : num [1:10, 1:2] 0.998 0.112 -0.774 -1.66 -2.545 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 10 10 10 10 10 10 10 10 10 10

# Type 1
ssm <- SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  theta = theta,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  time = time
)

str(ssm)
#> List of 10
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.854 -0.304 1.705 2.251 4.927 ...
#>   ..$ eta : num [1:10, 1:2] -0.161 0.833 1.724 2.65 3.604 ...
#>   ..$ x   : num [1:10, 1:2] 1.371 -0.565 0.363 0.633 0.404 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.541 -0.525 -0.72 0.721 2.734 ...
#>   ..$ eta : num [1:10, 1:2] -0.227 0.19 0.693 1.257 2.061 ...
#>   ..$ x   : num [1:10, 1:2] -0.307 -1.781 -0.172 1.215 1.895 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -3.1 -2.76 -3.64 -4.2 -3.44 ...
#>   ..$ eta : num [1:10, 1:2] -2.15 -2.68 -2.98 -3.29 -3.73 ...
#>   ..$ x   : num [1:10, 1:2] 0.206 -0.361 0.758 -0.727 -1.368 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.175 -0.135 0.19 -0.503 -3.475 ...
#>   ..$ eta : num [1:10, 1:2] 0.289 -0.27 -0.708 -1.102 -1.623 ...
#>   ..$ x   : num [1:10, 1:2] -0.367 0.185 0.582 1.4 -0.727 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -1.199 0.908 1.932 2.588 4.184 ...
#>   ..$ eta : num [1:10, 1:2] -1.066 0.138 1.417 2.823 3.998 ...
#>   ..$ x   : num [1:10, 1:2] 1.5127 0.2579 0.0884 -0.1209 -1.1943 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.72 3.89 6.4 7.17 10.91 ...
#>   ..$ eta : num [1:10, 1:2] 3.08 4.87 6.5 8.27 9.81 ...
#>   ..$ x   : num [1:10, 1:2] 1.201 1.045 -1.003 1.848 -0.667 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 6 6 6 6 6 6 6 6 6 6
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.219 0.404 2.449 3.117 3.752 ...
#>   ..$ eta : num [1:10, 1:2] 1.12 1.75 2.39 3.04 3.83 ...
#>   ..$ x   : num [1:10, 1:2] -1.49363 -1.47044 0.1247 -0.99664 -0.00182 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 7 7 7 7 7 7 7 7 7 7
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.132 -0.481 -0.476 -0.737 -2.755 ...
#>   ..$ eta : num [1:10, 1:2] -0.452 -0.709 -0.893 -1.108 -1.412 ...
#>   ..$ x   : num [1:10, 1:2] -0.0861 -0.8877 -0.4447 -0.0294 -0.4139 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 8 8 8 8 8 8 8 8 8 8
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.131 2.182 2.483 5.294 5.932 ...
#>   ..$ eta : num [1:10, 1:2] 0.71 2.1 3.32 4.68 6.19 ...
#>   ..$ x   : num [1:10, 1:2] -0.176 -1.072 0.163 -0.363 0.59 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 9 9 9 9 9 9 9 9 9 9
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.618 0.284 2.484 2.064 1.065 ...
#>   ..$ eta : num [1:10, 1:2] 0.117 0.732 1.41 1.988 2.441 ...
#>   ..$ x   : num [1:10, 1:2] 0.151 -0.584 0.369 0.295 -0.279 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 10 10 10 10 10 10 10 10 10 10

# Type 2
ssm <- SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0 = sigma0,
  theta = theta,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  time = time
)

str(ssm)
#> List of 10
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.353 4.682 3.016 5.827 6.738 ...
#>   ..$ eta : num [1:10, 1:2] 1.75 3.05 4.26 5.5 6.76 ...
#>   ..$ x   : num [1:10, 1:2] 1.371 -0.565 0.363 0.633 0.404 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.13 2.04 1.31 1.78 1.58 ...
#>   ..$ eta : num [1:10, 1:2] 1.12 1.13 1.23 1.39 1.79 ...
#>   ..$ x   : num [1:10, 1:2] -0.307 -1.781 -0.172 1.215 1.895 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -1.5 -1.31 -2.66 -0.933 -1.751 ...
#>   ..$ eta : num [1:10, 1:2] -0.827 -1.159 -1.257 -1.364 -1.598 ...
#>   ..$ x   : num [1:10, 1:2] 0.206 -0.361 0.758 -0.727 -1.368 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.6885 1.847 -0.0963 1.3571 0.0605 ...
#>   ..$ eta : num [1:10, 1:2] -0.115 0.155 0.545 0.98 1.288 ...
#>   ..$ x   : num [1:10, 1:2] -0.367 0.185 0.582 1.4 -0.727 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.18 4.59 3.83 7.09 8.89 ...
#>   ..$ eta : num [1:10, 1:2] 1.83 3.43 5.11 6.91 8.48 ...
#>   ..$ x   : num [1:10, 1:2] 1.5127 0.2579 0.0884 -0.1209 -1.1943 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 3.12 3.78 3.5 4.27 4.03 ...
#>   ..$ eta : num [1:10, 1:2] 2.89 3.23 3.43 3.76 3.85 ...
#>   ..$ x   : num [1:10, 1:2] 1.201 1.045 -1.003 1.848 -0.667 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 6 6 6 6 6 6 6 6 6 6
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.863 0.98 2.941 2.14 6.67 ...
#>   ..$ eta : num [1:10, 1:2] -0.245 1.335 2.927 4.532 6.28 ...
#>   ..$ x   : num [1:10, 1:2] -1.49363 -1.47044 0.1247 -0.99664 -0.00182 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 7 7 7 7 7 7 7 7 7 7
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.883 1.108 2.25 2.92 0.814 ...
#>   ..$ eta : num [1:10, 1:2] 0.896 1.363 1.902 2.41 2.831 ...
#>   ..$ x   : num [1:10, 1:2] -0.0861 -0.8877 -0.4447 -0.0294 -0.4139 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 8 8 8 8 8 8 8 8 8 8
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -1.79 -1.86 -4.97 -4.98 -4.16 ...
#>   ..$ eta : num [1:10, 1:2] -2.16 -2.95 -3.89 -4.7 -5.36 ...
#>   ..$ x   : num [1:10, 1:2] -0.176 -1.072 0.163 -0.363 0.59 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 9 9 9 9 9 9 9 9 9 9
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.94 2.18 1.41 2.55 1.7 ...
#>   ..$ eta : num [1:10, 1:2] -0.089 0.526 1.203 1.78 2.234 ...
#>   ..$ x   : num [1:10, 1:2] 0.151 -0.584 0.369 0.295 -0.279 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 10 10 10 10 10 10 10 10 10 10