Simulate Data from a Linear Growth Curve Model (Individual-Varying Parameters)

This function simulates data from a linear growth curve model for n > 1 individuals. In this model, the parameters can vary across individuals.

Usage

SimSSMLinGrowthIVary(
  n,
  mu0,
  sigma0_sqrt,
  theta_sqrt,
  gamma_y = NULL,
  gamma_eta = NULL,
  x = NULL,
  type = 0,
  time
)

Arguments

n

Positive integer. Number of individuals.

mu0

A list of numeric vectors. Each element of the list is a vector of length two. The first element is the mean of the intercept, and the second element is the mean of the slope.

sigma0_sqrt

A list of numeric matrices. Each element of the list is the Cholesky decomposition of the covariance matrix of the intercept and the slope.

theta_sqrt

A list numeric values. Each element of the list is the square root of the common measurement error variance.

gamma_y

Numeric matrix. Matrix relating the values of the covariate matrix at time t to y at time t (\(\boldsymbol{\Gamma}_{\mathbf{y}}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables (intercept and slope) at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

A list of length n of numeric matrices. Each element of the list is a matrix of observed covariates in type = 1 or type = 2. The number of rows in each matrix should be equal to time.

type

Integer. State space model type.

time

Positive integer. Number of time points to simulate.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

time: A vector of discrete time points from 1 to t.

id: A vector of ID numbers of length t.

Details

Parameters can vary across individuals by providing a list of parameter values. If the length of any of the parameters (mu0, sigma0_sqrt, mu, theta_sqrt, gamma_y, or gamma_eta) is less the n, the function will cycle through the available values.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
# In this example, the mean vector of the intercept and slope vary.
# Specifically, there are two sets of values representing two latent classes.
set.seed(42)
n <- 10
mu0_1 <- c(0.615, 1.006) # lower starting point, higher growth
mu0_2 <- c(1.000, 0.500) # higher starting point, lower growth
mu0 <- list(mu0_1, mu0_2)
sigma0 <- matrix(
  data = c(
    1.932,
    0.618,
    0.618,
    0.587
  ),
  nrow = 2
)
sigma0_sqrt <- list(chol(sigma0))
theta <- 0.6
theta_sqrt <- list(sqrt(theta))
time <- 10
gamma_y <- list(matrix(data = 0.10, nrow = 1, ncol = 2))
gamma_eta <- list(matrix(data = 0.10, nrow = 2, ncol = 2))
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = 2 * time),
        ncol = 2
      )
    )
  }
)

# Type 0
ssm <- SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  theta_sqrt = theta_sqrt,
  type = 0,
  time = time
)

str(ssm)
#> List of 10
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.452 0.462 1.272 2.337 3.977 ...
#>   ..$ eta : num [1:10, 1:2] 0.0238 1.0372 2.0506 3.0641 4.0775 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 4.53 4.3 4.62 4.83 6.33 ...
#>   ..$ eta : num [1:10, 1:2] 3.79 4.38 4.97 5.56 6.15 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.24 3.61 6.02 7.17 9.03 ...
#>   ..$ eta : num [1:10, 1:2] 2.86 4.51 6.16 7.82 9.47 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.6 3 3.99 3.5 4.78 ...
#>   ..$ eta : num [1:10, 1:2] 2.26 2.68 3.11 3.54 3.96 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.975 -1.162 -1.505 1.28 0.881 ...
#>   ..$ eta : num [1:10, 1:2] -1.156 -0.6158 -0.0756 0.4646 1.0048 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.809 0.74 -0.978 -0.106 -0.413 ...
#>   ..$ eta : num [1:10, 1:2] 0.521 0.206 -0.109 -0.423 -0.738 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 6 6 6 6 6 6 6 6 6 6
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.0868 2.6197 1.2629 2.6945 0.7698 ...
#>   ..$ eta : num [1:10, 1:2] 0.846 1.11 1.374 1.638 1.902 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 7 7 7 7 7 7 7 7 7 7
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.954 1.053 -0.737 -3.822 -4.129 ...
#>   ..$ eta : num [1:10, 1:2] 0.167 -0.473 -1.112 -1.751 -2.391 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 8 8 8 8 8 8 8 8 8 8
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -1.56 2.36 1.7 2.98 5.27 ...
#>   ..$ eta : num [1:10, 1:2] -0.314 0.98 2.274 3.567 4.861 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 9 9 9 9 9 9 9 9 9 9
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.572 0.618 -1.806 -2.865 -3.377 ...
#>   ..$ eta : num [1:10, 1:2] 0.0109 -0.8743 -1.7594 -2.6446 -3.5298 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 10 10 10 10 10 10 10 10 10 10

# Type 1
ssm <- SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  theta_sqrt = theta_sqrt,
  gamma_eta = gamma_eta,
  x = x,
  type = 1,
  time = time
)

str(ssm)
#> List of 10
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.9384 -0.0546 2.2886 3.1674 6.178 ...
#>   ..$ eta : num [1:10, 1:2] -0.245 1.082 2.307 3.567 4.855 ...
#>   ..$ x   : num [1:10, 1:2] 1.371 -0.565 0.363 0.633 0.404 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.83 0.162 0.365 2.203 4.613 ...
#>   ..$ eta : num [1:10, 1:2] 0.0632 0.8773 1.7777 2.7387 3.9397 ...
#>   ..$ x   : num [1:10, 1:2] -0.307 -1.781 -0.172 1.215 1.895 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -3.512 -2.273 -2.249 -1.914 -0.257 ...
#>   ..$ eta : num [1:10, 1:2] -2.56 -2.19 -1.59 -1 -0.54 ...
#>   ..$ x   : num [1:10, 1:2] 0.206 -0.361 0.758 -0.727 -1.368 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.7061 -0.4834 0.0235 -0.4874 -3.277 ...
#>   ..$ eta : num [1:10, 1:2] -0.242 -0.619 -0.875 -1.087 -1.425 ...
#>   ..$ x   : num [1:10, 1:2] -0.367 0.185 0.582 1.4 -0.727 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.8 1.94 3.59 4.88 7.11 ...
#>   ..$ eta : num [1:10, 1:2] -0.667 1.169 3.078 5.115 6.92 ...
#>   ..$ x   : num [1:10, 1:2] 1.5127 0.2579 0.0884 -0.1209 -1.1943 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 3.03 3.56 5.44 5.59 8.7 ...
#>   ..$ eta : num [1:10, 1:2] 3.38 4.54 5.55 6.69 7.6 ...
#>   ..$ x   : num [1:10, 1:2] 1.201 1.045 -1.003 1.848 -0.667 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 6 6 6 6 6 6 6 6 6 6
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.0181 0.0119 1.865 2.3416 2.785 ...
#>   ..$ eta : num [1:10, 1:2] 0.92 1.35 1.8 2.26 2.87 ...
#>   ..$ x   : num [1:10, 1:2] -1.49363 -1.47044 0.1247 -0.99664 -0.00182 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 7 7 7 7 7 7 7 7 7 7
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.1036 -0.0495 0.4162 0.6153 -0.9416 ...
#>   ..$ eta : num [1:10, 1:2] -0.48032 -0.27687 -0.00116 0.24423 0.40146 ...
#>   ..$ x   : num [1:10, 1:2] -0.0861 -0.8877 -0.4447 -0.0294 -0.4139 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 8 8 8 8 8 8 8 8 8 8
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.416 2.423 2.681 5.449 6.043 ...
#>   ..$ eta : num [1:10, 1:2] 0.995 2.337 3.521 4.838 6.3 ...
#>   ..$ x   : num [1:10, 1:2] -0.176 -1.072 0.163 -0.363 0.59 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 9 9 9 9 9 9 9 9 9 9
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.286 0.91 3.404 3.278 2.572 ...
#>   ..$ eta : num [1:10, 1:2] 0.449 1.359 2.33 3.202 3.949 ...
#>   ..$ x   : num [1:10, 1:2] 0.151 -0.584 0.369 0.295 -0.279 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 10 10 10 10 10 10 10 10 10 10

# Type 2
ssm <- SimSSMLinGrowthIVary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  theta_sqrt = theta_sqrt,
  gamma_y = gamma_y,
  gamma_eta = gamma_eta,
  x = x,
  type = 2,
  time = time
)

str(ssm)
#> List of 10
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.0564 4.1086 2.166 4.6995 5.3334 ...
#>   ..$ eta : num [1:10, 1:2] 1.45 2.48 3.41 4.37 5.36 ...
#>   ..$ x   : num [1:10, 1:2] 1.371 -0.565 0.363 0.633 0.404 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 1 1 1 1 1 1 1 1 1 1
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 0.827 1.703 0.942 1.371 1.137 ...
#>   ..$ eta : num [1:10, 1:2] 0.814 0.792 0.856 0.981 1.346 ...
#>   ..$ x   : num [1:10, 1:2] -0.307 -1.781 -0.172 1.215 1.895 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 2 2 2 2 2 2 2 2 2 2
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -2.0685 -1.4004 -2.2721 -0.0673 -0.4065 ...
#>   ..$ eta : num [1:10, 1:2] -1.396 -1.249 -0.869 -0.498 -0.254 ...
#>   ..$ x   : num [1:10, 1:2] 0.206 -0.361 0.758 -0.727 -1.368 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 3 3 3 3 3 3 3 3 3 3
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.537 2.31 0.678 2.443 1.458 ...
#>   ..$ eta : num [1:10, 1:2] 0.0368 0.6182 1.3201 2.0666 2.6862 ...
#>   ..$ x   : num [1:10, 1:2] -0.367 0.185 0.582 1.4 -0.727 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 4 4 4 4 4 4 4 4 4 4
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.2 4.32 3.26 6.23 7.73 ...
#>   ..$ eta : num [1:10, 1:2] 1.85 3.16 4.54 6.05 7.32 ...
#>   ..$ x   : num [1:10, 1:2] 1.5127 0.2579 0.0884 -0.1209 -1.1943 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 5 5 5 5 5 5 5 5 5 5
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 2.43 2.53 1.68 1.89 1.08 ...
#>   ..$ eta : num [1:10, 1:2] 2.208 1.984 1.612 1.374 0.903 ...
#>   ..$ x   : num [1:10, 1:2] 1.201 1.045 -1.003 1.848 -0.667 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 6 6 6 6 6 6 6 6 6 6
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -0.0736 2.0155 4.2214 3.6664 8.4418 ...
#>   ..$ eta : num [1:10, 1:2] 0.545 2.37 4.208 6.058 8.052 ...
#>   ..$ x   : num [1:10, 1:2] -1.49363 -1.47044 0.1247 -0.99664 -0.00182 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 7 7 7 7 7 7 7 7 7 7
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.06 1.32 2.49 3.19 1.11 ...
#>   ..$ eta : num [1:10, 1:2] 1.08 1.57 2.14 2.68 3.13 ...
#>   ..$ x   : num [1:10, 1:2] -0.0861 -0.8877 -0.4447 -0.0294 -0.4139 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 8 8 8 8 8 8 8 8 8 8
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] -2.4 -1.59 -3.82 -2.96 -1.26 ...
#>   ..$ eta : num [1:10, 1:2] -2.77 -2.68 -2.74 -2.67 -2.46 ...
#>   ..$ x   : num [1:10, 1:2] -0.176 -1.072 0.163 -0.363 0.59 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 9 9 9 9 9 9 9 9 9 9
#>  $ :List of 5
#>   ..$ y   : num [1:10, 1] 1.32 2.92 2.5 4.01 3.51 ...
#>   ..$ eta : num [1:10, 1:2] 0.29 1.26 2.3 3.24 4.05 ...
#>   ..$ x   : num [1:10, 1:2] 0.151 -0.584 0.369 0.295 -0.279 ...
#>   ..$ time: num [1:10, 1] 0 1 2 3 4 5 6 7 8 9
#>   ..$ id  : num [1:10, 1] 10 10 10 10 10 10 10 10 10 10