Simulate Data from an Ornstein–Uhlenbeck Model using a State Space Model Parameterization for n > 1 Individuals (Fixed Parameters)

This function simulates data from an Ornstein–Uhlenbeck model using a state space model parameterization for n > 1 individuals. In this model, the parameters are invariant across individuals. See details for more information.

Usage

SimSSMOUFixed(
  n,
  mu0,
  sigma0_sqrt,
  mu,
  phi,
  sigma_sqrt,
  nu,
  lambda,
  theta_sqrt,
  delta_t,
  time,
  burn_in
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_sqrt

Numeric matrix. Cholesky decomposition of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

mu

Numeric vector. The long-term mean or equilibrium level (\(\boldsymbol{\mu}\)).

phi

Numeric matrix. The rate of mean reversion, determining how quickly the variable returns to its mean (\(\boldsymbol{\Phi}\)).

sigma_sqrt

Numeric matrix. Cholesky decomposition of the matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

nu

Numeric vector. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta_sqrt

Numeric matrix. Cholesky decomposition of the measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

delta_t

Numeric. Time interval (\(\delta_t\)).

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

time: A vector of continuous time points of length t starting from 0 with delta_t increments.

id: A vector of ID numbers of length t.

n: Number of individuals.

Details

The measurement model is given by $$ \mathbf{y}_{i, t} = \boldsymbol{\nu} + \boldsymbol{\Lambda} \boldsymbol{\eta}_{i, t} + \boldsymbol{\varepsilon}_{i, t} \quad \mathrm{with} \quad \boldsymbol{\varepsilon}_{i, t} \sim \mathcal{N} \left( \mathbf{0}, \boldsymbol{\Theta} \right) $$ where \(\mathbf{y}_{i, t}\), \(\boldsymbol{\eta}_{i, t}\), and \(\boldsymbol{\varepsilon}_{i, t}\) are random variables and \(\boldsymbol{\nu}\), \(\boldsymbol{\Lambda}\), and \(\boldsymbol{\Theta}\) are model parameters. \(\mathbf{y}_{i, t}\) is a vector of observed random variables at time \(t\) and individual \(i\), \(\boldsymbol{\eta}_{i, t}\) is a vector of latent random variables at time \(t\) and individual \(i\), and \(\boldsymbol{\varepsilon}_{i, t}\) is a vector of random measurement errors at time \(t\) and individual \(i\), while \(\boldsymbol{\nu}\) is a vector of intercept, \(\boldsymbol{\Lambda}\) is a matrix of factor loadings, and \(\boldsymbol{\Theta}\) is the covariance matrix of \(\boldsymbol{\varepsilon}\).

The dynamic structure is given by $$ \mathrm{d} \boldsymbol{\eta}_{i, t} = \boldsymbol{\Phi} \left( \boldsymbol{\mu} - \boldsymbol{\eta}_{i, t} \right) \mathrm{d}t + \boldsymbol{\Sigma}^{\frac{1}{2}} \mathrm{d} \mathbf{W}_{i, t} $$ where \(\boldsymbol{\mu}\) is the long-term mean or equilibrium level, \(\boldsymbol{\Phi}\) is the rate of mean reversion, determining how quickly the variable returns to its mean, \(\boldsymbol{\Sigma}\) is the matrix of volatility or randomness in the process, and \(\mathrm{d}\boldsymbol{W}\) is a Wiener process or Brownian motion, which represents random fluctuations.

References

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Uhlenbeck, G. E., & Ornstein, L. S. (1930). On the theory of the brownian motion. Physical Review, 36(5), 823–841. doi:10.1103/physrev.36.823

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
p <- k <- 2
I <- diag(p)
I_sqrt <- chol(I)
n <- 5
mu0 <- c(-3.0, 1.5)
sigma0_sqrt <- I_sqrt
mu <- c(5.76, 5.18)
phi <- matrix(data = c(0.10, -0.05, -0.05, 0.10), nrow = p)
sigma_sqrt <- chol(
  matrix(data = c(2.79, 0.06, 0.06, 3.27), nrow = p)
)
nu <- rep(x = 0, times = k)
lambda <- diag(k)
theta_sqrt <- chol(diag(x = 0.50, nrow = k))
delta_t <- 0.10
time <- 50
burn_in <- 0

# generate data
ssm <- SimSSMOUFixed(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  mu = mu,
  phi = phi,
  sigma_sqrt = sigma_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -1.96 -2.55 -3.27 -4.04 -3.79 ...
#>   ..$ eta : num [1:50, 1:2] -3.53 -3.25 -3.33 -3.27 -3.63 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -2.73 -2.32 -2.85 -3.46 -2.49 ...
#>   ..$ eta : num [1:50, 1:2] -2.63 -2.58 -2.75 -2.28 -2.45 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.1 -4.49 -4.27 -4.06 -3.78 ...
#>   ..$ eta : num [1:50, 1:2] -4.85 -4.86 -4.16 -4.15 -4.05 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -3.917 -2.619 -1.685 -2.321 -0.666 ...
#>   ..$ eta : num [1:50, 1:2] -2.23 -2.28 -1.56 -1.32 -1.37 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -2.7 -3.27 -3.08 -4.15 -3.22 ...
#>   ..$ eta : num [1:50, 1:2] -3.03 -3.72 -3.47 -3.76 -2.53 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : num [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...