Simulate Data from an Ornstein–Uhlenbeck Model using a State Space Model Parameterization for n > 1 Individuals (Varying Parameters)

This function simulates data from an Ornstein–Uhlenbeck model using a state space model parameterization for n > 1 individuals. In this model, the parameters can vary across individuals.

Usage

SimSSMOUVary(
  n,
  mu0,
  sigma0_sqrt,
  mu,
  phi,
  sigma_sqrt,
  nu,
  lambda,
  theta_sqrt,
  delta_t,
  time,
  burn_in
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_sqrt

Numeric matrix. Cholesky decomposition of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

mu

List of numeric vectors. The long-term mean or equilibrium level (\(\boldsymbol{\mu}\)).

phi

List of numeric matrices. The rate of mean reversion, determining how quickly the variable returns to its mean (\(\boldsymbol{\Phi}\)).

sigma_sqrt

List of numeric matrices. Cholesky decomposition of the matrix of volatility or randomness in the process (\(\boldsymbol{\Sigma}\)).

nu

Numeric vector. Vector of intercepts for the measurement model (\(\boldsymbol{\nu}\)).

lambda

Numeric matrix. Factor loading matrix linking the latent variables to the observed variables (\(\boldsymbol{\Lambda}\)).

theta_sqrt

Numeric matrix. Cholesky decomposition of the measurement error covariance matrix (\(\boldsymbol{\Theta}\)).

delta_t

Numeric. Time interval (\(\delta_t\)).

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

time: A vector of discrete time points from 1 to t.

id: A vector of ID numbers of length t.

n: Number of individuals.

Details

Parameters can vary across individuals by providing a list of parameter values. If the length of any of the parameters (mu0, sigma0_sqrt, mu, phi, sigma_sqrt, nu, lambda, theta_sqrt) is less the n, the function will cycle through the available values.

References

Chow, S.-M., Losardo, D., Park, J., & Molenaar, P. C. M. (2023). Continuous-time dynamic models: Connections to structural equation models and other discrete-time models. In R. H. Hoyle (Ed.), Handbook of structural equation modeling (2nd ed.). The Guilford Press.

Uhlenbeck, G. E., & Ornstein, L. S. (1930). On the theory of the brownian motion. Physical Review, 36(5), 823–841. doi:10.1103/physrev.36.823

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
# In this example, phi varies across individuals
set.seed(42)
p <- k <- 2
iden <- diag(p)
iden_sqrt <- chol(iden)
n <- 5
mu0 <- list(c(-3.0, 1.5))
sigma0_sqrt <- list(iden_sqrt)
mu <- list(c(5.76, 5.18))
phi <- list(
  as.matrix(Matrix::expm(diag(x = -0.1, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.2, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.3, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.4, nrow = k))),
  as.matrix(Matrix::expm(diag(x = -0.5, nrow = k)))
)
sigma_sqrt <- list(
  chol(
    matrix(data = c(2.79, 0.06, 0.06, 3.27), nrow = p)
  )
)
nu <- list(rep(x = 0, times = k))
lambda <- list(diag(k))
theta_sqrt <- list(chol(diag(x = 0.50, nrow = k)))
delta_t <- 0.10
time <- 50
burn_in <- 0

ssm <- SimSSMOUVary(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  mu = mu,
  phi = phi,
  sigma_sqrt = sigma_sqrt,
  nu = nu,
  lambda = lambda,
  theta_sqrt = theta_sqrt,
  delta_t = delta_t,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -1.96 -1.83 -1.9 -2.09 -1.29 ...
#>   ..$ eta : num [1:50, 1:2] -3.53 -2.53 -1.96 -1.32 -1.13 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : int [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -2.725 -1.718 -1.691 -1.802 -0.365 ...
#>   ..$ eta : num [1:50, 1:2] -2.626 -1.98 -1.591 -0.623 -0.33 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : int [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -4.1 -3.82 -2.99 -2.25 -1.47 ...
#>   ..$ eta : num [1:50, 1:2] -4.85 -4.19 -2.88 -2.33 -1.73 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : int [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -3.917 -2.156 -0.812 -1.089 0.9 ...
#>   ..$ eta : num [1:50, 1:2] -2.2264 -1.8136 -0.6915 -0.0856 0.1976 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : int [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 4
#>   ..$ y   : num [1:50, 1:2] -2.7 -2.8 -2.17 -2.81 -1.52 ...
#>   ..$ eta : num [1:50, 1:2] -3.03 -3.25 -2.55 -2.42 -0.82 ...
#>   ..$ time: num [1:50, 1] 0 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9 ...
#>   ..$ id  : int [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...