Simulate Data from a Vector Autoregressive Model using a State Space Model Parameterization for n > 1 Individuals (Fixed Parameters)

This function simulates data from a vector autoregressive model using a state space model parameterization for n > 1 individuals. In this model, the parameters are invariant across individuals.

Usage

SimSSMVARFixed(
  n,
  mu0,
  sigma0_sqrt,
  alpha,
  beta,
  psi_sqrt,
  gamma_eta = NULL,
  x = NULL,
  time = 0,
  burn_in = 0
)

Arguments

n

Positive integer. Number of individuals.

mu0

Numeric vector. Mean of initial latent variable values (\(\boldsymbol{\mu}_{\boldsymbol{\eta} \mid 0}\)).

sigma0_sqrt

Numeric matrix. Cholesky decomposition of the covariance matrix of initial latent variable values (\(\boldsymbol{\Sigma}_{\boldsymbol{\eta} \mid 0}\)).

alpha

Numeric vector. Vector of intercepts for the dynamic model (\(\boldsymbol{\alpha}\)).

beta

Numeric matrix. Transition matrix relating the values of the latent variables at time t - 1 to those at time t (\(\boldsymbol{\beta}\)).

psi_sqrt

Numeric matrix. Cholesky decomposition of the process noise covariance matrix (\(\boldsymbol{\Psi}\)).

gamma_eta

Numeric matrix. Matrix relating the values of the covariate matrix at time t to the latent variables at time t (\(\boldsymbol{\Gamma}_{\boldsymbol{\eta}}\)).

x

A list of length n of numeric matrices. Each element of the list is a matrix of observed covariates in type = 1 or type = 2. The number of rows in each matrix should be equal to time + burn_in.

time

Positive integer. Number of time points to simulate.

burn_in

Positive integer. Number of burn-in points to exclude before returning the results.

Value

Returns a list of length n. Each element is a list with the following elements:

y: A t by k matrix of values for the manifest variables.

eta: A t by p matrix of values for the latent variables.

x: A t by j matrix of values for the covariates.

time: A vector of discrete time points from 1 to t.

id: A vector of ID numbers of length t.

References

Chow, S.-M., Ho, M. R., Hamaker, E. L., & Dolan, C. V. (2010). Equivalence and differences between structural equation modeling and state-space modeling techniques. Structural Equation Modeling: A Multidisciplinary Journal, 17(2), 303–332. doi:10.1080/10705511003661553

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# prepare parameters
set.seed(42)
k <- 3
iden <- diag(k)
iden_sqrt <- chol(iden)
null_vec <- rep(x = 0, times = k)
n <- 5
mu0 <- null_vec
sigma0_sqrt <- iden_sqrt
alpha <- null_vec
beta <- diag(x = 0.5, nrow = k)
psi_sqrt <- iden_sqrt
time <- 50
burn_in <- 0
gamma_eta <- 0.10 * diag(k)
x <- lapply(
  X = seq_len(n),
  FUN = function(i) {
    return(
      matrix(
        data = rnorm(n = k * (time + burn_in)),
        ncol = k
      )
    )
  }
)

# No covariates
ssm <- SimSSMVARFixed(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -0.9584 -0.6788 0.0758 1.5887 0.5055 ...
#>   ..$ eta : num [1:50, 1:3] -0.9584 -0.6788 0.0758 1.5887 0.5055 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -0.632 0.432 3.563 1.644 -0.633 ...
#>   ..$ eta : num [1:50, 1:3] -0.632 0.432 3.563 1.644 -0.633 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -0.3363 -0.3802 -0.0572 1.0263 1.6905 ...
#>   ..$ eta : num [1:50, 1:3] -0.3363 -0.3802 -0.0572 1.0263 1.6905 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -2.046 0.116 0.188 -0.62 -0.697 ...
#>   ..$ eta : num [1:50, 1:3] -2.046 0.116 0.188 -0.62 -0.697 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -0.308 -0.088 2.242 1.281 1.822 ...
#>   ..$ eta : num [1:50, 1:3] -0.308 -0.088 2.242 1.281 1.822 ...
#>   ..$ x   : int 0
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...

# With covariates
ssm <- SimSSMVARFixed(
  n = n,
  mu0 = mu0,
  sigma0_sqrt = sigma0_sqrt,
  alpha = alpha,
  beta = beta,
  psi_sqrt = psi_sqrt,
  gamma_eta = gamma_eta,
  x = x,
  time = time,
  burn_in = burn_in
)

str(ssm)
#> List of 5
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] 0.428 0.609 -0.0554 1.1467 -0.792 ...
#>   ..$ eta : num [1:50, 1:3] 0.428 0.609 -0.0554 1.1467 -0.792 ...
#>   ..$ x   : num [1:50, 1:3] 1.371 -0.565 0.363 0.633 0.404 ...
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] 0.4427 -0.0525 0.1017 1.1137 0.9603 ...
#>   ..$ eta : num [1:50, 1:3] 0.4427 -0.0525 0.1017 1.1137 0.9603 ...
#>   ..$ x   : num [1:50, 1:3] -0.0407 -1.5515 1.1672 -0.2736 -0.4678 ...
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 2 2 2 2 2 2 2 2 2 2 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -1.898 0.494 0.131 0.845 1.192 ...
#>   ..$ eta : num [1:50, 1:3] -1.898 0.494 0.131 0.845 1.192 ...
#>   ..$ x   : num [1:50, 1:3] -0.00462 0.76024 0.03899 0.73507 -0.14647 ...
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 3 3 3 3 3 3 3 3 3 3 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -0.92 0.934 0.382 -0.744 -1.393 ...
#>   ..$ eta : num [1:50, 1:3] -0.92 0.934 0.382 -0.744 -1.393 ...
#>   ..$ x   : num [1:50, 1:3] -1.304 -0.251 0.171 -0.403 0.105 ...
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 4 4 4 4 4 4 4 4 4 4 ...
#>  $ :List of 5
#>   ..$ y   : num [1:50, 1:3] -0.1266 0.5897 0.2575 -0.0341 -0.3956 ...
#>   ..$ eta : num [1:50, 1:3] -0.1266 0.5897 0.2575 -0.0341 -0.3956 ...
#>   ..$ x   : num [1:50, 1:3] -0.248 0.422 0.988 0.836 -0.661 ...
#>   ..$ time: num [1:50, 1] 0 1 2 3 4 5 6 7 8 9 ...
#>   ..$ id  : num [1:50, 1] 5 5 5 5 5 5 5 5 5 5 ...